Struct webrender::api::channel::Payload

pub struct Payload {
    pub epoch: Epoch,
    pub pipeline_id: PipelineId,
    pub display_list_data: Vec<u8>,
}

Fields
epoch: Epoch

An epoch used to get the proper payload for a pipeline id frame request.

TODO(emilio): Is this still relevant? We send the messages for the same pipeline in order, so we shouldn't need it. Seems like this was only wallpapering (in most cases) the underlying problem in #991.

pipeline_id: PipelineId

A pipeline id to key the payload with, along with the epoch.

display_list_data: Vec<u8>
Methods

impl Payload

fn to_data(&self) -> Vec<u8>

Convert the payload to a raw byte vector, in order for it to be efficiently shared via shmem, for example.

TODO(emilio, #1049): Consider moving the IPC boundary to the constellation in Servo and remove this complexity from WR.

fn from_data(data: &[u8]) -> Payload

Deserializes the given payload from a raw byte vector.