Struct webrender::api::DisplayListBuilder

pub struct DisplayListBuilder {
    pub data: Vec<u8>,
    pub pipeline_id: PipelineId,
    // some fields omitted
}

Fields
data: Vec<u8> pipeline_id: PipelineId
Methods

impl DisplayListBuilder

fn new(
pipeline_id: PipelineId,
content_size: TypedSize2D<f32, LayerPixel>
) -> DisplayListBuilder

fn with_capacity(
pipeline_id: PipelineId,
content_size: TypedSize2D<f32, LayerPixel>,
capacity: usize
) -> DisplayListBuilder

fn content_size(&self) -> TypedSize2D<f32, LayerPixel>

Return the content size for this display list

fn save(&mut self)

Saves the current display list state, so it may be restore()'d.

Conditions:

Doesn't support popping clips that were pushed before the save.

Doesn't support nested saves.

Must call clear_save() if the restore becomes unnecessary.

fn restore(&mut self)

Restores the state of the builder to when save() was last called.

fn clear_save(&mut self)

Discards the builder's save (indicating the attempted operation was sucessful).

fn print_display_list(&mut self)

fn push_rect(&mut self, info: &PrimitiveInfo<LayerPixel>, color: ColorF)

fn push_clear_rect(&mut self, info: &PrimitiveInfo<LayerPixel>)

fn push_line(
&mut self,
info: &PrimitiveInfo<LayerPixel>,
wavy_line_thickness: f32,
orientation: LineOrientation,
color: &ColorF,
style: LineStyle
)

fn push_image(
&mut self,
info: &PrimitiveInfo<LayerPixel>,
stretch_size: TypedSize2D<f32, LayerPixel>,
tile_spacing: TypedSize2D<f32, LayerPixel>,
image_rendering: ImageRendering,
key: ImageKey
)

fn push_yuv_image(
&mut self,
info: &PrimitiveInfo<LayerPixel>,
yuv_data: YuvData,
color_space: YuvColorSpace,
image_rendering: ImageRendering
)

Push a yuv image. All planar data in yuv image should use the same buffer type.

fn push_text(
&mut self,
info: &PrimitiveInfo<LayerPixel>,
glyphs: &[GlyphInstance],
font_key: FontInstanceKey,
color: ColorF,
glyph_options: Option<GlyphOptions>
)

fn create_gradient(
&mut self,
start_point: TypedPoint2D<f32, LayerPixel>,
end_point: TypedPoint2D<f32, LayerPixel>,
stops: Vec<GradientStop>,
extend_mode: ExtendMode
) -> Gradient

fn create_radial_gradient(
&mut self,
center: TypedPoint2D<f32, LayerPixel>,
radius: TypedSize2D<f32, LayerPixel>,
stops: Vec<GradientStop>,
extend_mode: ExtendMode
) -> RadialGradient

fn create_complex_radial_gradient(
&mut self,
start_center: TypedPoint2D<f32, LayerPixel>,
start_radius: f32,
end_center: TypedPoint2D<f32, LayerPixel>,
end_radius: f32,
ratio_xy: f32,
stops: Vec<GradientStop>,
extend_mode: ExtendMode
) -> RadialGradient

fn push_border(
&mut self,
info: &PrimitiveInfo<LayerPixel>,
widths: BorderWidths,
details: BorderDetails
)

fn push_box_shadow(
&mut self,
info: &PrimitiveInfo<LayerPixel>,
box_bounds: TypedRect<f32, LayerPixel>,
offset: TypedVector2D<f32, LayerPixel>,
color: ColorF,
blur_radius: f32,
spread_radius: f32,
border_radius: BorderRadius,
clip_mode: BoxShadowClipMode
)

fn push_gradient(
&mut self,
info: &PrimitiveInfo<LayerPixel>,
gradient: Gradient,
tile_size: TypedSize2D<f32, LayerPixel>,
tile_spacing: TypedSize2D<f32, LayerPixel>
)

Pushes a linear gradient to be displayed.

The gradient itself is described in the gradient parameter. It is drawn on a "tile" with the dimensions from tile_size. These tiles are now repeated to the right and to the bottom infinitly. If tile_spacing is not zero spacers with the given dimensions are inserted between the tiles as seams.

The origin of the tiles is given in info.rect.origin. If the gradient should only be displayed once limit the info.rect.size to a single tile. The gradient is only visible within the local clip.

fn push_radial_gradient(
&mut self,
info: &PrimitiveInfo<LayerPixel>,
gradient: RadialGradient,
tile_size: TypedSize2D<f32, LayerPixel>,
tile_spacing: TypedSize2D<f32, LayerPixel>
)

Pushes a radial gradient to be displayed.

See push_gradient for explanation.

fn push_stacking_context(
&mut self,
info: &PrimitiveInfo<LayerPixel>,
scroll_policy: ScrollPolicy,
transform: Option<PropertyBinding<TypedTransform3D<f32, LayerPixel, LayerPixel>>>,
transform_style: TransformStyle,
perspective: Option<TypedTransform3D<f32, LayerPixel, LayerPixel>>,
mix_blend_mode: MixBlendMode,
filters: Vec<FilterOp>
)

fn pop_stacking_context(&mut self)

fn push_stops(&mut self, stops: &[GradientStop])

fn define_scroll_frame<I>(
&mut self,
id: Option<ClipId>,
content_rect: TypedRect<f32, LayerPixel>,
clip_rect: TypedRect<f32, LayerPixel>,
complex_clips: I,
image_mask: Option<ImageMask>,
scroll_sensitivity: ScrollSensitivity
) -> ClipId where
I: IntoIterator<Item = ComplexClipRegion>,
<I as IntoIterator>::IntoIter: ExactSizeIterator,
<I as IntoIterator>::IntoIter: Clone,

fn define_scroll_frame_with_parent<I>(
&mut self,
id: Option<ClipId>,
parent: ClipId,
content_rect: TypedRect<f32, LayerPixel>,
clip_rect: TypedRect<f32, LayerPixel>,
complex_clips: I,
image_mask: Option<ImageMask>,
scroll_sensitivity: ScrollSensitivity
) -> ClipId where
I: IntoIterator<Item = ComplexClipRegion>,
<I as IntoIterator>::IntoIter: ExactSizeIterator,
<I as IntoIterator>::IntoIter: Clone,

fn define_clip<I>(
&mut self,
id: Option<ClipId>,
clip_rect: TypedRect<f32, LayerPixel>,
complex_clips: I,
image_mask: Option<ImageMask>
) -> ClipId where
I: IntoIterator<Item = ComplexClipRegion>,
<I as IntoIterator>::IntoIter: ExactSizeIterator,
<I as IntoIterator>::IntoIter: Clone,

fn define_clip_with_parent<I>(
&mut self,
id: Option<ClipId>,
parent: ClipId,
clip_rect: TypedRect<f32, LayerPixel>,
complex_clips: I,
image_mask: Option<ImageMask>
) -> ClipId where
I: IntoIterator<Item = ComplexClipRegion>,
<I as IntoIterator>::IntoIter: ExactSizeIterator,
<I as IntoIterator>::IntoIter: Clone,

fn define_sticky_frame(
&mut self,
id: Option<ClipId>,
frame_rect: TypedRect<f32, LayerPixel>,
margins: TypedSideOffsets2D<Option<f32>, UnknownUnit>,
vertical_offset_bounds: StickyOffsetBounds,
horizontal_offset_bounds: StickyOffsetBounds,
previously_applied_offset: TypedVector2D<f32, LayerPixel>
) -> ClipId

fn push_clip_id(&mut self, id: ClipId)

fn push_clip_and_scroll_info(&mut self, info: ClipAndScrollInfo)

fn pop_clip_id(&mut self)

fn push_iframe(
&mut self,
info: &PrimitiveInfo<LayerPixel>,
pipeline_id: PipelineId
)

fn push_shadow(&mut self, info: &PrimitiveInfo<LayerPixel>, shadow: Shadow)

fn pop_all_shadows(&mut self)

fn finalize(
self
) -> (PipelineId, TypedSize2D<f32, LayerPixel>, BuiltDisplayList)