Struct webrender::api::HitTestItem

pub struct HitTestItem {
    pub pipeline: PipelineId,
    pub tag: (u64, u16),
    pub point_in_viewport: TypedPoint2D<f32, LayerPixel>,
    pub point_relative_to_item: TypedPoint2D<f32, LayerPixel>,
}

Fields
pipeline: PipelineId

The pipeline that the display item that was hit belongs to.

tag: (u64, u16)

The tag of the hit display item.

point_in_viewport: TypedPoint2D<f32, LayerPixel>

The hit point in the coordinate space of the "viewport" of the display item. The viewport is the scroll node formed by the root reference frame of the display item's pipeline.

point_relative_to_item: TypedPoint2D<f32, LayerPixel>

The coordinates of the original hit test point relative to the origin of this item. This is useful for calculating things like text offsets in the client.