Struct webrender::api::StickyFrameDisplayItem

pub struct StickyFrameDisplayItem {
    pub id: ClipId,
    pub margins: TypedSideOffsets2D<Option<f32>, UnknownUnit>,
    pub vertical_offset_bounds: StickyOffsetBounds,
    pub horizontal_offset_bounds: StickyOffsetBounds,
    pub previously_applied_offset: TypedVector2D<f32, LayerPixel>,
}

Fields
id: ClipId margins: TypedSideOffsets2D<Option<f32>, UnknownUnit>

The margins that should be maintained between the edge of the parent viewport and this sticky frame. A margin of None indicates that the sticky frame should not stick at all to that particular edge of the viewport.

vertical_offset_bounds: StickyOffsetBounds

The minimum and maximum vertical offsets for this sticky frame. Ignoring these constraints, the sticky frame will continue to stick to the edge of the viewport as its original position is scrolled out of view. Constraints specify a maximum and minimum offset from the original position relative to non-sticky content within the same scrolling frame.

horizontal_offset_bounds: StickyOffsetBounds

The minimum and maximum horizontal offsets for this sticky frame. Ignoring these constraints, the sticky frame will continue to stick to the edge of the viewport as its original position is scrolled out of view. Constraints specify a maximum and minimum offset from the original position relative to non-sticky content within the same scrolling frame.

previously_applied_offset: TypedVector2D<f32, LayerPixel>

The amount of offset that has already been applied to the sticky frame. A positive y component this field means that a top-sticky item was in a scrollframe that has been scrolled down, such that the sticky item's position needed to be offset downwards by previously_applied_offset.y. A negative y component corresponds to the upward offset applied due to bottom-stickiness. The x-axis works analogously.