Trait webrender::api::BlobImageHandler

pub trait BlobImageHandler: Send {
    fn create_blob_rasterizer(
        &mut self
    ) -> Box<AsyncBlobImageRasterizer + 'static>;

    fn prepare_resources(
        &mut self, 
        services: &BlobImageResources, 
        requests: &[BlobImageParams]
    );

    fn add(&mut self, key: ImageKey, data: Arc<Vec<u8>>, tiling: Option<u16>);

    fn update(
        &mut self, 
        key: ImageKey, 
        data: Arc<Vec<u8>>, 
        dirty_rect: Option<TypedRect<u32, DevicePixel>>
    );

    fn delete(&mut self, key: ImageKey);

    fn delete_font(&mut self, key: FontKey);

    fn delete_font_instance(&mut self, key: FontInstanceKey);

    fn clear_namespace(&mut self, namespace: IdNamespace);
}

A handler on the render backend that can create rasterizer objects which will be sent to the scene builder thread to execute the rasterization.

The handler is responsible for collecting resources, managing/updating blob commands and creating the rasterizer objects, but isn't expected to do any rasterization itself.

Required Methods

fn create_blob_rasterizer(&mut self) -> Box<AsyncBlobImageRasterizer + 'static>

Creates a snapshot of the current state of blob images in the handler.

fn prepare_resources(
&mut self,
services: &BlobImageResources,
requests: &[BlobImageParams]
)

A hook to let the blob image handler update any state related to resources that are not bundled in the blob recording itself.

fn add(&mut self, key: ImageKey, data: Arc<Vec<u8>>, tiling: Option<u16>)

Register a blob image.

fn update(
&mut self,
key: ImageKey,
data: Arc<Vec<u8>>,
dirty_rect: Option<TypedRect<u32, DevicePixel>>
)

Update an already registered blob image.

fn delete(&mut self, key: ImageKey)

Delete an already registered blob image.

fn delete_font(&mut self, key: FontKey)

A hook to let the handler clean up any state related to a font which the resource cache is about to delete.

fn delete_font_instance(&mut self, key: FontInstanceKey)

A hook to let the handler clean up any state related to a font instance which the resource cache is about to delete.

fn clear_namespace(&mut self, namespace: IdNamespace)

A hook to let the handler clean up any state related a given namespace before the resource cache deletes them.