Struct webrender::batch::BatchBuilder

pub struct BatchBuilder {
    glyph_fetch_buffer: Vec<GlyphFetchResult>,
    pub batchers: Vec<AlphaBatchBuilder>,
}

Supports (recursively) adding a list of primitives and pictures to an alpha batch builder. In future, it will support multiple dirty regions / slices, allowing the contents of a picture to be spliced into multiple batch builders.

Fields
glyph_fetch_buffer: Vec<GlyphFetchResult>

A temporary buffer that is used during glyph fetching, stored here to reduce memory allocations.

batchers: Vec<AlphaBatchBuilder>
Implementations

impl BatchBuilder

pub fn new(batchers: Vec<AlphaBatchBuilder>) -> Self

pub fn finalize(self) -> Vec<AlphaBatchBuilder>

fn add_brush_instance_to_batches(
&mut self,
batch_key: BatchKey,
features: BatchFeatures,
bounding_rect: &PictureRect,
z_id: ZBufferId,
segment_index: i32,
edge_flags: EdgeAaSegmentMask,
clip_task_address: RenderTaskAddress,
brush_flags: BrushFlags,
prim_header_index: PrimitiveHeaderIndex,
resource_address: i32,
prim_vis_mask: PrimitiveVisibilityMask
)

fn add_split_composite_instance_to_batches(
&mut self,
batch_key: BatchKey,
bounding_rect: &PictureRect,
z_id: ZBufferId,
prim_header_index: PrimitiveHeaderIndex,
polygons_address: GpuCacheAddress,
prim_vis_mask: PrimitiveVisibilityMask
)

fn clear_batches(&mut self)

Clear all current batchers. This is typically used when a primitive is encountered that occludes all previous content in this batch list.

pub fn add_pic_to_batch(
&mut self,
pic: &PicturePrimitive,
ctx: &RenderTargetContext,
gpu_cache: &mut GpuCache,
render_tasks: &RenderTaskGraph,
deferred_resolves: &mut Vec<DeferredResolve>,
prim_headers: &mut PrimitiveHeaders,
transforms: &mut TransformPalette,
root_spatial_node_index: SpatialNodeIndex,
surface_spatial_node_index: SpatialNodeIndex,
z_generator: &mut ZBufferIdGenerator,
composite_state: &mut CompositeState
)

Add a picture to a given batch builder.

fn emit_placeholder(
&mut self,
prim_rect: LayoutRect,
prim_info: &PrimitiveVisibility,
z_id: ZBufferId,
transform_id: TransformPaletteId,
batch_features: BatchFeatures,
ctx: &RenderTargetContext,
gpu_cache: &mut GpuCache,
render_tasks: &RenderTaskGraph,
prim_headers: &mut PrimitiveHeaders
)

fn add_prim_to_batch(
&mut self,
prim_instance: &PrimitiveInstance,
prim_spatial_node_index: SpatialNodeIndex,
ctx: &RenderTargetContext,
gpu_cache: &mut GpuCache,
render_tasks: &RenderTaskGraph,
deferred_resolves: &mut Vec<DeferredResolve>,
prim_headers: &mut PrimitiveHeaders,
transforms: &mut TransformPalette,
root_spatial_node_index: SpatialNodeIndex,
surface_spatial_node_index: SpatialNodeIndex,
z_generator: &mut ZBufferIdGenerator,
composite_state: &mut CompositeState
)

fn add_segment_to_batch(
&mut self,
segment: &BrushSegment,
segment_data: &SegmentInstanceData,
segment_index: i32,
batch_kind: BrushBatchKind,
prim_header_index: PrimitiveHeaderIndex,
alpha_blend_mode: BlendMode,
features: BatchFeatures,
bounding_rect: &PictureRect,
transform_kind: TransformedRectKind,
render_tasks: &RenderTaskGraph,
z_id: ZBufferId,
prim_opacity: PrimitiveOpacity,
clip_task_index: ClipTaskIndex,
prim_vis_mask: PrimitiveVisibilityMask,
ctx: &RenderTargetContext
)

Add a single segment instance to a batch.

fn add_segmented_prim_to_batch(
&mut self,
brush_segments: Option<&[BrushSegment]>,
prim_opacity: PrimitiveOpacity,
params: &BrushBatchParameters,
alpha_blend_mode: BlendMode,
non_segmented_blend_mode: BlendMode,
features: BatchFeatures,
prim_header_index: PrimitiveHeaderIndex,
bounding_rect: &PictureRect,
transform_kind: TransformedRectKind,
render_tasks: &RenderTaskGraph,
z_id: ZBufferId,
clip_task_index: ClipTaskIndex,
prim_vis_mask: PrimitiveVisibilityMask,
ctx: &RenderTargetContext
)

Add any segment(s) from a brush to batches.

fn add_gradient_tiles(
&mut self,
visible_tiles: &[VisibleGradientTile],
stops_handle: &GpuCacheHandle,
kind: BrushBatchKind,
blend_mode: BlendMode,
bounding_rect: &PictureRect,
clip_task_address: RenderTaskAddress,
gpu_cache: &GpuCache,
base_prim_header: &PrimitiveHeader,
prim_headers: &mut PrimitiveHeaders,
z_id: ZBufferId,
prim_vis_mask: PrimitiveVisibilityMask
)