Module webrender::batch_builder

Structs

BoxShadowMetrics

GradientRect

Enums

ClipState

Constants

BORDER_DASH_SIZE

Traits

BorderSideHelpers

Functions

compute_box_shadow_rect

groove_ridge_border_colors
Returns the top/left and bottom/right colors respectively.

subdivide_border_corner
Subdivides the border corner into four quadrants and returns them in the order of outer corner, inner corner, color 0 and color 1, respectively. See the diagram in the documentation for add_border_corner for more information on what these values represent.