webrender

Module bump_allocator

Structs

BumpAllocator

A simple bump allocator, sub-allocating from fixed size chunks that are provided by a parent allocator.

Chunk

A Contiguous buffer of memory holding multiple sub-allocaions.

ChunkPool

A simple pool for allocating and recycling memory chunks of a fixed size, protected by a mutex.

ChunkPoolInner 🔒

RecycledChunk 🔒

Header at the beginning of recycled memory chunk.

Stats

Constants

CHUNK_ALIGNMENT 🔒

DEFAULT_CHUNK_SIZE 🔒

Functions

align 🔒