Struct webrender::clip_scroll_node::ClipScrollNode

pub struct ClipScrollNode {
    pub scrolling: ScrollingState,
    pub content_size: LayerSize,
    pub local_viewport_rect: LayerRect,
    pub local_clip_rect: LayerRect,
    pub combined_local_viewport_rect: LayerRect,
    pub world_viewport_transform: LayerToWorldTransform,
    pub world_content_transform: LayerToWorldTransform,
    pub pipeline_id: PipelineId,
    pub parent: Option<ScrollLayerId>,
    pub children: Vec<ScrollLayerId>,
    pub node_type: NodeType,
}

Contains scrolling and transform information stacking contexts.

Fields
scrolling: ScrollingState

Manages scrolling offset, overscroll state etc.

content_size: LayerSize

Size of the content inside the scroll region (in logical pixels)

local_viewport_rect: LayerRect

Viewing rectangle in the coordinate system of the parent reference frame.

local_clip_rect: LayerRect

Clip rect of this node - typically the same as viewport rect, except in overscroll cases.

combined_local_viewport_rect: LayerRect

Viewport rectangle clipped against parent layer(s) viewport rectangles. This is in the coordinate system of the parent reference frame.

world_viewport_transform: LayerToWorldTransform

World transform for the viewport rect itself. This is the parent reference frame transformation plus the scrolling offsets provided by the nodes in between the reference frame and this node.

world_content_transform: LayerToWorldTransform

World transform for content transformed by this node.

pipeline_id: PipelineId

Pipeline that this layer belongs to

parent: Option<ScrollLayerId>

Parent layer. If this is None, we are the root node.

children: Vec<ScrollLayerId>

Child layers

node_type: NodeType

Whether or not this node is a reference frame.

Methods

impl ClipScrollNode

fn new(pipeline_id: PipelineId,
parent_id: ScrollLayerId,
local_viewport_rect: &LayerRect,
content_size: LayerSize,
clip_info: ClipInfo)
-> ClipScrollNode

fn new_reference_frame(parent_id: Option<ScrollLayerId>,
local_viewport_rect: &LayerRect,
content_size: LayerSize,
local_transform: &LayerToScrollTransform,
pipeline_id: PipelineId)
-> ClipScrollNode

fn add_child(&mut self, child: ScrollLayerId)

fn finalize(&mut self, scrolling: &ScrollingState)

fn overscroll_amount(&self) -> LayerSize

fn set_scroll_origin(&mut self, origin: &LayerPoint) -> bool

fn update_transform(&mut self,
parent_reference_frame_transform: &LayerToWorldTransform,
parent_combined_viewport_rect: &ScrollLayerRect,
parent_accumulated_scroll_offset: LayerPoint)

fn scrollable_height(&self) -> f32

fn scrollable_width(&self) -> f32

fn scroll(&mut self,
scroll_location: ScrollLocation,
phase: ScrollEventPhase)
-> bool

fn stretch_overscroll_spring(&mut self)

fn tick_scrolling_bounce_animation(&mut self)

fn ray_intersects_node(&self, cursor: &WorldPoint) -> bool