Struct webrender::clip_scroll_tree::ClipScrollTree

pub struct ClipScrollTree {
    pub nodes: HashMap<ScrollLayerId, ClipScrollNode, BuildHasherDefault<FnvHasher>>,
    pub pending_scroll_offsets: HashMap<(PipelineId, ServoScrollRootId), LayerPoint>,
    pub current_scroll_layer_id: Option<ScrollLayerId>,
    current_reference_frame_id: usize,
    pub root_reference_frame_id: ScrollLayerId,
    pub topmost_scroll_layer_id: ScrollLayerId,
    pub pipelines_to_discard: HashSet<PipelineId>,
}

Fields
nodes: HashMap<ScrollLayerId, ClipScrollNode, BuildHasherDefault<FnvHasher>> pending_scroll_offsets: HashMap<(PipelineId, ServoScrollRootId), LayerPoint> current_scroll_layer_id: Option<ScrollLayerId>

The ScrollLayerId of the currently scrolling node. Used to allow the same node to scroll even if a touch operation leaves the boundaries of that node.

current_reference_frame_id: usize

The current reference frame id, used for giving a unique id to all new reference frames. The ClipScrollTree increments this by one every time a reference frame is created.

root_reference_frame_id: ScrollLayerId

The root reference frame, which is the true root of the ClipScrollTree. Initially this ID is not valid, which is indicated by node being empty.

topmost_scroll_layer_id: ScrollLayerId

The root scroll node which is the first child of the root reference frame. Initially this ID is not valid, which is indicated by nodes being empty.

pipelines_to_discard: HashSet<PipelineId>

A set of pipelines which should be discarded the next time this tree is drained.

Methods

impl ClipScrollTree

fn new() -> ClipScrollTree

fn root_reference_frame_id(&self) -> ScrollLayerId

fn topmost_scroll_layer_id(&self) -> ScrollLayerId

fn collect_nodes_bouncing_back(&self)
-> HashSet<ScrollLayerId, BuildHasherDefault<FnvHasher>>

fn find_scrolling_node_at_point_in_node(&self,
cursor: &WorldPoint,
scroll_layer_id: ScrollLayerId)
-> Option<ScrollLayerId>

fn find_scrolling_node_at_point(&self, cursor: &WorldPoint) -> ScrollLayerId

fn get_scroll_node_state(&self) -> Vec<ScrollLayerState>

fn drain(&mut self)
-> HashMap<ScrollLayerId, ScrollingState, BuildHasherDefault<FnvHasher>>

fn scroll_nodes(&mut self,
origin: LayerPoint,
pipeline_id: PipelineId,
scroll_root_id: ServoScrollRootId)
-> bool

fn scroll(&mut self,
scroll_location: ScrollLocation,
cursor: WorldPoint,
phase: ScrollEventPhase)
-> bool

fn update_all_node_transforms(&mut self)

fn update_node_transform(&mut self,
layer_id: ScrollLayerId,
parent_reference_frame_transform: &LayerToWorldTransform,
parent_viewport_rect: &ScrollLayerRect,
parent_accumulated_scroll_offset: LayerPoint)

fn tick_scrolling_bounce_animations(&mut self)

fn finalize_and_apply_pending_scroll_offsets(&mut self,
old_states: HashMap<ScrollLayerId, ScrollingState, BuildHasherDefault<FnvHasher>>)

fn add_reference_frame(&mut self,
rect: &LayerRect,
transform: &LayerToScrollTransform,
pipeline_id: PipelineId,
parent_id: Option<ScrollLayerId>)
-> ScrollLayerId

fn add_node(&mut self, node: ClipScrollNode, id: ScrollLayerId)

fn discard_frame_state_for_pipeline(&mut self, pipeline_id: PipelineId)