Module webrender::command_buffer

Structs

Command

A tightly packed command stored in a command buffer

CommandBuffer

A list of commands describing how to draw a primitive list.

CommandBufferBuilder

CommandBufferIndex

CommandBufferList

MaskFlags

Defines the space that a quad primitive is drawn in

QuadFlags

Flags related to quad primitives

Enums

CommandBufferBuilderKind

Abstracts whether a command buffer is being built for a tiled (picture cache) or simple (child surface).

PrimitiveCommand

The unpacked equivalent to a Command.