Struct webrender::command_buffer::Command

pub struct Command(u32);

A tightly packed command stored in a command buffer

Tuple Fields
0: u32
Implementations

impl Command

const CMD_DRAW_SIMPLE_PRIM: u32 = 0u32

Draw a simple primitive that needs prim instance index only.

const CMD_SET_SPATIAL_NODE: u32 = 268_435_456u32

Change the current spatial node.

const CMD_DRAW_COMPLEX_PRIM: u32 = 536_870_912u32

Draw a complex (3d-split) primitive, that has multiple GPU cache addresses.

const CMD_DRAW_INSTANCE: u32 = 805_306_368u32

Draw a primitive, that has a single GPU buffer addresses.

const CMD_DRAW_QUAD: u32 = 1_073_741_824u32

Draw a generic quad primitive

const CMD_SET_SEGMENTS: u32 = 1_342_177_280u32

Set a list of variable-length segments

const CMD_MASK: u32 = 4_026_531_840u32

Bitmask for command bits of the command.

const PARAM_MASK: u32 = 268_435_455u32

Bitmask for param bits of the command.

fn draw_simple_prim(prim_instance_index: PrimitiveInstanceIndex) -> Self

Encode drawing a simple primitive.

fn set_spatial_node(spatial_node_index: SpatialNodeIndex) -> Self

Encode changing spatial node.

fn set_segments(count: usize) -> Self

Encode a list of segments that follow

fn draw_complex_prim(prim_instance_index: PrimitiveInstanceIndex) -> Self

Encode drawing a complex prim.

fn draw_instance(prim_instance_index: PrimitiveInstanceIndex) -> Self

fn data(data: u32) -> Self

Encode arbitrary data word.

fn draw_quad(prim_instance_index: PrimitiveInstanceIndex) -> Self