Function webrender::composite::area_of_occluders

fn area_of_occluders(
    occluders: &[Occluder], 
    z_id: ZBufferId, 
    clip_rect: &DeviceIntRect
) -> i32

Return the total area covered by a set of occluders, accounting for overlapping areas between those rectangles.