Struct webrender::composite::ExternalSurfaceDescriptor

pub struct ExternalSurfaceDescriptor {
    pub local_rect: PictureRect,
    pub world_rect: WorldRect,
    pub device_rect: DeviceRect,
    pub local_clip_rect: PictureRect,
    pub clip_rect: DeviceRect,
    pub image_rendering: ImageRendering,
    pub z_id: ZBufferId,
    pub dependency: ExternalSurfaceDependency,
    pub native_surface_id: Option<NativeSurfaceId>,
    pub update_params: Option<DeviceIntSize>,
}

Describes information about drawing a primitive as a compositor surface. For now, we support only YUV images as compositor surfaces, but in future this will also support RGBA images.

Fields
local_rect: PictureRectworld_rect: WorldRectdevice_rect: DeviceRectlocal_clip_rect: PictureRectclip_rect: DeviceRectimage_rendering: ImageRenderingz_id: ZBufferIddependency: ExternalSurfaceDependencynative_surface_id: Option<NativeSurfaceId>

If native compositing is enabled, the native compositor surface handle. Otherwise, this will be None

update_params: Option<DeviceIntSize>

If the native surface needs to be updated, this will contain the size of the native surface as Some(size). If not dirty, this is None.