Trait webrender::composite::Compositor

pub trait Compositor {
    fn create_surface(
        &mut self, 
        id: NativeSurfaceId, 
        virtual_offset: DeviceIntPoint, 
        tile_size: DeviceIntSize, 
        is_opaque: bool
    );

    fn destroy_surface(&mut self, id: NativeSurfaceId);

    fn create_tile(&mut self, id: NativeTileId);

    fn destroy_tile(&mut self, id: NativeTileId);

    fn bind(
        &mut self, 
        id: NativeTileId, 
        dirty_rect: DeviceIntRect, 
        valid_rect: DeviceIntRect
    ) -> NativeSurfaceInfo;

    fn unbind(&mut self);

    fn begin_frame(&mut self);

    fn add_surface(
        &mut self, 
        id: NativeSurfaceId, 
        position: DeviceIntPoint, 
        clip_rect: DeviceIntRect
    );

    fn end_frame(&mut self);

    fn enable_native_compositor(&mut self, enable: bool);

    fn deinit(&mut self);

    fn get_capabilities(&self) -> CompositorCapabilities;
}

Defines an interface to a native (OS level) compositor. If supplied by the client application, then picture cache slices will be composited by the OS compositor, rather than drawn via WR batches.

Required methods

fn create_surface(
&mut self,
id: NativeSurfaceId,
virtual_offset: DeviceIntPoint,
tile_size: DeviceIntSize,
is_opaque: bool
)

Create a new OS compositor surface with the given properties.

fn destroy_surface(&mut self, id: NativeSurfaceId)

Destroy the surface with the specified id. WR may call this at any time the surface is no longer required (including during renderer deinit). It's the responsibility of the embedder to ensure that the surface is only freed once the GPU is no longer using the surface (if this isn't already handled by the operating system).

fn create_tile(&mut self, id: NativeTileId)

Create a new OS compositor tile with the given properties.

fn destroy_tile(&mut self, id: NativeTileId)

Destroy an existing compositor tile.

fn bind(
&mut self,
id: NativeTileId,
dirty_rect: DeviceIntRect,
valid_rect: DeviceIntRect
) -> NativeSurfaceInfo

Bind this surface such that WR can issue OpenGL commands that will target the surface. Returns an (x, y) offset where WR should draw into the surface. This can be set to (0, 0) if the OS doesn't use texture atlases. The dirty rect is a local surface rect that specifies which part of the surface needs to be updated. If max_update_rects in CompositeConfig is 0, this will always be the size of the entire surface. The returned offset is only relevant to compositors that store surfaces in a texture atlas (that is, WR expects that the dirty rect doesn't affect the coordinates of the returned origin).

fn unbind(&mut self)

Unbind the surface. This is called by WR when it has finished issuing OpenGL commands on the current surface.

fn begin_frame(&mut self)

Begin the frame

fn add_surface(
&mut self,
id: NativeSurfaceId,
position: DeviceIntPoint,
clip_rect: DeviceIntRect
)

Add a surface to the visual tree to be composited. Visuals must be added every frame, between the begin/end transaction call. The z-order of the surfaces is determined by the order they are added to the visual tree.

fn end_frame(&mut self)

Commit any changes in the compositor tree for this frame. WR calls this once when all surface and visual updates are complete, to signal that the OS composite transaction should be applied.

fn enable_native_compositor(&mut self, enable: bool)

Enable/disable native compositor usage

fn deinit(&mut self)

Safely deinitialize any remaining resources owned by the compositor.

fn get_capabilities(&self) -> CompositorCapabilities

Get the capabilities struct for this compositor. This is used to specify what features a compositor supports, depending on the underlying platform

Loading content...