webrender::device

Module gl

Structs

BoundPBO

Capabilities

CustomVAO

Device

ExternalTexture

FBOId

FormatDesc

GpuFrameId

Sequence number for frames, as tracked by the device layer.

IBOId 🔒

PBO

PixelBuffer 🔒

Program

ProgramBinary

ProgramCache

ProgramCacheEntry 🔒

ProgramSourceInfo

RBOId

SharedDepthTarget 🔒

A refcounted depth target, which may be shared by multiple textures across the device.

Stream

Texture

WebRender interface to an OpenGL texture.

TextureFlags

TextureFormatPair

A structure defining a particular workflow of texture transfers.

TextureSlot

TextureUploader

Used to perform a series of texture uploads. Create using Device::upload_texture(). Perform a series of uploads using either upload(), or stage() and upload_staged(), then call flush().

UniformLocation

UploadChunk 🔒

UploadPBO 🔒

A PBO for uploading texture data, managed by UploadPBOPool.

UploadPBOPool

Allocates and recycles PBOs used for uploading texture data. Tries to allocate and recycle PBOs of a fixed size, but will make exceptions when a larger buffer is required or to work around driver bugs.

UploadStagingBuffer

A buffer used to manually stage data to be uploaded to a texture. Created by calling TextureUploader::stage(), the data can then be written to via get_mapping().

VAO

VBO

VBOId

VertexAttribute

VertexDescriptor

Enums

DepthFunction

DrawTarget

Contains the parameters necessary to bind a draw target.

FBOTarget 🔒

PBOMapping 🔒

ProgramSourceType 🔒

ReadTarget

Contains the parameters necessary to bind a texture-backed read target.

ShaderError

StrideAlignment

Describes a required alignment for a stride, which can either be represented in bytes or pixels.

TexStorageUsage 🔒

Describes for which texture formats to use the glTexStorage* family of functions.

TextureFilter

UploadMethod

Method of uploading texel data from CPU to GPU.

VertexAttributeKind

VertexUsageHint

Constants

DEFAULT_TEXTURE 🔒

RESERVE_DEPTH_BITS 🔒

Traits

ProgramCacheObserver

The interfaces that an application can implement to handle ProgramCache update

Texel

Plain old data that can be used to initialize a texture.

Functions

depth_target_size_in_bytes 🔒

Returns the size in bytes of a depth target with the given dimensions.

from_gl_target

get_gl_target

get_shader_version 🔒

get_unoptimized_shader_source

gl_error_string 🔒

is_mali_bifrost 🔒

Returns whether this GPU belongs to the Mali Bifrost family

is_mali_midgard 🔒

Returns whether this GPU belongs to the Mali Midgard family

is_mali_valhall 🔒

Returns whether this GPU belongs to the Mali Valhall family

parse_mali_version 🔒

Parses the major, release, and patch versions from a GL_VERSION string on Mali devices. For example, for the version string “OpenGL ES 3.2 v1.r36p0-01eac0.28ab3a577f105e026887e2b4c93552fb” this returns Some((1, 36, 0)). Returns None if the version cannot be parsed.

supports_extension 🔒

texels_to_u8_slice 🔒