Struct webrender::frame::Frame

pub struct Frame {
    pub clip_scroll_tree: ClipScrollTree,
    pub pipeline_epoch_map: HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>>,
    pub pipeline_auxiliary_lists: HashMap<PipelineId, AuxiliaryLists, BuildHasherDefault<FnvHasher>>,
    id: FrameId,
    frame_builder_config: FrameBuilderConfig,
    frame_builder: Option<FrameBuilder>,
}

Fields
clip_scroll_tree: ClipScrollTree pipeline_epoch_map: HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>> pipeline_auxiliary_lists: HashMap<PipelineId, AuxiliaryLists, BuildHasherDefault<FnvHasher>> id: FrameId frame_builder_config: FrameBuilderConfig frame_builder: Option<FrameBuilder>
Methods

impl Frame

fn new(config: FrameBuilderConfig) -> Frame

fn reset(&mut self)
-> HashMap<ScrollLayerId, ScrollingState, BuildHasherDefault<FnvHasher>>

fn get_scroll_node_state(&self) -> Vec<ScrollLayerState>

fn scroll_nodes(&mut self,
origin: LayerPoint,
pipeline_id: PipelineId,
scroll_root_id: ServoScrollRootId)
-> bool

Returns true if any nodes actually changed position or false otherwise.

fn scroll(&mut self,
scroll_location: ScrollLocation,
cursor: WorldPoint,
phase: ScrollEventPhase)
-> bool

Returns true if any nodes actually changed position or false otherwise.

fn tick_scrolling_bounce_animations(&mut self)

fn discard_frame_state_for_pipeline(&mut self, pipeline_id: PipelineId)

fn create(&mut self, scene: &Scene, resource_cache: &mut ResourceCache)

fn flatten_scroll_layer<'a>(&mut self,
traversal: &mut DisplayListTraversal<'a>,
pipeline_id: PipelineId,
context: &mut FlattenContext,
current_reference_frame_id: ScrollLayerId,
parent_scroll_layer_id: ScrollLayerId,
layer_relative_transform: LayerToScrollTransform,
level: i32,
clip: &ClipRegion,
content_size: &LayerSize,
new_scroll_layer_id: ScrollLayerId)

fn flatten_stacking_context<'a>(&mut self,
traversal: &mut DisplayListTraversal<'a>,
pipeline_id: PipelineId,
context: &mut FlattenContext,
current_reference_frame_id: ScrollLayerId,
current_scroll_layer_id: ScrollLayerId,
layer_relative_transform: LayerToScrollTransform,
level: i32,
stacking_context: &StackingContext,
clip_region: &ClipRegion)

fn flatten_iframe<'a>(&mut self,
pipeline_id: PipelineId,
bounds: &LayerRect,
context: &mut FlattenContext,
current_scroll_layer_id: ScrollLayerId,
layer_relative_transform: LayerToScrollTransform)

fn flatten_items<'a>(&mut self,
traversal: &mut DisplayListTraversal<'a>,
pipeline_id: PipelineId,
context: &mut FlattenContext,
current_reference_frame_id: ScrollLayerId,
current_scroll_layer_id: ScrollLayerId,
layer_relative_transform: LayerToScrollTransform,
level: i32)

fn decompose_tiled_image(&mut self,
context: &mut FlattenContext,
item: &DisplayItem,
info: &ImageDisplayItem,
image_size: DeviceUintSize,
tile_size: u32)

fn add_tile_primitive(&mut self,
context: &mut FlattenContext,
item: &DisplayItem,
info: &ImageDisplayItem,
tile_offset: TileOffset,
stretched_tile_size: LayerSize,
tile_ratio_width: f32,
tile_ratio_height: f32,
repeat_x: bool,
repeat_y: bool)

fn build(&mut self,
resource_cache: &mut ResourceCache,
auxiliary_lists_map: &HashMap<PipelineId, AuxiliaryLists, BuildHasherDefault<FnvHasher>>,
device_pixel_ratio: f32)
-> RendererFrame

fn build_frame(&mut self,
resource_cache: &mut ResourceCache,
auxiliary_lists_map: &HashMap<PipelineId, AuxiliaryLists, BuildHasherDefault<FnvHasher>>,
device_pixel_ratio: f32)
-> RendererFrame