webrender

Module frame_builder

Structs

Frame

A rendering-oriented representation of the frame built by the render backend and presented to the renderer.

FrameBuilder

Produces the frames that are sent to the renderer.

FrameBuilderConfig

FrameBuildingContext

FrameBuildingState

FrameGlobalResources

A set of default / global resources that are re-built each frame.

FrameScratchBuffer

PictureContext

Immutable context of a picture when processing children.

PictureState

Mutable state of a picture that gets modified when the children are processed.

Functions

build_render_pass

Processes this pass to prepare it for rendering.