webrender::frame_builder

Struct FrameBuilder

pub struct FrameBuilder {
    prim_headers_prealloc: Preallocator,
    composite_state_prealloc: CompositeStatePreallocator,
    plane_splitters: Vec<BspSplitter<PlaneSplitAnchor>>,
}

Produces the frames that are sent to the renderer.

Fields
prim_headers_prealloc: Preallocatorcomposite_state_prealloc: CompositeStatePreallocatorplane_splitters: Vec<BspSplitter<PlaneSplitAnchor>>
Implementations

impl FrameBuilder

pub fn new() -> Self

fn build_layer_screen_rects_and_cull_layers( &mut self, scene: &mut BuiltScene, present: bool, global_screen_world_rect: WorldRect, resource_cache: &mut ResourceCache, rg_builder: &mut RenderTaskGraphBuilder, global_device_pixel_scale: DevicePixelScale, scene_properties: &SceneProperties, transform_palette: &mut TransformPalette, data_stores: &mut DataStores, scratch: &mut ScratchBuffer, debug_flags: DebugFlags, composite_state: &mut CompositeState, tile_caches: &mut FastHashMap<SliceId, Box<TileCacheInstance>>, spatial_tree: &SpatialTree, cmd_buffers: &mut CommandBufferList, frame_gpu_data: &mut GpuBufferBuilder, frame_memory: &FrameMemory, profile: &mut TransactionProfile, )

Compute the contribution (bounding rectangles, and resources) of layers and their primitives in screen space.

pub fn build( &mut self, scene: &mut BuiltScene, present: bool, resource_cache: &mut ResourceCache, rg_builder: &mut RenderTaskGraphBuilder, stamp: FrameStamp, device_origin: DeviceIntPoint, scene_properties: &SceneProperties, data_stores: &mut DataStores, scratch: &mut ScratchBuffer, debug_flags: DebugFlags, tile_caches: &mut FastHashMap<SliceId, Box<TileCacheInstance>>, spatial_tree: &mut SpatialTree, dirty_rects_are_valid: bool, profile: &mut TransactionProfile, minimap_data: FastHashMap<ExternalScrollId, MinimapData>, chunk_pool: Arc<ChunkPool>, ) -> Frame

fn render_minimap( &self, scratch: &mut PrimitiveScratchBuffer, spatial_tree: &SpatialTree, minimap_data_store: FastHashMap<ExternalScrollId, MinimapData>, )

fn skip_occluded_pictures_with_clips( &self, tile_cache_pictures: &Vec<PictureIndex>, pictures: &mut [PictureInstance], tile_caches: &mut FastHashMap<SliceId, Box<TileCacheInstance>>, frame_context: &FrameBuildingContext<'_>, composite_state: &mut CompositeState, )

Skip pictures that are fully covered by another opaque picture with the same rounded rect clip.

Since the clip adds some transparency to the occluder, the opaque rectangles occlusion detection fails to discard stacks of opaque rounded corners. This is done specifically to avoid conflation artifacts with the rounded rectangle around web content in Firefox.

fn build_composite_pass( &self, scene: &BuiltScene, ctx: &RenderTargetContext<'_, '_>, gpu_buffers: &mut GpuBufferBuilder, deferred_resolves: &mut Vec<DeferredResolve, FrameAllocator>, composite_state: &mut CompositeState, )