Struct webrender::frame_builder::LayerRectCalculationAndCullingPass

struct LayerRectCalculationAndCullingPass<'a> {
    frame_builder: &'a mut FrameBuilder,
    screen_rect: &'a DeviceIntRect,
    clip_scroll_tree: &'a ClipScrollTree,
    auxiliary_lists_map: &'a HashMap<PipelineId, AuxiliaryLists, BuildHasherDefault<FnvHasher>>,
    resource_cache: &'a mut ResourceCache,
    profile_counters: &'a mut FrameProfileCounters,
    device_pixel_ratio: f32,
    stacking_context_stack: Vec<StackingContextIndex>,
    scroll_layer_stack: Vec<ScrollLayerIndex>,
    current_clip_stack: Vec<(PackedLayerIndex, MaskCacheInfo)>,
    current_clip_stack_scroll_layer: Option<ScrollLayerIndex>,
}

Fields
frame_builder: &'a mut FrameBuilder screen_rect: &'a DeviceIntRect clip_scroll_tree: &'a ClipScrollTree auxiliary_lists_map: &'a HashMap<PipelineId, AuxiliaryLists, BuildHasherDefault<FnvHasher>> resource_cache: &'a mut ResourceCache profile_counters: &'a mut FrameProfileCounters device_pixel_ratio: f32 stacking_context_stack: Vec<StackingContextIndex> scroll_layer_stack: Vec<ScrollLayerIndex> current_clip_stack: Vec<(PackedLayerIndex, MaskCacheInfo)>

A cached clip info stack, which should handle the most common situation, which is that we are using the same clip info stack that we were using previously.

current_clip_stack_scroll_layer: Option<ScrollLayerIndex>

The scroll layer that defines the previous scroll layer info stack.

Methods

impl<'a> LayerRectCalculationAndCullingPass<'a>

fn create_and_run(frame_builder: &'a mut FrameBuilder,
screen_rect: &'a DeviceIntRect,
clip_scroll_tree: &'a ClipScrollTree,
auxiliary_lists_map: &'a HashMap<PipelineId, AuxiliaryLists, BuildHasherDefault<FnvHasher>>,
resource_cache: &'a mut ResourceCache,
profile_counters: &'a mut FrameProfileCounters,
device_pixel_ratio: f32)

fn run(&mut self)

fn recalculate_clip_scroll_groups(&mut self)

fn compute_stacking_context_visibility(&mut self)

fn handle_pop_stacking_context(&mut self)

fn handle_push_scroll_layer(&mut self, scroll_layer_index: ScrollLayerIndex)

fn handle_push_stacking_context(&mut self,
stacking_context_index: StackingContextIndex)

fn rebuild_clip_info_stack_if_necessary(&mut self,
scroll_layer_index: ScrollLayerIndex)

fn handle_primitive_run(&mut self,
prim_index: PrimitiveIndex,
prim_count: usize)

fn handle_pop_scroll_layer(&mut self)