Struct webrender::glyph_cache::GlyphCache

pub struct GlyphCache {
    glyph_key_caches: FastHashMap<FontInstance, ResourceClassCache<GlyphKey, GlyphCacheEntry, GlyphKeyCacheInfo>>,
    current_frame: FrameId,
    bytes_used: usize,
    max_bytes_used: usize,
}

Fields
glyph_key_caches: FastHashMap<FontInstance, ResourceClassCache<GlyphKey, GlyphCacheEntry, GlyphKeyCacheInfo>>current_frame: FrameIdbytes_used: usizemax_bytes_used: usize
Implementations

impl GlyphCache

pub const DEFAULT_MAX_BYTES_USED: usize

The default space usage threshold, in bytes, after which to start pruning away old fonts.

pub fn new(max_bytes_used: usize) -> Self

pub fn get_glyph_key_cache_for_font_mut(
&mut self,
font: FontInstance
) -> &mut ResourceClassCache<GlyphKey, GlyphCacheEntry, GlyphKeyCacheInfo>

pub fn get_glyph_key_cache_for_font(
&self,
font: &FontInstance
) -> &ResourceClassCache<GlyphKey, GlyphCacheEntry, GlyphKeyCacheInfo>

pub fn clear(&mut self)

pub fn clear_fonts<F>(&mut self, key_fun: F) where
F: Fn(&'r &FontInstance) -> bool,

fn clear_evicted(
&mut self,
texture_cache: &TextureCache,
render_task_cache: &RenderTaskCache
)

Clear out evicted entries from glyph key caches.

fn clear_empty_caches(&mut self, glyph_rasterizer: &mut GlyphRasterizer)

If possible, remove entirely any empty glyph key caches.

fn prune_excess_usage(
&mut self,
texture_cache: &mut TextureCache,
render_task_cache: &RenderTaskCache
)

Check the total space usage of the glyph cache. If it exceeds the maximum usage threshold, then start clearing the oldest glyphs until below the threshold.

pub fn begin_frame(
&mut self,
stamp: FrameStamp,
texture_cache: &mut TextureCache,
render_task_cache: &RenderTaskCache,
glyph_rasterizer: &mut GlyphRasterizer
)