Struct webrender::gpu_cache::GpuCache

pub struct GpuCache {
    now: FrameStamp,
    texture: Texture,
    saved_block_count: usize,
    debug_flags: DebugFlags,
    pending_clear: bool,
    prepared_for_frames: bool,
    requires_frame_build: bool,
    document_frames_to_build: HashSet<DocumentId, BuildHasherDefault<FxHasher>>,
}

The main LRU cache interface.

Fields
now: FrameStamp

Current FrameId.

texture: Texture

CPU-side texture allocator.

saved_block_count: usize

Number of blocks requested this frame that don’t need to be re-uploaded.

debug_flags: DebugFlags

The current debug flags for the system.

pending_clear: bool

Whether there is a pending clear to send with the next update.

prepared_for_frames: bool

Indicates that prepare_for_frames has been called for this group of frames. Used for sanity checks.

requires_frame_build: bool

This indicates that we performed a cleanup operation which requires all documents to build a frame.

document_frames_to_build: HashSet<DocumentId, BuildHasherDefault<FxHasher>>

The set of documents which have had frames built in this update. Used for sanity checks.

Implementations

impl GpuCache

pub fn new() -> Self

pub fn clear(&mut self)

Drops everything in the GPU cache. Must not be called once gpu cache entries for the next frame have already been requested.

pub fn requires_frame_build(&self) -> bool

pub fn prepare_for_frames(&mut self)

pub fn bookkeep_after_frames(&mut self)

pub fn begin_frame(&mut self, stamp: FrameStamp)

Begin a new frame.

pub fn invalidate(&mut self, handle: &GpuCacheHandle)

pub fn request<'a>(
&'a mut self,
handle: &'a mut GpuCacheHandle
) -> Option<GpuDataRequest<'a>>

Request a resource be added to the cache. If the resource is already in the cache, None will be returned.

pub fn push_per_frame_blocks(
&mut self,
blocks: &[GpuBlockData]
) -> GpuCacheHandle

pub fn push_deferred_per_frame_blocks(
&mut self,
block_count: usize
) -> GpuCacheHandle

pub fn end_frame(
&mut self,
profile_counters: &mut GpuCacheProfileCounters
) -> FrameStamp

End the frame. Return the list of updates to apply to the device specific cache texture.

pub fn should_reclaim_memory(&self) -> bool

Returns true if utilization has been low enough for long enough that we should blow the cache away and rebuild it.

pub fn extract_updates(&mut self) -> GpuCacheUpdateList

Extract the pending updates from the cache.

pub fn set_debug_flags(&mut self, flags: DebugFlags)

Sets the current debug flags for the system.

pub fn get_address(&self, id: &GpuCacheHandle) -> GpuCacheAddress

Get the actual GPU address in the texture for a given slot ID. It’s assumed at this point that the given slot has been requested and built for this frame. Attempting to get the address for a freed or pending slot will panic!