Module webrender::gpu_types

Structs

BlurInstance

BorderInstance

BoxShadowData

BrushFlags

Flags that define how the common brush shader code should process this instance.

BrushInstance

Convenience structure to encode into PrimitiveInstanceData.

ClearInstance

Vertex format for issuing colored quads.

ClipMaskInstance

A clipping primitive drawn into the clipping mask. Could be an image or a rectangle, which defines the way address is treated.

ClipMaskInstanceBoxShadow

ClipMaskInstanceCommon

ClipMaskInstanceRect

CompositeInstance

Vertex format for picture cache composite shader. When editing the members, update desc::COMPOSITE so its list of instance_attributes matches:

CompositorTransform

A GPU-friendly representation of the ScaleOffset type

CopyInstance

GlyphInstance

ImageBrushData

Convenience structure to encode into the image brush’s user data.

ImageSource

MaskInstance

PrimitiveHeader

PrimitiveHeaderF

PrimitiveHeaderI

PrimitiveHeaderIndex

PrimitiveHeaders

PrimitiveInstanceData

QuadInstance

QuadSegment

RelativeTransformKey 🔒

ScalingInstance

SplitCompositeInstance

SvgFilterInstance

TransformData

The GPU data payload for a transform palette entry.

TransformMetadata

TransformPalette

TransformPaletteId

ZBufferId

ZBufferIdGenerator

Enums

BlurDirection

BorderSegment

BoxShadowStretchMode

ClipSpace

RasterizationSpace

UvRectKind

Constants

UV_TYPE_NORMALIZED 🔒

Specifies that an RGB CompositeInstance’s UV coordinates are normalized.

UV_TYPE_UNNORMALIZED 🔒

Specifies that an RGB CompositeInstance’s UV coordinates are not normalized.

VECS_PER_TRANSFORM

Functions

get_shader_opacity

register_transform 🔒