webrender

Module hit_test

Structs

HitTest

HitTestClipNode 🔒

HitTestSpatialNode 🔒

A copy of important spatial node data to use during hit testing. This a copy of data from the SpatialTree that will persist as a new frame is under construction, allowing hit tests consistent with the currently rendered frame.

HitTester

HitTestingItem 🔒

HitTestingScene

Defines the immutable part of a hit tester for a given scene. The hit tester is recreated each time a frame is built, since it relies on the current values of the spatial tree. However, the clip chain and item definitions don’t change, so they are created once per scene, and shared between hit tester instances via Arc.

HitTestingSceneStats

Statistics about allocation sizes of current hit tester, used to pre-allocate size of the next hit tester.

SharedHitTester

Enums

HitTestRegion 🔒