webrender

Module image_source

This module contains the logic to obtain a primitive’s source texture and uv rect.

Currently this is a somewhat involved process because the code grew into having ad-hoc ways to store this information depending on how the image data is produced. The goal is for any textured primitive to be able to read from any source (texture cache, render tasks, etc.) without primitive-specific code.

Functions

resolve_cached_render_task

resolve_image

Resolve a resource cache’s imagre request into a texture cache item.