Struct webrender::intern::DataStore

pub struct DataStore<I: Internable> {
    items: Vec<Option<I::StoreData>>,
}

The data store lives in the frame builder thread. It contains a free-list of items for fast access.

Fields
items: Vec<Option<I::StoreData>>
Implementations

impl<I: Internable> DataStore<I>

pub fn apply_updates(
&mut self,
update_list: UpdateList<I::Key>,
profile_counter: &mut ResourceProfileCounter
)

Apply any updates from the scene builder thread to this data store.