webrender::intern

Struct Interner

pub struct Interner<I: Internable> {
    map: FastHashMap<I::Key, ItemDetails<I>>,
    free_list: Vec<usize>,
    update_list: UpdateList<I::Key>,
    current_epoch: Epoch,
    local_data: Vec<I::InternData>,
}

The main interning data structure. This lives in the scene builder thread, and handles hashing and interning unique data structures. It also manages a free-list for the items in the data store, which is synchronized via an update list of additions / removals.

Fields
map: FastHashMap<I::Key, ItemDetails<I>>

Uniquely map an interning key to a handle

free_list: Vec<usize>

List of free slots in the data store for re-use.

update_list: UpdateList<I::Key>

Pending list of updates that need to be applied.

current_epoch: Epoch

The current epoch for the interner.

local_data: Vec<I::InternData>

The information associated with each interned item that can be accessed by the interner.

Implementations

impl<I: Internable> Interner<I>

pub fn intern<F>(&mut self, data: &I::Key, fun: F) -> Handle<I>
where F: FnOnce() -> I::InternData,

Intern a data structure, and return a handle to that data. The handle can then be stored in the frame builder, and safely accessed via the data store that lives in the frame builder thread. The provided closure is invoked to build the local data about an interned structure if the key isn’t already interned.

pub fn end_frame_and_get_pending_updates(&mut self) -> UpdateList<I::Key>

Retrieve the pending list of updates for an interner that need to be applied to the data store. Also run a GC step that removes old entries.