Enum webrender::internal_types::FilterGraphOp

pub enum FilterGraphOp {

    SVGFESourceGraphic,
    SVGFESourceAlpha,
    SVGFEIdentity,
    SVGFEOpacity {
        valuebinding: PropertyBinding<f32>,
        value: f32,
    },
    SVGFEToAlpha,
    SVGFEBlendDarken,
    SVGFEBlendLighten,
    SVGFEBlendMultiply,
    SVGFEBlendNormal,
    SVGFEBlendScreen,
    SVGFEBlendOverlay,
    SVGFEBlendColorDodge,
    SVGFEBlendColorBurn,
    SVGFEBlendHardLight,
    SVGFEBlendSoftLight,
    SVGFEBlendDifference,
    SVGFEBlendExclusion,
    SVGFEBlendHue,
    SVGFEBlendSaturation,
    SVGFEBlendColor,
    SVGFEBlendLuminosity,
    SVGFEColorMatrix {
        values: [f32; 20],
    },
    SVGFEComponentTransfer,
    SVGFEComponentTransferInterned {
        handle: Handle<FilterDataIntern>,
        creates_pixels: bool,
    },
    SVGFECompositeArithmetic {
        k1: f32,
        k2: f32,
        k3: f32,
        k4: f32,
    },
    SVGFECompositeATop,
    SVGFECompositeIn,
    SVGFECompositeLighter,
    SVGFECompositeOut,
    SVGFECompositeOver,
    SVGFECompositeXOR,
    SVGFEConvolveMatrixEdgeModeDuplicate {
        order_x: i32,
        order_y: i32,
        kernel: [f32; 25],
        divisor: f32,
        bias: f32,
        target_x: i32,
        target_y: i32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        preserve_alpha: i32,
    },
    SVGFEConvolveMatrixEdgeModeNone {
        order_x: i32,
        order_y: i32,
        kernel: [f32; 25],
        divisor: f32,
        bias: f32,
        target_x: i32,
        target_y: i32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        preserve_alpha: i32,
    },
    SVGFEConvolveMatrixEdgeModeWrap {
        order_x: i32,
        order_y: i32,
        kernel: [f32; 25],
        divisor: f32,
        bias: f32,
        target_x: i32,
        target_y: i32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        preserve_alpha: i32,
    },
    SVGFEDiffuseLightingDistant {
        surface_scale: f32,
        diffuse_constant: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        azimuth: f32,
        elevation: f32,
    },
    SVGFEDiffuseLightingPoint {
        surface_scale: f32,
        diffuse_constant: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        x: f32,
        y: f32,
        z: f32,
    },
    SVGFEDiffuseLightingSpot {
        surface_scale: f32,
        diffuse_constant: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        x: f32,
        y: f32,
        z: f32,
        points_at_x: f32,
        points_at_y: f32,
        points_at_z: f32,
        cone_exponent: f32,
        limiting_cone_angle: f32,
    },
    SVGFEDisplacementMap {
        scale: f32,
        x_channel_selector: u32,
        y_channel_selector: u32,
    },
    SVGFEDropShadow {
        color: ColorF,
        dx: f32,
        dy: f32,
        std_deviation_x: f32,
        std_deviation_y: f32,
    },
    SVGFEFlood {
        color: ColorF,
    },
    SVGFEGaussianBlur {
        std_deviation_x: f32,
        std_deviation_y: f32,
    },
    SVGFEImage {
        sampling_filter: u32,
        matrix: [f32; 6],
    },
    SVGFEMorphologyDilate {
        radius_x: f32,
        radius_y: f32,
    },
    SVGFEMorphologyErode {
        radius_x: f32,
        radius_y: f32,
    },
    SVGFESpecularLightingDistant {
        surface_scale: f32,
        specular_constant: f32,
        specular_exponent: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        azimuth: f32,
        elevation: f32,
    },
    SVGFESpecularLightingPoint {
        surface_scale: f32,
        specular_constant: f32,
        specular_exponent: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        x: f32,
        y: f32,
        z: f32,
    },
    SVGFESpecularLightingSpot {

        surface_scale: f32,
        specular_constant: f32,
        specular_exponent: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        x: f32,
        y: f32,
        z: f32,
        points_at_x: f32,
        points_at_y: f32,
        points_at_z: f32,
        cone_exponent: f32,
        limiting_cone_angle: f32,
    
},
    SVGFETile,
    SVGFETurbulenceWithFractalNoiseWithNoStitching {
        base_frequency_x: f32,
        base_frequency_y: f32,
        num_octaves: u32,
        seed: u32,
    },
    SVGFETurbulenceWithFractalNoiseWithStitching {
        base_frequency_x: f32,
        base_frequency_y: f32,
        num_octaves: u32,
        seed: u32,
    },
    SVGFETurbulenceWithTurbulenceNoiseWithNoStitching {
        base_frequency_x: f32,
        base_frequency_y: f32,
        num_octaves: u32,
        seed: u32,
    },
    SVGFETurbulenceWithTurbulenceNoiseWithStitching {
        base_frequency_x: f32,
        base_frequency_y: f32,
        num_octaves: u32,
        seed: u32,
    },

}

Variants

SVGFESourceGraphic

Filter that copies the SourceGraphic image into the specified subregion, This is intentionally the only way to get SourceGraphic into the graph, as the filter region must be applied before it is used. parameters: FilterOpGraphNode SVG filter semantics - no inputs, no linear

SVGFESourceAlpha

Filter that copies the SourceAlpha image into the specified subregion, This is intentionally the only way to get SourceAlpha into the graph, as the filter region must be applied before it is used. parameters: FilterOpGraphNode SVG filter semantics - no inputs, no linear

SVGFEIdentity

Filter that does no transformation of the colors, used to implement a few things like SVGFEOffset, and this is the default value in impl_default_for_enums. parameters: FilterGraphNode SVG filter semantics - selectable input with offset

SVGFEOpacity

represents CSS opacity property as a graph node like the rest of the SVGFE* filters parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations

Fields

valuebinding: PropertyBinding<f32>

value: f32

SVGFEToAlpha

convert a color image to an alpha channel - internal use; generated by SVGFilterInstance::GetOrCreateSourceAlphaIndex().

SVGFEBlendDarken

combine 2 images with SVG_FEBLEND_MODE_DARKEN parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feBlendElement

SVGFEBlendLighten

combine 2 images with SVG_FEBLEND_MODE_LIGHTEN parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feBlendElement

SVGFEBlendMultiply

combine 2 images with SVG_FEBLEND_MODE_MULTIPLY parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feBlendElement

SVGFEBlendNormal

combine 2 images with SVG_FEBLEND_MODE_NORMAL parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feBlendElement

SVGFEBlendScreen

combine 2 images with SVG_FEBLEND_MODE_SCREEN parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feBlendElement

SVGFEBlendOverlay

combine 2 images with SVG_FEBLEND_MODE_OVERLAY parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

SVGFEBlendColorDodge

combine 2 images with SVG_FEBLEND_MODE_COLOR_DODGE parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

SVGFEBlendColorBurn

combine 2 images with SVG_FEBLEND_MODE_COLOR_BURN parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

SVGFEBlendHardLight

combine 2 images with SVG_FEBLEND_MODE_HARD_LIGHT parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

SVGFEBlendSoftLight

combine 2 images with SVG_FEBLEND_MODE_SOFT_LIGHT parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

SVGFEBlendDifference

combine 2 images with SVG_FEBLEND_MODE_DIFFERENCE parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

SVGFEBlendExclusion

combine 2 images with SVG_FEBLEND_MODE_EXCLUSION parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

SVGFEBlendHue

combine 2 images with SVG_FEBLEND_MODE_HUE parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

SVGFEBlendSaturation

combine 2 images with SVG_FEBLEND_MODE_SATURATION parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

SVGFEBlendColor

combine 2 images with SVG_FEBLEND_MODE_COLOR parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

SVGFEBlendLuminosity

combine 2 images with SVG_FEBLEND_MODE_LUMINOSITY parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

SVGFEColorMatrix

transform colors of image through 5x4 color matrix (transposed for efficiency) parameters: FilterGraphNode, matrix[5][4] SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feColorMatrixElement

Fields

values: [f32; 20]

SVGFEComponentTransfer

transform colors of image through configurable gradients with component swizzle parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feComponentTransferElement

SVGFEComponentTransferInterned

Processed version of SVGFEComponentTransfer with the FilterData replaced by an interned handle, this is made in wrap_prim_with_filters. Aside from the interned handle, creates_pixels indicates if the transfer parameters will probably fill the entire subregion with non-zero alpha.

Fields

handle: Handle<FilterDataIntern>

creates_pixels: bool

SVGFECompositeArithmetic

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterGraphNode, k1, k2, k3, k4 SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

Fields

k1: f32

k2: f32

k3: f32

k4: f32

SVGFECompositeATop

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

SVGFECompositeIn

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

SVGFECompositeLighter

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Docs: https://developer.mozilla.org/en-US/docs/Web/SVG/Element/feComposite

SVGFECompositeOut

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

SVGFECompositeOver

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

SVGFECompositeXOR

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

SVGFEConvolveMatrixEdgeModeDuplicate

transform image through convolution matrix of up to 25 values (spec allows more but for performance reasons we do not) parameters: FilterGraphNode, orderX, orderY, kernelValues[25], divisor, bias, targetX, targetY, kernelUnitLengthX, kernelUnitLengthY, preserveAlpha SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feConvolveMatrixElement

Fields

order_x: i32

order_y: i32

kernel: [f32; 25]

divisor: f32

bias: f32

target_x: i32

target_y: i32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

preserve_alpha: i32

SVGFEConvolveMatrixEdgeModeNone

transform image through convolution matrix of up to 25 values (spec allows more but for performance reasons we do not) parameters: FilterGraphNode, orderX, orderY, kernelValues[25], divisor, bias, targetX, targetY, kernelUnitLengthX, kernelUnitLengthY, preserveAlpha SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feConvolveMatrixElement

Fields

order_x: i32

order_y: i32

kernel: [f32; 25]

divisor: f32

bias: f32

target_x: i32

target_y: i32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

preserve_alpha: i32

SVGFEConvolveMatrixEdgeModeWrap

transform image through convolution matrix of up to 25 values (spec allows more but for performance reasons we do not) parameters: FilterGraphNode, orderX, orderY, kernelValues[25], divisor, bias, targetX, targetY, kernelUnitLengthX, kernelUnitLengthY, preserveAlpha SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feConvolveMatrixElement

Fields

order_x: i32

order_y: i32

kernel: [f32; 25]

divisor: f32

bias: f32

target_x: i32

target_y: i32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

preserve_alpha: i32

SVGFEDiffuseLightingDistant

calculate lighting based on heightmap image with provided values for a distant light source with specified direction parameters: FilterGraphNode, surfaceScale, diffuseConstant, kernelUnitLengthX, kernelUnitLengthY, azimuth, elevation SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDiffuseLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDistantLightElement

Fields

surface_scale: f32

diffuse_constant: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

azimuth: f32

elevation: f32

SVGFEDiffuseLightingPoint

calculate lighting based on heightmap image with provided values for a point light source at specified location parameters: FilterGraphNode, surfaceScale, diffuseConstant, kernelUnitLengthX, kernelUnitLengthY, x, y, z SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDiffuseLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEPointLightElement

Fields

surface_scale: f32

diffuse_constant: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

x: f32

y: f32

z: f32

SVGFEDiffuseLightingSpot

calculate lighting based on heightmap image with provided values for a spot light source at specified location pointing at specified target location with specified hotspot sharpness and cone angle parameters: FilterGraphNode, surfaceScale, diffuseConstant, kernelUnitLengthX, kernelUnitLengthY, x, y, z, pointsAtX, pointsAtY, pointsAtZ, specularExponent, limitingConeAngle SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDiffuseLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFESpotLightElement

Fields

surface_scale: f32

diffuse_constant: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

x: f32

y: f32

z: f32

points_at_x: f32

points_at_y: f32

points_at_z: f32

cone_exponent: f32

limiting_cone_angle: f32

SVGFEDisplacementMap

calculate a distorted version of first input image using offset values from second input image at specified intensity parameters: FilterGraphNode, scale, xChannelSelector, yChannelSelector SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDisplacementMapElement

Fields

scale: f32

x_channel_selector: u32

y_channel_selector: u32

SVGFEDropShadow

create and merge a dropshadow version of the specified image’s alpha channel with specified offset and blur radius parameters: FilterGraphNode, flood_color, flood_opacity, dx, dy, stdDeviationX, stdDeviationY SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDropShadowElement

Fields

color: ColorF

dx: f32

dy: f32

std_deviation_x: f32

std_deviation_y: f32

SVGFEFlood

synthesize a new image of specified size containing a solid color parameters: FilterGraphNode, color SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEFloodElement

Fields

color: ColorF

SVGFEGaussianBlur

create a blurred version of the input image parameters: FilterGraphNode, stdDeviationX, stdDeviationY SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEGaussianBlurElement

Fields

std_deviation_x: f32

std_deviation_y: f32

SVGFEImage

synthesize a new image based on a url (i.e. blob image source) parameters: FilterGraphNode, samplingFilter (see SamplingFilter in Types.h), transform SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEImageElement

Fields

sampling_filter: u32

matrix: [f32; 6]

SVGFEMorphologyDilate

create a new image based on the input image with the contour stretched outward (dilate operator) parameters: FilterGraphNode, radiusX, radiusY SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEMorphologyElement

Fields

radius_x: f32

radius_y: f32

SVGFEMorphologyErode

create a new image based on the input image with the contour shrunken inward (erode operator) parameters: FilterGraphNode, radiusX, radiusY SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEMorphologyElement

Fields

radius_x: f32

radius_y: f32

SVGFESpecularLightingDistant

calculate lighting based on heightmap image with provided values for a distant light source with specified direction parameters: FilerData, surfaceScale, specularConstant, specularExponent, kernelUnitLengthX, kernelUnitLengthY, azimuth, elevation SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFESpecularLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDistantLightElement

Fields

surface_scale: f32

specular_constant: f32

specular_exponent: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

azimuth: f32

elevation: f32

SVGFESpecularLightingPoint

calculate lighting based on heightmap image with provided values for a point light source at specified location parameters: FilterGraphNode, surfaceScale, specularConstant, specularExponent, kernelUnitLengthX, kernelUnitLengthY, x, y, z SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFESpecularLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEPointLightElement

Fields

surface_scale: f32

specular_constant: f32

specular_exponent: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

x: f32

y: f32

z: f32

SVGFESpecularLightingSpot

calculate lighting based on heightmap image with provided values for a spot light source at specified location pointing at specified target location with specified hotspot sharpness and cone angle parameters: FilterGraphNode, surfaceScale, specularConstant, specularExponent, kernelUnitLengthX, kernelUnitLengthY, x, y, z, pointsAtX, pointsAtY, pointsAtZ, specularExponent, limitingConeAngle SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFESpecularLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFESpotLightElement

Fields

surface_scale: f32

specular_constant: f32

specular_exponent: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

x: f32

y: f32

z: f32

points_at_x: f32

points_at_y: f32

points_at_z: f32

cone_exponent: f32

limiting_cone_angle: f32

SVGFETile

create a new image based on the input image, repeated throughout the output rectangle parameters: FilterGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFETileElement

SVGFETurbulenceWithFractalNoiseWithNoStitching

synthesize a new image based on Fractal Noise (Perlin) with the chosen stitching mode parameters: FilterGraphNode, baseFrequencyX, baseFrequencyY, numOctaves, seed SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFETurbulenceElement

Fields

base_frequency_x: f32

base_frequency_y: f32

num_octaves: u32

seed: u32

SVGFETurbulenceWithFractalNoiseWithStitching

synthesize a new image based on Fractal Noise (Perlin) with the chosen stitching mode parameters: FilterGraphNode, baseFrequencyX, baseFrequencyY, numOctaves, seed SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFETurbulenceElement

Fields

base_frequency_x: f32

base_frequency_y: f32

num_octaves: u32

seed: u32

SVGFETurbulenceWithTurbulenceNoiseWithNoStitching

synthesize a new image based on Turbulence Noise (offset vectors) parameters: FilterGraphNode, baseFrequencyX, baseFrequencyY, numOctaves, seed SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFETurbulenceElement

Fields

base_frequency_x: f32

base_frequency_y: f32

num_octaves: u32

seed: u32

SVGFETurbulenceWithTurbulenceNoiseWithStitching

synthesize a new image based on Turbulence Noise (offset vectors) parameters: FilterGraphNode, baseFrequencyX, baseFrequencyY, numOctaves, seed SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFETurbulenceElement

Fields

base_frequency_x: f32

base_frequency_y: f32

num_octaves: u32

seed: u32

Implementations

impl FilterGraphOp

pub fn kind(&self) -> &'static str