Enum webrender::internal_types::TextureUpdateDetails

pub enum TextureUpdateDetails {
    Raw,
    Blit(Vec<u8>),
    Blur(Vec<u8>, Size2D<u32>, Au, TextureImage, TextureImage, BorderType),
    BorderRadius(DevicePixel, DevicePixel, DevicePixel, DevicePixel, Option<u32>, bool, BorderType),
    BoxShadow(DevicePixel, DevicePixel, Size2D<DevicePixel>, Point2D<DevicePixel>, bool, BorderType),
}

Variants
RawBlit(Vec<u8>)Blur(Vec<u8>, Size2D<u32>, Au, TextureImage, TextureImage, BorderType)BorderRadius(DevicePixel, DevicePixel, DevicePixel, DevicePixel, Option<u32>, bool, BorderType)

All four corners, the tessellation index, and whether inverted, respectively.

BoxShadow(DevicePixel, DevicePixel, Size2D<DevicePixel>, Point2D<DevicePixel>, bool, BorderType)

Blur radius, border radius, box size, raster origin, and whether inverted, respectively.