Struct webrender::internal_types::RendererFrame

pub struct RendererFrame {
    pub pipeline_epoch_map: HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>>,
    pub layers_bouncing_back: HashSet<ScrollLayerId, BuildHasherDefault<FnvHasher>>,
    pub frame: Option<Frame>,
}

Mostly wraps a tiling::Frame, adding a bit of extra information.

Fields
pipeline_epoch_map: HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>>

The last rendered epoch for each pipeline present in the frame. This information is used to know if a certain transformation on the layout has been rendered, which is necessary for reftests.

layers_bouncing_back: HashSet<ScrollLayerId, BuildHasherDefault<FnvHasher>>

The layers that are currently affected by the over-scrolling animation.

frame: Option<Frame>
Methods

impl RendererFrame

fn new(pipeline_epoch_map: HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>>,
layers_bouncing_back: HashSet<ScrollLayerId, BuildHasherDefault<FnvHasher>>,
frame: Option<Frame>)
-> RendererFrame