webrender

Module invalidation

Invalidation tracking and dirty region management

This module contains types and logic for tracking dirty regions and dependencies used to determine what needs to be redrawn each frame.

Modules

cached_surface

compare

Dependency tracking for tile invalidation

quadtree

Quadtree-based dirty region tracking for tiles

vert_buffer

Quantized raster-space vertex buffer for output-space tile invalidation.

Structs

DirtyRegion

Represents the dirty region of a tile cache picture, relative to a “visibility” spatial node. At the moment the visibility node is world space, but the plan is to switch to raster space.

Enums

InvalidationReason

Debugging information about why a tile was invalidated

PrimitiveCompareResult

The result of a primitive dependency comparison. Size is a u8 since this is a hot path in the code, and keeping the data small is a performance win.