webrender::invalidation::cached_surface

Struct CachedSurface

pub struct CachedSurface {
    pub current_descriptor: CachedSurfaceDescriptor,
    pub prev_descriptor: CachedSurfaceDescriptor,
    pub is_valid: bool,
    pub local_valid_rect: PictureBox2D,
    pub local_dirty_rect: PictureRect,
    pub local_rect: PictureRect,
    pub root: TileNode,
    pub background_color: Option<ColorF>,
    pub invalidation_reason: Option<InvalidationReason>,
    pub sub_graphs: Vec<(PictureRect, Vec<(PictureCompositeMode, SurfaceIndex)>)>,
}

Fields
current_descriptor: CachedSurfaceDescriptorprev_descriptor: CachedSurfaceDescriptoris_valid: boollocal_valid_rect: PictureBox2Dlocal_dirty_rect: PictureRectlocal_rect: PictureRectroot: TileNodebackground_color: Option<ColorF>invalidation_reason: Option<InvalidationReason>sub_graphs: Vec<(PictureRect, Vec<(PictureCompositeMode, SurfaceIndex)>)>
Implementations

impl CachedSurface

pub fn new() -> Self

pub fn print(&self, pt: &mut dyn PrintTreePrinter)

pub fn pre_update( &mut self, background_color: Option<ColorF>, local_tile_rect: PictureRect, frame_id: FrameId, is_visible: bool, )

Setup state before primitive dependency calculation.

pub fn add_prim_dependency( &mut self, info: &PrimitiveDependencyInfo, corners_cache: &CornersCache, prim_clamp_to_tile: bool, local_raster_rect: &RasterRect, local_tile_rect: PictureRect, )

fn update_dirty_rects( &mut self, ctx: &TileUpdateDirtyContext<'_>, state: &mut TileUpdateDirtyState<'_>, invalidation_reason: &mut Option<InvalidationReason>, frame_context: &FrameVisibilityContext<'_>, ) -> PictureRect

Check if the content of the previous and current tile descriptors match

pub fn update_content_validity( &mut self, ctx: &TileUpdateDirtyContext<'_>, state: &mut TileUpdateDirtyState<'_>, frame_context: &FrameVisibilityContext<'_>, )

Invalidate a tile based on change in content. This must be called even if the tile is not currently visible on screen. We might be able to improve this later by changing how ComparableVec is used.

pub fn invalidate( &mut self, invalidation_rect: Option<PictureRect>, reason: InvalidationReason, )

Invalidate this tile. If invalidation_rect is None, the entire tile is invalidated.