Struct webrender::layer::Layer

pub struct Layer {
    pub scrolling: ScrollingState,
    pub content_size: LayerSize,
    pub local_viewport_rect: LayerRect,
    pub combined_local_viewport_rect: LayerRect,
    pub world_viewport_transform: LayerToWorldTransform,
    pub world_content_transform: LayerToWorldTransform,
    pub local_transform: LayerToScrollTransform,
    pub pipeline_id: PipelineId,
    pub children: Vec<ScrollLayerId>,
}

Contains scrolling and transform information stacking contexts.

Fields
scrolling: ScrollingState

Manages scrolling offset, overscroll state etc.

content_size: LayerSize

Size of the content inside the scroll region (in logical pixels)

local_viewport_rect: LayerRect

Viewing rectangle

combined_local_viewport_rect: LayerRect

Viewing rectangle clipped against parent layer(s)

world_viewport_transform: LayerToWorldTransform

World transform for the viewport rect itself.

world_content_transform: LayerToWorldTransform

World transform for content within this layer

local_transform: LayerToScrollTransform

Transform for this layer, relative to parent scrollable layer.

pipeline_id: PipelineId

Pipeline that this layer belongs to

children: Vec<ScrollLayerId>

Child layers

Methods

impl Layer

fn new(local_viewport_rect: &LayerRect,
content_size: LayerSize,
local_transform: &LayerToScrollTransform,
pipeline_id: PipelineId)
-> Layer

fn add_child(&mut self, child: ScrollLayerId)

fn finalize(&mut self, scrolling: &ScrollingState)

fn overscroll_amount(&self) -> LayerSize

fn set_scroll_origin(&mut self, origin: &LayerPoint) -> bool

fn update_transform(&mut self,
parent_world_transform: &ScrollToWorldTransform,
parent_viewport_rect: &ScrollLayerRect)

fn scrollable_height(&self) -> f32

fn scrollable_width(&self) -> f32

fn scroll(&mut self,
scroll_location: ScrollLocation,
phase: ScrollEventPhase)
-> bool

fn stretch_overscroll_spring(&mut self)

fn tick_scrolling_bounce_animation(&mut self)

fn ray_intersects_layer(&self, cursor: &WorldPoint) -> bool