webrender

Module lru_cache

Structs

Item 🔒

Stores a strong handle controlling the lifetime of the data in the LRU cache, and a doubly-linked list node specifying where in the current LRU order this element exists. These items are themselves backed by a freelist to minimize heap allocations and improve cache access patterns.

ItemIndex 🔒

Index of an LRU tracking element

LRUCache

The main public interface to the LRU cache

LRUCacheEntry 🔒

Stores the data supplied by the user to be cached, and an index into the LRU tracking freelist for this element.

LRUTracker 🔒

Internal implementation of the LRU tracking list