Enum webrender::picture::PictureCompositeMode

pub enum PictureCompositeMode {
    MixBlend(MixBlendMode),
    Filter(Filter),
    ComponentTransferFilter(Handle<FilterDataIntern>),
    Blit(BlitReason),
    TileCache {
        slice_id: SliceId,
    },
    SvgFilter(Vec<FilterPrimitive>, Vec<SFilterData>),
    IntermediateSurface,
}

Specifies how this Picture should be composited onto the target it belongs to.

Variants

MixBlend(MixBlendMode)

Apply CSS mix-blend-mode effect.

Filter(Filter)

Apply a CSS filter (except component transfer).

ComponentTransferFilter(Handle<FilterDataIntern>)

Apply a component transfer filter.

Blit(BlitReason)

Draw to intermediate surface, copy straight across. This is used for CSS isolation, and plane splitting.

TileCache

Fields

slice_id: SliceId

Used to cache a picture as a series of tiles.

SvgFilter(Vec<FilterPrimitive>, Vec<SFilterData>)

Apply an SVG filter

IntermediateSurface

A surface that is used as an input to another primitive

Implementations

impl PictureCompositeMode

pub fn get_rect( &self, surface: &SurfaceInfo, sub_rect: Option<LayoutRect> ) -> LayoutRect

pub fn get_coverage( &self, surface: &SurfaceInfo, sub_rect: Option<LayoutRect> ) -> LayoutRect