Struct webrender::picture::CompareHelper

struct CompareHelper<'a, T> {
    offset_curr: usize,
    offset_prev: usize,
    curr_items: &'a [T],
    prev_items: &'a [T],
}

A small helper to compare two arrays of primitive dependencies.

Fields
offset_curr: usizeoffset_prev: usizecurr_items: &'a [T]prev_items: &'a [T]
Methods

impl<'a, T> CompareHelper<'a, T> where
T: PartialEq,

fn new(prev_items: &'a [T], curr_items: &'a [T]) -> Self

Construct a new compare helper for a current / previous set of dependency information.

fn reset(&mut self)

Reset the current position in the dependency array to the start

fn is_same<F>(&self, prev_count: u8, curr_count: u8, f: F) -> bool where
F: Fn(&T) -> bool,

Test if two sections of the dependency arrays are the same, by checking both item equality, and a user closure to see if the content of the item changed.

fn advance_prev(&mut self, count: u8)

fn advance_curr(&mut self, count: u8)