Struct webrender::picture::PictureCacheState

pub struct PictureCacheState {
    pub tiles: FastHashMap<TileOffset, Box<Tile>>,
    spatial_node_comparer: SpatialNodeComparer,
    opacity_bindings: FastHashMap<PropertyBindingId, BindingInfo<f32>>,
    color_bindings: FastHashMap<PropertyBindingId, BindingInfo<ColorU>>,
    root_transform: TransformKey,
    current_tile_size: DeviceIntSize,
    allocations: PictureCacheRecycledAllocations,
    pub native_surface: Option<NativeSurface>,
    pub external_native_surface_cache: FastHashMap<ExternalNativeSurfaceKey, ExternalNativeSurface>,
    virtual_offset: DeviceIntPoint,
    frame_id: FrameId,
}

Picture-caching state to keep between scenes.

Fields
tiles: FastHashMap<TileOffset, Box<Tile>>

The tiles retained by this picture cache.

spatial_node_comparer: SpatialNodeComparer

State of the spatial nodes from previous frame

opacity_bindings: FastHashMap<PropertyBindingId, BindingInfo<f32>>

State of opacity bindings from previous frame

color_bindings: FastHashMap<PropertyBindingId, BindingInfo<ColorU>>

State of color bindings from previous frame

root_transform: TransformKey

The current transform of the picture cache root spatial node

current_tile_size: DeviceIntSize

The current tile size in device pixels

allocations: PictureCacheRecycledAllocations

Various allocations we want to avoid re-doing.

native_surface: Option<NativeSurface>

Currently allocated native compositor surface for this picture cache.

external_native_surface_cache: FastHashMap<ExternalNativeSurfaceKey, ExternalNativeSurface>

A cache of compositor surfaces that are retained between display lists

virtual_offset: DeviceIntPoint

The retained virtual offset for this slice between display lists.

frame_id: FrameId

Current frame ID of this picture cache