Struct webrender::picture::PicturePrimitive

pub struct PicturePrimitive {
    pub prim_list: PrimitiveList,
    pub state: Option<PictureState>,
    pub apply_local_clip_rect: bool,
    pub is_backface_visible: bool,
    pub secondary_render_task_id: Option<RenderTaskId>,
    pub requested_composite_mode: Option<PictureCompositeMode>,
    pub requested_raster_space: RasterSpace,
    pub raster_config: Option<RasterConfig>,
    pub context_3d: Picture3DContext<OrderedPictureChild>,
    pub frame_output_pipeline_id: Option<PipelineId>,
    pub extra_gpu_data_handles: SmallVec<[GpuCacheHandle; 1]>,
    pub spatial_node_index: SpatialNodeIndex,
    pub estimated_local_rect: LayoutRect,
    pub precise_local_rect: LayoutRect,
    pub segments_are_valid: bool,
    pub tile_cache: Option<Box<TileCacheInstance>>,
    pub options: PictureOptions,
    num_render_tasks: usize,
}

Fields
prim_list: PrimitiveList

List of primitives, and associated info for this picture.

state: Option<PictureState>apply_local_clip_rect: bool

If true, apply the local clip rect to primitive drawn in this picture.

is_backface_visible: bool

If false and transform ends up showing the back of the picture, it will be considered invisible.

secondary_render_task_id: Option<RenderTaskId>requested_composite_mode: Option<PictureCompositeMode>

How this picture should be composited. If None, don't composite - just draw directly on parent surface.

requested_raster_space: RasterSpace

Requested rasterization space for this picture. It is a performance hint only.

raster_config: Option<RasterConfig>context_3d: Picture3DContext<OrderedPictureChild>frame_output_pipeline_id: Option<PipelineId>extra_gpu_data_handles: SmallVec<[GpuCacheHandle; 1]>spatial_node_index: SpatialNodeIndex

The spatial node index of this picture when it is composited into the parent picture.

estimated_local_rect: LayoutRect

The conservative local rect of this picture. It is built dynamically during the first picture traversal. It is composed of already snapped primitives.

precise_local_rect: LayoutRect

The local rect of this picture. It is built dynamically during the frame visibility update. It differs from the estimated_local_rect because it will not contain culled primitives, takes into account surface inflation and the whole clip chain. It is frequently the same, but may be quite different depending on how much was culled.

segments_are_valid: bool

If false, this picture needs to (re)build segments if it supports segment rendering. This can occur if the local rect of the picture changes due to transform animation and/or scrolling.

tile_cache: Option<Box<TileCacheInstance>>

If Some(..) the tile cache that is associated with this picture.

options: PictureOptions

The config options for this picture.

num_render_tasks: usize

Keep track of the number of render tasks dependencies to pre-allocate the dependency array next frame.

Implementations

impl PicturePrimitive

pub fn print<T: PrintTreePrinter>(
&self,
pictures: &[Self],
self_index: PictureIndex,
pt: &mut T
)

pub fn can_use_segments(&self) -> bool

Returns true if this picture supports segmented rendering.

fn resolve_scene_properties(&mut self, properties: &SceneProperties) -> bool

pub fn is_visible(&self) -> bool

pub fn destroy(&mut self, retained_tiles: &mut RetainedTiles)

Destroy an existing picture. This is called just before a frame builder is replaced with a newly built scene. It gives a picture a chance to retain any cached tiles that may be useful during the next scene build.

pub fn new_image(
requested_composite_mode: Option<PictureCompositeMode>,
context_3d: Picture3DContext<OrderedPictureChild>,
frame_output_pipeline_id: Option<PipelineId>,
apply_local_clip_rect: bool,
flags: PrimitiveFlags,
requested_raster_space: RasterSpace,
prim_list: PrimitiveList,
spatial_node_index: SpatialNodeIndex,
tile_cache: Option<Box<TileCacheInstance>>,
options: PictureOptions
) -> Self

pub fn get_raster_space(&self, spatial_tree: &SpatialTree) -> RasterSpace

Gets the raster space to use when rendering the picture. Usually this would be the requested raster space. However, if the picture's spatial node or one of its ancestors is being pinch zoomed then we round it. This prevents us rasterizing glyphs for every minor change in zoom level, as that would be too expensive.

pub fn take_context(
&mut self,
pic_index: PictureIndex,
clipped_prim_bounding_rect: WorldRect,
surface_spatial_node_index: SpatialNodeIndex,
raster_spatial_node_index: SpatialNodeIndex,
parent_surface_index: SurfaceIndex,
parent_subpixel_mode: &SubpixelMode,
frame_state: &mut FrameBuildingState,
frame_context: &FrameBuildingContext,
scratch: &mut PrimitiveScratchBuffer,
tile_cache_logger: &mut TileCacheLogger
) -> Option<(PictureContext, PictureState, PrimitiveList)>

pub fn restore_context(
&mut self,
parent_surface_index: SurfaceIndex,
prim_list: PrimitiveList,
context: PictureContext,
state: PictureState,
frame_state: &mut FrameBuildingState
)

pub fn take_state(&mut self) -> PictureState

pub fn add_split_plane(
splitter: &mut BspSplitter<f64, WorldPixel, PlaneSplitAnchor>,
spatial_tree: &SpatialTree,
prim_spatial_node_index: SpatialNodeIndex,
original_local_rect: LayoutRect,
combined_local_clip_rect: &LayoutRect,
world_rect: WorldRect,
plane_split_anchor: PlaneSplitAnchor
) -> bool

Add a primitive instance to the plane splitter. The function would generate an appropriate polygon, clip it against the frustum, and register with the given plane splitter.

pub fn resolve_split_planes(
&mut self,
splitter: &mut BspSplitter<f64, WorldPixel, PlaneSplitAnchor>,
gpu_cache: &mut GpuCache,
spatial_tree: &SpatialTree
)

fn pre_update(
&mut self,
state: &mut PictureUpdateState,
frame_context: &FrameBuildingContext
) -> Option<PrimitiveList>

Called during initial picture traversal, before we know the bounding rect of children. It is possible to determine the surface / raster config now though.

fn post_update(
&mut self,
prim_list: PrimitiveList,
state: &mut PictureUpdateState,
frame_context: &FrameBuildingContext,
data_stores: &mut DataStores
)

Called after updating child pictures during the initial picture traversal.

pub fn prepare_for_render(
&mut self,
frame_context: &FrameBuildingContext,
frame_state: &mut FrameBuildingState,
data_stores: &mut DataStores
) -> bool