Struct webrender::picture::SpatialNodeComparer

struct SpatialNodeComparer {
    ref_spatial_node_index: SpatialNodeIndex,
    spatial_nodes: FastHashMap<SpatialNodeKey, TransformKey>,
    compare_cache: FastHashMap<(SpatialNodeKey, SpatialNodeKey), bool>,
    referenced_frames: HashSet<FrameId, BuildHasherDefault<FxHasher>>,
}

A helper for comparing spatial nodes between frames. The comparisons are done by-value, so that if the shape of the spatial node tree changes, invalidations aren’t done simply due to the spatial node index changing between display lists.

Fields
ref_spatial_node_index: SpatialNodeIndex

The root spatial node index of the tile cache

spatial_nodes: FastHashMap<SpatialNodeKey, TransformKey>

Maintains a map of currently active transform keys

compare_cache: FastHashMap<(SpatialNodeKey, SpatialNodeKey), bool>

A cache of recent comparisons between prev and current spatial nodes

referenced_frames: HashSet<FrameId, BuildHasherDefault<FxHasher>>

A set of frames that we need to retain spatial node entries for

Implementations

impl SpatialNodeComparer

fn new() -> Self

Construct a new comparer

fn next_frame(&mut self, ref_spatial_node_index: SpatialNodeIndex)

Advance to the next frame

fn register_used_transform(
&mut self,
spatial_node_index: SpatialNodeIndex,
frame_id: FrameId,
spatial_tree: &SpatialTree
)

Register a transform that is used, and build the transform key for it if new.

fn are_transforms_equivalent(
&mut self,
prev_spatial_node_key: &SpatialNodeKey,
curr_spatial_node_key: &SpatialNodeKey
) -> bool

Return true if the transforms for two given spatial nodes are considered equivalent

fn retain_for_frame(&mut self, frame_id: FrameId)

Ensure that the comparer won’t GC any nodes for a given frame id