Struct webrender::picture::TileCacheInstance

pub struct TileCacheInstance {
    pub slice: usize,
    pub current_tile_size: DeviceIntSize,
    pub spatial_node_index: SpatialNodeIndex,
    pub tiles: HashMap<Point2D<i32, TileCoordinate>, Tile, BuildHasherDefault<FxHasher>>,
    map_local_to_surface: SpaceMapper<LayoutPixel, PicturePixel>,
    map_child_pic_to_surface: SpaceMapper<PicturePixel, PicturePixel>,
    opacity_bindings: HashMap<PropertyBindingId, OpacityBindingInfo, BuildHasherDefault<FxHasher>>,
    spatial_nodes: HashMap<SpatialNodeIndex, SpatialNodeDependency, BuildHasherDefault<FxHasher>>,
    used_spatial_nodes: HashSet<SpatialNodeIndex, BuildHasherDefault<FxHasher>>,
    pub dirty_region: DirtyRegion,
    tile_size: PictureSize,
    tile_rect: Rect<i32, TileCoordinate>,
    tile_bounds_p0: Point2D<i32, TileCoordinate>,
    tile_bounds_p1: Point2D<i32, TileCoordinate>,
    pub local_rect: PictureRect,
    local_clip_rect: PictureRect,
    pub tiles_to_draw: Vec<Point2D<i32, TileCoordinate>>,
    surface_index: SurfaceIndex,
    pub background_color: Option<ColorF>,
    backdrop: BackdropInfo,
    pub subpixel_mode: SubpixelMode,
    pub shared_clips: Vec<Handle<ClipIntern>>,
    shared_clip_chain: ClipChainId,
    root_transform: TransformKey,
    frames_until_size_eval: usize,
    fract_offset: PictureVector2D,
}

Represents a cache of tiles that make up a picture primitives.

Fields
slice: usize

Index of the tile cache / slice for this frame builder. It's determined by the setup_picture_caching method during flattening, which splits the picture tree into multiple slices. It's used as a simple input to the tile keys. It does mean we invalidate tiles if a new layer gets inserted / removed between display lists - this seems very unlikely to occur on most pages, but can be revisited if we ever notice that.

current_tile_size: DeviceIntSize

The currently selected tile size to use for this cache

spatial_node_index: SpatialNodeIndex

The positioning node for this tile cache.

tiles: HashMap<Point2D<i32, TileCoordinate>, Tile, BuildHasherDefault<FxHasher>>

Hash of tiles present in this picture.

map_local_to_surface: SpaceMapper<LayoutPixel, PicturePixel>

A helper struct to map local rects into surface coords.

map_child_pic_to_surface: SpaceMapper<PicturePixel, PicturePixel>

A helper struct to map child picture rects into picture cache surface coords.

opacity_bindings: HashMap<PropertyBindingId, OpacityBindingInfo, BuildHasherDefault<FxHasher>>

List of opacity bindings, with some extra information about whether they changed since last frame.

spatial_nodes: HashMap<SpatialNodeIndex, SpatialNodeDependency, BuildHasherDefault<FxHasher>>

List of spatial nodes, with some extra information about whether they changed since last frame.

used_spatial_nodes: HashSet<SpatialNodeIndex, BuildHasherDefault<FxHasher>>

A set of spatial nodes that primitives / clips depend on found during dependency creation. This is used to avoid trying to calculate invalid relative transforms when building the spatial nodes hash above.

dirty_region: DirtyRegion

The current dirty region tracker for this picture.

tile_size: PictureSize

Current size of tiles in picture units.

tile_rect: Rect<i32, TileCoordinate>

Tile coords of the currently allocated grid.

tile_bounds_p0: Point2D<i32, TileCoordinate>

Pre-calculated versions of the tile_rect above, used to speed up the calculations in get_tile_coords_for_rect.

tile_bounds_p1: Point2D<i32, TileCoordinate>local_rect: PictureRect

Local rect (unclipped) of the picture this cache covers.

local_clip_rect: PictureRect

The local clip rect, from the shared clips of this picture.

tiles_to_draw: Vec<Point2D<i32, TileCoordinate>>

A list of tiles that are valid and visible, which should be drawn to the main scene.

surface_index: SurfaceIndex

The surface index that this tile cache will be drawn into.

background_color: Option<ColorF>

The background color from the renderer. If this is set opaque, we know it's fine to clear the tiles to this and allow subpixel text on the first slice.

backdrop: BackdropInfo

Information about the calculated backdrop content of this cache.

subpixel_mode: SubpixelMode

The allowed subpixel mode for this surface, which depends on the detected opacity of the background.

shared_clips: Vec<Handle<ClipIntern>>

A list of clip handles that exist on every (top-level) primitive in this picture. It's often the case that these are root / fixed position clips. By handling them here, we can avoid applying them to the items, which reduces work, but more importantly reduces invalidations.

shared_clip_chain: ClipChainId

The clip chain that represents the shared_clips above. Used to build the local clip rect for this tile cache.

root_transform: TransformKey

The current transform of the picture cache root spatial node

frames_until_size_eval: usize

The number of frames until this cache next evaluates what tile size to use. If a picture rect size is regularly changing just around a size threshold, we don't want to constantly invalidate and reallocate different tile size configuration each frame.

fract_offset: PictureVector2D

The current fractional offset of the cached picture

Methods

impl TileCacheInstance

pub fn new(
slice: usize,
spatial_node_index: SpatialNodeIndex,
background_color: Option<ColorF>,
shared_clips: Vec<Handle<ClipIntern>>,
shared_clip_chain: ClipChainId
) -> Self

pub fn is_opaque(&self) -> bool

Returns true if this tile cache is considered opaque.

fn get_tile_coords_for_rect(
&self,
rect: &PictureRect
) -> (Point2D<i32, TileCoordinate>, Point2D<i32, TileCoordinate>)

Get the tile coordinates for a given rectangle.

pub fn pre_update(
&mut self,
pic_rect: PictureRect,
surface_index: SurfaceIndex,
frame_context: &FrameVisibilityContext,
frame_state: &mut FrameVisibilityState
) -> WorldRect

Update transforms, opacity bindings and tile rects.

pub fn update_prim_dependencies(
&mut self,
prim_instance: &PrimitiveInstance,
prim_spatial_node_index: SpatialNodeIndex,
prim_clip_chain: Option<&ClipChainInstance>,
local_prim_rect: LayoutRect,
clip_scroll_tree: &ClipScrollTree,
data_stores: &DataStores,
clip_store: &ClipStore,
pictures: &[PicturePrimitive],
resource_cache: &ResourceCache,
opacity_binding_store: &Storage<OpacityBinding>,
image_instances: &Storage<ImageInstance>,
surface_index: SurfaceIndex,
surface_spatial_node_index: SpatialNodeIndex
) -> bool

Update the dependencies for each tile for a given primitive instance.

fn print(&self)

Print debug information about this picture cache to a tree printer.

pub fn post_update(
&mut self,
frame_context: &FrameVisibilityContext,
frame_state: &mut FrameVisibilityState
)

Apply any updates after prim dependency updates. This applies any late tile invalidations, and sets up the dirty rect and set of tile blits.