Struct webrender::picture::TileNode

struct TileNode {
    kind: TileNodeKind,
    rect: PictureRect,
}

A node in the dirty rect tracking quadtree.

Fields
kind: TileNodeKind

Leaf or internal node

rect: PictureRect

Rect of this node in the same space as the tile cache picture

Methods

impl TileNode

fn new_leaf(curr_indices: Vec<PrimitiveDependencyIndex>) -> Self

Construct a new leaf node, with the given primitive dependency index buffer

fn draw_debug_rects(
&self,
pic_to_world_mapper: &SpaceMapper<PicturePixel, WorldPixel>,
is_opaque: bool,
scratch: &mut PrimitiveScratchBuffer,
global_device_pixel_scale: DevicePixelScale
)

Draw debug information about this tile node

fn get_child_rects(rect: &PictureRect) -> Vec<PictureRect>

Calculate the four child rects for a given node

fn clear(&mut self, rect: PictureRect)

Called during pre_update, to clear the current dependencies

fn add_prim(&mut self, index: PrimitiveDependencyIndex, prim_rect: &PictureRect)

Add a primitive dependency to this node

fn get_preference(
&self,
level: i32,
can_merge: bool,
max_split_levels: i32
) -> Option<TileModification>

Return whether this tile wants to merge, split or do nothing.

fn maybe_merge_or_split(
&mut self,
level: i32,
curr_prims: &[PrimitiveDescriptor],
max_split_levels: i32
)

Apply a merge or split operation to this tile, if desired

fn update_dirty_rects(
&mut self,
prev_prims: &[PrimitiveDescriptor],
curr_prims: &[PrimitiveDescriptor],
prim_comparer: &mut PrimitiveComparer,
dirty_rect: &mut PictureRect,
compare_cache: &mut HashMap<PrimitiveComparisonKey, PrimitiveCompareResult, BuildHasherDefault<FxHasher>>,
invalidation_reason: &mut Option<InvalidationReason>
)

Update the dirty state of this node, building the overall dirty rect