Struct webrender::picture::TilePostUpdateContext

struct TilePostUpdateContext<'a> {
    pic_to_world_mapper: SpaceMapper<PicturePixel, WorldPixel>,
    global_device_pixel_scale: DevicePixelScale,
    local_clip_rect: PictureRect,
    backdrop: BackdropInfo,
    opacity_bindings: &'a FastHashMap<PropertyBindingId, BindingInfo<f32>>,
    color_bindings: &'a FastHashMap<PropertyBindingId, BindingInfo<ColorU>>,
    current_tile_size: DeviceIntSize,
    local_rect: PictureRect,
    external_surfaces: &'a [ExternalSurfaceDescriptor],
    z_id_opaque: ZBufferId,
    z_id_alpha: ZBufferId,
}

Fields
pic_to_world_mapper: SpaceMapper<PicturePixel, WorldPixel>

Maps from picture cache coords -> world space coords.

global_device_pixel_scale: DevicePixelScale

Global scale factor from world -> device pixels.

local_clip_rect: PictureRect

The local clip rect (in picture space) of the entire picture cache

backdrop: BackdropInfo

The calculated backdrop information for this cache instance.

opacity_bindings: &'a FastHashMap<PropertyBindingId, BindingInfo<f32>>

Information about opacity bindings from the picture cache.

color_bindings: &'a FastHashMap<PropertyBindingId, BindingInfo<ColorU>>

Information about color bindings from the picture cache.

current_tile_size: DeviceIntSize

Current size in device pixels of tiles for this cache

local_rect: PictureRect

The local rect of the overall picture cache

external_surfaces: &'a [ExternalSurfaceDescriptor]

A list of the external surfaces that are present on this slice

z_id_opaque: ZBufferId

Pre-allocated z-id to assign to opaque tiles during post_update. We use a different z-id for opaque/alpha tiles, so that compositor surfaces (such as videos) can have a z-id between these values, which allows compositor surfaces to occlude opaque tiles, but not alpha tiles.

z_id_alpha: ZBufferId

Pre-allocated z-id to assign to alpha tiles during post_update