webrender

Module picture_composite_mode

Structs

SurfaceAllocInfo

Information from get_surface_rects about the allocated size, UV sampling parameters etc for an off-screen surface

Enums

PictureCompositeKey

Represents a hashable description of how a picture primitive will be composited into its parent.

PictureCompositeMode

Specifies how this Picture should be composited onto the target it belongs to.

Functions

calculate_uv_rect_kind

get_surface_rects

prepare_composite_mode

request_render_task 🔒