webrender

Module prepare

Prepare pass

TODO: document this!

Structs

LinearGradientSegmentPattern 🔒

Pattern builder for a single fast-path two-stop segment emitted by decompose_axis_aligned_gradient. Holds the segment’s gradient line and stop colors (in segment-local coords); build translates start/end into the prim’s spatial-node space by adding ctx.prim_origin.

Constants

MAX_MASK_SIZE 🔒

MIN_BRUSH_SPLIT_AREA 🔒

Functions

adjust_mask_scale_for_max_size 🔒

build_segments_if_needed 🔒

can_use_clip_chain_for_quad_path 🔒

prepare_interned_prim_for_render 🔒

Prepare an interned primitive for rendering, by requesting resources, render tasks etc. This is equivalent to the prepare_prim_for_render_inner call for old style primitives.

prepare_picture

The entry point of the preapre pass.

prepare_prim_for_render 🔒

prepare_primitives 🔒

update_brush_segment_clip_task

Write out to the clip mask instances array the correct clip mask config for this segment.

update_clip_task

update_clip_task_for_brush 🔒

write_brush_segment_description 🔒

write_segment 🔒