webrender::prim_store::gradient

Module conic

Conic gradients

Specification: https://drafts.csswg.org/css-images-4/#conic-gradients

Conic gradients are rendered via cached render tasks and composited with the image brush.

Structs

ConicGradient

ConicGradientKey

Identifying key for a line decoration.

ConicGradientParams

Hashable conic gradient parameters, for use during prim interning.

ConicGradientTemplate

Type Aliases

ConicGradientDataHandle