webrender::prim_store::gradient

Module radial

Radial gradients

Specification: https://drafts.csswg.org/css-images-4/#radial-gradients

Radial gradients are rendered via cached render tasks and composited with the image brush.

Structs

RadialGradient

RadialGradientKey

Identifying key for a radial gradient.

RadialGradientParams

Hashable radial gradient parameters, for use during prim interning.

RadialGradientTemplate

Functions

optimize_radial_gradient

Avoid invoking the radial gradient shader on large areas where the color is constant.

Type Aliases

RadialGradientDataHandle