webrender::prim_store::image

Struct ImageData

pub struct ImageData {
    pub key: ImageKey,
    pub stretch_size: StretchSize,
    pub tile_spacing: LayoutSize,
    pub color: ColorF,
    pub image_rendering: ImageRendering,
    pub alpha_type: AlphaType,
}

Fields
key: ImageKeystretch_size: StretchSizetile_spacing: LayoutSizecolor: ColorFimage_rendering: ImageRenderingalpha_type: AlphaType
Implementations

impl ImageData

pub fn update( &mut self, common: &mut PrimTemplateCommonData, prim_instance_index: PrimitiveInstanceIndex, prim_spatial_node_index: SpatialNodeIndex, frame_state: &mut FrameBuildingState<'_>, frame_context: &FrameBuildingContext<'_>, prim_rect: LayoutRect, scratch: &mut PrimitiveScratchBuffer, ) -> Index<ImageScratch>

Update the GPU cache for a given primitive template. This may be called multiple times per frame, by each primitive reference that refers to this interned template. The initial request call to the GPU cache ensures that work is only done if the cache entry is invalid (due to first use or eviction).

pub fn write_prim_gpu_blocks( &self, adjustment: &AdjustedImageSource, stretch_size: LayoutSize, writer: &mut GpuBufferWriter<'_, GpuBufferBlockF>, )