Struct webrender::prim_store::DeferredResolve

pub struct DeferredResolve {
    pub resource_address: GpuStoreAddress,
    pub image_properties: ImageProperties,
}

For external images, it's not possible to know the UV coords of the image (or the image data itself) until the render thread receives the frame and issues callbacks to the client application. For external images that are visible, a DeferredResolve is created that is stored in the frame. This allows the render thread to iterate this list and update any changed texture data and update the UV rect.

Fields
resource_address: GpuStoreAddress image_properties: ImageProperties