Struct webrender::prim_store::PrimitiveInstance

pub struct PrimitiveInstance {
    pub kind: PrimitiveInstanceKind,
    pub local_clip_rect: LayoutRect,
    pub id: PrimitiveDebugId,
    pub prepared_frame_id: FrameId,
    pub visibility_info: PrimitiveVisibilityIndex,
    pub clip_chain_id: ClipChainId,
}

Fields
kind: PrimitiveInstanceKind

Identifies the kind of primitive this instance is, and references to where the relevant information for the primitive can be found.

local_clip_rect: LayoutRect

Local space clip rect for this instance

id: PrimitiveDebugIdprepared_frame_id: FrameId

The last frame ID (of the RenderTaskGraph) this primitive was prepared for rendering in.

visibility_info: PrimitiveVisibilityIndex

If this primitive is visible, an index into the instance visibility scratch buffer. If not visible, INVALID.

clip_chain_id: ClipChainId

ID of the clip chain that this primitive is clipped by.

Implementations

impl PrimitiveInstance

pub fn new(
local_clip_rect: LayoutRect,
kind: PrimitiveInstanceKind,
clip_chain_id: ClipChainId
) -> Self

pub fn reset(&mut self)

pub fn is_chased(&self) -> bool

pub fn uid(&self) -> ItemUid

impl PrimitiveInstance

fn build_segments_if_needed(
&mut self,
prim_info: &PrimitiveVisibility,
frame_state: &mut FrameBuildingState,
prim_store: &mut PrimitiveStore,
data_stores: &DataStores,
segments_store: &mut Storage<BrushSegment>,
segment_instances_store: &mut Storage<SegmentedInstance>
)

fn update_clip_task_for_brush(
&self,
prim_origin: &LayoutPoint,
prim_info: &mut PrimitiveVisibility,
prim_spatial_node_index: SpatialNodeIndex,
root_spatial_node_index: SpatialNodeIndex,
pic_context: &PictureContext,
pic_state: &mut PictureState,
frame_context: &FrameBuildingContext,
frame_state: &mut FrameBuildingState,
prim_store: &PrimitiveStore,
data_stores: &mut DataStores,
segments_store: &mut Storage<BrushSegment>,
segment_instances_store: &mut Storage<SegmentedInstance>,
clip_mask_instances: &mut Vec<ClipMaskKind>,
unclipped: &DeviceRect,
device_pixel_scale: DevicePixelScale
) -> bool

fn update_clip_task(
&mut self,
prim_origin: &LayoutPoint,
prim_spatial_node_index: SpatialNodeIndex,
root_spatial_node_index: SpatialNodeIndex,
pic_context: &PictureContext,
pic_state: &mut PictureState,
frame_context: &FrameBuildingContext,
frame_state: &mut FrameBuildingState,
prim_store: &mut PrimitiveStore,
data_stores: &mut DataStores,
scratch: &mut PrimitiveScratchBuffer
)