Struct webrender::prim_store::PrimitiveStore

pub struct PrimitiveStore {
    pub pictures: Vec<PicturePrimitive>,
    pub text_runs: Storage<TextRunPrimitive>,
    pub linear_gradients: Storage<LinearGradientPrimitive>,
    pub images: Storage<ImageInstance>,
    pub opacity_bindings: Storage<OpacityBinding>,
    pub color_bindings: Storage<PropertyBinding<ColorU>>,
}

Fields
pictures: Vec<PicturePrimitive>text_runs: Storage<TextRunPrimitive>linear_gradients: Storage<LinearGradientPrimitive>images: Storage<ImageInstance>

A list of image instances. These are stored separately as storing them inline in the instance makes the structure bigger for other types.

opacity_bindings: Storage<OpacityBinding>

List of animated opacity bindings for a primitive.

color_bindings: Storage<PropertyBinding<ColorU>>

animated color bindings for this primitive.

Implementations

impl PrimitiveStore

pub fn new(stats: &PrimitiveStoreStats) -> PrimitiveStore

pub fn get_stats(&self) -> PrimitiveStoreStats

pub fn print_picture_tree(&self, root: PictureIndex)

pub fn destroy(&mut self, retained_tiles: &mut RetainedTiles)

Destroy an existing primitive store. This is called just before a primitive store is replaced with a newly built scene.

pub fn prim_count(&self) -> usize

Returns the total count of primitive instances contained in pictures.

pub fn update_visibility(
&mut self,
pic_index: PictureIndex,
parent_surface_index: SurfaceIndex,
world_culling_rect: &WorldRect,
frame_context: &FrameVisibilityContext<'_>,
frame_state: &mut FrameVisibilityState<'_>
) -> Option<PictureRect>

Update visibility pass - update each primitive visibility struct, and build the clip chain instance if appropriate.

fn request_resources_for_prim(
&mut self,
prim_instance: &mut PrimitiveInstance,
prim_spatial_node_index: SpatialNodeIndex,
prim_world_rect: WorldRect,
frame_context: &FrameVisibilityContext<'_>,
frame_state: &mut FrameVisibilityState<'_>
)

pub fn get_opacity_binding(
&self,
opacity_binding_index: Index<OpacityBinding>
) -> f32

fn get_opacity_collapse_prim(
&self,
pic_index: PictureIndex
) -> Option<PictureIndex>

pub fn optimize_picture_if_possible(&mut self, pic_index: PictureIndex)

fn prepare_prim_for_render(
&mut self,
prim_instance: &mut PrimitiveInstance,
prim_spatial_node_index: SpatialNodeIndex,
pic_context: &PictureContext,
pic_state: &mut PictureState,
frame_context: &FrameBuildingContext<'_>,
frame_state: &mut FrameBuildingState<'_>,
plane_split_anchor: PlaneSplitAnchor,
data_stores: &mut DataStores,
scratch: &mut PrimitiveScratchBuffer,
tile_cache_log: &mut TileCacheLogger
) -> bool

pub fn prepare_primitives(
&mut self,
prim_list: &mut PrimitiveList,
pic_context: &PictureContext,
pic_state: &mut PictureState,
frame_context: &FrameBuildingContext<'_>,
frame_state: &mut FrameBuildingState<'_>,
data_stores: &mut DataStores,
scratch: &mut PrimitiveScratchBuffer,
tile_cache_log: &mut TileCacheLogger
)

fn prepare_interned_prim_for_render(
&mut self,
prim_instance: &mut PrimitiveInstance,
prim_spatial_node_index: SpatialNodeIndex,
plane_split_anchor: PlaneSplitAnchor,
pic_context: &PictureContext,
pic_state: &mut PictureState,
frame_context: &FrameBuildingContext<'_>,
frame_state: &mut FrameBuildingState<'_>,
data_stores: &mut DataStores,
scratch: &mut PrimitiveScratchBuffer
)

Prepare an interned primitive for rendering, by requesting resources, render tasks etc. This is equivalent to the prepare_prim_for_render_inner call for old style primitives.