webrender::print_tree

Trait PrintTreePrinter

pub trait PrintTreePrinter {
    // Required methods
    fn new_level(&mut self, title: String);
    fn end_level(&mut self);
    fn add_item(&mut self, text: String);
}

A trait that makes it easy to describe a pretty tree of data, regardless of the printing destination, to either print it directly to stdout, or serialize it as in the debugger

Required Methods

fn new_level(&mut self, title: String)

fn end_level(&mut self)

fn add_item(&mut self, text: String)