webrender

Module quad

Structs

QuadCacheKey

QuadTileClassifier

A helper struct for classifying a set of tiles within a quad depending on what strategy they can be used to draw them.

QuadTileInfo

QuadTileIterator

QuadTransformState

Contains some transform-related information that is computed per primitive cluster.

Enums

QuadRenderStrategy

Describes how clipping affects the rendering of a quad primitive.

QuadTileKind

Classification result for a tile within a quad

Constants

MAX_TILES_PER_QUAD_X 🔒

MAX_TILES_PER_QUAD_Y 🔒

MIN_AA_SEGMENTS_SIZE 🔒

MIN_QUAD_SPLIT_SIZE 🔒

Functions

add_composite_prim 🔒

add_pattern_prim 🔒

add_render_task_with_mask 🔒

add_to_batch

adjust_indirect_pattern_resolution 🔒

Adjust the transform and device rect until the latter fits the provided maximum size. Also ensure that near-zero size tasks do are at least

cache_key

create_quad_primitive 🔒

get_prim_render_strategy 🔒

prepare_border_image_nine_patch

prepare_clip_range

prepare_clip_task

prepare_indirect_pattern 🔒

prepare_nine_patch 🔒

prepare_quad

prepare_quad_impl 🔒

prepare_repeatable_quad

prepare_tiles 🔒

write_device_prim_blocks

Write the gpu blocks for a primitive in device space.

write_layout_prim_blocks

Write the gpu blocks for a primitive in layout space.

write_prim_blocks 🔒

Write the GPU blocks, either in local or device space

write_prim_blocks_impl 🔒

Type Aliases

LayoutOrDeviceRect

This type reflects the unfortunate situation with quad coordinates where we sometimes use layout and sometimes device coordinates.