webrender::quad

Function write_prim_blocks

fn write_prim_blocks(
    builder: &mut GpuBufferBuilderImpl<GpuBufferBlockF>,
    local_rect: &LayoutRect,
    local_clip_rect: &LayoutRect,
    device_clip_rect: &DeviceRect,
    local_to_device: Option<&ScaleOffset>,
    round_edges: EdgeMask,
    pattern: &Pattern,
) -> GpuBufferAddress

Write the GPU blocks, either in local or device space

If a local-to-device transform is provided, then the primitive is written in device space, otherwise it is written in layout space.