webrender::render_target

Function add_svg_filter_node_instances

fn add_svg_filter_node_instances(
    instances: &mut Vec<(BatchTextures, Vec<SVGFEFilterInstance, FrameAllocator>), FrameAllocator>,
    render_tasks: &RenderTaskGraph,
    task_info: &SVGFEFilterTask,
    target_task: &RenderTask,
    input_1_task: Option<RenderTaskId>,
    input_2_task: Option<RenderTaskId>,
    extra_data_address: Option<GpuBufferAddress>,
    memory: &FrameMemory,
)

Generates SVGFEFilterInstances from a single SVGFEFilterTask, this is what prepares vertex data for the shader, and adds it to the appropriate batch.

The interesting parts of the handling of SVG filters are:

scene_building.rs : wrap_prim_with_filters

picture.rs : get_coverage_svgfe

render_task.rs : new_svg_filter_graph

render_target.rs : add_svg_filter_node_instances (you are here)