Struct webrender::render_target::RenderTargetContext

pub struct RenderTargetContext<'a, 'rc> {
    pub global_device_pixel_scale: DevicePixelScale,
    pub prim_store: &'a PrimitiveStore,
    pub resource_cache: &'rc mut ResourceCache,
    pub use_dual_source_blending: bool,
    pub use_advanced_blending: bool,
    pub break_advanced_blend_batches: bool,
    pub batch_lookback_count: usize,
    pub spatial_tree: &'a SpatialTree,
    pub data_stores: &'a DataStores,
    pub surfaces: &'a [SurfaceInfo],
    pub scratch: &'a PrimitiveScratchBuffer,
    pub screen_world_rect: WorldRect,
    pub globals: &'a FrameGlobalResources,
}

Fields
global_device_pixel_scale: DevicePixelScaleprim_store: &'a PrimitiveStoreresource_cache: &'rc mut ResourceCacheuse_dual_source_blending: booluse_advanced_blending: boolbreak_advanced_blend_batches: boolbatch_lookback_count: usizespatial_tree: &'a SpatialTreedata_stores: &'a DataStoressurfaces: &'a [SurfaceInfo]scratch: &'a PrimitiveScratchBufferscreen_world_rect: WorldRectglobals: &'a FrameGlobalResources
Implementations

impl<'a, 'rc> RenderTargetContext<'a, 'rc>

fn get_clip_task_address(
&self,
clip_task_index: ClipTaskIndex,
offset: i32,
render_tasks: &RenderTaskGraph
) -> Option<RenderTaskAddress>

Retrieve the GPU task address for a given clip task instance. Returns None if the segment was completely clipped out. Returns Some(OPAQUE_TASK_ADDRESS) if no clip mask is needed. Returns Some(task_address) if there was a valid clip mask.

fn get_prim_clip_task_address(
&self,
clip_task_index: ClipTaskIndex,
render_tasks: &RenderTaskGraph
) -> Option<RenderTaskAddress>

Helper function to get the clip task address for a non-segmented primitive.