webrender

Module render_task

Structs

BlitTask

BlurTask

BorderTask

CacheMaskTask

CachedTask

ClipRegionTask

EmptyTask

ImageClipSubTask

An clip applied to a render task using the multiply blend mode on top of the content being clipped.

ImageRequestTask

LineDecorationTask

PictureTask

PrimTask

ReadbackTask

RectangleClipSubTask

A (rounded) rectangle clip applied to a render task using the multiply blend mode on top of the content being clipped.

RenderTask

RenderTaskAddress

RenderTaskData

SVGFEFilterTask

ScalingTask

TileCompositeTask

Enums

BlurTaskKey

Since we only use it within a single primitive, the key only needs to contain the down-scaling level and the blur std deviation.

RenderTaskKind

RenderTaskLocation

Identifies the output buffer location for a given RenderTask.

StaticRenderTaskSurface

A render task location that targets a persistent output buffer which will be retained over multiple frames.

SubTask

A rendering operation applied on top of a render task.

Constants

FLOATS_PER_RENDER_TASK_INFO 🔒

MAX_BLUR_STD_DEVIATION

MIN_DOWNSCALING_RT_SIZE

Functions

render_task_sanity_check 🔒

Type Aliases

BlurTaskCache

In order to avoid duplicating the down-scaling and blur passes when a picture has several blurs, we use a local (primitive-level) cache of the render tasks generated for a single shadowed primitive in a single frame.

TaskDependencies