Struct webrender::render_task_graph::RenderTaskGraph

pub struct RenderTaskGraph {
    pub tasks: Vec<RenderTask>,
    pub passes: Vec<Pass>,
    frame_id: FrameId,
    pub task_data: Vec<RenderTaskData>,
}

The RenderTaskGraph is the immutable representation of the render task graph. It is built by the RenderTaskGraphBuilder, and is constructed once per frame.

Fields
tasks: Vec<RenderTask>

List of tasks added to the graph

passes: Vec<Pass>

The passes that were created, based on dependencies between tasks

frame_id: FrameId

Current frame id, used for debug validation

task_data: Vec<RenderTaskData>

GPU specific data for each task that is made available to shaders

Implementations

impl RenderTaskGraph

pub fn print(&self)

Print the render task graph to console

pub fn resolve_texture( &self, task_id: impl Into<Option<RenderTaskId>> ) -> Option<TextureSource>

pub fn resolve_location( &self, task_id: impl Into<Option<RenderTaskId>>, gpu_cache: &GpuCache ) -> Option<(GpuCacheAddress, TextureSource)>

fn resolve_impl( &self, task_id: RenderTaskId, gpu_cache: &GpuCache ) -> Option<(GpuCacheAddress, TextureSource)>

pub fn frame_id(&self) -> FrameId

Return current frame id, used for validation