webrender::renderer

Module gpu_buffer

Structs

DeferredBlock 🔒

Record a patch to the GPU buffer for a render task

DeferredUvCopy

Records that the uv rect block at dst must be overwritten with the block at src once the renderer has resolved external images.

GpuBuffer

GpuBufferAddress

GpuBufferBlockF

A single texel in RGBAF32 texture - 16 bytes.

GpuBufferBlockI

A single texel in RGBAI32 texture - 16 bytes.

GpuBufferBuilder

GpuBufferBuilderImpl

GpuBufferHandle

GpuBuffer handle is similar to GpuBufferAddress with additional checks to avoid accidentally using the same handle in multiple frames.

GpuBufferWriter

Interface to allow writing multiple GPU blocks, possibly of different types

Traits

GpuBufferDataF

GpuBufferDataI

Functions

ensure_row_capacity 🔒

finish_row 🔒

Type Aliases

GpuBufferBuilderF

GpuBufferBuilderI

GpuBufferF

GpuBufferI

GpuBufferWriterF

GpuBufferWriterI