Module webrender::renderer::gpu_cache

Structs

CacheRow 🔒

Tracks the state of each row in the GPU cache texture.

GpuCacheTexture

The device-specific representation of the cache texture in gpu_cache.rs

Enums

GpuCacheBus 🔒

The bus over which CPU and GPU versions of the GPU cache get synchronized.

Constants

GPU_CACHE_RESIZE_TEST 🔒

Enabling this toggle would force the GPU cache scattered texture to be resized every frame, which enables GPU debuggers to see if this is performed correctly.