Struct webrender::renderer::init::DebugFlags

#[repr(C)]
pub struct DebugFlags(pub(crate) u32);

Flags to enable/disable various builtin debugging tools.

Tuple Fields
0: u32
Implementations

impl DebugFlags

pub const PROFILER_DBG: DebugFlags = _

Display the frame profiler on screen.

pub const RENDER_TARGET_DBG: DebugFlags = _

Display intermediate render targets on screen.

pub const TEXTURE_CACHE_DBG: DebugFlags = _

Display all texture cache pages on screen.

pub const GPU_TIME_QUERIES: DebugFlags = _

Display GPU timing results.

pub const GPU_SAMPLE_QUERIES: DebugFlags = _

Query the number of pixels that pass the depth test divided and show it in the profiler as a percentage of the number of pixels in the screen (window width times height).

pub const DISABLE_BATCHING: DebugFlags = _

Render each quad with their own draw call.

Terrible for performance but can help with understanding the drawing order when inspecting renderdoc or apitrace recordings.

pub const EPOCHS: DebugFlags = _

Display the pipeline epochs.

pub const ECHO_DRIVER_MESSAGES: DebugFlags = _

Print driver messages to stdout.

pub const SHOW_OVERDRAW: DebugFlags = _

Show an overlay displaying overdraw amount.

pub const GPU_CACHE_DBG: DebugFlags = _

Display the contents of GPU cache.

pub const TEXTURE_CACHE_DBG_CLEAR_EVICTED: DebugFlags = _

Clear evicted parts of the texture cache for debugging purposes.

pub const PICTURE_CACHING_DBG: DebugFlags = _

Show picture caching debug overlay

pub const PRIMITIVE_DBG: DebugFlags = _

Highlight all primitives with colors based on kind.

pub const ZOOM_DBG: DebugFlags = _

Draw a zoom widget showing part of the framebuffer zoomed in.

pub const SMALL_SCREEN: DebugFlags = _

Scale the debug renderer down for a smaller screen. This will disrupt any mapping between debug display items and page content, so shouldn’t be used with overlays like the picture caching or primitive display.

pub const DISABLE_OPAQUE_PASS: DebugFlags = _

Disable various bits of the WebRender pipeline, to help narrow down where slowness might be coming from.

pub const DISABLE_ALPHA_PASS: DebugFlags = _

pub const DISABLE_CLIP_MASKS: DebugFlags = _

pub const DISABLE_TEXT_PRIMS: DebugFlags = _

pub const DISABLE_GRADIENT_PRIMS: DebugFlags = _

pub const OBSCURE_IMAGES: DebugFlags = _

pub const GLYPH_FLASHING: DebugFlags = _

Taint the transparent area of the glyphs with a random opacity to easily see when glyphs are re-rasterized.

pub const SMART_PROFILER: DebugFlags = _

The profiler only displays information that is out of the ordinary.

pub const INVALIDATION_DBG: DebugFlags = _

If set, dump picture cache invalidation debug to console.

pub const PROFILER_CAPTURE: DebugFlags = _

Collect and dump profiler statistics to captures.

pub const FORCE_PICTURE_INVALIDATION: DebugFlags = _

Invalidate picture tiles every frames (useful when inspecting GPU work in external tools).

pub const WINDOW_VISIBILITY_DBG: DebugFlags = _

Display window visibility on screen.

pub const RESTRICT_BLOB_SIZE: DebugFlags = _

Render large blobs with at a smaller size (incorrectly). This is a temporary workaround for fuzzing.

pub const SURFACE_PROMOTION_LOGGING: DebugFlags = _

Enable surface promotion logging.

impl DebugFlags

pub const fn empty() -> DebugFlags

Get a flags value with all bits unset.

pub const fn all() -> DebugFlags

Get a flags value with all known bits set.

pub const fn bits(&self) -> u32

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u32) -> Option<DebugFlags>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u32) -> DebugFlags

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u32) -> DebugFlags

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<DebugFlags>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: DebugFlags) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: DebugFlags) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: DebugFlags)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: DebugFlags)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: DebugFlags)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: DebugFlags, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: DebugFlags) -> DebugFlags

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: DebugFlags) -> DebugFlags

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: DebugFlags) -> DebugFlags

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: DebugFlags) -> DebugFlags

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> DebugFlags

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl DebugFlags

pub const fn iter(&self) -> Iter<DebugFlags>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<DebugFlags>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.