webrender::renderer::shade

Struct Shaders

pub struct Shaders {

    loader: ShaderLoader,
    cs_blur_rgba8: ShaderHandle,
    cs_border_segment: ShaderHandle,
    cs_border_solid: ShaderHandle,
    cs_scale: Vec<Option<ShaderHandle>>,
    cs_line_decoration: ShaderHandle,
    cs_svg_filter_node: ShaderHandle,
    brush_solid: BrushShader,
    brush_image: Vec<Option<BrushShader>>,
    brush_fast_image: Vec<Option<BrushShader>>,
    brush_blend: BrushShader,
    brush_mix_blend: BrushShader,
    brush_yuv_image: Vec<Option<BrushShader>>,
    brush_opacity: BrushShader,
    brush_opacity_aa: BrushShader,
    cs_clip_rectangle_slow: ShaderHandle,
    cs_clip_rectangle_fast: ShaderHandle,
    ps_text_run: TextShader,
    ps_text_run_dual_source: Option<TextShader>,
    ps_split_composite: ShaderHandle,
    ps_quad_textured: ShaderHandle,
    ps_quad_textured_external: Option<ShaderHandle>,
    ps_quad_textured_external_bt709: Option<ShaderHandle>,
    ps_quad_textured_rect: Option<ShaderHandle>,
    ps_quad_repeat: ShaderHandle,
    ps_quad_gradient: ShaderHandle,
    ps_quad_box_shadow: ShaderHandle,
    ps_mask: ShaderHandle,
    ps_mask_fast: ShaderHandle,
    ps_clear: ShaderHandle,
    ps_copy: ShaderHandle,
    composite: CompositorShaders,

}

Fields
loader: ShaderLoadercs_blur_rgba8: ShaderHandlecs_border_segment: ShaderHandlecs_border_solid: ShaderHandlecs_scale: Vec<Option<ShaderHandle>>cs_line_decoration: ShaderHandlecs_svg_filter_node: ShaderHandlebrush_solid: BrushShaderbrush_image: Vec<Option<BrushShader>>brush_fast_image: Vec<Option<BrushShader>>brush_blend: BrushShaderbrush_mix_blend: BrushShaderbrush_yuv_image: Vec<Option<BrushShader>>brush_opacity: BrushShaderbrush_opacity_aa: BrushShadercs_clip_rectangle_slow: ShaderHandle

These are “cache clip shaders”. These shaders are used to draw clip instances into the cached clip mask. The results of these shaders are also used by the primitive shaders.

cs_clip_rectangle_fast: ShaderHandleps_text_run: TextShaderps_text_run_dual_source: Option<TextShader>ps_split_composite: ShaderHandleps_quad_textured: ShaderHandleps_quad_textured_external: Option<ShaderHandle>ps_quad_textured_external_bt709: Option<ShaderHandle>ps_quad_textured_rect: Option<ShaderHandle>ps_quad_repeat: ShaderHandleps_quad_gradient: ShaderHandleps_quad_box_shadow: ShaderHandleps_mask: ShaderHandleps_mask_fast: ShaderHandleps_clear: ShaderHandleps_copy: ShaderHandlecomposite: CompositorShaders
Implementations

impl Shaders

pub fn new( device: &mut Device, gl_type: GlType, options: &WebRenderOptions, ) -> Result<Self, ShaderError>

pub fn precache_all( &mut self, precache_flags: ShaderPrecacheFlags, ) -> PendingShadersToPrecache

pub fn resume_precache( &mut self, device: &mut Device, pending_shaders: &mut PendingShadersToPrecache, ) -> Result<bool, ShaderError>

Returns true if another call is needed, false if precaching is finished.

fn get_compositing_shader_index(buffer_kind: ImageBufferKind) -> usize

pub fn get_composite_shader( &mut self, format: CompositeSurfaceFormat, buffer_kind: ImageBufferKind, features: CompositeFeatures, ) -> &mut LazilyCompiledShader

pub fn get_scale_shader( &mut self, buffer_kind: ImageBufferKind, ) -> &mut LazilyCompiledShader

pub fn get_quad_shader( &mut self, pattern: PatternKind, ) -> &mut LazilyCompiledShader

pub fn get( &mut self, key: &BatchKey, features: BatchFeatures, debug_flags: DebugFlags, device: &Device, ) -> &mut LazilyCompiledShader

pub fn get_handle( &mut self, key: &BatchKey, features: BatchFeatures, debug_flags: DebugFlags, device: &Device, ) -> ShaderHandle

pub fn cs_blur_rgba8(&mut self) -> &mut LazilyCompiledShader

pub fn cs_border_segment(&mut self) -> &mut LazilyCompiledShader

pub fn cs_border_solid(&mut self) -> &mut LazilyCompiledShader

pub fn cs_line_decoration(&mut self) -> &mut LazilyCompiledShader

pub fn cs_svg_filter_node(&mut self) -> &mut LazilyCompiledShader

pub fn cs_clip_rectangle_slow(&mut self) -> &mut LazilyCompiledShader

pub fn cs_clip_rectangle_fast(&mut self) -> &mut LazilyCompiledShader

pub fn ps_quad_textured(&mut self) -> &mut LazilyCompiledShader

pub fn ps_mask(&mut self) -> &mut LazilyCompiledShader

pub fn ps_mask_fast(&mut self) -> &mut LazilyCompiledShader

pub fn ps_clear(&mut self) -> &mut LazilyCompiledShader

pub fn ps_copy(&mut self) -> &mut LazilyCompiledShader

pub fn deinit(self, device: &mut Device)