Struct webrender::renderer::RendererOptions

pub struct RendererOptions {
    pub device_pixel_ratio: f32,
    pub resource_override_path: Option<PathBuf>,
    pub enable_aa: bool,
    pub enable_dithering: bool,
    pub max_recorded_profiles: usize,
    pub precache_flags: ShaderPrecacheFlags,
    pub renderer_kind: RendererKind,
    pub enable_subpixel_aa: bool,
    pub force_subpixel_aa: bool,
    pub clear_color: Option<ColorF>,
    pub enable_clear_scissor: bool,
    pub max_texture_size: Option<i32>,
    pub max_glyph_cache_size: Option<usize>,
    pub scatter_gpu_cache_updates: bool,
    pub upload_method: UploadMethod,
    pub workers: Option<Arc<ThreadPool>>,
    pub blob_image_handler: Option<Box<dyn BlobImageHandler>>,
    pub recorder: Option<Box<dyn ApiRecordingReceiver>>,
    pub thread_listener: Option<Box<dyn ThreadListener + Send + Sync>>,
    pub size_of_op: Option<VoidPtrToSizeFn>,
    pub enclosing_size_of_op: Option<VoidPtrToSizeFn>,
    pub cached_programs: Option<Rc<ProgramCache>>,
    pub debug_flags: DebugFlags,
    pub renderer_id: Option<u64>,
    pub scene_builder_hooks: Option<Box<dyn SceneBuilderHooks + Send>>,
    pub sampler: Option<Box<dyn AsyncPropertySampler + Send>>,
    pub chase_primitive: ChasePrimitive,
    pub support_low_priority_transactions: bool,
    pub namespace_alloc_by_client: bool,
    pub enable_picture_caching: bool,
    pub testing: bool,
    pub gpu_supports_fast_clears: bool,
    pub allow_dual_source_blending: bool,
    pub allow_advanced_blend_equation: bool,
    pub allow_pixel_local_storage_support: bool,
    pub allow_texture_storage_support: bool,
    pub allow_texture_swizzling: bool,
    pub batch_lookback_count: usize,
    pub start_debug_server: bool,
    pub dump_shader_source: Option<String>,
    pub surface_origin_is_top_left: bool,
    pub compositor_config: CompositorConfig,
    pub enable_gpu_markers: bool,
}

Fields
device_pixel_ratio: f32resource_override_path: Option<PathBuf>enable_aa: boolenable_dithering: boolmax_recorded_profiles: usizeprecache_flags: ShaderPrecacheFlagsrenderer_kind: RendererKindenable_subpixel_aa: bool

Enable sub-pixel anti-aliasing if a fast implementation is available.

force_subpixel_aa: bool

Enable sub-pixel anti-aliasing if it requires a slow implementation.

clear_color: Option<ColorF>enable_clear_scissor: boolmax_texture_size: Option<i32>max_glyph_cache_size: Option<usize>scatter_gpu_cache_updates: boolupload_method: UploadMethodworkers: Option<Arc<ThreadPool>>blob_image_handler: Option<Box<dyn BlobImageHandler>>recorder: Option<Box<dyn ApiRecordingReceiver>>thread_listener: Option<Box<dyn ThreadListener + Send + Sync>>size_of_op: Option<VoidPtrToSizeFn>enclosing_size_of_op: Option<VoidPtrToSizeFn>cached_programs: Option<Rc<ProgramCache>>debug_flags: DebugFlagsrenderer_id: Option<u64>scene_builder_hooks: Option<Box<dyn SceneBuilderHooks + Send>>sampler: Option<Box<dyn AsyncPropertySampler + Send>>chase_primitive: ChasePrimitivesupport_low_priority_transactions: boolnamespace_alloc_by_client: boolenable_picture_caching: booltesting: boolgpu_supports_fast_clears: bool

Set to true if this GPU supports hardware fast clears as a performance optimization. Likely requires benchmarking on various GPUs to see if it is a performance win. The default is false, which tends to be best performance on lower end / integrated GPUs.

allow_dual_source_blending: boolallow_advanced_blend_equation: boolallow_pixel_local_storage_support: bool

If true, allow WR to use pixel local storage if the device supports it. For now, this defaults to false since the code is still experimental and not complete. This option will probably be removed once support is complete, and WR can implicitly choose whether to make use of PLS.

allow_texture_storage_support: bool

If true, allow textures to be initialized with glTexStorage. This affects VRAM consumption and data upload paths.

allow_texture_swizzling: bool

If true, we allow the data uploaded in a different format from the one expected by the driver, pretending the format is matching, and swizzling the components on all the shader sampling.

batch_lookback_count: usize

Number of batches to look back in history for adding the current transparent instance into.

start_debug_server: bool

Start the debug server for this renderer.

dump_shader_source: Option<String>

Output the source of the shader with the given name.

surface_origin_is_top_left: boolcompositor_config: CompositorConfig

The configuration options defining how WR composites the final scene.

enable_gpu_markers: bool