Struct webrender::renderer::upload::UploadTexturePool

pub struct UploadTexturePool {
    textures: [VecDeque<(Texture, u64)>; 4],
    delay_texture_deallocation: [u64; 4],
    current_frame: u64,
    temporary_buffers: Vec<Vec<MaybeUninit<u8>>>,
    min_temporary_buffers: usize,
    delay_buffer_deallocation: u64,
}

A very basic pool to avoid reallocating staging textures as well as staging CPU side buffers.

Fields
textures: [VecDeque<(Texture, u64)>; 4]

The textures in the pool associated with a last used frame index.

The outer array corresponds to each of teh three supported texture formats.

delay_texture_deallocation: [u64; 4]current_frame: u64temporary_buffers: Vec<Vec<MaybeUninit<u8>>>

Temporary buffers that are used when using staging uploads + glTexImage2D.

Temporary buffers aren’t used asynchronously so they can be reused every frame. To keep things simple we always allocate enough memory for formats with four bytes per pixel (more than we need for alpha-only textures but it works just as well).

min_temporary_buffers: usizedelay_buffer_deallocation: u64
Implementations

impl UploadTexturePool

pub fn new() -> Self

fn format_index(&self, format: ImageFormat) -> usize

pub fn begin_frame(&mut self)

pub fn get_texture( &mut self, device: &mut Device, format: ImageFormat ) -> Texture

Create or reuse a staging texture.

See also return_texture.

pub fn return_texture(&mut self, texture: Texture)

Hand the staging texture back to the pool after being done with uploads.

The texture must have been obtained from this pool via get_texture.

pub fn get_temporary_buffer(&mut self) -> Vec<MaybeUninit<u8>>

Create or reuse a temporary CPU buffer.

These buffers are used in the batched upload path when PBOs are not supported. Content is first written to the temporary buffer and uploaded via a single glTexSubImage2D call.

pub fn return_temporary_buffer(&mut self, buffer: Vec<MaybeUninit<u8>>)

Return memory that was obtained from this pool via get_temporary_buffer.

pub fn delete_textures(&mut self, device: &mut Device)

Deallocate this pool’s CPU and GPU memory.

pub fn end_frame(&mut self, device: &mut Device)

Deallocate some textures if there are too many for a long time.

pub fn report_memory_to( &self, report: &mut MemoryReport, size_op_funs: &MallocSizeOfOps )