webrender

Module resource_cache

Structs

CacheItem

CachedImageInfo 🔒

CachedImageKey 🔒

GlyphFetchResult

ImageGeneration

ImageProperties

ImageRequest

ImageResource 🔒

ImageTemplates 🔒

RenderTarget 🔒

Internal information about allocated render targets in the pool

ResourceCache

High-level container for resources managed by the RenderBackend.

ResourceClassCache

Resources 🔒

Enums

CachedImageData

Represents the backing store of an image in the cache. This storage can take several forms.

ImageCacheError

ImageResult 🔒

State 🔒

Statics

NEXT_NATIVE_SURFACE_ID 🔒

Functions

to_image_dirty_rect 🔒

Type Aliases

GlyphDimensionsCache

ImageCache 🔒

RasterizedBlob 🔒

Post scene building state.