Enum webrender::resource_cache::GlyphCacheMsg

enum GlyphCacheMsg {
    BeginFrame(FrameId, ResourceClassCache<RenderedGlyphKey, Option<FreeListItemId>>),
    AddFont(FontKey, FontTemplate),
    RequestGlyphs(FontKey, Au, ColorF, Vec<GlyphInstance>, FontRenderMode, Option<GlyphOptions>),
    EndFrame,
}

Message sent from the resource cache to the glyph cache thread.

Variants
BeginFrame(FrameId, ResourceClassCache<RenderedGlyphKey, Option<FreeListItemId>>)

Begin the frame - pass ownership of the glyph cache to the thread.

AddFont(FontKey, FontTemplate)

Add a new font.

RequestGlyphs(FontKey, Au, ColorF, Vec<GlyphInstance>, FontRenderMode, Option<GlyphOptions>)

Request glyphs for a text run.

EndFrame

Finished requesting glyphs. Reply with new glyphs.