webrender::scene

Struct BuiltScene

pub struct BuiltScene {

    pub has_root_pipeline: bool,
    pub pipeline_epochs: FastHashMap<PipelineId, Epoch>,
    pub output_rect: DeviceIntRect,
    pub prim_store: PrimitiveStore,
    pub clip_store: ClipStore,
    pub config: FrameBuilderConfig,
    pub hit_testing_scene: Arc<HitTestingScene>,
    pub tile_cache_config: TileCacheConfig,
    pub snapshot_pictures: Vec<PictureIndex>,
    pub tile_cache_pictures: Vec<PictureIndex>,
    pub picture_graph: PictureGraph,
    pub num_plane_splitters: usize,
    pub prim_instances: Vec<PrimitiveInstance>,
    pub surfaces: Vec<SurfaceInfo>,
    pub clip_tree: ClipTree,
    pub recycler_tx: Option<Sender<BuiltScene>>,

}

WebRender’s internal representation of the scene.

Fields
has_root_pipeline: boolpipeline_epochs: FastHashMap<PipelineId, Epoch>output_rect: DeviceIntRectprim_store: PrimitiveStoreclip_store: ClipStoreconfig: FrameBuilderConfighit_testing_scene: Arc<HitTestingScene>tile_cache_config: TileCacheConfigsnapshot_pictures: Vec<PictureIndex>tile_cache_pictures: Vec<PictureIndex>picture_graph: PictureGraphnum_plane_splitters: usizeprim_instances: Vec<PrimitiveInstance>surfaces: Vec<SurfaceInfo>clip_tree: ClipTreerecycler_tx: Option<Sender<BuiltScene>>

Deallocating memory outside of the thread that allocated it causes lock contention in jemalloc. To avoid this we send the built scene back to the scene builder thread when we don’t need it anymore, and in the process, also reuse some allocations.

Implementations

impl BuiltScene

pub fn empty() -> Self

pub fn recycle(self)

Send the scene back to the scene builder thread so that recycling/deallocations can happen there.

pub fn get_stats(&self) -> SceneStats

Get the memory usage statistics to pre-allocate for the next scene.

pub fn create_hit_tester(&mut self, spatial_tree: &SpatialTree) -> HitTester