Struct webrender::scene::SceneProperties

pub struct SceneProperties {
    transform_properties: HashMap<PropertyBindingId, LayoutTransform>,
    float_properties: HashMap<PropertyBindingId, f32>,
}

Stores a map of the animated property bindings for the current display list. These can be used to animate the transform and/or opacity of a display list without re-submitting the display list itself.

Fields
transform_properties: HashMap<PropertyBindingId, LayoutTransform> float_properties: HashMap<PropertyBindingId, f32>
Methods

impl SceneProperties

fn new() -> SceneProperties

fn set_properties(&mut self, properties: DynamicProperties)

Set the current property list for this display list.

fn resolve_layout_transform(&self,
property: Option<&PropertyBinding<LayoutTransform>>)
-> LayoutTransform

Get the current value for a transform property.

fn resolve_float(&self,
property: &PropertyBinding<f32>,
default_value: f32)
-> f32

Get the current value for a float property.