webrender

Module scene_builder_thread

Macros

declare_interners 🔒

Structs

BuiltTransaction

Represent the remaining work associated to a transaction after the scene building phase as well as the result of scene building itself if applicable.

Document 🔒

InternerUpdates

Interners

This struct contains all items that can be shared between display lists. We want to intern and share the same clips, primitives and other things between display lists so that:

LowPrioritySceneBuilderThread

A scene builder thread which executes expensive operations such as blob rasterization with a lower priority than the normal scene builder thread.

OffscreenBuiltScene

SceneBuilderThread

SceneBuilderThreadChannels

Enums

SceneBuilderRequest

Message to the scene builder thread.

SceneBuilderResult

SceneSwapResult

Functions

rasterize_blobs 🔒