webrender::scene_builder_thread

Struct BuiltTransaction

pub struct BuiltTransaction {

    pub document_id: DocumentId,
    pub built_scene: Option<BuiltScene>,
    pub offscreen_scenes: Vec<OffscreenBuiltScene>,
    pub view: SceneView,
    pub resource_updates: Vec<ResourceUpdate>,
    pub rasterized_blobs: Vec<(BlobImageRequest, BlobImageResult)>,
    pub blob_rasterizer: Option<Box<dyn AsyncBlobImageRasterizer>>,
    pub frame_ops: Vec<FrameMsg>,
    pub removed_pipelines: Vec<(PipelineId, DocumentId)>,
    pub notifications: Vec<NotificationRequest>,
    pub interner_updates: Option<InternerUpdates>,
    pub spatial_tree_updates: Option<SpatialTreeUpdates>,
    pub render_frame: bool,
    pub present: bool,
    pub tracked: bool,
    pub invalidate_rendered_frame: bool,
    pub profile: TransactionProfile,
    pub frame_stats: FullFrameStats,

}

Represent the remaining work associated to a transaction after the scene building phase as well as the result of scene building itself if applicable.

Fields
document_id: DocumentIdbuilt_scene: Option<BuiltScene>offscreen_scenes: Vec<OffscreenBuiltScene>view: SceneViewresource_updates: Vec<ResourceUpdate>rasterized_blobs: Vec<(BlobImageRequest, BlobImageResult)>blob_rasterizer: Option<Box<dyn AsyncBlobImageRasterizer>>frame_ops: Vec<FrameMsg>removed_pipelines: Vec<(PipelineId, DocumentId)>notifications: Vec<NotificationRequest>interner_updates: Option<InternerUpdates>spatial_tree_updates: Option<SpatialTreeUpdates>render_frame: boolpresent: booltracked: boolinvalidate_rendered_frame: boolprofile: TransactionProfileframe_stats: FullFrameStats