Struct webrender::scene_builder_thread::Interners

pub struct Interners {

    pub clip: Interner<ClipIntern>,
    pub prim: Interner<PrimitiveKeyKind>,
    pub normal_border: Interner<NormalBorderPrim>,
    pub image_border: Interner<ImageBorder>,
    pub image: Interner<Image>,
    pub yuv_image: Interner<YuvImage>,
    pub line_decoration: Interner<LineDecoration>,
    pub linear_grad: Interner<LinearGradient>,
    pub radial_grad: Interner<RadialGradient>,
    pub conic_grad: Interner<ConicGradient>,
    pub picture: Interner<Picture>,
    pub text_run: Interner<TextRun>,
    pub filter_data: Interner<FilterDataIntern>,
    pub backdrop_capture: Interner<BackdropCapture>,
    pub backdrop_render: Interner<BackdropRender>,
    pub polygon: Interner<PolygonIntern>,
    pub box_shadow: Interner<BoxShadow>,

}

This struct contains all items that can be shared between display lists. We want to intern and share the same clips, primitives and other things between display lists so that:

GPU cache handles remain valid, reducing GPU cache updates.

Comparison of primitives and pictures between two display lists is (a) fast (b) done during scene building.

Fields
clip: Interner<ClipIntern>prim: Interner<PrimitiveKeyKind>normal_border: Interner<NormalBorderPrim>image_border: Interner<ImageBorder>image: Interner<Image>yuv_image: Interner<YuvImage>line_decoration: Interner<LineDecoration>linear_grad: Interner<LinearGradient>radial_grad: Interner<RadialGradient>conic_grad: Interner<ConicGradient>picture: Interner<Picture>text_run: Interner<TextRun>filter_data: Interner<FilterDataIntern>backdrop_capture: Interner<BackdropCapture>backdrop_render: Interner<BackdropRender>polygon: Interner<PolygonIntern>box_shadow: Interner<BoxShadow>
Implementations

impl Interners

fn report_memory(&self, ops: &mut MallocSizeOfOps, r: &mut MemoryReport)

Reports CPU heap memory used by the interners.

fn end_frame_and_get_pending_updates(&mut self) -> InternerUpdates