Struct webrender::scene_builder_thread::SceneBuilderThread

pub struct SceneBuilderThread {
    documents: FastHashMap<DocumentId, Document>,
    rx: Receiver<SceneBuilderRequest>,
    tx: Sender<ApiMsg>,
    config: FrameBuilderConfig,
    fonts: SharedFontResources,
    size_of_ops: Option<MallocSizeOfOps>,
    hooks: Option<Box<dyn SceneBuilderHooks + Send>>,
    simulate_slow_ms: u32,
    removed_pipelines: HashSet<PipelineId, BuildHasherDefault<FxHasher>>,
    capture_config: Option<CaptureConfig>,
}

Fields
documents: FastHashMap<DocumentId, Document>rx: Receiver<SceneBuilderRequest>tx: Sender<ApiMsg>config: FrameBuilderConfigfonts: SharedFontResourcessize_of_ops: Option<MallocSizeOfOps>hooks: Option<Box<dyn SceneBuilderHooks + Send>>simulate_slow_ms: u32removed_pipelines: HashSet<PipelineId, BuildHasherDefault<FxHasher>>capture_config: Option<CaptureConfig>
Implementations

impl SceneBuilderThread

pub fn new( config: FrameBuilderConfig, fonts: SharedFontResources, size_of_ops: Option<MallocSizeOfOps>, hooks: Option<Box<dyn SceneBuilderHooks + Send>>, channels: SceneBuilderThreadChannels ) -> Self

pub fn send(&self, msg: SceneBuilderResult)

Send a message to the render backend thread.

We first put something in the result queue and then send a wake-up message to the api queue that the render backend is blocking on.

pub fn run(&mut self)

The scene builder thread’s event loop.

fn save_scene(&mut self, config: CaptureConfig)

fn save_capture_sequence(&mut self)

fn start_capture_sequence(&mut self, config: CaptureConfig)

fn process_transaction(&mut self, txn: TransactionMsg) -> Box<BuiltTransaction>

Do the bulk of the work of the scene builder thread.

fn forward_built_transactions(&mut self, txns: Vec<Box<BuiltTransaction>>)

Send the results of process_transaction back to the render backend.

fn report_memory(&mut self) -> MemoryReport

Reports CPU heap memory used by the SceneBuilder.