Struct webrender::spatial_tree::SpatialTree

pub struct SpatialTree {
    spatial_nodes: Vec<SpatialNode>,
    coord_systems: Vec<CoordinateSystem>,
    root_reference_frame_index: SpatialNodeIndex,
    update_state_stack: Vec<TransformUpdateState>,
}

Represents the spatial tree during frame building, which is mostly read-only, apart from the tree update at the start of the frame

Fields
spatial_nodes: Vec<SpatialNode>

Nodes which determine the positions (offsets and transforms) for primitives and clips.

coord_systems: Vec<CoordinateSystem>

A list of transforms that establish new coordinate systems. Spatial nodes only establish a new coordinate system when they have a transform that is not a simple 2d translation.

root_reference_frame_index: SpatialNodeIndexupdate_state_stack: Vec<TransformUpdateState>

Stack of current state for each parent node while traversing and updating tree

Implementations

impl SpatialTree

pub fn new() -> Self

fn visit_node_impl_mut<F>(&mut self, index: SpatialNodeIndex, f: &mut F)where F: FnMut(SpatialNodeIndex, &mut SpatialNode),

fn visit_node_impl<F>(&self, index: SpatialNodeIndex, f: &mut F)where F: FnMut(SpatialNodeIndex, &SpatialNode),

pub fn visit_nodes<F>(&self, f: F)where F: FnMut(SpatialNodeIndex, &SpatialNode),

Visit all nodes from the root of the tree, invoking a closure on each one

pub fn visit_nodes_mut<F>(&mut self, f: F)where F: FnMut(SpatialNodeIndex, &mut SpatialNode),

Visit all nodes from the root of the tree, invoking a closure on each one

pub fn apply_updates(&mut self, updates: SpatialTreeUpdates)

Apply updates from a new scene to the frame spatial tree

pub fn get_last_sampled_scroll_offsets( &self ) -> FastHashMap<ExternalScrollId, Vec<SampledScrollOffset>>

pub fn apply_last_sampled_scroll_offsets( &mut self, last_sampled_offsets: FastHashMap<ExternalScrollId, Vec<SampledScrollOffset>> )

pub fn get_spatial_node(&self, index: SpatialNodeIndex) -> &SpatialNode

pub fn get_spatial_node_mut( &mut self, index: SpatialNodeIndex ) -> &mut SpatialNode

pub fn spatial_node_count(&self) -> usize

Get total number of spatial nodes

pub fn find_spatial_node_by_anim_id( &self, id: PropertyBindingId ) -> Option<SpatialNodeIndex>

pub fn get_relative_transform( &self, child_index: SpatialNodeIndex, parent_index: SpatialNodeIndex ) -> CoordinateSpaceMapping<LayoutPixel, LayoutPixel>

Calculate the relative transform from child_index to parent_index. This method will panic if the nodes are not connected!

pub fn get_relative_transform_with_face( &self, child_index: SpatialNodeIndex, parent_index: SpatialNodeIndex, visible_face: Option<&mut VisibleFace> ) -> CoordinateSpaceMapping<LayoutPixel, LayoutPixel>

Calculate the relative transform from child_index to parent_index. This method will panic if the nodes are not connected! Also, switch the visible face to Back if at any stage where the combined transform is flattened, we see the back face.

pub fn is_matching_coord_system( &self, index0: SpatialNodeIndex, index1: SpatialNodeIndex ) -> bool

Returns true if both supplied spatial nodes are in the same coordinate system (implies the relative transform produce axis-aligned rects).

fn get_world_transform_impl( &self, index: SpatialNodeIndex, scroll: TransformScroll ) -> CoordinateSpaceMapping<LayoutPixel, WorldPixel>

pub fn get_world_transform( &self, index: SpatialNodeIndex ) -> CoordinateSpaceMapping<LayoutPixel, WorldPixel>

Calculate the relative transform from index to the root.

pub fn get_world_viewport_transform( &self, index: SpatialNodeIndex ) -> CoordinateSpaceMapping<LayoutPixel, WorldPixel>

Calculate the relative transform from index to the root. Unlike get_world_transform, this variant doesn’t account for the local scroll offset.

pub fn root_reference_frame_index(&self) -> SpatialNodeIndex

The root reference frame, which is the true root of the SpatialTree.

pub fn set_scroll_offsets( &mut self, id: ExternalScrollId, offsets: Vec<SampledScrollOffset> ) -> bool

pub fn update_tree(&mut self, scene_properties: &SceneProperties)

fn update_node( &mut self, node_index: SpatialNodeIndex, scene_properties: &SceneProperties )

pub fn build_transform_palette(&self) -> TransformPalette

fn print_node<T: PrintTreePrinter>(&self, index: SpatialNodeIndex, pt: &mut T)

pub fn get_local_visible_face( &self, node_index: SpatialNodeIndex ) -> VisibleFace

Get the visible face of the transfrom from the specified node to its parent.

pub fn print(&self)