webrender::spatial_tree

Trait SpatialNodeContainer

pub trait SpatialNodeContainer {
    // Required method
    fn get_node_info(&self, index: SpatialNodeIndex) -> SpatialNodeInfo<'_>;

    // Provided method
    fn get_snapping_info(
        &self,
        parent_index: Option<SpatialNodeIndex>,
    ) -> Option<ScaleOffset> { ... }
}

Allows functions and methods to retrieve common information about a spatial node, whether during scene or frame building

Required Methods

fn get_node_info(&self, index: SpatialNodeIndex) -> SpatialNodeInfo<'_>

Get the common information for a given spatial node

Provided Methods

fn get_snapping_info( &self, parent_index: Option<SpatialNodeIndex>, ) -> Option<ScaleOffset>