Struct webrender::TileCacheLoggerUpdateLists

pub struct TileCacheLoggerUpdateLists {
    pub clip: (Vec<String>, Vec<UpdateList<<ClipIntern as Internable>::Key>>),
    pub prim: (Vec<String>, Vec<UpdateList<<PrimitiveKeyKind as Internable>::Key>>),
    pub normal_border: (Vec<String>, Vec<UpdateList<<NormalBorderPrim as Internable>::Key>>),
    pub image_border: (Vec<String>, Vec<UpdateList<<ImageBorder as Internable>::Key>>),
    pub image: (Vec<String>, Vec<UpdateList<<Image as Internable>::Key>>),
    pub yuv_image: (Vec<String>, Vec<UpdateList<<YuvImage as Internable>::Key>>),
    pub line_decoration: (Vec<String>, Vec<UpdateList<<LineDecoration as Internable>::Key>>),
    pub linear_grad: (Vec<String>, Vec<UpdateList<<LinearGradient as Internable>::Key>>),
    pub radial_grad: (Vec<String>, Vec<UpdateList<<RadialGradient as Internable>::Key>>),
    pub conic_grad: (Vec<String>, Vec<UpdateList<<ConicGradient as Internable>::Key>>),
    pub picture: (Vec<String>, Vec<UpdateList<<Picture as Internable>::Key>>),
    pub text_run: (Vec<String>, Vec<UpdateList<<TextRun as Internable>::Key>>),
    pub filter_data: (Vec<String>, Vec<UpdateList<<FilterDataIntern as Internable>::Key>>),
    pub backdrop: (Vec<String>, Vec<UpdateList<<Backdrop as Internable>::Key>>),
}

Fields
clip: (Vec<String>, Vec<UpdateList<<ClipIntern as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

prim: (Vec<String>, Vec<UpdateList<<PrimitiveKeyKind as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

normal_border: (Vec<String>, Vec<UpdateList<<NormalBorderPrim as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

image_border: (Vec<String>, Vec<UpdateList<<ImageBorder as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

image: (Vec<String>, Vec<UpdateList<<Image as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

yuv_image: (Vec<String>, Vec<UpdateList<<YuvImage as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

line_decoration: (Vec<String>, Vec<UpdateList<<LineDecoration as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

linear_grad: (Vec<String>, Vec<UpdateList<<LinearGradient as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

radial_grad: (Vec<String>, Vec<UpdateList<<RadialGradient as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

conic_grad: (Vec<String>, Vec<UpdateList<<ConicGradient as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

picture: (Vec<String>, Vec<UpdateList<<Picture as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

text_run: (Vec<String>, Vec<UpdateList<<TextRun as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

filter_data: (Vec<String>, Vec<UpdateList<<FilterDataIntern as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

backdrop: (Vec<String>, Vec<UpdateList<<Backdrop as Internable>::Key>>)

Generate storage, one per interner. the tuple is a workaround to avoid the need for multiple fields that start with $name (macro concatenation). the string is .ron serialized updatelist at capture time; the updates is the list of DataStore updates (avoid UpdateList due to Default() requirements on the Keys) reconstructed at load time.

Implementations

impl TileCacheLoggerUpdateLists

pub fn new() -> Self

fn serialize_updates(&mut self, updates: &InternerUpdates)

serialize all interners in updates to .ron

fn is_empty(&self) -> bool

fn to_ron(&self) -> String