webrender

Module surface

Contains functionality to help building the render task graph from a series of off-screen surfaces that are created during the prepare pass, and other surface related types and helpers.

Re-exports

pub use crate::picture_composite_mode::get_surface_rects;

Structs

CommandBufferTargets 🔒

SurfaceBuilder

SurfaceDescriptor

SurfaceIndex

An index into the surface array

SurfaceInfo

Information about an offscreen surface. For now, it contains information about the size and coordinate system of the surface. In the future, it will contain information about the contents of the surface, which will allow surfaces to be cached / retained between frames and display lists.

SurfaceTileDescriptor

Enums

SubpixelMode

Specify whether a surface allows subpixel AA text rendering.

SurfaceDescriptorKind

Constants

MAX_BLUR_RADIUS 🔒

Maximum blur radius for blur filter

Functions

calculate_screen_uv