webrender::surface

Struct SurfaceInfo

pub struct SurfaceInfo {

    pub unclipped_local_rect: PictureRect,
    pub clipped_local_rect: PictureRect,
    pub clipping_rect: PictureRect,
    pub culling_rect: VisRect,
    pub map_local_to_picture: SpaceMapper<LayoutPixel, PicturePixel>,
    pub surface_spatial_node_index: SpatialNodeIndex,
    pub raster_spatial_node_index: SpatialNodeIndex,
    pub visibility_spatial_node_index: SpatialNodeIndex,
    pub device_pixel_scale: DevicePixelScale,
    pub world_scale_factors: (f32, f32),
    pub local_scale: (f32, f32),
    pub is_opaque: bool,
    pub allow_snapping: bool,
    pub force_scissor_rect: bool,

}

Information about an offscreen surface. For now, it contains information about the size and coordinate system of the surface. In the future, it will contain information about the contents of the surface, which will allow surfaces to be cached / retained between frames and display lists.

Fields
unclipped_local_rect: PictureRect

A local rect defining the size of this surface, in the coordinate system of the parent surface. This contains the unclipped bounding rect of child primitives.

SNAPTODO: This rect is built by mapping per-cluster bounding rects (and child-surface coverage rects) into this surface’s picture space via map_local_to_picture. Even once the source cluster bound is a true union of per-prim snapped local rects, the resulting unclipped_local_rect is not guaranteed to be snapped: any 2D transform that isn’t an axis-aligned, integer-pixel translation between the cluster/child-surface spatial node and this surface’s spatial node will produce sub-pixel edges in picture space. Float blur-margin inflation inside composite_mode.get_coverage can also break the snap. Consumers that need a snapped value will either need to re-snap in surface space or restrict the snap path to surfaces where the cross-space mapping preserves grid alignment (see SurfaceInfo.allow_snapping).

clipped_local_rect: PictureRect

The local space coverage of child primitives after they are are clipped to their owning clip-chain.

clipping_rect: PictureRect

The (conservative) valid part of this surface rect. Used to reduce the size of render target allocation.

culling_rect: VisRect

The rectangle to use for culling and clipping.

map_local_to_picture: SpaceMapper<LayoutPixel, PicturePixel>

Helper structs for mapping local rects in different coordinate systems into the picture coordinates.

surface_spatial_node_index: SpatialNodeIndex

The positioning node for the surface itself,

raster_spatial_node_index: SpatialNodeIndex

The rasterization root for this surface.

visibility_spatial_node_index: SpatialNodeIndex

The spatial node for culling and clipping (anything using VisPixel). TODO: Replace with the raster spatial node.

device_pixel_scale: DevicePixelScale

The device pixel ratio specific to this surface.

world_scale_factors: (f32, f32)

The scale factors of the surface to world transform.

local_scale: (f32, f32)

Local scale factors surface to raster transform

is_opaque: bool

If true, we know this surface is completely opaque.

allow_snapping: bool

If true, allow snapping on this and child surfaces

force_scissor_rect: bool

If true, the scissor rect must be set when drawing this surface

Implementations

impl SurfaceInfo

pub fn new( surface_spatial_node_index: SpatialNodeIndex, raster_spatial_node_index: SpatialNodeIndex, world_rect: WorldRect, spatial_tree: &SpatialTree, device_pixel_scale: DevicePixelScale, world_scale_factors: (f32, f32), local_scale: (f32, f32), allow_snapping: bool, force_scissor_rect: bool, ) -> Self

pub fn clamp_blur_radius( &self, x_blur_radius: f32, y_blur_radius: f32, ) -> (f32, f32)

Clamps the blur radius depending on scale factors.

pub fn update_culling_rect( &mut self, parent_culling_rect: VisRect, composite_mode: &PictureCompositeMode, frame_context: &FrameVisibilityContext<'_>, )

pub fn map_to_device_rect( &self, picture_rect: &PictureRect, spatial_tree: &SpatialTree, ) -> DeviceRect

pub fn get_surface_rect( &self, local_rect: &PictureRect, spatial_tree: &SpatialTree, ) -> Option<DeviceIntRect>

Clip and transform a local rect to a device rect suitable for allocating a child off-screen surface of this surface (e.g. for clip-masks)