webrender

Module svg_filter

Structs

FilterGraphNode

FilterGraphNodeKey

FilterGraphPictureReference

FilterGraphPictureReferenceKey

Enums

FilterGraphOp

FilterGraphOpKey

FilterGraphPictureBufferIdKey

Constants

SVGFE_CONVOLVE_DIAMETER_LIMIT

SVGFE_CONVOLVE_VALUES_LIMIT

Functions

get_coverage_source_svgfe

Here we transform target rect to source rect for SVGFEGraph by walking the whole graph and propagating subregions based on the provided invalidation rect, and we want it to be a tight fit so we don’t waste time applying multiple filters to pixels that do not contribute to the invalidated rect.

get_coverage_target_svgfe

Here we transform source rect to target rect for SVGFEGraph by walking the whole graph and propagating subregions based on the provided invalidation rect, and we want it to be a tight fit so we don’t waste time applying multiple filters to pixels that do not contribute to the invalidated rect.