webrender::svg_filter

Function get_coverage_target_svgfe

pub fn get_coverage_target_svgfe(
    filters: &[(FilterGraphNode, FilterGraphOp)],
    surface_rect: LayoutRect,
) -> LayoutRect

Here we transform source rect to target rect for SVGFEGraph by walking the whole graph and propagating subregions based on the provided invalidation rect, and we want it to be a tight fit so we don’t waste time applying multiple filters to pixels that do not contribute to the invalidated rect.

The interesting parts of the handling of SVG filters are:

scene_building.rs : wrap_prim_with_filters

picture.rs : get_coverage_target_svgfe (you are here)

picture.rs : get_coverage_source_svgfe

render_task.rs : new_svg_filter_graph

render_target.rs : add_svg_filter_node_instances