Module webrender::texture_allocator

Structs

ArrayAllocationTracker
A texture allocator using the guillotine algorithm with the rectangle merge improvement. See sections 2.2 and 2.2.5 in "A Thousand Ways to Pack the Bin - A Practical Approach to Two- Dimensional Rectangle Bin Packing":

FreeListBin

FreeListIndex

FreeRect

FreeRectSlice

Constants

FIND_SMALLEST_AREA
This is an optimization tweak to enable looking through all the free rectangles in a bin and choosing the smallest, as opposed to picking the first match.

MIN_RECT_AXIS_SIZES
The minimum number of pixels on each side that we require for rects to be classified as particular bin of freelists.

NUM_BINS