webrender

Module texture_cache

Structs

CacheAllocParams 🔒

Container struct for the various parameters used in cache allocation.

CacheEntry

EvictionNotice

SharedTextures 🔒

A set of lazily allocated, fixed size, texture arrays for each format the texture cache supports.

TextureCache

General-purpose manager for images in GPU memory. This includes images, rasterized glyphs, rasterized blobs, cached render tasks, etc.

TextureCacheConfig

Startup parameters for the texture cache.

TextureParameters

Enums

AutoCacheEntryMarker

BudgetType 🔒

The different budget types for the texture cache. Each type has its own memory budget. Once the budget is exceeded, entries with automatic eviction are evicted. Entries with manual eviction share the same budget but are not evicted once the budget is exceeded. Keeping separate budgets ensures that we don’t evict entries from unrelated textures if one texture gets full.

EntryDetails

Items in the texture cache can either be standalone textures, or a sub-rect inside the shared cache.

Eviction

Describes the eviction policy for a given entry in the texture cache.

ManualCacheEntryMarker

TargetShader

Information about which shader will use the entry.

TextureCacheHandle

A texture cache handle is a weak reference to a cache entry.

Constants

TEXTURE_REGION_DIMENSIONS

The size of each region in shared cache texture arrays.