Enum webrender::texture_cache::BudgetType

#[repr(u8)]
enum BudgetType {
    SharedColor8Linear,
    SharedColor8Nearest,
    SharedColor8Glyphs,
    SharedAlpha8,
    SharedAlpha8Glyphs,
    SharedAlpha16,
    Standalone,
}

The different budget types for the texture cache. Each type has its own memory budget. Once the budget is exceeded, entries with automatic eviction are evicted. Entries with manual eviction share the same budget but are not evicted once the budget is exceeded. Keeping separate budgets ensures that we don’t evict entries from unrelated textures if one texture gets full.

Variants

SharedColor8Linear

SharedColor8Nearest

SharedColor8Glyphs

SharedAlpha8

SharedAlpha8Glyphs

SharedAlpha16

Standalone

Implementations

impl BudgetType

pub const COUNT: usize = 7usize

pub const VALUES: [BudgetType; 7] = _

pub const PRESSURE_COUNTERS: [usize; 7] = _

pub fn iter() -> impl Iterator<Item = BudgetType>