Struct webrender::texture_cache::CacheEntry

pub struct CacheEntry {
    pub size: DeviceIntSize,
    pub details: EntryDetails,
    pub user_data: [f32; 4],
    pub last_access: FrameStamp,
    pub uv_rect_handle: GpuCacheHandle,
    pub input_format: ImageFormat,
    pub filter: TextureFilter,
    pub swizzle: Swizzle,
    pub texture_id: CacheTextureId,
    pub eviction_notice: Option<EvictionNotice>,
    pub uv_rect_kind: UvRectKind,
    pub shader: TargetShader,
}

Fields
size: DeviceIntSize

Size of the requested item, in device pixels. Does not include any padding for alignment that the allocator may have added to this entry’s allocation.

details: EntryDetails

Details specific to standalone or shared items.

user_data: [f32; 4]

Arbitrary user data associated with this item.

last_access: FrameStamp

The last frame this item was requested for rendering.

uv_rect_handle: GpuCacheHandle

Handle to the resource rect in the GPU cache.

input_format: ImageFormat

Image format of the data that the entry expects.

filter: TextureFilterswizzle: Swizzletexture_id: CacheTextureId

The actual device texture ID this is part of.

eviction_notice: Option<EvictionNotice>

Optional notice when the entry is evicted from the cache.

uv_rect_kind: UvRectKind

The type of UV rect this entry specifies.

shader: TargetShader
Implementations

impl CacheEntry

fn new_standalone( texture_id: CacheTextureId, last_access: FrameStamp, params: &CacheAllocParams, swizzle: Swizzle, size_in_bytes: usize ) -> Self

fn update_gpu_cache(&mut self, gpu_cache: &mut GpuCache)

fn evict(&self)

fn alternative_input_format(&self) -> ImageFormat