Struct webrender::texture_cache::TextureCache

pub struct TextureCache {
    shared_textures: SharedTextures,
    picture_textures: Vec<WholeTextureArray>,
    max_texture_size: i32,
    max_texture_layers: usize,
    swizzle: Option<SwizzleSettings>,
    debug_flags: DebugFlags,
    next_id: CacheTextureId,
    pending_updates: TextureUpdateList,
    now: FrameStamp,
    reached_reclaim_threshold: Option<SystemTime>,
    entries: FreeList<CacheEntry, CacheEntryMarker>,
    per_doc_data: HashMap<DocumentId, PerDocumentData, BuildHasherDefault<FxHasher>>,
    doc_data: PerDocumentData,
    require_frame_build: bool,
}

General-purpose manager for images in GPU memory. This includes images, rasterized glyphs, rasterized blobs, cached render tasks, etc.

The texture cache is owned and managed by the RenderBackend thread, and produces a series of commands to manipulate the textures on the Renderer thread. These commands are executed before any rendering is performed for a given frame.

Entries in the texture cache are not guaranteed to live past the end of the frame in which they are requested, and may be evicted. The API supports querying whether an entry is still available.

The TextureCache is different from the GpuCache in that the former stores images, whereas the latter stores data and parameters for use in the shaders. This means that the texture cache can be visualized, which is a good way to understand how it works. Enabling gfx.webrender.debug.texture-cache shows a live view of its contents in Firefox.

Fields
shared_textures: SharedTextures

Set of texture arrays in different formats used for the shared cache.

picture_textures: Vec<WholeTextureArray>

A single texture array for picture caching.

max_texture_size: i32

Maximum texture size supported by hardware.

max_texture_layers: usize

Maximum number of texture layers supported by hardware.

swizzle: Option<SwizzleSettings>

Settings on using texture unit swizzling.

debug_flags: DebugFlags

The current set of debug flags.

next_id: CacheTextureId

The next unused virtual texture ID. Monotonically increasing.

pending_updates: TextureUpdateList

A list of allocations and updates that need to be applied to the texture cache in the rendering thread this frame.

now: FrameStamp

The current FrameStamp. Used for cache eviction policies.

reached_reclaim_threshold: Option<SystemTime>

The time at which we first reached the byte threshold for reclaiming cache memory. `None if we haven't reached the threshold.

entries: FreeList<CacheEntry, CacheEntryMarker>

Maintains the list of all current items in the texture cache.

per_doc_data: HashMap<DocumentId, PerDocumentData, BuildHasherDefault<FxHasher>>

Holds items that need to be maintained on a per-document basis. If we modify this data for a document without also building a frame for that document, then we might end up erroneously evicting items out from under that document.

doc_data: PerDocumentData

The current document's data. This is moved out of per_doc_data in begin_frame and moved back in end_frame to solve borrow checker issues. We should try removing this when we require a rustc with NLL.

require_frame_build: bool

This indicates that we performed a cleanup operation which requires all documents to build a frame.

Methods

impl TextureCache

pub fn new(
max_texture_size: i32,
max_texture_layers: usize,
picture_tile_sizes: &[DeviceIntSize],
initial_size: DeviceIntSize,
color_formats: TextureFormatPair<ImageFormat>,
swizzle: Option<SwizzleSettings>
) -> Self

pub fn set_debug_flags(&mut self, flags: DebugFlags)

fn clear_kind(&mut self, kind: EntryKind)

Clear all entries of the specified kind.

fn clear_standalone(&mut self)

fn clear_picture(&mut self)

fn clear_shared(&mut self)

pub fn clear_all(&mut self)

Clear all entries in the texture cache. This is a fairly drastic step that should only be called very rarely.

fn set_doc_data(&mut self)

fn unset_doc_data(&mut self)

pub fn prepare_for_frames(&mut self, time: SystemTime)

pub fn bookkeep_after_frames(&mut self)

pub fn requires_frame_build(&self) -> bool

pub fn begin_frame(&mut self, stamp: FrameStamp)

Called at the beginning of each frame.

fn maybe_reclaim_shared_memory(&mut self, time: SystemTime)

fn maybe_do_periodic_gc(&mut self)

Called at the beginning of each frame to periodically GC by expiring old shared entries. If necessary, the shared memory opened up as a result of expiring these entries will be reclaimed before the next group of document frames.

pub fn end_frame(
&mut self,
texture_cache_profile: &mut TextureCacheProfileCounters
)

pub fn request(
&mut self,
handle: &WeakFreeListHandle<CacheEntryMarker>,
gpu_cache: &mut GpuCache
) -> bool

pub fn needs_upload(
&self,
handle: &WeakFreeListHandle<CacheEntryMarker>
) -> bool

pub fn max_texture_size(&self) -> i32

pub fn pending_updates(&mut self) -> TextureUpdateList

pub fn update(
&mut self,
handle: &mut WeakFreeListHandle<CacheEntryMarker>,
descriptor: ImageDescriptor,
filter: TextureFilter,
data: Option<CachedImageData>,
user_data: [f32; 3],
dirty_rect: ImageDirtyRect,
gpu_cache: &mut GpuCache,
eviction_notice: Option<&EvictionNotice>,
uv_rect_kind: UvRectKind,
eviction: Eviction
)

pub fn is_allocated(
&self,
handle: &WeakFreeListHandle<CacheEntryMarker>
) -> bool

pub fn is_recently_used(
&self,
handle: &WeakFreeListHandle<CacheEntryMarker>,
margin: usize
) -> bool

pub fn get_allocated_size(
&self,
handle: &WeakFreeListHandle<CacheEntryMarker>
) -> Option<usize>

pub fn get(&self, handle: &WeakFreeListHandle<CacheEntryMarker>) -> CacheItem

pub fn get_cache_location(
&self,
handle: &WeakFreeListHandle<CacheEntryMarker>
) -> (CacheTextureId, usize, DeviceIntRect, Swizzle, GpuCacheHandle)

A more detailed version of get(). This allows access to the actual device rect of the cache allocation.

Returns a tuple identifying the texture, the layer, the region, and its GPU handle.

pub fn mark_unused(&mut self, handle: &WeakFreeListHandle<CacheEntryMarker>)

fn default_eviction(&self) -> EvictionThreshold

Returns the default eviction policy.

These parameters come from very rough instrumentation of hits in the shared cache, with simple browsing on a few pages. In rough terms, more than 99.5% of cache hits occur for entries that were used in the previous frame. This is obviously the dominant case, but we still want good behavior in long-tail cases (i.e. a large image is scrolled off-screen and on again). If we exclude immediately-reused (first frame) entries, 70% of the remaining hits happen within the first 200 frames. So we can be relatively agressive about eviction without sacrificing much in terms of cache performance. The one wrinkle is that animation-heavy pages do tend to extend the distribution, presumably because they churn through FrameIds faster than their more-static counterparts. As such, we also provide a time floor (which was not measured with the same degree of rigour).

fn expire_old_entries(&mut self, kind: EntryKind, threshold: EvictionThreshold)

Shared eviction code for standalone and shared entries.

See EvictionThreshold for more details on policy.

fn maybe_expire_old_shared_entries(
&mut self,
threshold: EvictionThreshold
) -> bool

Expires old shared entries, if we haven't done so this frame.

Returns true if any entries were expired.

fn free(&mut self, entry: &CacheEntry)

fn has_space_in_shared_cache(&mut self, params: &CacheAllocParams) -> bool

Check if we can allocate this entry without growing any of the texture cache arrays.

fn allocate_from_shared_cache(
&mut self,
params: &CacheAllocParams
) -> CacheEntry

Allocate a block from the shared cache.

pub fn is_allowed_in_shared_cache(
&self,
filter: TextureFilter,
descriptor: &ImageDescriptor
) -> bool

fn allocate_standalone_entry(&mut self, params: &CacheAllocParams) -> CacheEntry

Allocates a new standalone cache entry.

fn allocate_cache_entry(&mut self, params: &CacheAllocParams) -> CacheEntry

Allocates a cache entry appropriate for the given parameters.

This allocates from the shared cache unless the parameters do not meet the shared cache requirements, in which case a standalone texture is used.

fn upsert_entry(
&mut self,
cache_entry: CacheEntry,
handle: &mut WeakFreeListHandle<CacheEntryMarker>
)

fn allocate(
&mut self,
params: &CacheAllocParams,
handle: &mut WeakFreeListHandle<CacheEntryMarker>
)

Allocates a cache entry for the given parameters, and updates the provided handle to point to the new entry.

pub fn update_picture_cache(
&mut self,
tile_size: DeviceIntSize,
handle: &mut WeakFreeListHandle<CacheEntryMarker>,
gpu_cache: &mut GpuCache
)

pub fn shared_alpha_expected_format(&self) -> ImageFormat

pub fn shared_color_expected_format(&self) -> ImageFormat