Struct webrender::texture_cache::TexturePage

pub struct TexturePage {
    texture_id: TextureId,
    texture_size: u32,
    free_list: FreeRectList,
    allocations: u32,
    dirty: bool,
}

A texture allocator using the guillotine algorithm with the rectangle merge improvement. See sections 2.2 and 2.2.5 in "A Thousand Ways to Pack the Bin - A Practical Approach to Two- Dimensional Rectangle Bin Packing":

http://clb.demon.fi/files/RectangleBinPack.pdf

This approach was chosen because of its simplicity, good performance, and easy support for dynamic texture deallocation.

Fields

texture_id

texture_size

free_list

allocations

dirty

Methods

impl TexturePage

fn new(texture_id: TextureId, texture_size: u32) -> TexturePage

fn texture_id(&self) -> TextureId

fn find_index_of_best_rect_in_bin(&self, bin: FreeListBin, requested_dimensions: &Size2D<u32>) -> Option<FreeListIndex>

fn find_index_of_best_rect(&self, requested_dimensions: &Size2D<u32>) -> Option<FreeListIndex>

fn allocate(&mut self, requested_dimensions: &Size2D<u32>, filter: TextureFilter) -> Option<Point2D<u32>>

fn coalesce(&mut self)

fn clear(&mut self)

fn free(&mut self, rect: &Rect<u32>)

fn grow(&mut self, new_texture_size: u32)

fn can_grow(&self) -> bool