webrender

Module texture_pack

Re-exports

pub use guillotine::*;

Modules

guillotine 🔒

Structs

AllocId

ID of an allocation within a given allocator.

AllocatorList

BucketedShelfAllocator

A faster but less precise Shelf-packing dynamic texture atlas allocator, inspired by https://github.com/mapbox/shelf-pack/

CompactionChange

ShelfAllocator

A shelf-packing dynamic texture atlas allocator tracking each allocation individually and with support for coalescing empty shelves.

ShelfAllocatorOptions

Options to tweak the behavior of the atlas allocator.

TextureUnit 🔒

A number of 2D textures (single layer), with their own atlas allocator.

Traits

AtlasAllocator

AtlasAllocatorList