webrender::texture_pack

Module guillotine

Structs

Bin 🔒

FreeListBin 🔒

FreeListIndex 🔒

FreeRect 🔒

FreeRectSize 🔒

FreeRectSlice

GuillotineAllocator

A texture allocator using the guillotine algorithm.

Constants

MIN_RECT_AXIS_SIZES 🔒

The minimum number of pixels on each side that we require for rects to be classified as particular bin of freelists.

NUM_BINS 🔒