webrender::texture_pack

Trait AtlasAllocator

pub trait AtlasAllocator {
    type Parameters;

    // Required methods
    fn new(size: i32, parameters: &Self::Parameters) -> Self;
    fn allocate(
        &mut self,
        size: DeviceIntSize,
    ) -> Option<(AllocId, DeviceIntRect)>;
    fn deallocate(&mut self, id: AllocId);
    fn is_empty(&self) -> bool;
    fn allocated_space(&self) -> i32;
    fn dump_into_svg(
        &self,
        rect: &Box2D<f32>,
        output: &mut dyn Write,
    ) -> Result<()>;
}

Required Associated Types

type Parameters

Specific parameters of the allocator.

Required Methods

fn new(size: i32, parameters: &Self::Parameters) -> Self

Constructor

fn allocate(&mut self, size: DeviceIntSize) -> Option<(AllocId, DeviceIntRect)>

Allocate a rectangle.

fn deallocate(&mut self, id: AllocId)

Deallocate a rectangle and return its size.

fn is_empty(&self) -> bool

Return true if there is no live allocations.

fn allocated_space(&self) -> i32

Allocated area in pixels.

fn dump_into_svg(&self, rect: &Box2D<f32>, output: &mut dyn Write) -> Result<()>

Write a debug visualization of the atlas fitting in the provided rectangle.

This is inserted in a larger dump so it shouldn’t contain the xml start/end tags.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.