webrender::tile_cache::slice_builder

Struct TileCacheBuilder

pub struct TileCacheBuilder {
    primary_slices: Vec<PrimarySlice>,
    prev_scroll_root_cache: (SpatialNodeIndex, SpatialNodeIndex),
    root_spatial_node_index: SpatialNodeIndex,
    debug_flags: DebugFlags,
}

Used during scene building to construct the list of pending tile caches.

Fields
primary_slices: Vec<PrimarySlice>

List of tile caches that have been created so far (last in the list is currently active).

prev_scroll_root_cache: (SpatialNodeIndex, SpatialNodeIndex)

Cache the previous scroll root search for a spatial node, since they are often the same.

root_spatial_node_index: SpatialNodeIndex

Handle to the root reference frame

debug_flags: DebugFlags

Debug flags to provide to our TileCacheInstances.

Implementations

impl TileCacheBuilder

pub fn new( root_spatial_node_index: SpatialNodeIndex, background_color: Option<ColorF>, debug_flags: DebugFlags, ) -> Self

Construct a new tile cache builder.

pub fn make_current_slice_atomic(&mut self)

pub fn is_current_slice_empty(&self) -> bool

Returns true if the current slice has no primitives added yet

pub fn add_tile_cache_barrier( &mut self, slice_flags: SliceFlags, iframe_clip: Option<ClipId>, )

Set a barrier that forces a new tile cache next time a prim is added.

fn build_tile_cache( &mut self, prim_list: PrimitiveList, spatial_tree: &SceneSpatialTree, ) -> Option<SliceDescriptor>

Create a new tile cache for an existing prim_list

pub fn add_prim( &mut self, prim_instance: PrimitiveInstance, prim_rect: LayoutRect, spatial_node_index: SpatialNodeIndex, prim_flags: PrimitiveFlags, spatial_tree: &SceneSpatialTree, quality_settings: &QualitySettings, prim_instances: &mut Vec<PrimitiveInstance>, clip_tree_builder: &ClipTreeBuilder, )

Add a primitive, either to the current tile cache, or a new one, depending on various conditions.

pub fn build( self, config: &FrameBuilderConfig, prim_store: &mut PrimitiveStore, spatial_tree: &SceneSpatialTree, prim_instances: &[PrimitiveInstance], clip_tree_builder: &mut ClipTreeBuilder, interners: &Interners, ) -> (TileCacheConfig, Vec<PictureIndex>)

Consume this object and build the list of tile cache primitives