Struct webrender::tiling::ClipBatcher

pub struct ClipBatcher {
    pub rectangles: Vec<CacheClipInstance>,
    pub images: HashMap<SourceTexture, Vec<CacheClipInstance>>,
}

Batcher managing draw calls into the clip mask (in the RT cache).

Fields
rectangles: Vec<CacheClipInstance>

Rectangle draws fill up the rectangles with rounded corners.

images: HashMap<SourceTexture, Vec<CacheClipInstance>>

Image draws apply the image masking.

Methods

impl ClipBatcher

fn new() -> ClipBatcher

fn add<'a>(&mut self,
task_index: RenderTaskIndex,
clips: &[(PackedLayerIndex, MaskCacheInfo)],
resource_cache: &ResourceCache,
geometry_kind: MaskGeometryKind)