Struct webrender::tiling::RenderPass

pub struct RenderPass {
    pass_index: RenderPassIndex,
    pub is_framebuffer: bool,
    tasks: Vec<RenderTask>,
    pub targets: Vec<RenderTarget>,
}

A render pass represents a set of rendering operations that don't depend on one another.

A render pass can have several render targets if there wasn't enough space in one target to do all of the rendering for that pass.

Fields
pass_index: RenderPassIndex is_framebuffer: bool tasks: Vec<RenderTask> targets: Vec<RenderTarget>
Methods

impl RenderPass

fn new(pass_index: isize, is_framebuffer: bool) -> RenderPass

fn add_render_task(&mut self, task: RenderTask)

fn build(&mut self,
ctx: &RenderTargetContext,
render_tasks: &mut RenderTaskCollection)