webrender

Trait PartialPresentCompositor

pub trait PartialPresentCompositor {
    // Required method
    fn set_buffer_damage_region(&mut self, rects: &[DeviceIntRect]);
}

Defines an interface to a non-native (application-level) Compositor which handles partial present. This is required if webrender must query the backbuffer’s age. TODO: Use the Compositor trait for native and non-native compositors, and integrate this functionality there.

Required Methods

fn set_buffer_damage_region(&mut self, rects: &[DeviceIntRect])

Allows webrender to specify the total region that will be rendered to this frame, ie the frame’s dirty region and some previous frames’ dirty regions, if applicable (calculated using the buffer age). Must be called before anything has been rendered to the main framebuffer.