webrender::util

Function extract_inner_rect_safe

pub fn extract_inner_rect_safe<U>(
    rect: &Box2D<f32, U>,
    radii: &BorderRadius,
) -> Option<Box2D<f32, U>>

Return an aligned rectangle that is inside the clip region and doesn’t intersect any of the bounding rectangles of the rounded corners.