Crate webrender

A GPU based Webrender.

It serves as an experimental render backend for Servo, but it can also be used as such in a standalone application.

External dependencies

Webrender currently depends on FreeType

Api Structure

The main entry point to webrender is the webrender::renderer::Renderer.

By calling Renderer::new(...) you get a Renderer, as well as a RenderApiSender. Your Renderer is responsible to render the previously processed frames onto the screen.

By calling yourRenderApiSenderInstance.create_api(), you'll get a RenderApi instance, which is responsible for the processing of new frames. A worker thread is used internally to untie the workload from the application thread and therefore be able to make better use of multicore systems.

What is referred to as a frame, is the current geometry on the screen. A new Frame is created by calling set_root_stacking_context() on the RenderApi. When the geometry is processed, the application will be informed via a RenderNotifier, a callback which you employ with set_render_notifier on the Renderer More information about stacking contexts.

set_root_stacking_context() also needs to be supplied with BuiltDisplayLists. These are obtained by finalizing a DisplayListBuilder. These are used to draw your geometry. But it doesn't only contain trivial geometry, it can also store another StackingContext, as they're nestable.

Remember to insert the DisplayListId into the StackingContext as well, as they'll be referenced from there. An Id for your DisplayList can be obtained by calling yourRenderApiInstance.next_display_list_id();

Reexports

pub use renderer::ExternalImage;

pub use renderer::ExternalImageSource;

pub use renderer::ExternalImageHandler;

pub use renderer::Renderer;

pub use renderer::RendererOptions;

Modules

renderer
The webrender API.

Statics

WEBRENDER_RECORDING_HEADER

Traits

ApiRecordingReceiver

Functions

set_recording_detour