Struct webrender::batch::BatchBuilder

pub struct BatchBuilder<'a> {
    vertex_buffer: &'a mut VertexBuffer,
    batches: Vec<Batch>,
    current_matrix_index: u8,
    clip_offset: Point2D<f32>,
    clip_in_rect_stack: Vec<Rect<f32>>,
    cached_clip_in_rect: Option<Rect<f32>>,
    clip_out_rect: Option<Rect<f32>>,
    pub complex_clip: Option<ComplexClipRegion>,
    pub device_pixel_ratio: f32,
}

Fields

vertex_buffer

batches

current_matrix_index

clip_offset

clip_in_rect_stack

cached_clip_in_rect

clip_out_rect

complex_clip

device_pixel_ratio

Methods

impl<'a> BatchBuilder<'a>

fn new(vertex_buffer: &mut VertexBuffer, device_pixel_ratio: f32) -> BatchBuilder

fn finalize(self) -> Vec<Batch>

fn set_current_clip_rect_offset(&mut self, offset: Point2D<f32>)

fn next_draw_list(&mut self)

fn update_clip_in_rect(&mut self)

fn push_clip_in_rect(&mut self, rect: &Rect<f32>)

fn pop_clip_in_rect(&mut self)

fn set_clip_out_rect(&mut self, rect: Option<Rect<f32>>) -> Option<Rect<f32>>

fn push_complex_clip(&mut self, clip: &[ComplexClipRegion])

fn pop_complex_clip(&mut self)

fn add_rectangle(&mut self, color_texture_id: TextureId, mask_texture_id: TextureId, pos_rect: &Rect<f32>, uv_rect: &RectUv<f32>, muv_rect: &RectUv<DevicePixel>, colors: &[ColorF; 4], color_mode: PackedVertexColorMode, tile_params: Option<TileParams>)

fn push_vertex_for_rectangle(&mut self, vertex: PackedVertexForQuad)

impl<'a> BatchBuilder<'a>

fn add_simple_rectangle(&mut self, color_texture_id: TextureId, pos_rect: &Rect<f32>, uv_rect: &RectUv<f32>, mask_texture_id: TextureId, muv_rect: &RectUv<DevicePixel>, colors: &[ColorF; 4], tile_params: Option<TileParams>)

fn snap_value_to_device_pixel(&self, value: &mut f32)

fn add_complex_clipped_rectangle(&mut self, color_texture_id: TextureId, pos_rect: &Rect<f32>, uv_rect: &RectUv<f32>, colors: &[ColorF; 4], tile_params: Option<TileParams>, resource_cache: &ResourceCache, frame_id: FrameId)

fn add_color_rectangle(&mut self, rect: &Rect<f32>, color: &ColorF, resource_cache: &ResourceCache, frame_id: FrameId)

fn add_webgl_rectangle(&mut self, rect: &Rect<f32>, resource_cache: &ResourceCache, webgl_context_id: &WebGLContextId, frame_id: FrameId)

fn add_image(&mut self, rect: &Rect<f32>, stretch_size: &Size2D<f32>, image_key: ImageKey, image_rendering: ImageRendering, resource_cache: &ResourceCache, frame_id: FrameId)

fn add_text(&mut self, _rect: &Rect<f32>, font_key: FontKey, size: Au, blur_radius: Au, color: &ColorF, glyphs: &[GlyphInstance], resource_cache: &ResourceCache, frame_id: FrameId, device_pixel_ratio: f32)

fn add_axis_aligned_gradient_with_stops(&mut self, rect: &Rect<f32>, direction: AxisDirection, stops: &[GradientStop], resource_cache: &ResourceCache, frame_id: FrameId)

fn add_gradient(&mut self, rect: &Rect<f32>, start_point: &Point2D<f32>, end_point: &Point2D<f32>, stops: &ItemRange, auxiliary_lists: &AuxiliaryLists, resource_cache: &ResourceCache, frame_id: FrameId)

fn add_box_shadow(&mut self, box_bounds: &Rect<f32>, box_offset: &Point2D<f32>, color: &ColorF, blur_radius: f32, spread_radius: f32, border_radius: f32, clip_mode: BoxShadowClipMode, resource_cache: &ResourceCache, frame_id: FrameId)

fn add_box_shadow_corners(&mut self, box_bounds: &Rect<f32>, box_offset: &Point2D<f32>, color: &ColorF, blur_radius: f32, spread_radius: f32, border_radius: f32, clip_mode: BoxShadowClipMode, resource_cache: &ResourceCache, frame_id: FrameId)

fn add_box_shadow_sides(&mut self, box_bounds: &Rect<f32>, box_offset: &Point2D<f32>, color: &ColorF, blur_radius: f32, spread_radius: f32, border_radius: f32, clip_mode: BoxShadowClipMode, resource_cache: &ResourceCache, frame_id: FrameId)

fn fill_outside_area_of_inset_box_shadow(&mut self, box_bounds: &Rect<f32>, box_offset: &Point2D<f32>, color: &ColorF, blur_radius: f32, spread_radius: f32, border_radius: f32, resource_cache: &ResourceCache, frame_id: FrameId)

fn undo_clip_state(&mut self, clip_state: ClipState)

fn adjust_clip_for_box_shadow_clip_mode(&mut self, box_bounds: &Rect<f32>, _border_radius: f32, clip_mode: BoxShadowClipMode) -> ClipState

fn add_border_edge(&mut self, rect: &Rect<f32>, side: RectSide, color: &ColorF, border_style: BorderStyle, resource_cache: &ResourceCache, frame_id: FrameId)

fn add_border_corner(&mut self, border_style: BorderStyle, corner_bounds: &Rect<f32>, radius_extent: &Point2D<f32>, color0: &ColorF, color1: &ColorF, outer_radius: &Size2D<f32>, inner_radius: &Size2D<f32>, resource_cache: &ResourceCache, frame_id: FrameId, can_tessellate: bool, rotation_angle: BasicRotationAngle, device_pixel_ratio: f32)

Draws a border corner.

The following diagram attempts to explain the parameters to this function. It's an enlarged version of a border corner that looks like this:

╭─
│

The parameters are as follows:

⤹ corner_bounds.origin
∙┈┈┈┈┈┬┈┈┈┈┈┬┈┈┈┈┈
┊   ╱ ┊     ┊
┊  ╱  ┊     ┊
┊ ╱╲  ┊   ←─┼── color1
┊╱  ╲ ┊     ┊
┊    ╲┊     ┊
├┈┈┈┈┈∙←────┼── radius_extent
┊     ┊╲    ┊
┊     ┊ ╲   ┊
┊     ┊  ╲  ┊
┊  ↑  ┊   ╲ ┊
┊  │  ┊    ╲┊
├┈┈┼┈┈┴┈┈┈┈┈∙┈┈┈┈┈
┊  │        ┊↖︎
┊           ┊  corner_bounds.bottom_right()
┊color0     ┊

fn add_solid_border_corner(&mut self, corner_bounds: &Rect<f32>, radius_extent: &Point2D<f32>, color0: &ColorF, color1: &ColorF, outer_radius: &Size2D<f32>, inner_radius: &Size2D<f32>, resource_cache: &ResourceCache, frame_id: FrameId, can_tessellate: bool, rotation_angle: BasicRotationAngle, _device_pixel_ratio: f32)

fn add_border_corner_piece(&mut self, rect_pos_uv: RectPolygon<RectUv<DevicePixel>>, mask_image: &TextureCacheItem, color0: &ColorF, color1: &ColorF, resource_cache: &ResourceCache, rotation_angle: BasicRotationAngle)

Draws one rectangle making up a border corner.

fn add_color_image_rectangle(&mut self, v0: &Point2D<f32>, v1: &Point2D<f32>, color0: &ColorF, color1: &ColorF, color_image: &TextureCacheItem, resource_cache: &ResourceCache, rotation_angle: BasicRotationAngle)

fn add_border(&mut self, rect: &Rect<f32>, info: &BorderDisplayItem, resource_cache: &ResourceCache, frame_id: FrameId, device_pixel_ratio: f32)

fn add_box_shadow_corner(&mut self, top_left: &Point2D<f32>, bottom_right: &Point2D<f32>, corner_area_top_left: &Point2D<f32>, corner_area_bottom_right: &Point2D<f32>, box_rect: &Rect<f32>, color: &ColorF, blur_radius: f32, border_radius: f32, clip_mode: BoxShadowClipMode, resource_cache: &ResourceCache, frame_id: FrameId, rotation_angle: BasicRotationAngle)

fn add_box_shadow_edge(&mut self, top_left: &Point2D<f32>, bottom_right: &Point2D<f32>, box_rect: &Rect<f32>, color: &ColorF, blur_radius: f32, border_radius: f32, clip_mode: BoxShadowClipMode, resource_cache: &ResourceCache, frame_id: FrameId, rotation_angle: BasicRotationAngle)