Module webrender::border

Structs

BorderRadiusAu

BorderSegmentCacheKey

Cache key that uniquely identifies a border segment in the render task cache.

BorderSideAu

DotInfo

EdgeInfo

Information needed to place and draw a border edge.

NormalBorderAu

Enums

BorderClipKind

The kind of border corner clip.

Constants

MAX_BORDER_RESOLUTION

Maximum resolution in device pixels at which borders are rasterized.

MAX_DASH_COUNT

Maximum number of dots or dashes per segment to avoid freezing and filling up memory with unreasonable inputs. It would be better to address this by not building a list of per-dot information in the first place.

Traits

BorderSideHelpers

Functions

add_corner_segment

Add a corner segment (if valid) to the list of border segments for this primitive.

add_edge_segment

Add an edge segment (if valid) to the list of border segments for this primitive.

add_segment

build_border_instances

Build the set of border instances needed to draw a border segment into the render task cache.

compute_half_dash

compute_outer_and_clip_sign

create_border_segments

Create the set of border segments and render task cache keys for a given CSS border.

ensure_no_corner_overlap

get_edge_info

get_max_scale_for_border

Computes the maximum scale that we allow for this set of border parameters. capping the scale will result in rendering very large corners at a lower resolution and stretching them, so they will have the right shape, but blurrier.

write_dashed_corner_instances

write_dotted_corner_instances