Module webrender::clip

Internal representation of clips in WebRender.

Data structures

There are a number of data structures involved in the clip module:

ClipStore - Main interface used by other modules.

ClipItem - A single clip item (e.g. a rounded rect, or a box shadow). These are an exposed API type, stored inline in a ClipNode.

ClipNode - A ClipItem with an attached GPU handle. The GPU handle is populated when a ClipNodeInstance is built from this node (which happens while preparing primitives for render).

ClipNodeInstance - A ClipNode with attached positioning information (a spatial node index). This is stored as a contiguous array of nodes within the ClipStore.

+-----------------------+-----------------------+-----------------------+
| ClipNodeInstance      | ClipNodeInstance      | ClipNodeInstance      |
+-----------------------+-----------------------+-----------------------+
| ClipItem              | ClipItem              | ClipItem              |
| Spatial Node Index    | Spatial Node Index    | Spatial Node Index    |
| GPU cache handle      | GPU cache handle      | GPU cache handle      |
| ...                   | ...                   | ...                   |
+-----------------------+-----------------------+-----------------------+
           0                        1                       2
   +----------------+    |                                              |
   | ClipNodeRange  |____|                                              |
   |    index: 1    |                                                   |
   |    count: 2    |___________________________________________________|
   +----------------+

ClipNodeRange - A clip item range identifies a range of clip nodes instances. It is stored as an (index, count).

ClipChainNode - A clip chain node contains a handle to an interned clip item, positioning information (from where the clip was defined), and an optional parent link to another ClipChainNode. ClipChainId is an index into an array, or ClipChainId::NONE for no parent.

+----------------+    ____+----------------+    ____+----------------+   /---> ClipChainId::NONE
| ClipChainNode  |   |    | ClipChainNode  |   |    | ClipChainNode  |   |
+----------------+   |    +----------------+   |    +----------------+   |
| ClipDataHandle |   |    | ClipDataHandle |   |    | ClipDataHandle |   |
| Spatial index  |   |    | Spatial index  |   |    | Spatial index  |   |
| Parent Id      |___|    | Parent Id      |___|    | Parent Id      |___|
| ...            |        | ...            |        | ...            |
+----------------+        +----------------+        +----------------+

ClipChainInstance - A ClipChain that has been built for a specific primitive + positioning node.

When given a clip chain ID, and a local primitive rect and its spatial node, the clip module creates a clip chain instance. This is a struct with various pieces of useful information (such as a local clip rect). It also contains a (index, count) range specifier into an index buffer of the ClipNodeInstance structures that are actually relevant for this clip chain instance. The index buffer structure allows a single array to be used for all of the clip-chain instances built in a single frame. Each entry in the index buffer also stores some flags relevant to the clip node in this positioning context.

+----------------------+
| ClipChainInstance    |
+----------------------+
| ...                  |
| local_clip_rect      |________________________________________________________________________
| clips_range          |_______________                                                        |
+----------------------+              |                                                        |
                                      |                                                        |
+------------------+------------------+------------------+------------------+------------------+
| ClipNodeInstance | ClipNodeInstance | ClipNodeInstance | ClipNodeInstance | ClipNodeInstance |
+------------------+------------------+------------------+------------------+------------------+
| flags            | flags            | flags            | flags            | flags            |
| ...              | ...              | ...              | ...              | ...              |
+------------------+------------------+------------------+------------------+------------------+

Rendering clipped primitives

See the segment module documentation.

Structs

ClipChainBuilder

A helper used during scene building to construct (internal) clip chains from the public API definitions (a hierarchy of ClipIds)

ClipChainId

ClipChainInstance

ClipChainLevel

Maintains a (flattened) list of clips for a given level in the surface level stack.

ClipChainNode

ClipChainStack

Maintains a stack of clip chain ids that are currently active, when a clip exists on a picture that has no surface, and is passed on down to the child primitive(s).

ClipInstance

Defines a clip that is positioned by a specific spatial node

ClipInternData

The data available about an interned clip node during scene building

ClipItem

ClipItemKey

ClipNode

ClipNodeFlags

ClipNodeInfo

ClipNodeInstance

ClipNodeRange

ClipRegion

ClipStore

The main clipping public interface that other modules access.

ClipTemplate

A clip template defines clips in terms of the public API. Specifically, this is a parent ClipId and some number of clip instances. See the CLIPPING_AND_POSITIONING.md document in doc/ for more information.

ComplexTranslateIter

Geometry

Represents a local rect and a device space rectangles that are either outside or inside bounds.

Enums

ClipItemKeyKind

ClipItemKind

ClipNodeKind

Helper to identify simple clips (normal rects) from other kinds of clips, which can often be handled via fast code paths.

ClipResult

ClipSpaceConversion

A helper struct for converting between coordinate systems of clip sources and primitives.

Functions

add_clip_node_to_current_chain

compute_box_shadow_parameters

projected_rect_contains

rounded_rectangle_contains_point

Type Definitions

ClipDataHandle

ClipDataStore