Struct webrender::clip::ClipChainStack

pub struct ClipChainStack {
    levels: Vec<ClipChainLevel>,
    clips: Vec<ClipChainId>,
    clip_counts: Vec<usize>,
}

Maintains a stack of clip chain ids that are currently active, when a clip exists on a picture that has no surface, and is passed on down to the child primitive(s).

In order to avoid many small vector allocations, all clip chain ids are stored in a single vector instead of per-level. Since we only work with the top-most level of the stack, we only need to know the first index in the clips vector that belongs to each level. The last index for the top-most level is always the end of the clips array.

Likewise, we push several clip chain ids to the clips array at each push_clip, and the number of clip chain ids removed during pop_clip must match. This is done by having a separate stack of clip counts in the clip-stack rather than per-level to avoid vector allocations.

             +----+----+---
     levels: |    |    | ...
             +----+----+---
              |first   \
              |         \
              |          \
             +--+--+--+--+--+--+--
      clips: |  |  |  |  |  |  | ...
             +--+--+--+--+--+--+--
              |     /     /
              |    /    /
              |   /   /
             +--+--+--+--
clip_counts: | 1| 2| 2| ...
             +--+--+--+--

Fields
levels: Vec<ClipChainLevel>

A stack of clip chain lists. Each time a new surface is pushed, a new level is added. Each time a new picture without surface is pushed, it adds the picture clip chain to the clips vector in the range belonging to the level (always the top-most level, so always at the end of the clips array).

clips: Vec<ClipChainId>

The actual stack of clip ids.

clip_counts: Vec<usize>

How many clip ids to pop from the vector each time we call pop_clip.

Implementations

impl ClipChainStack

pub fn new() -> Self

pub fn push_clip(&mut self, clip_chain_id: ClipChainId, clip_store: &ClipStore)

Push a clip chain root onto the currently active list.

pub fn pop_clip(&mut self)

Pop a clip chain root from the currently active list.

pub fn push_surface(
&mut self,
maybe_shared_clips: &[ClipInstance],
spatial_tree: &SpatialTree
)

When a surface is created, it takes all clips and establishes a new stack of clips to be propagated.

pub fn pop_surface(&mut self)

Pop a surface from the clip chain stack

pub fn current_clips_array(&self) -> &[ClipChainId]

Get the list of currently active clip chains