Struct webrender::clip::ClipStore

pub struct ClipStore {
    pub clip_chain_nodes: Vec<ClipChainNode>,
    pub clip_node_instances: Vec<ClipNodeInstance>,
    active_clip_node_info: Vec<ClipNodeInfo>,
    active_local_clip_rect: Option<LayoutRect>,
    templates: FastHashMap<ClipId, ClipTemplate>,
    chain_builder_stack: Vec<ClipChainBuilder>,
}

The main clipping public interface that other modules access.

Fields
clip_chain_nodes: Vec<ClipChainNode>clip_node_instances: Vec<ClipNodeInstance>active_clip_node_info: Vec<ClipNodeInfo>active_local_clip_rect: Option<LayoutRect>templates: FastHashMap<ClipId, ClipTemplate>

Map of all clip templates defined by the public API to templates

chain_builder_stack: Vec<ClipChainBuilder>

A stack of current clip-chain builders. A new clip-chain builder is typically created each time a clip root (such as an iframe or stacking context) is defined.

Implementations

impl ClipStore

pub fn new() -> Self

pub fn register_clip_template(
&mut self,
clip_id: ClipId,
parent: ClipId,
instances: &[ClipInstance]
)

Register a new clip template for the clip_id defined in the display list.

pub fn get_template(&self, clip_id: ClipId) -> &ClipTemplate

pub fn get_or_build_clip_chain_id(&mut self, clip_id: ClipId) -> ClipChainId

The main method used to build a clip-chain for a given ClipId on a primitive

pub fn push_clip_root(&mut self, clip_id: Option<ClipId>, link_to_parent: bool)

Push a new clip root. This is used at boundaries of clips (such as iframes and stacking contexts). This means that any clips on the existing clip chain builder will not be added to clip-chains defined within this level, since the clips will be applied by the parent.

pub fn pop_clip_root(&mut self)

On completion of a stacking context or iframe, pop the current clip root.

pub fn get_clip_chain(&self, clip_chain_id: ClipChainId) -> &ClipChainNode

pub fn add_clip_chain_node(
&mut self,
handle: Handle<ClipIntern>,
spatial_node_index: SpatialNodeIndex,
parent_clip_chain_id: ClipChainId
) -> ClipChainId

pub fn get_instance_from_range(
&self,
node_range: &ClipNodeRange,
index: u32
) -> &ClipNodeInstance

pub fn set_active_clips(
&mut self,
local_prim_clip_rect: LayoutRect,
spatial_node_index: SpatialNodeIndex,
clip_chains: &[ClipChainId],
spatial_tree: &SpatialTree,
clip_data_store: &DataStore<ClipIntern>
)

Setup the active clip chains for building a clip chain instance.

pub fn set_active_clips_from_clip_chain(
&mut self,
prim_clip_chain: &ClipChainInstance,
prim_spatial_node_index: SpatialNodeIndex,
spatial_tree: &SpatialTree
)

Setup the active clip chains, based on an existing primitive clip chain instance.

pub fn build_clip_chain_instance(
&mut self,
local_prim_rect: LayoutRect,
prim_to_pic_mapper: &SpaceMapper<LayoutPixel, PicturePixel>,
pic_to_world_mapper: &SpaceMapper<PicturePixel, WorldPixel>,
spatial_tree: &SpatialTree,
gpu_cache: &mut GpuCache,
resource_cache: &mut ResourceCache,
device_pixel_scale: DevicePixelScale,
world_rect: &WorldRect,
clip_data_store: &mut DataStore<ClipIntern>,
request_resources: bool,
is_chased: bool
) -> Option<ClipChainInstance>

The main interface external code uses. Given a local primitive, positioning information, and a clip chain id, build an optimized clip chain instance.

pub fn clear_old_instances(&mut self)