Struct webrender::clip_scroll_tree::ClipScrollTree

pub struct ClipScrollTree {
    pub spatial_nodes: Vec<SpatialNode>,
    coord_systems: Vec<CoordinateSystem>,
    pub pending_scroll_offsets: HashMap<ExternalScrollId, (LayoutPoint, ScrollClamping), BuildHasherDefault<FxHasher>>,
    pub pipelines_to_discard: HashSet<PipelineId, BuildHasherDefault<FxHasher>>,
    nodes_to_update: Vec<(SpatialNodeIndex, TransformUpdateState)>,
}

Fields
spatial_nodes: Vec<SpatialNode>

Nodes which determine the positions (offsets and transforms) for primitives and clips.

coord_systems: Vec<CoordinateSystem>

A list of transforms that establish new coordinate systems. Spatial nodes only establish a new coordinate system when they have a transform that is not a simple 2d translation.

pending_scroll_offsets: HashMap<ExternalScrollId, (LayoutPoint, ScrollClamping), BuildHasherDefault<FxHasher>>pipelines_to_discard: HashSet<PipelineId, BuildHasherDefault<FxHasher>>

A set of pipelines which should be discarded the next time this tree is drained.

nodes_to_update: Vec<(SpatialNodeIndex, TransformUpdateState)>

Temporary stack of nodes to update when traversing the tree.

Methods

impl ClipScrollTree

pub fn new() -> Self

pub fn external_scroll_offset(
&self,
node_index: SpatialNodeIndex
) -> LayoutVector2D

Calculate the accumulated external scroll offset for a given spatial node.

pub fn get_relative_transform(
&self,
child_index: SpatialNodeIndex,
parent_index: SpatialNodeIndex
) -> CoordinateSpaceMapping<LayoutPixel, LayoutPixel>

Calculate the relative transform from child_index to parent_index. This method will panic if the nodes are not connected!

fn get_world_transform_impl(
&self,
index: SpatialNodeIndex,
scroll: TransformScroll
) -> CoordinateSpaceMapping<LayoutPixel, WorldPixel>

pub fn get_world_transform(
&self,
index: SpatialNodeIndex
) -> CoordinateSpaceMapping<LayoutPixel, WorldPixel>

Calculate the relative transform from index to the root.

pub fn get_world_viewport_transform(
&self,
index: SpatialNodeIndex
) -> CoordinateSpaceMapping<LayoutPixel, WorldPixel>

Calculate the relative transform from index to the root. Unlike get_world_transform, this variant doesn't account for the local scroll offset.

pub fn root_reference_frame_index(&self) -> SpatialNodeIndex

The root reference frame, which is the true root of the ClipScrollTree. Initially this ID is not valid, which is indicated by spatial_nodes being empty.

pub fn topmost_scroll_node_index(&self) -> SpatialNodeIndex

The root scroll node which is the first child of the root reference frame. Initially this ID is not valid, which is indicated by spatial_nodes being empty.

pub fn get_scroll_node_state(&self) -> Vec<ScrollNodeState>

pub fn drain(
&mut self
) -> HashMap<ExternalScrollId, ScrollFrameInfo, BuildHasherDefault<FxHasher>>

pub fn scroll_node(
&mut self,
origin: LayoutPoint,
id: ExternalScrollId,
clamp: ScrollClamping
) -> bool

fn find_nearest_scrolling_ancestor(
&self,
index: Option<SpatialNodeIndex>
) -> SpatialNodeIndex

pub fn scroll_nearest_scrolling_ancestor(
&mut self,
scroll_location: ScrollLocation,
node_index: Option<SpatialNodeIndex>
) -> bool

pub fn update_tree(
&mut self,
pan: WorldPoint,
global_device_pixel_scale: DevicePixelScale,
scene_properties: &SceneProperties
)

pub fn build_transform_palette(&self) -> TransformPalette

pub fn finalize_and_apply_pending_scroll_offsets(
&mut self,
old_states: HashMap<ExternalScrollId, ScrollFrameInfo, BuildHasherDefault<FxHasher>>
)

pub fn add_scroll_frame(
&mut self,
parent_index: SpatialNodeIndex,
external_id: Option<ExternalScrollId>,
pipeline_id: PipelineId,
frame_rect: &LayoutRect,
content_size: &LayoutSize,
scroll_sensitivity: ScrollSensitivity,
frame_kind: ScrollFrameKind,
external_scroll_offset: LayoutVector2D
) -> SpatialNodeIndex

pub fn add_reference_frame(
&mut self,
parent_index: Option<SpatialNodeIndex>,
transform_style: TransformStyle,
source_transform: PropertyBinding<LayoutTransform>,
kind: ReferenceFrameKind,
origin_in_parent_reference_frame: LayoutVector2D,
pipeline_id: PipelineId
) -> SpatialNodeIndex

pub fn add_sticky_frame(
&mut self,
parent_index: SpatialNodeIndex,
sticky_frame_info: StickyFrameInfo,
pipeline_id: PipelineId
) -> SpatialNodeIndex

pub fn add_spatial_node(&mut self, node: SpatialNode) -> SpatialNodeIndex

pub fn discard_frame_state_for_pipeline(&mut self, pipeline_id: PipelineId)

pub fn find_scroll_root(
&self,
spatial_node_index: SpatialNodeIndex
) -> SpatialNodeIndex

Find the spatial node that is the scroll root for a given spatial node. A scroll root is the first spatial node when found travelling up the spatial node tree that is an explicit scroll frame.

fn print_node<T: PrintTreePrinter>(&self, index: SpatialNodeIndex, pt: &mut T)

pub fn get_local_visible_face(
&self,
node_index: SpatialNodeIndex
) -> VisibleFace

Get the visible face of the transfrom from the specified node to its parent.

pub fn print(&self)