Module webrender::composite

Structs

CompositeState
The list of tiles to be drawn this frame

CompositeTile
Describes the geometry and surface of a tile to be composited

NativeSurfaceId
An arbitrary identifier for a native (OS compositor) surface

NativeSurfaceInfo
Information about a bound surface that the native compositor returns to WR.

NativeSurfaceOperation
Describes an operation to apply to a native surface

Occluder
Information about an opaque surface used to occlude tiles.

Enums

CompositeTileSurface
Describes the source surface information for a tile to be composited. This is the analog of the TileSurface type, with target surface information resolved such that it can be used by the renderer.

CompositorConfig
Public interface specified in RendererOptions that configures how WR compositing will operate.

CompositorKind
This is a representation of CompositorConfig without the Compositor trait present. This allows it to be freely copied to other threads, such as the render backend where the frame builder can access it.

NativeSurfaceOperationDetails
Describes details of an operation to apply to a native surface

Constants

_IMPL_SERIALIZE_FOR_CompositeState

_IMPL_SERIALIZE_FOR_CompositeTile

_IMPL_SERIALIZE_FOR_CompositeTileSurface

_IMPL_SERIALIZE_FOR_CompositorKind

_IMPL_SERIALIZE_FOR_NativeSurfaceId

_IMPL_SERIALIZE_FOR_NativeSurfaceOperation

_IMPL_SERIALIZE_FOR_NativeSurfaceOperationDetails

_IMPL_SERIALIZE_FOR_Occluder

Statics

NEXT_NATIVE_SURFACE_ID

Traits

Compositor
Defines an interface to a native (OS level) compositor. If supplied by the client application, then picture cache slices will be composited by the OS compositor, rather than drawn via WR batches.

Functions

area_of_occluders
Return the total area covered by a set of occluders, accounting for overlapping areas between those rectangles.