Module webrender::composite

Structs

CompositeDescriptor

Describes surface properties used to composite a frame. This is used to compare compositions between frames.

CompositeState

The list of tiles to be drawn this frame

CompositeSurfaceDescriptor

Describes the properties that identify a surface composition uniquely.

CompositeTile

Describes the geometry and surface of a tile to be composited

CompositeTileDescriptor

Describes properties that identify a tile composition uniquely. The backing surface for this tile.

CompositorCapabilities

ExternalPlaneDescriptor

Information about a plane in a YUV or RGB surface.

ExternalSurfaceDescriptor

Describes information about drawing a primitive as a compositor surface. For now, we support only YUV images as compositor surfaces, but in future this will also support RGBA images.

NativeSurfaceId

An arbitrary identifier for a native (OS compositor) surface

NativeSurfaceInfo

Information about a bound surface that the native compositor returns to WR.

NativeSurfaceOperation

Describes an operation to apply to a native surface

NativeTileId

Occluder

Information about an opaque surface used to occlude tiles.

ResolvedExternalSurface

An ExternalSurfaceDescriptor that has had image keys resolved to texture handles. This contains all the information that the compositor step in renderer needs to know.

ResolvedExternalSurfaceIndex

Enums

CompositeSurfaceFormat

The surface format for a tile being composited.

CompositeTileSurface

Describes the source surface information for a tile to be composited. This is the analog of the TileSurface type, with target surface information resolved such that it can be used by the renderer.

CompositorConfig

Public interface specified in RendererOptions that configures how WR compositing will operate.

CompositorKind

This is a representation of CompositorConfig without the Compositor trait present. This allows it to be freely copied to other threads, such as the render backend where the frame builder can access it.

ExternalSurfaceDependency

NativeSurfaceOperationDetails

Describes details of an operation to apply to a native surface

ResolvedExternalSurfaceColorData

TileSurfaceKind

The backing surface kind for a tile. Same as TileSurface, minus the texture cache handles, visibility masks etc.

Traits

Compositor

Defines an interface to a native (OS level) compositor. If supplied by the client application, then picture cache slices will be composited by the OS compositor, rather than drawn via WR batches.

Functions

area_of_occluders

Return the total area covered by a set of occluders, accounting for overlapping areas between those rectangles.