Struct webrender::composite::CompositeState

pub struct CompositeState {
    pub native_tiles: Vec<CompositeTile>,
    pub opaque_tiles: Vec<CompositeTile>,
    pub alpha_tiles: Vec<CompositeTile>,
    pub clear_tiles: Vec<CompositeTile>,
    pub z_generator: ZBufferIdGenerator,
    pub dirty_rects_are_valid: bool,
    pub native_surface_updates: Vec<NativeSurfaceOperation>,
    pub compositor_kind: CompositorKind,
    pub picture_caching_is_enabled: bool,
    global_device_pixel_scale: DevicePixelScale,
    occluders: Vec<Occluder>,
}

The list of tiles to be drawn this frame

Fields
native_tiles: Vec<CompositeTile>

List of tiles to be drawn by the native compositor. These are added in draw order and not separated by kind (opacity is a property of the native surface).

opaque_tiles: Vec<CompositeTile>

List of opaque tiles to be drawn by the Draw compositor.

alpha_tiles: Vec<CompositeTile>

List of alpha tiles to be drawn by the Draw compositor.

clear_tiles: Vec<CompositeTile>

List of clear tiles to be drawn by the Draw compositor.

z_generator: ZBufferIdGenerator

Used to generate z-id values for tiles in the Draw compositor mode.

dirty_rects_are_valid: boolnative_surface_updates: Vec<NativeSurfaceOperation>

List of OS native surface create / destroy operations to apply when render occurs.

compositor_kind: CompositorKind

The kind of compositor for picture cache tiles (e.g. drawn by WR, or OS compositor)

picture_caching_is_enabled: bool

Picture caching may be disabled dynamically, based on debug flags, pinch zoom etc.

global_device_pixel_scale: DevicePixelScale

The overall device pixel scale, used for tile occlusion conversions.

occluders: Vec<Occluder>

List of registered occluders

Methods

impl CompositeState

pub fn new(
compositor_kind: CompositorKind,
picture_caching_is_enabled: bool,
global_device_pixel_scale: DevicePixelScale
) -> Self

Construct a new state for compositing picture tiles. This is created during each frame construction and passed to the renderer.

pub fn create_surface(
&mut self,
size: DeviceIntSize,
is_opaque: bool
) -> NativeSurfaceId

Queue up allocation of a new OS native compositor surface with the specified id and dimensions.

pub fn destroy_surface(&mut self, id: NativeSurfaceId)

Queue up destruction of an existing native OS surface. This is used when a picture cache tile is dropped or resized.

pub fn register_occluder(&mut self, slice: usize, rect: WorldRect)

Register an occluder during picture cache updates that can be used during frame building to occlude tiles.

pub fn is_tile_occluded(&self, slice: usize, rect: WorldRect) -> bool

Returns true if a tile with the specified rectangle and slice is occluded by an opaque surface in front of it.

pub fn push_tile(&mut self, tile: CompositeTile, is_opaque: bool)

Add a tile to the appropriate array, depending on tile properties and compositor mode.

impl CompositeState

pub fn destroy_native_surfaces<'a, I: Iterator<Item = &'a Tile>>(
&mut self,
tiles_iter: I
)

impl CompositeState

fn composite_native(&self, compositor: &mut dyn Compositor)

Use the client provided native compositor interface to add all picture cache tiles to the OS compositor