Struct webrender::composite::CompositeState

pub struct CompositeState {
    pub opaque_tiles: Vec<CompositeTile>,
    pub alpha_tiles: Vec<CompositeTile>,
    pub clear_tiles: Vec<CompositeTile>,
    pub external_surfaces: Vec<ResolvedExternalSurface>,
    pub z_generator: ZBufferIdGenerator,
    pub dirty_rects_are_valid: bool,
    pub compositor_kind: CompositorKind,
    pub picture_caching_is_enabled: bool,
    global_device_pixel_scale: DevicePixelScale,
    occluders: Vec<Occluder>,
    pub descriptor: CompositeDescriptor,
}

The list of tiles to be drawn this frame

Fields
opaque_tiles: Vec<CompositeTile>

List of opaque tiles to be drawn by the Draw compositor.

alpha_tiles: Vec<CompositeTile>

List of alpha tiles to be drawn by the Draw compositor.

clear_tiles: Vec<CompositeTile>

List of clear tiles to be drawn by the Draw compositor.

external_surfaces: Vec<ResolvedExternalSurface>

List of primitives that were promoted to be compositor surfaces.

z_generator: ZBufferIdGenerator

Used to generate z-id values for tiles in the Draw compositor mode.

dirty_rects_are_valid: boolcompositor_kind: CompositorKind

The kind of compositor for picture cache tiles (e.g. drawn by WR, or OS compositor)

picture_caching_is_enabled: bool

Picture caching may be disabled dynamically, based on debug flags, pinch zoom etc.

global_device_pixel_scale: DevicePixelScale

The overall device pixel scale, used for tile occlusion conversions.

occluders: Vec<Occluder>

List of registered occluders

descriptor: CompositeDescriptor

Description of the surfaces and properties that are being composited.

Implementations

impl CompositeState

pub fn new(
compositor_kind: CompositorKind,
picture_caching_is_enabled: bool,
global_device_pixel_scale: DevicePixelScale,
max_depth_ids: i32
) -> Self

Construct a new state for compositing picture tiles. This is created during each frame construction and passed to the renderer.

pub fn register_occluder(&mut self, z_id: ZBufferId, rect: WorldRect)

Register an occluder during picture cache updates that can be used during frame building to occlude tiles.

pub fn is_tile_occluded(&self, z_id: ZBufferId, device_rect: DeviceRect) -> bool

Returns true if a tile with the specified rectangle and z_id is occluded by an opaque surface in front of it.

pub fn push_surface(
&mut self,
tile_cache: &TileCacheInstance,
device_clip_rect: DeviceRect,
global_device_pixel_scale: DevicePixelScale,
resource_cache: &ResourceCache,
gpu_cache: &mut GpuCache,
deferred_resolves: &mut Vec<DeferredResolve>
)

Add a picture cache to be composited

fn push_tile(&mut self, tile: CompositeTile, is_opaque: bool)

Add a tile to the appropriate array, depending on tile properties and compositor mode.

impl CompositeState

pub fn destroy_native_tiles<'a, I: Iterator<Item = &'a mut Box<Tile>>>(
&mut self,
tiles_iter: I,
resource_cache: &mut ResourceCache
)

impl CompositeState

fn composite_native(&self, compositor: &mut dyn Compositor)

Use the client provided native compositor interface to add all picture cache tiles to the OS compositor