Module webrender::device::gl

Structs

BoundPBO

Capabilities

CustomVAO

Device

ExternalTexture

FBOId

FormatDesc

GpuFrameId
Sequence number for frames, as tracked by the device layer.

IBOId

PBO

PixelBuffer

Program

ProgramBinary

ProgramCache

ProgramCacheEntry

ProgramSourceInfo

RBOId

SharedDepthTarget
A refcounted depth target, which may be shared by multiple textures across the device.

Stream

Texture
WebRender interface to an OpenGL texture.

TextureFlags

TextureFormatPair
A structure defining a particular workflow of texture transfers.

TextureSlot

TextureUploader

UniformLocation

UploadChunk

UploadTarget

VAO

VBO

VBOId

VertexAttribute

VertexDescriptor

Enums

DepthFunction

DrawTarget
Contains the parameters necessary to bind a draw target.

FBOTarget

ProgramSourceType

ReadTarget
Contains the parameters necessary to bind a texture-backed read target.

ShaderError

StrideAlignment
Describes a required alignment for a stride, which can either be represented in bytes or pixels.

TexStorageUsage
Describes for which texture formats to use the glTexStorage* family of functions.

TextureFilter

TextureUploaderType

UploadMethod
Method of uploading texel data from CPU to GPU.

VertexAttributeKind

VertexUsageHint

Constants

DEFAULT_TEXTURE

RESERVE_DEPTH_BITS

Traits

FileWatcherHandler

ProgramCacheObserver
The interfaces that an application can implement to handle ProgramCache update

Texel
Plain old data that can be used to initialize a texture.

Functions

depth_target_size_in_bytes
Returns the size in bytes of a depth target with the given dimensions.

get_gl_target

get_shader_version

get_unoptimized_shader_source

supports_extension

texels_to_u8_slice