Struct webrender::device::gl::Texture

pub struct Texture {
    id: GLuint,
    target: GLuint,
    layer_count: i32,
    format: ImageFormat,
    size: DeviceIntSize,
    filter: TextureFilter,
    flags: TextureFlags,
    active_swizzle: Cell<Swizzle>,
    fbos: Vec<FBOId>,
    fbos_with_depth: Vec<FBOId>,
    blit_workaround_buffer: Option<(RBOId, FBOId)>,
    last_frame_used: GpuFrameId,
}

WebRender interface to an OpenGL texture.

Because freeing a texture requires various device handles that are not reachable from this struct, manual destruction via Device is required. Our Drop implementation asserts that this has happened.

Fields
id: GLuinttarget: GLuintlayer_count: i32format: ImageFormatsize: DeviceIntSizefilter: TextureFilterflags: TextureFlagsactive_swizzle: Cell<Swizzle>

An internally mutable swizzling state that may change between batches.

fbos: Vec<FBOId>

Framebuffer Objects, one for each layer of the texture, allowing this texture to be rendered to. Empty if this texture is not used as a render target.

fbos_with_depth: Vec<FBOId>

Same as the above, but with a depth buffer attached.

FBOs are cheap to create but expensive to reconfigure (since doing so invalidates framebuffer completeness caching). Moreover, rendering with a depth buffer attached but the depth write+test disabled relies on the driver to optimize it out of the rendering pass, which most drivers probably do but, according to jgilbert, is best not to rely on.

So we lazily generate a second list of FBOs with depth. This list is empty if this texture is not used as a render target or if it is, but the depth buffer has never been requested.

Note that we always fill fbos, and then lazily create fbos_with_depth when needed. We could make both lazy (i.e. render targets would have one or the other, but not both, unless they were actually used in both configurations). But that would complicate a lot of logic in this module, and FBOs are cheap enough to create.

blit_workaround_buffer: Option<(RBOId, FBOId)>

If we are unable to blit directly to a texture array then we need an intermediate renderbuffer.

last_frame_used: GpuFrameId
Implementations

impl Texture

pub fn get_dimensions(&self) -> DeviceIntSize

pub fn get_layer_count(&self) -> i32

pub fn get_format(&self) -> ImageFormat

pub fn get_filter(&self) -> TextureFilter

pub fn supports_depth(&self) -> bool

pub fn last_frame_used(&self) -> GpuFrameId

pub fn used_in_frame(&self, frame_id: GpuFrameId) -> bool

pub fn used_recently(
&self,
current_frame_id: GpuFrameId,
threshold: usize
) -> bool

Returns true if this texture was used within threshold frames of the current frame.

pub fn flags(&self) -> &TextureFlags

Returns the flags for this texture.

pub fn flags_mut(&mut self) -> &mut TextureFlags

Returns a mutable borrow of the flags for this texture.

pub fn layer_size_in_bytes(&self) -> usize

Returns the number of bytes (generally in GPU memory) that each layer of this texture consumes.

pub fn size_in_bytes(&self) -> usize

Returns the number of bytes (generally in GPU memory) that this texture consumes.