webrender

Module ellipse

Structs

Ellipse

Represents an ellipse centred at a local space origin.

Constants

STEP_COUNT 🔒

Number of steps to integrate arc length over.

Functions

get_simpson_length 🔒

Use Simpsons rule to approximate the arc length of part of an ellipse. Note that this only works over the range of [0, pi/2].