Module webrender::frame_builder

Structs

Frame
A rendering-oriented representation of the frame built by the render backend and presented to the renderer.

FrameBuilder
Produces the frames that are sent to the renderer.

FrameBuilderConfig

FrameBuildingContext

FrameBuildingState

FrameGlobalResources
A set of common / global resources that are retained between new display lists, such that any GPU cache handles can be persisted even when a new display list arrives.

FrameVisibilityContext

FrameVisibilityState

PictureContext
Immutable context of a picture when processing children.

PictureState
Mutable state of a picture that gets modified when the children are processed.

Enums

ChasePrimitive

Functions

build_render_pass
Processes this pass to prepare it for rendering.