Struct webrender::frame_builder::Frame

pub struct Frame {
    pub content_origin: DeviceIntPoint,
    pub device_rect: DeviceIntRect,
    pub layer: DocumentLayer,
    pub passes: Vec<RenderPass>,
    pub profile_counters: FrameProfileCounters,
    pub transform_palette: Vec<TransformData>,
    pub render_tasks: RenderTaskGraph,
    pub prim_headers: PrimitiveHeaders,
    pub gpu_cache_frame_id: FrameId,
    pub deferred_resolves: Vec<DeferredResolve>,
    pub has_texture_cache_tasks: bool,
    pub has_been_rendered: bool,
    pub recorded_dirty_regions: Vec<RecordedDirtyRegion>,
    pub debug_items: Vec<DebugItem>,
    pub composite_state: CompositeState,
}

A rendering-oriented representation of the frame built by the render backend and presented to the renderer.

Fields
content_origin: DeviceIntPoint

The origin on content produced by the render tasks.

device_rect: DeviceIntRect

The rectangle to show the frame in, on screen.

layer: DocumentLayerpasses: Vec<RenderPass>profile_counters: FrameProfileCounterstransform_palette: Vec<TransformData>render_tasks: RenderTaskGraphprim_headers: PrimitiveHeadersgpu_cache_frame_id: FrameId

The GPU cache frame that the contents of Self depend on

deferred_resolves: Vec<DeferredResolve>

List of textures that we don't know about yet from the backend thread. The render thread will use a callback to resolve these and patch the data structures.

has_texture_cache_tasks: bool

True if this frame contains any render tasks that write to the texture cache.

has_been_rendered: bool

True if this frame has been drawn by the renderer.

recorded_dirty_regions: Vec<RecordedDirtyRegion>

Dirty regions recorded when generating this frame. Empty when not in testing.

debug_items: Vec<DebugItem>

Debugging information to overlay for this frame.

composite_state: CompositeState

Contains picture cache tiles, and associated information. Used by the renderer to composite tiles into the framebuffer, or hand them off to an OS compositor.

Methods

impl Frame

pub fn must_be_drawn(&self) -> bool