Struct webrender::frame_builder::Frame

pub struct Frame {

    pub device_rect: DeviceIntRect,
    pub passes: Vec<RenderPass>,
    pub transform_palette: Vec<TransformData>,
    pub render_tasks: RenderTaskGraph,
    pub prim_headers: PrimitiveHeaders,
    pub gpu_cache_frame_id: FrameId,
    pub deferred_resolves: Vec<DeferredResolve>,
    pub has_texture_cache_tasks: bool,
    pub has_been_rendered: bool,
    pub debug_items: Vec<DebugItem>,
    pub composite_state: CompositeState,
    pub gpu_buffer_f: GpuBuffer<GpuBufferBlockF>,
    pub gpu_buffer_i: GpuBuffer<GpuBufferBlockI>,

}

A rendering-oriented representation of the frame built by the render backend and presented to the renderer.

Fields
device_rect: DeviceIntRect

The rectangle to show the frame in, on screen.

passes: Vec<RenderPass>transform_palette: Vec<TransformData>render_tasks: RenderTaskGraphprim_headers: PrimitiveHeadersgpu_cache_frame_id: FrameId

The GPU cache frame that the contents of Self depend on

deferred_resolves: Vec<DeferredResolve>

List of textures that we don’t know about yet from the backend thread. The render thread will use a callback to resolve these and patch the data structures.

has_texture_cache_tasks: bool

True if this frame contains any render tasks that write to the texture cache.

has_been_rendered: bool

True if this frame has been drawn by the renderer.

debug_items: Vec<DebugItem>

Debugging information to overlay for this frame.

composite_state: CompositeState

Contains picture cache tiles, and associated information. Used by the renderer to composite tiles into the framebuffer, or hand them off to an OS compositor.

gpu_buffer_f: GpuBuffer<GpuBufferBlockF>

Main GPU data buffer constructed (primarily) during the prepare pass for primitives that were visible and dirty.

gpu_buffer_i: GpuBuffer<GpuBufferBlockI>
Implementations

impl Frame

pub fn must_be_drawn(&self) -> bool

pub fn is_nop(&self) -> bool