Struct webrender::frame_builder::FrameBuilder

pub struct FrameBuilder {
    screen_size: DeviceUintSize,
    background_color: Option<ColorF>,
    prim_store: PrimitiveStore,
    cmds: Vec<PrimitiveRunCmd>,
    config: FrameBuilderConfig,
    stacking_context_store: Vec<StackingContext>,
    clip_scroll_group_store: Vec<ClipScrollGroup>,
    packed_layers: Vec<PackedLayer>,
    scrollbar_prims: Vec<ScrollbarPrimitive>,
    clip_scroll_node_stack: Vec<ScrollLayerId>,
    stacking_context_stack: Vec<StackingContextIndex>,
}

Fields
screen_size: DeviceUintSize background_color: Option<ColorF> prim_store: PrimitiveStore cmds: Vec<PrimitiveRunCmd> config: FrameBuilderConfig stacking_context_store: Vec<StackingContext> clip_scroll_group_store: Vec<ClipScrollGroup> packed_layers: Vec<PackedLayer> scrollbar_prims: Vec<ScrollbarPrimitive> clip_scroll_node_stack: Vec<ScrollLayerId>

A stack of scroll nodes used during display list processing to properly parent new scroll nodes.

stacking_context_stack: Vec<StackingContextIndex>

A stack of stacking contexts used for creating ClipScrollGroups as primitives are added to the frame.

Methods

impl FrameBuilder

fn new(screen_size: DeviceUintSize,
background_color: Option<ColorF>,
config: FrameBuilderConfig)
-> FrameBuilder

fn add_primitive(&mut self,
scroll_layer_id: ScrollLayerId,
rect: &LayerRect,
clip_region: &ClipRegion,
container: PrimitiveContainer)
-> PrimitiveIndex

fn create_clip_scroll_group(&mut self,
stacking_context_index: StackingContextIndex,
scroll_layer_id: ScrollLayerId)
-> ClipScrollGroupIndex

fn push_stacking_context(&mut self,
reference_frame_offset: LayerPoint,
rect: LayerRect,
pipeline_id: PipelineId,
is_page_root: bool,
composite_ops: CompositeOps)

fn pop_stacking_context(&mut self)

fn push_reference_frame(&mut self,
parent_id: Option<ScrollLayerId>,
pipeline_id: PipelineId,
rect: &LayerRect,
transform: &LayerToScrollTransform,
clip_scroll_tree: &mut ClipScrollTree)
-> ScrollLayerId

fn current_reference_frame_id(&self) -> ScrollLayerId

fn setup_viewport_offset(&mut self,
window_size: DeviceUintSize,
inner_rect: DeviceUintRect,
device_pixel_ratio: f32,
clip_scroll_tree: &mut ClipScrollTree)

fn push_root(&mut self,
pipeline_id: PipelineId,
viewport_size: &LayerSize,
content_size: &LayerSize,
clip_scroll_tree: &mut ClipScrollTree)
-> ScrollLayerId

fn push_clip_scroll_node(&mut self,
new_node_id: ScrollLayerId,
parent_id: ScrollLayerId,
pipeline_id: PipelineId,
local_viewport_rect: &LayerRect,
content_size: &LayerSize,
scroll_root_id: Option<ServoScrollRootId>,
clip_region: &ClipRegion,
clip_scroll_tree: &mut ClipScrollTree)

fn pop_clip_scroll_node(&mut self)

fn current_clip_scroll_node_id(&mut self) -> ScrollLayerId

fn add_solid_rectangle(&mut self,
scroll_layer_id: ScrollLayerId,
rect: &LayerRect,
clip_region: &ClipRegion,
color: &ColorF,
flags: PrimitiveFlags)

fn supported_style(&mut self, border: &BorderSide) -> bool

fn add_border(&mut self,
scroll_layer_id: ScrollLayerId,
rect: LayerRect,
clip_region: &ClipRegion,
border_item: &BorderDisplayItem)

fn add_gradient(&mut self,
scroll_layer_id: ScrollLayerId,
rect: LayerRect,
clip_region: &ClipRegion,
start_point: LayerPoint,
end_point: LayerPoint,
stops: ItemRange,
extend_mode: ExtendMode)

fn add_radial_gradient(&mut self,
scroll_layer_id: ScrollLayerId,
rect: LayerRect,
clip_region: &ClipRegion,
start_center: LayerPoint,
start_radius: f32,
end_center: LayerPoint,
end_radius: f32,
stops: ItemRange,
extend_mode: ExtendMode)

fn add_text(&mut self,
scroll_layer_id: ScrollLayerId,
rect: LayerRect,
clip_region: &ClipRegion,
font_key: FontKey,
size: Au,
blur_radius: Au,
color: &ColorF,
glyph_range: ItemRange,
glyph_options: Option<GlyphOptions>)

fn add_box_shadow(&mut self,
scroll_layer_id: ScrollLayerId,
box_bounds: &LayerRect,
clip_region: &ClipRegion,
box_offset: &LayerPoint,
color: &ColorF,
blur_radius: f32,
spread_radius: f32,
border_radius: f32,
clip_mode: BoxShadowClipMode)

fn add_webgl_rectangle(&mut self,
scroll_layer_id: ScrollLayerId,
rect: LayerRect,
clip_region: &ClipRegion,
context_id: WebGLContextId)

fn add_image(&mut self,
scroll_layer_id: ScrollLayerId,
rect: LayerRect,
clip_region: &ClipRegion,
stretch_size: &LayerSize,
tile_spacing: &LayerSize,
sub_rect: Option<TexelRect>,
image_key: ImageKey,
image_rendering: ImageRendering,
tile: Option<TileOffset>)

fn add_yuv_image(&mut self,
scroll_layer_id: ScrollLayerId,
rect: LayerRect,
clip_region: &ClipRegion,
y_image_key: ImageKey,
u_image_key: ImageKey,
v_image_key: ImageKey,
color_space: YuvColorSpace)

fn build_layer_screen_rects_and_cull_layers(&mut self,
screen_rect: &DeviceIntRect,
clip_scroll_tree: &mut ClipScrollTree,
auxiliary_lists_map: &HashMap<PipelineId, AuxiliaryLists, BuildHasherDefault<FnvHasher>>,
resource_cache: &mut ResourceCache,
profile_counters: &mut FrameProfileCounters,
device_pixel_ratio: f32)

Compute the contribution (bounding rectangles, and resources) of layers and their primitives in screen space.

fn update_scroll_bars(&mut self, clip_scroll_tree: &ClipScrollTree)

fn build_render_task(&self) -> (RenderTask, usize)

fn build(&mut self,
resource_cache: &mut ResourceCache,
frame_id: FrameId,
clip_scroll_tree: &mut ClipScrollTree,
auxiliary_lists_map: &HashMap<PipelineId, AuxiliaryLists, BuildHasherDefault<FnvHasher>>,
device_pixel_ratio: f32,
texture_cache_profile: &mut TextureCacheProfileCounters)
-> Frame