Struct webrender::frame_builder::FrameBuilder

pub struct FrameBuilder {
    pending_retained_tiles: RetainedTiles,
    pub globals: FrameGlobalResources,
}

Produces the frames that are sent to the renderer.

Fields
pending_retained_tiles: RetainedTiles

Cache of surface tiles from the previous frame builder that can optionally be consumed by this frame builder.

globals: FrameGlobalResources
Implementations

impl FrameBuilder

pub fn new() -> Self

pub fn set_retained_resources(&mut self, retained_tiles: RetainedTiles)

Provide any cached surface tiles from the previous frame builder to a new frame builder. These will be consumed or dropped the first time a new frame builder creates a frame.

fn build_layer_screen_rects_and_cull_layers(
&mut self,
scene: &mut BuiltScene,
global_screen_world_rect: WorldRect,
resource_cache: &mut ResourceCache,
gpu_cache: &mut GpuCache,
render_tasks: &mut RenderTaskGraph,
profile_counters: &mut FrameProfileCounters,
global_device_pixel_scale: DevicePixelScale,
scene_properties: &SceneProperties,
transform_palette: &mut TransformPalette,
data_stores: &mut DataStores,
surfaces: &mut Vec<SurfaceInfo>,
scratch: &mut PrimitiveScratchBuffer,
debug_flags: DebugFlags,
texture_cache_profile: &mut TextureCacheProfileCounters,
composite_state: &mut CompositeState,
tile_cache_logger: &mut TileCacheLogger
) -> Option<RenderTaskId>

Compute the contribution (bounding rectangles, and resources) of layers and their primitives in screen space.

pub fn build(
&mut self,
scene: &mut BuiltScene,
resource_cache: &mut ResourceCache,
gpu_cache: &mut GpuCache,
stamp: FrameStamp,
global_device_pixel_scale: DevicePixelScale,
layer: DocumentLayer,
device_origin: DeviceIntPoint,
pan: WorldPoint,
resource_profile: &mut ResourceProfileCounters,
scene_properties: &SceneProperties,
data_stores: &mut DataStores,
scratch: &mut PrimitiveScratchBuffer,
render_task_counters: &mut RenderTaskGraphCounters,
debug_flags: DebugFlags,
tile_cache_logger: &mut TileCacheLogger
) -> Frame