Struct webrender::frame_builder::FrameBuildingState

pub struct FrameBuildingState<'a> {

    pub rg_builder: &'a mut RenderTaskGraphBuilder,
    pub clip_store: &'a mut ClipStore,
    pub resource_cache: &'a mut ResourceCache,
    pub gpu_cache: &'a mut GpuCache,
    pub transforms: &'a mut TransformPalette,
    pub segment_builder: SegmentBuilder,
    pub surfaces: &'a mut Vec<SurfaceInfo>,
    pub dirty_region_stack: Vec<DirtyRegion>,
    pub composite_state: &'a mut CompositeState,
    pub num_visible_primitives: u32,
    pub plane_splitters: &'a mut [BspSplitter<PlaneSplitAnchor>],
    pub surface_builder: SurfaceBuilder,
    pub cmd_buffers: &'a mut CommandBufferList,
    pub clip_tree: &'a ClipTree,
    pub frame_gpu_data: &'a mut GpuBufferBuilder,

}

Fields
rg_builder: &'a mut RenderTaskGraphBuilderclip_store: &'a mut ClipStoreresource_cache: &'a mut ResourceCachegpu_cache: &'a mut GpuCachetransforms: &'a mut TransformPalettesegment_builder: SegmentBuildersurfaces: &'a mut Vec<SurfaceInfo>dirty_region_stack: Vec<DirtyRegion>composite_state: &'a mut CompositeStatenum_visible_primitives: u32plane_splitters: &'a mut [BspSplitter<PlaneSplitAnchor>]surface_builder: SurfaceBuildercmd_buffers: &'a mut CommandBufferListclip_tree: &'a ClipTreeframe_gpu_data: &'a mut GpuBufferBuilder
Implementations

impl<'a> FrameBuildingState<'a>

pub fn current_dirty_region(&self) -> &DirtyRegion

Retrieve the current dirty region during primitive traversal.

pub fn push_dirty_region(&mut self, region: DirtyRegion)

Push a new dirty region for child primitives to cull / clip against.

pub fn pop_dirty_region(&mut self)

Pop the top dirty region from the stack.

pub fn push_prim( &mut self, cmd: &PrimitiveCommand, spatial_node_index: SpatialNodeIndex, targets: &[CommandBufferIndex], )

Push a primitive command to a set of command buffers

pub fn push_cmd( &mut self, cmd: &PrimitiveCommand, targets: &[CommandBufferIndex], )

Push a command to a set of command buffers

pub fn set_segments( &mut self, segments: &[QuadSegment], targets: &[CommandBufferIndex], )

Set the active list of segments in a set of command buffers