Module webrender::gpu_types

Structs

BlurInstance

BorderInstance

BrushFlags
Flags that define how the common brush shader code should process this instance.

BrushInstance

ClipMaskBorderCornerDotDash
A border corner dot or dash drawn into the clipping mask.

ClipMaskInstance
A clipping primitive drawn into the clipping mask. Could be an image or a rectangle, which defines the way address is treated.

CompositeInstance
Vertex format for picture cache composite shader.

GlyphInstance

ImageSource

PrimitiveHeader

PrimitiveHeaderF

PrimitiveHeaderI

PrimitiveHeaderIndex

PrimitiveHeaders

PrimitiveInstanceData

RelativeTransformKey

ResolveInstanceData
Vertex format for resolve style operations with pixel local storage.

ScalingInstance

SplitCompositeInstance

SvgFilterInstance

TransformData
The GPU data payload for a transform palette entry.

TransformMetadata

TransformPalette

TransformPaletteId

ZBufferId

ZBufferIdGenerator

Enums

BlurDirection

BorderSegment

BoxShadowStretchMode

RasterizationSpace

UvRectKind

Constants

MAX_DOCUMENT_LAYERS

MAX_DOCUMENT_LAYER_VALUE

MAX_ITEMS_PER_DOCUMENT_LAYER

MIN_DOCUMENT_LAYER_VALUE

VECS_PER_TRANSFORM

_IMPL_SERIALIZE_FOR_BlurDirection

_IMPL_SERIALIZE_FOR_BlurInstance

_IMPL_SERIALIZE_FOR_BorderInstance

_IMPL_SERIALIZE_FOR_BorderSegment

_IMPL_SERIALIZE_FOR_BoxShadowStretchMode

_IMPL_SERIALIZE_FOR_BrushFlags

_IMPL_SERIALIZE_FOR_ClipMaskBorderCornerDotDash

_IMPL_SERIALIZE_FOR_ClipMaskInstance

_IMPL_SERIALIZE_FOR_ImageSource

_IMPL_SERIALIZE_FOR_PrimitiveHeaderF

_IMPL_SERIALIZE_FOR_PrimitiveHeaderI

_IMPL_SERIALIZE_FOR_PrimitiveHeaderIndex

_IMPL_SERIALIZE_FOR_PrimitiveHeaders

_IMPL_SERIALIZE_FOR_PrimitiveInstanceData

_IMPL_SERIALIZE_FOR_RasterizationSpace

_IMPL_SERIALIZE_FOR_ScalingInstance

_IMPL_SERIALIZE_FOR_SvgFilterInstance

_IMPL_SERIALIZE_FOR_TransformData

_IMPL_SERIALIZE_FOR_TransformPaletteId

_IMPL_SERIALIZE_FOR_UvRectKind

_IMPL_SERIALIZE_FOR_ZBufferId

_IMPL_SERIALIZE_FOR_ZBufferIdGenerator

Functions

register_transform