Module webrender::gpu_types

Structs

BlurInstance

BorderInstance

BrushFlags
Flags that define how the common brush shader code should process this instance.

BrushInstance
Convenience structure to encode into PrimitiveInstanceData.

ClearInstance
Vertex format for issuing colored quads.

ClipMaskBorderCornerDotDash
A border corner dot or dash drawn into the clipping mask.

ClipMaskInstance
A clipping primitive drawn into the clipping mask. Could be an image or a rectangle, which defines the way address is treated.

CompositeInstance
Vertex format for picture cache composite shader. When editing the members, update desc::COMPOSITE so its list of instance_attributes matches:

GlyphInstance

ImageBrushData
Convenience structure to encode into the image brush's user data.

ImageSource

PrimitiveHeader

PrimitiveHeaderF

PrimitiveHeaderI

PrimitiveHeaderIndex

PrimitiveHeaders

PrimitiveInstanceData

RelativeTransformKey

ResolveInstanceData
Vertex format for resolve style operations with pixel local storage.

ScalingInstance

SplitCompositeInstance

SvgFilterInstance

TransformData
The GPU data payload for a transform palette entry.

TransformMetadata

TransformPalette

TransformPaletteId

ZBufferId

ZBufferIdGenerator

Enums

BlurDirection

BorderSegment

BoxShadowStretchMode

BrushShaderKind
A shader kind identifier that can be used by a generic-shader to select the behavior at runtime.

RasterizationSpace

UvRectKind

Constants

MAX_DOCUMENT_LAYERS

MAX_DOCUMENT_LAYER_VALUE

MIN_DOCUMENT_LAYER_VALUE

VECS_PER_TRANSFORM

Functions

get_shader_opacity

register_transform