webrender

Module image_tiling

Structs

Repetition

RepetitionIterator

Tile

TileIterator

TileIteratorExtent

Functions

compute_tile_origin_1d 🔒

compute_tile_range

compute_tile_rect

compute_tile_size

compute_tile_size_1d 🔒

compute_valid_tiles_if_bounds_change

first_tile_size_1d 🔒

for_each_tile_in_range

last_tile_size_1d 🔒

repetitions

simplify_repeated_primitive

If repetitions are far enough apart that only one is within the primitive rect, then we can simplify the parameters and treat the primitive as not repeated. This can let us avoid unnecessary work later to handle some of the parameters.

tile_range_1d 🔒

Compute the range of tiles (in number of tiles) that intersect the provided image range (in pixels) in an arbitrary dimension.

tiles

tiles_1d 🔒

Decompose tiles along an arbitrary axis.