Module webrender::internal_types

Structs

BatchTextures
Optional textures that can be used as a source in the shaders. Textures that are not used by the batch are equal to TextureId::invalid().

CacheTextureId

DebugColorVertex

DebugFontVertex

PackedColor

PackedTexel

PackedVertex

RectUv

RendererFrame
Mostly wraps a tiling::Frame, adding a bit of extra information.

StackingContextIndex

TextureUpdate

TextureUpdateList

Enums

AxisDirection

BlurAttribute

ClearAttribute

ClipAttribute

FontTemplate

GLContextHandleWrapper

GLContextWrapper

HardwareCompositeOp

LowLevelFilterOp

RenderTargetMode

ResultMsg

SourceTexture

TextureSampler

TextureUpdateOp

VertexAttribute

Constants

ANGLE_FLOAT_TO_FIXED

COLOR_FLOAT_TO_FIXED

DEFAULT_TEXTURE

ORTHO_FAR_PLANE

ORTHO_NEAR_PLANE

Type Definitions

ExternalImageUpdateList