webrender

Module internal_types

Re-exports

pub use crate::frame_allocator::FrameAllocator;

pub use crate::frame_allocator::FrameMemory;

Structs

CacheTextureId

An ID for a texture that is owned by the texture_cache module.

DeferredResolveIndex

FilterGraphNode

FilterGraphPictureReference

FrameId

FrameStamp

Identifier to track a sequence of frames.

LayoutPrimitiveInfo

Primitive metadata we pass around in a bunch of places

PipelineInstanceId

PlaneSplitAnchor

Custom field embedded inside the Polygon struct of the plane-split crate.

PlaneSplitterIndex

An index into the scene’s list of plane splitters

RenderTargetInfo

RenderedDocument

Wraps a frame_builder::Frame, but conceptually could hold more information

ResourceUpdateList

A list of updates built by the render backend that should be applied by the renderer thread.

SwizzleSettings

Swizzle settings of the texture cache.

TextureCacheAllocInfo

Information used when allocating / reallocating.

TextureCacheAllocation

Command to allocate, reallocate, or free a texture for the texture cache.

TextureCacheCopy

Command to update the contents of the texture cache.

TextureCacheUpdate

Command to update the contents of the texture cache.

TextureSourceExternal

TextureUpdateList

Atomic set of commands to manipulate the texture cache, generated on the RenderBackend thread and executed on the Renderer thread.

Enums

DebugOutput

Filter

Equivalent to api::FilterOp with added internal information

FilterGraphOp

ResultMsg

Swizzle

TextureCacheAllocationKind

Sub-operation-specific information for allocation operations.

TextureCacheCategory

A little bit of extra information to make memory reports more useful

TextureSource

Identifies the source of an input texture to a shader.

TextureUpdateSource

Constants

OPACITY_EPSILON 🔒

An arbitrary number which we assume opacity is invisible below.

SVGFE_CONVOLVE_DIAMETER_LIMIT

SVGFE_CONVOLVE_VALUES_LIMIT

Functions

size_of_frame_vec

Type Aliases

FastHashMap

FastHashSet

FrameVec

PlaneSplitter

A concrete plane splitter type used in WebRender.