Module webrender::internal_types

Structs

CacheTextureId

An ID for a texture that is owned by the texture_cache module.

DeferredResolveIndex

FrameId

FrameStamp

Identifier to track a sequence of frames.

LayoutPrimitiveInfo

Primitive metadata we pass around in a bunch of places

PipelineInstanceId

PlaneSplitAnchor

Custom field embedded inside the Polygon struct of the plane-split crate.

PlaneSplitterIndex

An index into the scene’s list of plane splitters

RenderTargetInfo

RenderedDocument

Wraps a frame_builder::Frame, but conceptually could hold more information

ResourceUpdateList

A list of updates built by the render backend that should be applied by the renderer thread.

SwizzleSettings

Swizzle settings of the texture cache.

TextureCacheAllocInfo

Information used when allocating / reallocating.

TextureCacheAllocation

Command to allocate, reallocate, or free a texture for the texture cache.

TextureCacheCopy

Command to update the contents of the texture cache.

TextureCacheUpdate

Command to update the contents of the texture cache.

TextureUpdateList

Atomic set of commands to manipulate the texture cache, generated on the RenderBackend thread and executed on the Renderer thread.

Enums

DebugOutput

Filter

Equivalent to api::FilterOp with added internal information

ResultMsg

Swizzle

TextureCacheAllocationKind

Sub-operation-specific information for allocation operations.

TextureCacheCategory

A little bit of extra information to make memory reports more useful

TextureSource

Identifies the source of an input texture to a shader.

TextureUpdateSource

Constants

OPACITY_EPSILON 🔒

An arbitrary number which we assume opacity is invisible below.

Type Aliases

FastHashMap

FastHashSet

PlaneSplitter

A concrete plane splitter type used in WebRender.