Struct webrender::internal_types::TextureUpdateList

pub struct TextureUpdateList {
    pub clears_shared_cache: bool,
    pub allocations: Vec<TextureCacheAllocation>,
    pub updates: FastHashMap<CacheTextureId, Vec<TextureCacheUpdate>>,
    pub copies: FastHashMap<(CacheTextureId, CacheTextureId), Vec<TextureCacheCopy>>,
}

Atomic set of commands to manipulate the texture cache, generated on the RenderBackend thread and executed on the Renderer thread.

The list of allocation operations is processed before the updates. This is important to allow coalescing of certain allocation operations.

Fields
clears_shared_cache: bool

Indicates that there was some kind of cleanup clear operation. Used for sanity checks.

allocations: Vec<TextureCacheAllocation>

Commands to alloc/realloc/free the textures. Processed first.

updates: FastHashMap<CacheTextureId, Vec<TextureCacheUpdate>>

Commands to update the contents of the textures. Processed second.

copies: FastHashMap<(CacheTextureId, CacheTextureId), Vec<TextureCacheCopy>>

Commands to move items within the cache, these are applied before everything else in the update list.

Implementations

impl TextureUpdateList

pub fn new() -> Self

Mints a new TextureUpdateList.

pub fn is_nop(&self) -> bool

Returns true if this is a no-op (no updates to be applied).

pub fn note_clear(&mut self)

Sets the clears_shared_cache flag for renderer-side sanity checks.

pub fn push_update(&mut self, id: CacheTextureId, update: TextureCacheUpdate)

Pushes an update operation onto the list.

pub fn push_debug_clear( &mut self, id: CacheTextureId, origin: DeviceIntPoint, width: i32, height: i32 )

Sends a command to the Renderer to clear the portion of the shared region we just freed. Used when the texture cache debugger is enabled.

pub fn push_alloc(&mut self, id: CacheTextureId, info: TextureCacheAllocInfo)

Pushes an allocation operation onto the list.

pub fn push_reset(&mut self, id: CacheTextureId, info: TextureCacheAllocInfo)

Pushes a reallocation operation onto the list, potentially coalescing with previous operations.

pub fn push_free(&mut self, id: CacheTextureId)

Pushes a free operation onto the list, potentially coalescing with previous operations.

pub fn push_copy( &mut self, src_id: CacheTextureId, src_rect: &DeviceIntRect, dst_id: CacheTextureId, dst_rect: &DeviceIntRect )

Push a copy operation from a texture to another.

The source and destination rectangles must have the same size. The copies are applied before every other operations in the texture update list.

fn debug_assert_coalesced(&self, id: CacheTextureId)