Struct webrender::layer::Layer

pub struct Layer {
    pub aabb_tree: AABBTree,
    pub scrolling: ScrollingState,
    pub viewport_rect: Rect<f32>,
    pub viewport_transform: Matrix4D<f32>,
    pub layer_size: Size2D<f32>,
    pub world_origin: Point2D<f32>,
    pub local_transform: Matrix4D<f32>,
    pub local_perspective: Matrix4D<f32>,
    pub world_transform: Option<Matrix4D<f32>>,
    pub pipeline_id: PipelineId,
    pub stacking_context_id: ServoStackingContextId,
    pub children: Vec<ScrollLayerId>,
}

Fields
aabb_tree: AABBTree scrolling: ScrollingState viewport_rect: Rect<f32>

The viewable region, in world coordinates.

viewport_transform: Matrix4D<f32>

The transform to apply to the viewable region, in world coordinates.

TODO(pcwalton): These should really be a stack of clip regions and transforms.

layer_size: Size2D<f32> world_origin: Point2D<f32> local_transform: Matrix4D<f32> local_perspective: Matrix4D<f32> world_transform: Option<Matrix4D<f32>> pipeline_id: PipelineId stacking_context_id: ServoStackingContextId children: Vec<ScrollLayerId>
Methods

impl Layer

fn new(world_origin: Point2D<f32>, layer_size: Size2D<f32>, viewport_rect: &Rect<f32>, viewport_transform: &Matrix4D<f32>, local_transform: &Matrix4D<f32>, local_perspective: &Matrix4D<f32>, pipeline_id: PipelineId, stacking_context_id: ServoStackingContextId) -> Layer

fn add_child(&mut self, child: ScrollLayerId)

fn reset(&mut self, pending_updates: &mut BatchUpdateList)

fn insert(&mut self, rect: Rect<f32>, draw_list_group_id: DrawListGroupId, draw_list_id: DrawListId, item_index: DrawListItemIndex)

fn finalize(&mut self, scrolling: &ScrollingState)

fn cull(&mut self)

fn print(&self)

fn overscroll_amount(&self) -> Size2D<f32>

fn stretch_overscroll_spring(&mut self)

fn tick_scrolling_bounce_animation(&mut self)