Struct webrender::picture::TileCacheInstance

pub struct TileCacheInstance {

    pub slice: usize,
    pub slice_flags: SliceFlags,
    pub current_tile_size: DeviceIntSize,
    pub spatial_node_index: SpatialNodeIndex,
    pub tiles: FastHashMap<TileOffset, Box<Tile>>,
    map_local_to_surface: SpaceMapper<LayoutPixel, PicturePixel>,
    map_child_pic_to_surface: SpaceMapper<PicturePixel, PicturePixel>,
    opacity_bindings: FastHashMap<PropertyBindingId, BindingInfo<f32>>,
    old_opacity_bindings: FastHashMap<PropertyBindingId, BindingInfo<f32>>,
    spatial_node_comparer: SpatialNodeComparer,
    color_bindings: FastHashMap<PropertyBindingId, BindingInfo<ColorU>>,
    old_color_bindings: FastHashMap<PropertyBindingId, BindingInfo<ColorU>>,
    pub dirty_region: DirtyRegion,
    tile_size: PictureSize,
    tile_rect: Rect<i32, TileCoordinate>,
    tile_bounds_p0: TileOffset,
    tile_bounds_p1: TileOffset,
    pub local_rect: PictureRect,
    local_clip_rect: PictureRect,
    surface_index: SurfaceIndex,
    pub background_color: Option<ColorF>,
    pub backdrop: BackdropInfo,
    pub subpixel_mode: SubpixelMode,
    pub shared_clips: Vec<ClipInstance>,
    shared_clip_chain: ClipChainId,
    root_transform: TransformKey,
    frames_until_size_eval: usize,
    fract_offset: PictureVector2D,
    virtual_offset: DeviceIntPoint,
    compare_cache: FastHashMap<PrimitiveComparisonKey, PrimitiveCompareResult>,
    pub native_surface: Option<NativeSurface>,
    pub device_position: DevicePoint,
    tile_size_override: Option<DeviceIntSize>,
    pub external_surfaces: Vec<ExternalSurfaceDescriptor>,
    pub z_id_opaque: ZBufferId,
    pub external_native_surface_cache: FastHashMap<ExternalNativeSurfaceKey, ExternalNativeSurface>,
    frame_id: FrameId,

}

Represents a cache of tiles that make up a picture primitives.

Fields
slice: usize

Index of the tile cache / slice for this frame builder. It’s determined by the setup_picture_caching method during flattening, which splits the picture tree into multiple slices. It’s used as a simple input to the tile keys. It does mean we invalidate tiles if a new layer gets inserted / removed between display lists - this seems very unlikely to occur on most pages, but can be revisited if we ever notice that.

slice_flags: SliceFlags

Propagated information about the slice

current_tile_size: DeviceIntSize

The currently selected tile size to use for this cache

spatial_node_index: SpatialNodeIndex

The positioning node for this tile cache.

tiles: FastHashMap<TileOffset, Box<Tile>>

Hash of tiles present in this picture.

map_local_to_surface: SpaceMapper<LayoutPixel, PicturePixel>

A helper struct to map local rects into surface coords.

map_child_pic_to_surface: SpaceMapper<PicturePixel, PicturePixel>

A helper struct to map child picture rects into picture cache surface coords.

opacity_bindings: FastHashMap<PropertyBindingId, BindingInfo<f32>>

List of opacity bindings, with some extra information about whether they changed since last frame.

old_opacity_bindings: FastHashMap<PropertyBindingId, BindingInfo<f32>>

Switch back and forth between old and new bindings hashmaps to avoid re-allocating.

spatial_node_comparer: SpatialNodeComparer

A helper to compare transforms between previous and current frame.

color_bindings: FastHashMap<PropertyBindingId, BindingInfo<ColorU>>

List of color bindings, with some extra information about whether they changed since last frame.

old_color_bindings: FastHashMap<PropertyBindingId, BindingInfo<ColorU>>

Switch back and forth between old and new bindings hashmaps to avoid re-allocating.

dirty_region: DirtyRegion

The current dirty region tracker for this picture.

tile_size: PictureSize

Current size of tiles in picture units.

tile_rect: Rect<i32, TileCoordinate>

Tile coords of the currently allocated grid.

tile_bounds_p0: TileOffset

Pre-calculated versions of the tile_rect above, used to speed up the calculations in get_tile_coords_for_rect.

tile_bounds_p1: TileOffsetlocal_rect: PictureRect

Local rect (unclipped) of the picture this cache covers.

local_clip_rect: PictureRect

The local clip rect, from the shared clips of this picture.

surface_index: SurfaceIndex

The surface index that this tile cache will be drawn into.

background_color: Option<ColorF>

The background color from the renderer. If this is set opaque, we know it’s fine to clear the tiles to this and allow subpixel text on the first slice.

backdrop: BackdropInfo

Information about the calculated backdrop content of this cache.

subpixel_mode: SubpixelMode

The allowed subpixel mode for this surface, which depends on the detected opacity of the background.

shared_clips: Vec<ClipInstance>

A list of clip handles that exist on every (top-level) primitive in this picture. It’s often the case that these are root / fixed position clips. By handling them here, we can avoid applying them to the items, which reduces work, but more importantly reduces invalidations.

shared_clip_chain: ClipChainId

The clip chain that represents the shared_clips above. Used to build the local clip rect for this tile cache.

root_transform: TransformKey

The current transform of the picture cache root spatial node

frames_until_size_eval: usize

The number of frames until this cache next evaluates what tile size to use. If a picture rect size is regularly changing just around a size threshold, we don’t want to constantly invalidate and reallocate different tile size configuration each frame.

fract_offset: PictureVector2D

The current fractional offset of the cached picture

virtual_offset: DeviceIntPoint

For DirectComposition, virtual surfaces don’t support negative coordinates. However, picture cache tile coordinates can be negative. To handle this, we apply an offset to each tile in DirectComposition. We want to change this as little as possible, to avoid invalidating tiles. However, if we have a picture cache tile coordinate which is outside the virtual surface bounds, we must change this to allow correct remapping of the coordinates passed to BeginDraw in DC.

compare_cache: FastHashMap<PrimitiveComparisonKey, PrimitiveCompareResult>

keep around the hash map used as compare_cache to avoid reallocating it each frame.

native_surface: Option<NativeSurface>

The allocated compositor surfaces for this picture cache. May be None if not using native compositor, or if the surface was destroyed and needs to be reallocated next time this surface contains valid tiles.

device_position: DevicePoint

The current device position of this cache. Used to set the compositor offset of the surface when building the visual tree.

tile_size_override: Option<DeviceIntSize>

The currently considered tile size override. Used to check if we should re-evaluate tile size, even if the frame timer hasn’t expired.

external_surfaces: Vec<ExternalSurfaceDescriptor>

List of external surfaces that have been promoted from primitives in this tile cache.

z_id_opaque: ZBufferId

z-buffer ID assigned to opaque tiles in this slice

external_native_surface_cache: FastHashMap<ExternalNativeSurfaceKey, ExternalNativeSurface>

A cache of compositor surfaces that are retained between frames

frame_id: FrameId

Current frame ID of this tile cache instance. Used for book-keeping / garbage collecting

Implementations

impl TileCacheInstance

pub fn new(
slice: usize,
slice_flags: SliceFlags,
spatial_node_index: SpatialNodeIndex,
background_color: Option<ColorF>,
shared_clips: Vec<ClipInstance>,
shared_clip_chain: ClipChainId,
fb_config: &FrameBuilderConfig
) -> Self

pub fn is_opaque(&self) -> bool

Returns true if this tile cache is considered opaque.

fn get_tile_coords_for_rect(
&self,
rect: &PictureRect
) -> (TileOffset, TileOffset)

Get the tile coordinates for a given rectangle.

pub fn pre_update(
&mut self,
pic_rect: PictureRect,
surface_index: SurfaceIndex,
frame_context: &FrameVisibilityContext<'_>,
frame_state: &mut FrameVisibilityState<'_>
) -> WorldRect

Update transforms, opacity, color bindings and tile rects.

fn can_promote_to_surface(
&mut self,
flags: PrimitiveFlags,
prim_clip_chain: &ClipChainInstance,
prim_spatial_node_index: SpatialNodeIndex,
on_picture_surface: bool,
frame_context: &FrameVisibilityContext<'_>
) -> SurfacePromotionResult

fn setup_compositor_surfaces_yuv(
&mut self,
prim_info: &mut PrimitiveDependencyInfo,
prim_rect: PictureRect,
frame_context: &FrameVisibilityContext<'_>,
image_dependencies: &[ImageDependency; 3],
api_keys: &[ImageKey; 3],
resource_cache: &mut ResourceCache,
composite_state: &mut CompositeState,
image_rendering: ImageRendering,
color_depth: ColorDepth,
color_space: YuvColorSpace,
format: YuvFormat
) -> bool

fn setup_compositor_surfaces_rgb(
&mut self,
prim_info: &mut PrimitiveDependencyInfo,
prim_rect: PictureRect,
frame_context: &FrameVisibilityContext<'_>,
image_dependency: ImageDependency,
api_key: ImageKey,
resource_cache: &mut ResourceCache,
composite_state: &mut CompositeState,
image_rendering: ImageRendering,
flip_y: bool
) -> bool

fn setup_compositor_surfaces_impl(
&mut self,
prim_info: &mut PrimitiveDependencyInfo,
prim_rect: PictureRect,
frame_context: &FrameVisibilityContext<'_>,
dependency: ExternalSurfaceDependency,
api_keys: &[ImageKey; 3],
resource_cache: &mut ResourceCache,
composite_state: &mut CompositeState,
image_rendering: ImageRendering
) -> bool

pub fn update_prim_dependencies(
&mut self,
prim_instance: &mut PrimitiveInstance,
prim_spatial_node_index: SpatialNodeIndex,
prim_clip_chain: Option<&ClipChainInstance>,
local_prim_rect: LayoutRect,
frame_context: &FrameVisibilityContext<'_>,
data_stores: &DataStores,
clip_store: &ClipStore,
pictures: &[PicturePrimitive],
resource_cache: &mut ResourceCache,
opacity_binding_store: &Storage<OpacityBinding>,
color_bindings: &Storage<PropertyBinding<ColorU>>,
image_instances: &Storage<ImageInstance>,
surface_stack: &[SurfaceIndex],
composite_state: &mut CompositeState
) -> Option<PrimitiveVisibilityFlags>

Update the dependencies for each tile for a given primitive instance.

fn print(&self)

Print debug information about this picture cache to a tree printer.

fn calculate_subpixel_mode(&self) -> SubpixelMode

pub fn post_update(
&mut self,
frame_context: &FrameVisibilityContext<'_>,
frame_state: &mut FrameVisibilityState<'_>
)

Apply any updates after prim dependency updates. This applies any late tile invalidations, and sets up the dirty rect and set of tile blits.