Module webrender::prepare

Prepare pass

TODO: document this!

Enums

QuadRenderStrategy

Describes how clipping affects the rendering of a quad primitive.

Constants

MAX_MASK_SIZE 🔒

MIN_AA_SEGMENTS_SIZE 🔒

MIN_BRUSH_SPLIT_AREA 🔒

MIN_BRUSH_SPLIT_SIZE 🔒

Functions

add_composite_prim 🔒

add_segment 🔒

adjust_mask_scale_for_max_size 🔒

build_segments_if_needed 🔒

can_use_clip_chain_for_quad_path 🔒

decompose_repeated_gradient 🔒

get_prim_render_strategy 🔒

prepare_interned_prim_for_render 🔒

Prepare an interned primitive for rendering, by requesting resources, render tasks etc. This is equivalent to the prepare_prim_for_render_inner call for old style primitives.

prepare_prim_for_render 🔒

prepare_primitives

update_brush_segment_clip_task

Write out to the clip mask instances array the correct clip mask config for this segment.

update_clip_task

update_clip_task_for_brush 🔒

write_brush_segment_description 🔒

write_prim_blocks

write_segment 🔒