Module webrender::prepare

Prepare pass

TODO: document this!

Constants

MAX_MASK_SIZE ğŸ”’

MIN_BRUSH_SPLIT_AREA ğŸ”’

Functions

adjust_mask_scale_for_max_size ğŸ”’

build_segments_if_needed ğŸ”’

can_use_clip_chain_for_quad_path ğŸ”’

decompose_repeated_gradient ğŸ”’

prepare_interned_prim_for_render ğŸ”’

Prepare an interned primitive for rendering, by requesting resources, render tasks etc. This is equivalent to the prepare_prim_for_render_inner call for old style primitives.

prepare_prim_for_render ğŸ”’

prepare_primitives

update_brush_segment_clip_task

Write out to the clip mask instances array the correct clip mask config for this segment.

update_clip_task

update_clip_task_for_brush ğŸ”’

write_brush_segment_description ğŸ”’

write_segment ğŸ”’