webrender

Module prim_store

Modules

backdrop

borders

gradient

image

interned

line_dec

picture

storage 🔒

text_run

Structs

BorderSegmentInfo

Information about how to cache a border segment, along with the current render task cache entry.

BrushSegment

ClipCorner 🔒

ClipData

ClipRect 🔒

ClipTaskIndex

DeferredResolve

For external images, it’s not possible to know the UV coords of the image (or the image data itself) until the render thread receives the frame and issues callbacks to the client application. For external images that are visible, a DeferredResolve is created that is stored in the frame. This allows the render thread to iterate this list and update any changed texture data and update the UV rect. Any filtering is handled externally for NativeTexture external images.

FloatKey

A hashable float for using as a key during primitive interning.

NinePatchDescriptor

A hashable descriptor for nine-patches, used by image and gradient borders.

PictureIndex

PointKey

A hashable point for using as a key during primitive interning.

PolygonKey

To create a fixed-size representation of a polygon, we use a fixed number of points. Our initialization method restricts us to values <= 32. If our constant POLYGON_CLIP_VERTEX_MAX is > 32, the Rust compiler will complain.

PrimKey

PrimKeyCommonData

PrimTemplate

PrimTemplateCommonData

PrimitiveInstance

PrimitiveInstanceIndex

PrimitiveKey

PrimitiveOpacity

PrimitiveScratchBuffer

Contains various vecs of data that is used only during frame building, where we want to recycle the memory each new display list, to avoid constantly re-allocating and moving memory around. Written during primitive preparation, and read during batching.

PrimitiveStore

PrimitiveStoreStats

PrimitiveTemplate

RectangleKey

SegmentedInstance

SideOffsetsKey

A hashable SideOffset2D that can be used in primitive keys.

SizeKey

A hashable size for using as a key during primitive interning.

VectorKey

A hashable vec for using as a key during primitive interning.

VisibleGradientTile

VisibleMaskImageTile

Enums

ClipMaskKind

Represents the visibility state of a segment (wrt clip masks).

PrimitiveInstanceKind

PrimitiveTemplateKind

The shared information for a given primitive. This is interned and retained both across frames and display lists, by comparing the matching PrimitiveKey.

Constants

VECS_PER_SEGMENT

Traits

InternablePrimitive

Trait for primitives that are directly internable. see SceneBuilder::add_primitive

Type Aliases

BorderHandleStorage

ColorBindingIndex

ColorBindingStorage

GlyphKeyStorage

GradientTileRange

GradientTileStorage

ImageInstanceIndex

ImageInstanceStorage

LinearGradientStorage

PrimitiveDataHandle 🔒

SegmentInstanceIndex

SegmentInstanceStorage

SegmentStorage

SegmentsRange

TextRunIndex

TextRunStorage