Module webrender::prim_store::gradient

Structs

GradientCacheKey

GradientDataEntry
An entry in a gradient data table representing a segment of the gradient color space.

GradientGpuBlockBuilder

GradientStopKey
A hashable gradient stop that can be used in primitive keys.

LinearGradient

LinearGradientKey
Identifying key for a line decoration.

LinearGradientPrimitive

LinearGradientTemplate

RadialGradient

RadialGradientKey
Identifying key for a line decoration.

RadialGradientParams
Hashable radial gradient parameters, for use during prim interning.

RadialGradientTemplate

Constants

GRADIENT_DATA_FIRST_STOP

GRADIENT_DATA_LAST_STOP

GRADIENT_DATA_SIZE

GRADIENT_DATA_TABLE_BEGIN

GRADIENT_DATA_TABLE_END

GRADIENT_DATA_TABLE_SIZE

GRADIENT_FP_STOPS
The maximum number of stops a gradient may have to use the fast path.

_IMPL_SERIALIZE_FOR_GradientCacheKey

_IMPL_SERIALIZE_FOR_GradientStopKey

_IMPL_SERIALIZE_FOR_LinearGradient

_IMPL_SERIALIZE_FOR_LinearGradientKey

_IMPL_SERIALIZE_FOR_LinearGradientPrimitive

_IMPL_SERIALIZE_FOR_LinearGradientTemplate

_IMPL_SERIALIZE_FOR_RadialGradient

_IMPL_SERIALIZE_FOR_RadialGradientKey

_IMPL_SERIALIZE_FOR_RadialGradientParams

_IMPL_SERIALIZE_FOR_RadialGradientTemplate

Type Definitions

LinearGradientDataHandle

RadialGradientDataHandle