Module webrender::prim_store::gradient

Structs

ConicGradient

ConicGradientKey

Identifying key for a line decoration.

ConicGradientParams

Conic gradients Hashable conic gradient parameters, for use during prim interning.

ConicGradientTemplate

GradientCacheKey

GradientDataEntry

An entry in a gradient data table representing a segment of the gradient color space.

GradientGpuBlockBuilder

GradientStopKey

A hashable gradient stop that can be used in primitive keys.

LinearGradient

LinearGradientKey

Identifying key for a linear gradient.

LinearGradientPrimitive

LinearGradientTemplate

RadialGradient

RadialGradientKey

Identifying key for a radial gradient.

RadialGradientParams

Hashable radial gradient parameters, for use during prim interning.

RadialGradientTemplate

Constants

GRADIENT_DATA_FIRST_STOP

GRADIENT_DATA_LAST_STOP

GRADIENT_DATA_SIZE

GRADIENT_DATA_TABLE_BEGIN

GRADIENT_DATA_TABLE_END

GRADIENT_DATA_TABLE_SIZE

GRADIENT_FP_STOPS

The maximum number of stops a gradient may have to use the fast path.

Functions

stops_and_min_alpha

Type Definitions

ConicGradientDataHandle

LinearGradientDataHandle

RadialGradientDataHandle