webrender::prim_store::gradient

Module conic

Conic gradients

Specification: https://drafts.csswg.org/css-images-4/#conic-gradients

Conic gradients are rendered via cached render tasks and composited with the image brush.

Structs

ConicGradient

ConicGradientCacheKey

ConicGradientInstance

The per-instance shader input of a radial gradient render task.

ConicGradientKey

Identifying key for a line decoration.

ConicGradientParams

Hashable conic gradient parameters, for use during prim interning.

ConicGradientTask

ConicGradientTemplate

Functions

conic_gradient_pattern

conic_gradient_pattern_with_table

Type Aliases

ConicGradientDataHandle