Module webrender::prim_store::gradient::linear

Linear gradients

Specification: https://drafts.csswg.org/css-images-4/#linear-gradients

Linear gradients are rendered via cached render tasks and composited with the image brush.

Structs

CachedGradientSegment

FastLinearGradientInstance

The per-instance shader input of a fast-path linear gradient render task.

FastLinearGradientTask

LinearGradient

LinearGradientCacheKey

LinearGradientInstance

The per-instance shader input of a linear gradient render task.

LinearGradientKey

Identifying key for a linear gradient.

LinearGradientPrimitive

LinearGradientTask

LinearGradientTemplate

Constants

MAX_CACHED_SIZE

Functions

optimize_linear_gradient

Perform a few optimizations to the gradient that are relevant to scene building.

Type Aliases

FastLinearGradientCacheKey

LinearGradientDataHandle