webrender::prim_store

Module image

Structs

AdjustedImageSource

Represents an adjustment to apply to an image primitive. This can be used to compensate for a difference between the bounds of the images expected by the primitive and the bounds that were actually drawn in the texture cache.

Image

ImageCacheKey

ImageData

ImageInstance

Instance specific fields for an image primitive. These are currently stored in a separate array to avoid bloating the size of PrimitiveInstance. In the future, we should be able to remove this and store the information inline, by: (a) Removing opacity collapse / binding support completely. Once we have general picture caching, we don’t need this. (b) Change visible_tiles to use Storage in the primitive scratch buffer. This will reduce the size of the visible_tiles field here, and save memory allocation when image tiling is used. I’ve left it as a Vec for now to reduce the number of changes, and because image tiling is very rare on real pages.

VisibleImageTile

YuvImage

YuvImageData

Functions

edge_flags_for_tile_spacing 🔒

Type Aliases

ImageDataHandle

ImageKey

ImageTemplate

YuvImageDataHandle

YuvImageKey

YuvImageTemplate