Struct webrender::prim_store::text_run::TextRunTemplate

pub struct TextRunTemplate {
    pub common: PrimTemplateCommonData,
    pub font: FontInstance,
    pub glyphs: Arc<Vec<GlyphInstance>>,
}

Fields
common: PrimTemplateCommonDatafont: FontInstanceglyphs: Arc<Vec<GlyphInstance>>
Implementations

impl TextRunTemplate

pub fn update(&mut self, frame_state: &mut FrameBuildingState<'_>)

Update the GPU cache for a given primitive template. This may be called multiple times per frame, by each primitive reference that refers to this interned template. The initial request call to the GPU cache ensures that work is only done if the cache entry is invalid (due to first use or eviction).

fn write_prim_gpu_blocks(&mut self, frame_state: &mut FrameBuildingState<'_>)