Module webrender::profiler

Modules

expected

Structs

AverageIntProfileCounter

A profile counter recording average and maximum integer values over time slices of half a second.

AverageTimeProfileCounter

BackendProfileCounters

ChangeIndicator

DrawState

FrameProfileCounters

GpuCacheProfileCounters

GpuFrame

GpuFrameCollection

GpuProfileTag

GraphStats

IntProfileCounter

InternProfileCounters

PercentageProfileCounter

ProfileGraph

ProfileScope

A simple RAII style struct to manage a profile scope.

Profiler

RendererProfileCounters

RendererProfileTimers

ResourceProfileCounter

ResourceProfileCounters

TextureCacheProfileCounters

TimeProfileCounter

Timer

TransactionProfileCounters

Enums

ProfileStyle

Constants

AVERAGE_OVER_NS

GRAPH_FRAME_HEIGHT

GRAPH_HEIGHT

GRAPH_PADDING

GRAPH_WIDTH

ONE_SECOND_NS

PROFILE_PADDING

Statics

PROFILER_HOOKS

The current global profiler callbacks, if set by embedder.

Traits

ProfileCounter

ProfilerHooks

Defines the interface for hooking up an external profiler to WR.

Functions

add_event_marker

Records a marker of the given duration that just ended.

add_text_marker

Records a marker of the given duration that just ended.

set_profiler_hooks

Set the profiler callbacks, or None to disable the profiler. This function must only ever be called before any WR instances have been created, or the hooks will not be set.

thread_is_being_profiled

Returns true if the current thread is being profiled.