Struct webrender::profiler::Profiler

pub struct Profiler {
    counters: Vec<Counter>,
    gpu_frames: ProfilerFrameCollection,
    frame_stats: ProfilerFrameCollection,
    start: u64,
    avg_over_period: u64,
    num_graph_samples: usize,
    frame_timestamps_within_last_second: Vec<u64>,
    ui: Vec<Item>,
}

Fields
counters: Vec<Counter>gpu_frames: ProfilerFrameCollectionframe_stats: ProfilerFrameCollectionstart: u64avg_over_period: u64num_graph_samples: usizeframe_timestamps_within_last_second: Vec<u64>ui: Vec<Item>
Implementations

impl Profiler

pub fn new() -> Self

fn update_slow_event( &mut self, dst_counter: usize, counters: &[usize], threshold: f64 )

Sum a few counters and if the total amount is larger than a threshold, update a specific counter.

This is useful to monitor slow frame and slow transactions.

pub fn update(&mut self)

pub fn update_frame_stats(&mut self, stats: FullFrameStats)

pub fn set_gpu_time_queries(&mut self, gpu_queries: Vec<GpuTimer>)

pub fn index_of(&self, name: &str) -> Option<usize>

pub fn set_ui(&mut self, names: &str)

fn append_to_ui(&mut self, selection: &mut Vec<Item>, names: &str)

pub fn set_counters(&mut self, counters: &mut TransactionProfile)

pub fn get(&self, id: usize) -> Option<f64>

fn draw_counters( counters: &[Counter], selected: &[usize], x: f32, y: f32, text_buffer: &mut String, debug_renderer: &mut DebugRenderer ) -> Rect<f32>

fn draw_graph( counter: &Counter, x: f32, y: f32, text_buffer: &mut String, debug_renderer: &mut DebugRenderer ) -> Rect<f32>

fn draw_change_indicator( counter: &Counter, x: f32, y: f32, debug_renderer: &mut DebugRenderer ) -> Rect<f32>

fn draw_bar( label: &str, label_color: ColorU, counters: &[(ColorU, usize)], x: f32, y: f32, debug_renderer: &mut DebugRenderer ) -> Rect<f32>

fn draw_gpu_cache_bars( &self, x: f32, y: f32, text_buffer: &mut String, debug_renderer: &mut DebugRenderer ) -> Rect<f32>

fn draw_frame_graph( frame_collection: &ProfilerFrameCollection, x: f32, y: f32, debug_renderer: &mut DebugRenderer ) -> Rect<f32>

pub fn draw_profile( &mut self, _frame_index: u64, debug_renderer: &mut DebugRenderer, device_size: DeviceIntSize )

pub fn dump_stats(&self, sink: &mut dyn Write) -> Result<()>