webrender

Module quad

Structs

QuadTileClassifier

A helper struct for classifying a set of tiles within a quad depending on what strategy they can be used to draw them.

QuadTileInfo

Enums

QuadRenderStrategy

Describes how clipping affects the rendering of a quad primitive.

QuadTileKind

Classification result for a tile within a quad

Constants

MAX_TILES_PER_QUAD 🔒

MIN_AA_SEGMENTS_SIZE 🔒

MIN_QUAD_SPLIT_SIZE 🔒

Functions

add_composite_prim 🔒

add_pattern_prim 🔒

add_render_task_with_mask 🔒

add_to_batch

get_prim_render_strategy 🔒

prepare_quad

write_prim_blocks