Module webrender::render_backend

The high-level module responsible for managing the pipeline and preparing commands to be issued by the Renderer.

See the comment at the top of the renderer module for a description of how these two pieces interact.

Structs

DataStores
A collection of resources that are shared by clips, primitives between display lists.

Document

DocumentOps

DocumentView

FrameId

FrameStamp
Identifier to track a sequence of frames.

FrameView
Some rendering parameters applying at the frame level.

PlainRenderBackend

RenderBackend
The render backend is responsible for transforming high level display lists into GPU-friendly work which is then submitted to the renderer in the form of a frame::Frame.

SceneView
Some rendering parameters applying at the scene level.

Enums

RenderBackendStatus

Statics

NEXT_NAMESPACE_ID
The unique id for WR resource identification. The namespace_id should start from 1.