webrender::render_backend

Struct DataStores

pub struct DataStores {

    pub clip: DataStore<ClipIntern>,
    pub prim: DataStore<PrimitiveKeyKind>,
    pub normal_border: DataStore<NormalBorderPrim>,
    pub image_border: DataStore<ImageBorder>,
    pub image: DataStore<Image>,
    pub yuv_image: DataStore<YuvImage>,
    pub line_decoration: DataStore<LineDecoration>,
    pub linear_grad: DataStore<LinearGradient>,
    pub radial_grad: DataStore<RadialGradient>,
    pub conic_grad: DataStore<ConicGradient>,
    pub picture: DataStore<Picture>,
    pub text_run: DataStore<TextRun>,
    pub filter_data: DataStore<FilterDataIntern>,
    pub backdrop_capture: DataStore<BackdropCapture>,
    pub backdrop_render: DataStore<BackdropRender>,
    pub polygon: DataStore<PolygonIntern>,
    pub box_shadow: DataStore<BoxShadow>,

}

A collection of resources that are shared by clips, primitives between display lists.

Fields
clip: DataStore<ClipIntern>prim: DataStore<PrimitiveKeyKind>normal_border: DataStore<NormalBorderPrim>image_border: DataStore<ImageBorder>image: DataStore<Image>yuv_image: DataStore<YuvImage>line_decoration: DataStore<LineDecoration>linear_grad: DataStore<LinearGradient>radial_grad: DataStore<RadialGradient>conic_grad: DataStore<ConicGradient>picture: DataStore<Picture>text_run: DataStore<TextRun>filter_data: DataStore<FilterDataIntern>backdrop_capture: DataStore<BackdropCapture>backdrop_render: DataStore<BackdropRender>polygon: DataStore<PolygonIntern>box_shadow: DataStore<BoxShadow>
Implementations

impl DataStores

fn report_memory(&self, ops: &mut MallocSizeOfOps, r: &mut MemoryReport)

Reports CPU heap usage.

fn apply_updates( &mut self, updates: InternerUpdates, profile: &mut TransactionProfile, )

impl DataStores

pub fn get_local_prim_rect( &self, prim_instance: &PrimitiveInstance, pictures: &[PicturePrimitive], surfaces: &[SurfaceInfo], ) -> LayoutRect

Returns the local rect for a primitive. For most primitives, this is stored in the template. For pictures, this is stored inside the picture primitive instance itself, since this is determined during frame building.

pub fn get_local_prim_coverage_rect( &self, prim_instance: &PrimitiveInstance, pictures: &[PicturePrimitive], surfaces: &[SurfaceInfo], ) -> LayoutRect

Returns the local coverage (space occupied) for a primitive. For most primitives, this is stored in the template. For pictures, this is stored inside the picture primitive instance itself, since this is determined during frame building.

pub fn prim_may_need_repetition( &self, prim_instance: &PrimitiveInstance, ) -> bool

Returns true if this primitive might need repition.

pub fn prim_has_anti_aliasing(&self, prim_instance: &PrimitiveInstance) -> bool

Returns true if this primitive has anti-aliasing enabled.

pub fn as_common_data( &self, prim_inst: &PrimitiveInstance, ) -> &PrimTemplateCommonData