Struct webrender::render_backend::Document

struct Document {
    id: DocumentId,
    removed_pipelines: Vec<(PipelineId, DocumentId)>,
    view: DocumentView,
    stamp: FrameStamp,
    scene: BuiltScene,
    frame_builder: FrameBuilder,
    hit_tester: Option<Arc<HitTester>>,
    shared_hit_tester: Arc<SharedHitTester>,
    dynamic_properties: SceneProperties,
    frame_is_valid: bool,
    hit_tester_is_valid: bool,
    rendered_frame_is_valid: bool,
    has_built_scene: bool,
    data_stores: DataStores,
    scratch: PrimitiveScratchBuffer,
    render_task_counters: RenderTaskGraphCounters,
    prev_composite_descriptor: CompositeDescriptor,
}

Fields
id: DocumentId

The id of this document

removed_pipelines: Vec<(PipelineId, DocumentId)>

Temporary list of removed pipelines received from the scene builder thread and forwarded to the renderer.

view: DocumentViewstamp: FrameStamp

The id and time of the current frame.

scene: BuiltScene

The latest built scene, usable to build frames. received from the scene builder thread.

frame_builder: FrameBuilder

The builder object that prodces frames, kept around to preserve some retained state.

hit_tester: Option<Arc<HitTester>>

A data structure to allow hit testing against rendered frames. This is updated every time we produce a fully rendered frame.

shared_hit_tester: Arc<SharedHitTester>

To avoid synchronous messaging we update a shared hit-tester that other threads can query.

dynamic_properties: SceneProperties

Properties that are resolved during frame building and can be changed at any time without requiring the scene to be re-built.

frame_is_valid: bool

Track whether the last built frame is up to date or if it will need to be re-built before rendering again.

hit_tester_is_valid: boolrendered_frame_is_valid: boolhas_built_scene: bool

We track this information to be able to display debugging information from the renderer.

data_stores: DataStoresscratch: PrimitiveScratchBuffer

Contains various vecs of data that is used only during frame building, where we want to recycle the memory each new display list, to avoid constantly re-allocating and moving memory around.

render_task_counters: RenderTaskGraphCounters

Keep track of the size of render task graph to pre-allocate memory up-front the next frame.

prev_composite_descriptor: CompositeDescriptor

Tracks the state of the picture cache tiles that were composited on the previous frame.

Implementations

impl Document

pub fn new(
id: DocumentId,
size: DeviceIntSize,
layer: DocumentLayer,
default_device_pixel_ratio: f32
) -> Self

fn can_render(&self) -> bool

fn has_pixels(&self) -> bool

fn process_frame_msg(&mut self, message: FrameMsg) -> DocumentOps

fn build_frame(
&mut self,
resource_cache: &mut ResourceCache,
gpu_cache: &mut GpuCache,
resource_profile: &mut ResourceProfileCounters,
debug_flags: DebugFlags,
tile_cache_logger: &mut TileCacheLogger
) -> RenderedDocument

fn rebuild_hit_tester(&mut self)

pub fn updated_pipeline_info(&mut self) -> PipelineInfo

pub fn scroll_nearest_scrolling_ancestor(
&mut self,
scroll_location: ScrollLocation,
scroll_node_index: Option<SpatialNodeIndex>
) -> bool

Returns true if any nodes actually changed position or false otherwise.

pub fn scroll_node(
&mut self,
origin: LayoutPoint,
id: ExternalScrollId,
clamp: ScrollClamping
) -> bool

Returns true if the node actually changed position or false otherwise.

pub fn new_async_scene_ready(
&mut self,
built_scene: BuiltScene,
recycler: &mut Recycler
)