Struct webrender::render_backend::RenderBackend

pub struct RenderBackend {
    api_rx: MsgReceiver<ApiMsg>,
    payload_rx: PayloadReceiver,
    payload_tx: PayloadSender,
    result_tx: Sender<ResultMsg>,
    device_pixel_ratio: f32,
    next_namespace_id: IdNamespace,
    resource_cache: ResourceCache,
    scene: Scene,
    frame: Frame,
    notifier: Arc<Mutex<Option<Box<RenderNotifier>>>>,
    webrender_context_handle: Option<GLContextHandleWrapper>,
    webgl_contexts: HashMap<WebGLContextId, GLContextWrapper>,
    current_bound_webgl_context_id: Option<WebGLContextId>,
    recorder: Option<Box<ApiRecordingReceiver>>,
    main_thread_dispatcher: Arc<Mutex<Option<Box<RenderDispatcher>>>>,
    next_webgl_id: usize,
    vr_compositor_handler: Arc<Mutex<Option<Box<VRCompositorHandler>>>>,
}

The render backend is responsible for transforming high level display lists into GPU-friendly work which is then submitted to the renderer in the form of a frame::Frame.

The render backend operates on its own thread.

Fields
api_rx: MsgReceiver<ApiMsg> payload_rx: PayloadReceiver payload_tx: PayloadSender result_tx: Sender<ResultMsg> device_pixel_ratio: f32 next_namespace_id: IdNamespace resource_cache: ResourceCache scene: Scene frame: Frame notifier: Arc<Mutex<Option<Box<RenderNotifier>>>> webrender_context_handle: Option<GLContextHandleWrapper> webgl_contexts: HashMap<WebGLContextId, GLContextWrapper> current_bound_webgl_context_id: Option<WebGLContextId> recorder: Option<Box<ApiRecordingReceiver>> main_thread_dispatcher: Arc<Mutex<Option<Box<RenderDispatcher>>>> next_webgl_id: usize vr_compositor_handler: Arc<Mutex<Option<Box<VRCompositorHandler>>>>
Methods

impl RenderBackend

fn new(api_rx: MsgReceiver<ApiMsg>,
payload_rx: PayloadReceiver,
payload_tx: PayloadSender,
result_tx: Sender<ResultMsg>,
device_pixel_ratio: f32,
texture_cache: TextureCache,
enable_aa: bool,
workers: Arc<Mutex<ThreadPool>>,
notifier: Arc<Mutex<Option<Box<RenderNotifier>>>>,
webrender_context_handle: Option<GLContextHandleWrapper>,
config: FrameBuilderConfig,
recorder: Option<Box<ApiRecordingReceiver>>,
main_thread_dispatcher: Arc<Mutex<Option<Box<RenderDispatcher>>>>,
vr_compositor_handler: Arc<Mutex<Option<Box<VRCompositorHandler>>>>)
-> RenderBackend

fn run(&mut self)

fn discard_frame_state_for_pipeline(&mut self, pipeline_id: PipelineId)

fn build_scene(&mut self)

fn render(&mut self) -> RendererFrame

fn publish_frame(&mut self,
frame: RendererFrame,
profile_counters: &mut BackendProfileCounters)

fn publish_frame_and_notify_compositor(&mut self,
frame: RendererFrame,
profile_counters: &mut BackendProfileCounters)

fn notify_compositor_of_new_scroll_frame(&mut self, composite_needed: bool)

fn handle_vr_compositor_command(&mut self,
ctx_id: WebGLContextId,
cmd: VRCompositorCommand)