Struct webrender::render_backend::RenderBackend

pub struct RenderBackend {
    api_rx: MsgReceiver<ApiMsg>,
    payload_rx: Receiver<Payload>,
    result_tx: Sender<ResultMsg>,
    scene_tx: Sender<SceneBuilderRequest>,
    low_priority_scene_tx: Sender<SceneBuilderRequest>,
    scene_rx: Receiver<SceneBuilderResult>,
    payload_buffer: Vec<Payload>,
    default_device_pixel_ratio: f32,
    gpu_cache: GpuCache,
    resource_cache: ResourceCache,
    frame_config: FrameBuilderConfig,
    documents: HashMap<DocumentId, Document, BuildHasherDefault<FxHasher>>,
    notifier: Box<dyn RenderNotifier>,
    recorder: Option<Box<dyn ApiRecordingReceiver>>,
    logrecorder: Option<Box<LogRecorder>>,
    sampler: Option<Box<dyn AsyncPropertySampler + Send>>,
    size_of_ops: Option<MallocSizeOfOps>,
    debug_flags: DebugFlags,
    namespace_alloc_by_client: bool,
    recycler: Recycler,
}

The render backend is responsible for transforming high level display lists into GPU-friendly work which is then submitted to the renderer in the form of a frame::Frame.

The render backend operates on its own thread.

Fields
api_rx: MsgReceiver<ApiMsg>payload_rx: Receiver<Payload>result_tx: Sender<ResultMsg>scene_tx: Sender<SceneBuilderRequest>low_priority_scene_tx: Sender<SceneBuilderRequest>scene_rx: Receiver<SceneBuilderResult>payload_buffer: Vec<Payload>default_device_pixel_ratio: f32gpu_cache: GpuCacheresource_cache: ResourceCacheframe_config: FrameBuilderConfigdocuments: HashMap<DocumentId, Document, BuildHasherDefault<FxHasher>>notifier: Box<dyn RenderNotifier>recorder: Option<Box<dyn ApiRecordingReceiver>>logrecorder: Option<Box<LogRecorder>>sampler: Option<Box<dyn AsyncPropertySampler + Send>>size_of_ops: Option<MallocSizeOfOps>debug_flags: DebugFlagsnamespace_alloc_by_client: boolrecycler: Recycler
Methods

impl RenderBackend

pub fn new(
api_rx: MsgReceiver<ApiMsg>,
payload_rx: Receiver<Payload>,
result_tx: Sender<ResultMsg>,
scene_tx: Sender<SceneBuilderRequest>,
low_priority_scene_tx: Sender<SceneBuilderRequest>,
scene_rx: Receiver<SceneBuilderResult>,
default_device_pixel_ratio: f32,
resource_cache: ResourceCache,
notifier: Box<dyn RenderNotifier>,
frame_config: FrameBuilderConfig,
recorder: Option<Box<dyn ApiRecordingReceiver>>,
sampler: Option<Box<dyn AsyncPropertySampler + Send>>,
size_of_ops: Option<MallocSizeOfOps>,
debug_flags: DebugFlags,
namespace_alloc_by_client: bool
) -> RenderBackend

fn process_scene_msg(
&mut self,
document_id: DocumentId,
message: SceneMsg,
frame_counter: u32,
txn: &mut Transaction,
ipc_profile_counters: &mut IpcProfileCounters
)

fn next_namespace_id(&self) -> IdNamespace

pub fn run(&mut self, profile_counters: BackendProfileCounters)

fn process_api_msg(
&mut self,
msg: ApiMsg,
profile_counters: &mut BackendProfileCounters,
frame_counter: &mut u32
) -> RenderBackendStatus

fn prepare_for_frames(&mut self)

fn bookkeep_after_frames(&mut self)

fn requires_frame_build(&mut self) -> bool

fn prepare_transactions(
&mut self,
document_ids: Vec<DocumentId>,
transaction_msgs: Vec<TransactionMsg>,
frame_counter: &mut u32,
profile_counters: &mut BackendProfileCounters
)

fn maybe_force_nop_documents<F>(
&mut self,
frame_counter: &mut u32,
profile_counters: &mut BackendProfileCounters,
document_already_present: F
) where
F: Fn(DocumentId) -> bool,

In certain cases, resources shared by multiple documents have to run maintenance operations, like cleaning up unused cache items. In those cases, we are forced to build frames for all documents, however we may not have a transaction ready for every document - this method calls update_document with the details of a fake, nop transaction just to force a frame build.

fn update_document(
&mut self,
document_id: DocumentId,
resource_updates: Vec<ResourceUpdate>,
interner_updates: Option<InternerUpdates>,
frame_ops: Vec<FrameMsg>,
notifications: Vec<NotificationRequest>,
render_frame: bool,
invalidate_rendered_frame: bool,
frame_counter: &mut u32,
profile_counters: &mut BackendProfileCounters,
has_built_scene: bool
)

fn get_docs_for_debugger(&self) -> String

fn get_clip_scroll_tree_for_debugger(&self) -> String

fn report_memory(&mut self, tx: MsgSender<MemoryReport>)

impl RenderBackend

fn save_capture(
&mut self,
root: PathBuf,
bits: CaptureBits,
profile_counters: &mut BackendProfileCounters
) -> DebugOutput