Struct webrender::render_backend::RenderBackend

pub struct RenderBackend {
    api_rx: Receiver<ApiMsg>,
    result_tx: Sender<ResultMsg>,
    scene_tx: Sender<SceneBuilderRequest>,
    low_priority_scene_tx: Sender<SceneBuilderRequest>,
    backend_scene_tx: Sender<BackendSceneBuilderRequest>,
    scene_rx: Receiver<SceneBuilderResult>,
    default_device_pixel_ratio: f32,
    gpu_cache: GpuCache,
    resource_cache: ResourceCache,
    frame_config: FrameBuilderConfig,
    default_compositor_kind: CompositorKind,
    documents: FastHashMap<DocumentId, Document>,
    notifier: Box<dyn RenderNotifier>,
    tile_cache_logger: TileCacheLogger,
    sampler: Option<Box<dyn AsyncPropertySampler + Send>>,
    size_of_ops: Option<MallocSizeOfOps>,
    debug_flags: DebugFlags,
    namespace_alloc_by_client: bool,
    blob_image_handler: Option<Box<dyn BlobImageHandler>>,
    recycler: Recycler,
    capture_config: Option<CaptureConfig>,
}

The render backend is responsible for transforming high level display lists into GPU-friendly work which is then submitted to the renderer in the form of a frame::Frame.

The render backend operates on its own thread.

Fields
api_rx: Receiver<ApiMsg>result_tx: Sender<ResultMsg>scene_tx: Sender<SceneBuilderRequest>low_priority_scene_tx: Sender<SceneBuilderRequest>backend_scene_tx: Sender<BackendSceneBuilderRequest>scene_rx: Receiver<SceneBuilderResult>default_device_pixel_ratio: f32gpu_cache: GpuCacheresource_cache: ResourceCacheframe_config: FrameBuilderConfigdefault_compositor_kind: CompositorKinddocuments: FastHashMap<DocumentId, Document>notifier: Box<dyn RenderNotifier>tile_cache_logger: TileCacheLoggersampler: Option<Box<dyn AsyncPropertySampler + Send>>size_of_ops: Option<MallocSizeOfOps>debug_flags: DebugFlagsnamespace_alloc_by_client: boolblob_image_handler: Option<Box<dyn BlobImageHandler>>recycler: Recyclercapture_config: Option<CaptureConfig>
Implementations

impl RenderBackend

pub fn new(
api_rx: Receiver<ApiMsg>,
result_tx: Sender<ResultMsg>,
scene_tx: Sender<SceneBuilderRequest>,
low_priority_scene_tx: Sender<SceneBuilderRequest>,
backend_scene_tx: Sender<BackendSceneBuilderRequest>,
scene_rx: Receiver<SceneBuilderResult>,
default_device_pixel_ratio: f32,
resource_cache: ResourceCache,
notifier: Box<dyn RenderNotifier>,
blob_image_handler: Option<Box<dyn BlobImageHandler>>,
frame_config: FrameBuilderConfig,
sampler: Option<Box<dyn AsyncPropertySampler + Send>>,
size_of_ops: Option<MallocSizeOfOps>,
debug_flags: DebugFlags,
namespace_alloc_by_client: bool
) -> RenderBackend

fn next_namespace_id(&self) -> IdNamespace

pub fn run(&mut self, profile_counters: BackendProfileCounters)

fn process_transaction(
&mut self,
txns: Vec<Box<BuiltTransaction>>,
result_tx: Option<Sender<SceneSwapResult>>,
frame_counter: &mut u32,
profile_counters: &mut BackendProfileCounters
) -> bool

fn process_api_msg(
&mut self,
msg: ApiMsg,
profile_counters: &mut BackendProfileCounters,
frame_counter: &mut u32
) -> RenderBackendStatus

fn update_frame_builder_config(&self)

fn prepare_for_frames(&mut self)

fn bookkeep_after_frames(&mut self)

fn requires_frame_build(&mut self) -> bool

fn prepare_transactions(
&mut self,
txns: Vec<Box<TransactionMsg>>,
frame_counter: &mut u32,
profile_counters: &mut BackendProfileCounters
)

fn maybe_force_nop_documents<F>(
&mut self,
frame_counter: &mut u32,
profile_counters: &mut BackendProfileCounters,
document_already_present: F
) where
F: Fn(DocumentId) -> bool,

In certain cases, resources shared by multiple documents have to run maintenance operations, like cleaning up unused cache items. In those cases, we are forced to build frames for all documents, however we may not have a transaction ready for every document - this method calls update_document with the details of a fake, nop transaction just to force a frame build.

fn update_document(
&mut self,
document_id: DocumentId,
resource_updates: Vec<ResourceUpdate>,
frame_ops: Vec<FrameMsg>,
notifications: Vec<NotificationRequest>,
render_frame: bool,
invalidate_rendered_frame: bool,
frame_counter: &mut u32,
profile_counters: &mut BackendProfileCounters,
has_built_scene: bool
) -> bool

fn send_backend_message(&self, msg: BackendSceneBuilderRequest)

fn get_spatial_tree_for_debugger(&self) -> String

fn report_memory(&mut self, tx: Sender<Box<MemoryReport>>)

fn save_capture_sequence(&mut self)

impl RenderBackend

fn save_capture(
&mut self,
root: PathBuf,
bits: CaptureBits,
profile_counters: &mut BackendProfileCounters
) -> DebugOutput

fn start_capture_sequence(&mut self, root: PathBuf, bits: CaptureBits)

fn stop_capture_sequence(&mut self)