webrender

Module render_target

Structs

BlitJob

LineDecorationJob

PictureCacheTarget

RenderTarget

Contains the work (in the form of instance arrays) needed to fill a color color (RGBA8) or alpha output surface.

RenderTargetContext

RenderTargetList

A series of RenderTarget instances, serving as the high-level container into which RenderTasks are assigned.

ResolveOp

Enums

PictureCacheTargetKind

RenderTargetKind

A tag used to identify the output format of a RenderTarget.

Constants

NUM_PATTERNS ğŸ”’

STYLE_MASK ğŸ”’

STYLE_SOLID ğŸ”’

Functions

add_blur_instances ğŸ”’

add_scaling_instances ğŸ”’

add_svg_filter_instances ğŸ”’

add_svg_filter_node_instances ğŸ”’

Generates SVGFEFilterInstances from a single SVGFEFilterTask, this is what prepares vertex data for the shader, and adds it to the appropriate batch.

build_mask_tasks ğŸ”’

build_sub_pass ğŸ”’