Module webrender::render_task

Structs

BlitTask

BlurTask

BorderTask

CacheMaskTask

ClipRegionTask

GradientTask

LineDecorationTask

PictureTask

RenderTask

RenderTaskAddress

RenderTaskData

ScalingTask

SvgFilterTask

Enums

BlitSource

BlurTaskKey
Since we only use it within a single primitive, the key only needs to contain the down-scaling level and the blur std deviation.

ClearMode

RenderTaskKind

RenderTaskLocation
Identifies the output buffer location for a given RenderTask.

SvgFilterInfo

Constants

FLOATS_PER_RENDER_TASK_INFO

MAX_BLUR_STD_DEVIATION

MIN_DOWNSCALING_RT_SIZE

RENDER_TASK_SIZE_SANITY_CHECK

Functions

render_task_sanity_check

Type Definitions

BlurTaskCache
In order to avoid duplicating the down-scaling and blur passes when a picture has several blurs, we use a local (primitive-level) cache of the render tasks generated for a single shadowed primitive in a single frame.

TaskDependencies