Module webrender::render_task_graph

This module contains the render task graph.

Code associated with creating specific render tasks is in the render_task module.

Structs

RenderPass
A render pass represents a set of rendering operations that don't depend on one another.

RenderTaskAllocation
Allows initializing a render task directly into the render task buffer.

RenderTaskGraph

RenderTaskGraphCounters

RenderTaskId

Enums

RenderPassKind
Contains the set of RenderTargets specific to the kind of pass.

Functions

dump_render_tasks_as_svg

dump_task_dependency_link