Struct webrender::render_task_graph::RenderTaskGraph

pub struct RenderTaskGraph {
    pub tasks: Vec<RenderTask>,
    pub task_data: Vec<RenderTaskData>,
    pub cacheable_render_tasks: Vec<RenderTaskId>,
    next_saved: SavedTargetIndex,
    frame_id: FrameId,
}

Fields
tasks: Vec<RenderTask>task_data: Vec<RenderTaskData>cacheable_render_tasks: Vec<RenderTaskId>

Tasks that don't have dependencies, and that may be shared between picture tasks.

We render these unconditionally before-rendering the rest of the tree.

next_saved: SavedTargetIndexframe_id: FrameId
Implementations

impl RenderTaskGraph

fn resolve_surface(
&self,
task_id: RenderTaskId,
gpu_cache: &GpuCache
) -> (GpuCacheAddress, BatchTextures)

impl RenderTaskGraph

pub fn new(frame_id: FrameId, counters: &RenderTaskGraphCounters) -> Self

pub fn counters(&self) -> RenderTaskGraphCounters

pub fn add(&mut self) -> RenderTaskAllocation

pub fn add_dependency(
&mut self,
parent_id: RenderTaskId,
child_id: RenderTaskId
)

Express a render task dependency between a parent and child task. This is used to assign tasks to render passes.

pub fn generate_passes(
&mut self,
main_render_task: Option<RenderTaskId>,
screen_size: DeviceIntSize,
gpu_supports_fast_clears: bool
) -> Vec<RenderPass>

Assign this frame's render tasks to render passes ordered so that passes appear earlier than the ones that depend on them.

fn generate_passes_impl(
&self,
root_tasks: &[RenderTaskId],
screen_size: DeviceIntSize,
gpu_supports_fast_clears: bool,
for_main_framebuffer: bool,
passes: &mut Vec<RenderPass>
)

Assign the render tasks from the tree rooted at root_task to render passes and append them to the passes vector so that the passes that we depend on end up earlier in the pass list.

fn resolve_target_conflicts(&mut self, passes: &mut [RenderPass])

Resolve conflicts between the generated passes and the limitiations of our target allocation scheme.

The render task graph operates with a ping-pong target allocation scheme where a set of targets is written to by even passes and a different set of targets is written to by odd passes. Since tasks cannot read and write the same target, we can run into issues if a task pass in N + 2 reads the result of a task in pass N. To avoid such cases have to insert blit tasks to copy the content of the task into pass N + 1 which is readable by pass N + 2.

In addition, allocated rects of pass N are currently not tracked and can be overwritten by allocations in later passes on the same target, unless the task has been marked for saving, which perserves the allocated rect until the end of the frame. This is a big hammer, hopefully we won't need to mark many passes for saving. A better solution would be to track allocations through the entire graph, there is a prototype of that in https://github.com/nical/toy-render-graph/

pub fn get_task_address(&self, id: RenderTaskId) -> RenderTaskAddress

pub fn write_task_data(&mut self)

pub fn save_target(&mut self) -> SavedTargetIndex

pub fn frame_id(&self) -> FrameId