Struct webrender::render_task_graph::RenderTaskGraphBuilder

pub struct RenderTaskGraphBuilder {
    tasks: Vec<RenderTask>,
    roots: HashSet<RenderTaskId, BuildHasherDefault<FxHasher>>,
    frame_id: FrameId,
    textures_to_free: HashSet<CacheTextureId, BuildHasherDefault<FxHasher>>,
    active_surfaces: FastHashMap<CacheTextureId, Surface>,
}

The persistent interface that is used during frame building to construct the frame graph.

Fields
tasks: Vec<RenderTask>

List of tasks added to the builder

roots: HashSet<RenderTaskId, BuildHasherDefault<FxHasher>>

List of task roots

frame_id: FrameId

Current frame id, used for debug validation

textures_to_free: HashSet<CacheTextureId, BuildHasherDefault<FxHasher>>

A list of texture surfaces that can be freed at the end of a pass. Retained here to reduce heap allocations.

active_surfaces: FastHashMap<CacheTextureId, Surface>
Implementations

impl RenderTaskGraphBuilder

pub fn new() -> Self

Construct a new graph builder. Typically constructed once and maintained over many frames, to avoid extra heap allocations where possible.

pub fn frame_id(&self) -> FrameId

pub fn begin_frame(&mut self, frame_id: FrameId)

Begin a new frame

pub fn get_task(&self, task_id: RenderTaskId) -> &RenderTask

Get immutable access to a task

pub fn get_task_mut(&mut self, task_id: RenderTaskId) -> &mut RenderTask

Get mutable access to a task

pub fn add(&mut self) -> RenderTaskAllocation<'_>

Add a new task to the graph.

pub fn add_dependency(&mut self, task_id: RenderTaskId, input: RenderTaskId)

Express a dependency, such that task_id depends on input as a texture source.

pub fn end_frame( &mut self, resource_cache: &mut ResourceCache, gpu_cache: &mut GpuCache, deferred_resolves: &mut Vec<DeferredResolve>, max_shared_surface_size: i32 ) -> RenderTaskGraph

End the graph building phase and produce the immutable task graph for this frame