Module webrender::renderer

The webrender API.

The webrender::renderer module provides the interface to webrender, which is accessible through Renderer

Structs

ExternalImage
The data that an external client should provide about an external image. The timestamp is used to test if the renderer should upload new texture data this frame. For instance, if providing video frames, the application could call wr.render() whenever a new video frame is ready. If the callback increments the returned timestamp for a given image, the renderer will know to re-upload the image data to the GPU. Note that the UV coords are supplied in texel-space!

Renderer
The renderer is responsible for submitting to the GPU the work prepared by the RenderBackend.

RendererOptions

Enums

BlendMode

ExternalImageSource

Constants

MAX_VERTEX_TEXTURE_WIDTH

VERTEX_TEXTURE_POOL

Traits

ExternalImageHandler
The interfaces that an application can implement to support providing external image buffers. When the the application passes an external image to WR, it should kepp that external image life time untile the release() call.