Module webrender::renderer::init

Structs

DebugFlags

Flags to enable/disable various builtin debugging tools.

WebRenderOptions

Constants

ONE_TIME_USAGE_HINT

Use this hint for all vertex data re-initialization. This allows the driver to better re-use RBOs internally.

Statics

HAS_BEEN_INITIALIZED 🔒

Is only false if no WR instances have ever been created.

Traits

AsyncPropertySampler

Allows callers to hook into the main render_backend loop and provide additional frame ops for generate_frame transactions. These functions are all called from the render backend thread.

SceneBuilderHooks

Allows callers to hook in at certain points of the async scene build. These functions are all called from the scene builder thread.

Functions

create_webrender_instance

Initializes WebRender and creates a Renderer and RenderApiSender.

wr_has_been_initialized

Returns true if a WR instance has ever been initialized in this process.