Struct webrender::renderer::Renderer

pub struct Renderer {
    result_rx: Receiver<ResultMsg>,
    device: Device,
    pending_texture_updates: Vec<TextureUpdateList>,
    pending_shader_updates: Vec<PathBuf>,
    current_frame: Option<RendererFrame>,
    device_pixel_ratio: f32,
    cs_box_shadow: LazilyCompiledShader,
    cs_clip_clear: LazilyCompiledShader,
    cs_clip_rectangle: LazilyCompiledShader,
    cs_clip_image: LazilyCompiledShader,
    cs_text_run: LazilyCompiledShader,
    cs_blur: LazilyCompiledShader,
    ps_rectangle: PrimitiveShader,
    ps_rectangle_clip: PrimitiveShader,
    ps_text_run: PrimitiveShader,
    ps_text_run_subpixel: PrimitiveShader,
    ps_image: PrimitiveShader,
    ps_yuv_image: PrimitiveShader,
    ps_border: PrimitiveShader,
    ps_gradient: PrimitiveShader,
    ps_angle_gradient: PrimitiveShader,
    ps_box_shadow: PrimitiveShader,
    ps_cache_image: PrimitiveShader,
    ps_blend: LazilyCompiledShader,
    ps_composite: LazilyCompiledShader,
    tile_clear_shader: LazilyCompiledShader,
    max_clear_tiles: usize,
    max_prim_blends: usize,
    max_prim_composites: usize,
    max_cache_instances: usize,
    max_blurs: usize,
    notifier: Arc<Mutex<Option<Box<RenderNotifier>>>>,
    flush_notifier: Arc<Mutex<Option<Box<FlushNotifier>>>>,
    enable_profiler: bool,
    debug: DebugRenderer,
    backend_profile_counters: BackendProfileCounters,
    profile_counters: RendererProfileCounters,
    profiler: Profiler,
    last_time: u64,
    render_targets: Vec<TextureId>,
    gpu_profile: GpuProfiler<GpuProfileTag>,
    quad_vao_id: VAOId,
    layer_texture: VertexDataTexture,
    render_task_texture: VertexDataTexture,
    prim_geom_texture: VertexDataTexture,
    data16_texture: VertexDataTexture,
    data32_texture: VertexDataTexture,
    data64_texture: VertexDataTexture,
    data128_texture: VertexDataTexture,
    resource_rects_texture: VertexDataTexture,
    pipeline_epoch_map: HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>>,
    main_thread_dispatcher: Arc<Mutex<Option<Box<RenderDispatcher>>>>,
    cache_texture_id_map: Vec<Option<TextureId>>,
    external_image_handler: Option<Box<ExternalImageHandler>>,
    external_images: HashMap<ExternalImageId, TextureId, BuildHasherDefault<FnvHasher>>,
}

The renderer is responsible for submitting to the GPU the work prepared by the RenderBackend.

Fields
result_rx: Receiver<ResultMsg> device: Device pending_texture_updates: Vec<TextureUpdateList> pending_shader_updates: Vec<PathBuf> current_frame: Option<RendererFrame> device_pixel_ratio: f32 cs_box_shadow: LazilyCompiledShader cs_clip_clear: LazilyCompiledShader cs_clip_rectangle: LazilyCompiledShader cs_clip_image: LazilyCompiledShader cs_text_run: LazilyCompiledShader cs_blur: LazilyCompiledShader ps_rectangle: PrimitiveShader ps_rectangle_clip: PrimitiveShader ps_text_run: PrimitiveShader ps_text_run_subpixel: PrimitiveShader ps_image: PrimitiveShader ps_yuv_image: PrimitiveShader ps_border: PrimitiveShader ps_gradient: PrimitiveShader ps_angle_gradient: PrimitiveShader ps_box_shadow: PrimitiveShader ps_cache_image: PrimitiveShader ps_blend: LazilyCompiledShader ps_composite: LazilyCompiledShader tile_clear_shader: LazilyCompiledShader max_clear_tiles: usize max_prim_blends: usize max_prim_composites: usize max_cache_instances: usize max_blurs: usize notifier: Arc<Mutex<Option<Box<RenderNotifier>>>> flush_notifier: Arc<Mutex<Option<Box<FlushNotifier>>>> enable_profiler: bool debug: DebugRenderer backend_profile_counters: BackendProfileCounters profile_counters: RendererProfileCounters profiler: Profiler last_time: u64 render_targets: Vec<TextureId> gpu_profile: GpuProfiler<GpuProfileTag> quad_vao_id: VAOId layer_texture: VertexDataTexture render_task_texture: VertexDataTexture prim_geom_texture: VertexDataTexture data16_texture: VertexDataTexture data32_texture: VertexDataTexture data64_texture: VertexDataTexture data128_texture: VertexDataTexture resource_rects_texture: VertexDataTexture pipeline_epoch_map: HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>> main_thread_dispatcher: Arc<Mutex<Option<Box<RenderDispatcher>>>>

Used to dispatch functions to the main thread's event loop. Required to allow GLContext sharing in some implementations like WGL.

cache_texture_id_map: Vec<Option<TextureId>>

A vector for fast resolves of texture cache IDs to native texture IDs. This maps to a free-list managed by the backend thread / texture cache. Because of this, items in this array may be None if they have been freed by the backend thread. This saves having to use a hashmap, and allows a flat vector for performance.

external_image_handler: Option<Box<ExternalImageHandler>>

Optional trait object that allows the client application to provide external buffers for image data.

external_images: HashMap<ExternalImageId, TextureId, BuildHasherDefault<FnvHasher>>

Map of external image IDs to native textures.

Methods

impl Renderer

fn new(options: RendererOptions) -> (Renderer, RenderApiSender)

Initializes webrender and creates a Renderer and RenderApiSender.

Examples

Initializes a Renderer with some reasonable values. For more information see RendererOptions.

let opts = webrender::RendererOptions {
   device_pixel_ratio: 1.0,
   resource_override_path: None,
   enable_aa: false,
   enable_msaa: false,
   enable_profiler: false,
};
let (renderer, sender) = Renderer::new(opts);

fn set_render_notifier(&self, notifier: Box<RenderNotifier>)

Sets the new RenderNotifier.

The RenderNotifier will be called when processing e.g. of a (scrolling) frame is done, and therefore the screen should be updated.

fn set_flush_notifier(&self, flush_notifier: Box<FlushNotifier>)

Sets the new FlushNotifier

The FlushNotifier is called after all messages to the backend thread have been processed. This should be used when we need to sync wait for rendering to happen, such as with reftests.

fn set_main_thread_dispatcher(&self, dispatcher: Box<RenderDispatcher>)

Sets the new MainThreadDispatcher.

Allows to dispatch functions to the main thread's event loop.

fn current_epoch(&self, pipeline_id: PipelineId) -> Option<Epoch>

Returns the Epoch of the current frame in a pipeline.

fn update(&mut self)

Processes the result queue.

Should be called before render(), as texture cache updates are done here.

fn resolve_source_texture(&mut self, texture_id: &SourceTexture) -> TextureId

fn set_external_image_handler(&mut self, handler: Box<ExternalImageHandler>)

Set a callback for handling external images.

fn render(&mut self, framebuffer_size: Size2D<u32>)

Renders the current frame.

A Frame is supplied by calling set_root_stacking_context().

fn layers_are_bouncing_back(&self) -> bool

fn update_texture_cache(&mut self)

fn add_debug_rect(&mut self,
p0: DeviceIntPoint,
p1: DeviceIntPoint,
label: &str,
c: &ColorF)

fn draw_ubo_batch<T>(&mut self,
ubo_data: &[T],
shader: ProgramId,
quads_per_item: usize,
textures: &BatchTextures,
max_prim_items: usize,
projection: &Matrix4D<f32>)

fn draw_target(&mut self,
render_target: Option<(TextureId, i32)>,
target: &RenderTarget,
target_size: &DeviceSize,
cache_texture: Option<TextureId>,
should_clear: bool)

fn update_deferred_resolves(&mut self, frame: &mut Frame)

fn release_external_textures(&mut self)

fn draw_tile_frame(&mut self,
frame: &mut Frame,
framebuffer_size: &DeviceUintSize)

fn debug_renderer<'a>(&'a mut self) -> &'a mut DebugRenderer

fn get_profiler_enabled(&mut self) -> bool

fn set_profiler_enabled(&mut self, enabled: bool)