Struct webrender::renderer::Renderer

pub struct Renderer {
    result_rx: Receiver<ResultMsg>,
    device: Device,
    pending_texture_updates: Vec<TextureUpdateList>,
    pending_shader_updates: Vec<PathBuf>,
    current_frame: Option<RendererFrame>,
    device_pixel_ratio: f32,
    raster_batches: Vec<RasterBatch>,
    raster_op_vao: Option<VAOId>,
    blur_program_id: ProgramId,
    u_direction: UniformLocation,
    cs_box_shadow: LazilyCompiledShader,
    ps_rectangle: PrimitiveShader,
    ps_text_run: PrimitiveShader,
    ps_image: PrimitiveShader,
    ps_border: PrimitiveShader,
    ps_gradient: PrimitiveShader,
    ps_gradient_clip: PrimitiveShader,
    ps_angle_gradient: PrimitiveShader,
    ps_box_shadow: PrimitiveShader,
    ps_rectangle_clip: PrimitiveShader,
    ps_image_clip: PrimitiveShader,
    ps_blend: LazilyCompiledShader,
    ps_composite: LazilyCompiledShader,
    tile_clear_shader: LazilyCompiledShader,
    max_clear_tiles: usize,
    max_prim_blends: usize,
    max_prim_composites: usize,
    max_cache_instances: usize,
    notifier: Arc<Mutex<Option<Box<RenderNotifier>>>>,
    enable_profiler: bool,
    debug: DebugRenderer,
    backend_profile_counters: BackendProfileCounters,
    profile_counters: RendererProfileCounters,
    profiler: Profiler,
    last_time: u64,
    max_raster_op_size: u32,
    raster_op_target_a8: TextureId,
    raster_op_target_rgba8: TextureId,
    render_targets: Vec<TextureId>,
    gpu_profile: GpuProfiler<GpuProfileTag>,
    quad_vao_id: VAOId,
    layer_texture: VertexDataTexture,
    render_task_texture: VertexDataTexture,
    prim_geom_texture: VertexDataTexture,
    data16_texture: VertexDataTexture,
    data32_texture: VertexDataTexture,
    data64_texture: VertexDataTexture,
    data128_texture: VertexDataTexture,
    pipeline_epoch_map: HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>>,
    main_thread_dispatcher: Arc<Mutex<Option<Box<RenderDispatcher>>>>,
}

The renderer is responsible for submitting to the GPU the work prepared by the RenderBackend.

Fields
result_rx: Receiver<ResultMsg> device: Device pending_texture_updates: Vec<TextureUpdateList> pending_shader_updates: Vec<PathBuf> current_frame: Option<RendererFrame> device_pixel_ratio: f32 raster_batches: Vec<RasterBatch> raster_op_vao: Option<VAOId> blur_program_id: ProgramId u_direction: UniformLocation cs_box_shadow: LazilyCompiledShader ps_rectangle: PrimitiveShader ps_text_run: PrimitiveShader ps_image: PrimitiveShader ps_border: PrimitiveShader ps_gradient: PrimitiveShader ps_gradient_clip: PrimitiveShader ps_angle_gradient: PrimitiveShader ps_box_shadow: PrimitiveShader ps_rectangle_clip: PrimitiveShader ps_image_clip: PrimitiveShader ps_blend: LazilyCompiledShader ps_composite: LazilyCompiledShader tile_clear_shader: LazilyCompiledShader max_clear_tiles: usize max_prim_blends: usize max_prim_composites: usize max_cache_instances: usize notifier: Arc<Mutex<Option<Box<RenderNotifier>>>> enable_profiler: bool debug: DebugRenderer backend_profile_counters: BackendProfileCounters profile_counters: RendererProfileCounters profiler: Profiler last_time: u64 max_raster_op_size: u32 raster_op_target_a8: TextureId raster_op_target_rgba8: TextureId render_targets: Vec<TextureId> gpu_profile: GpuProfiler<GpuProfileTag> quad_vao_id: VAOId layer_texture: VertexDataTexture render_task_texture: VertexDataTexture prim_geom_texture: VertexDataTexture data16_texture: VertexDataTexture data32_texture: VertexDataTexture data64_texture: VertexDataTexture data128_texture: VertexDataTexture pipeline_epoch_map: HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>> main_thread_dispatcher: Arc<Mutex<Option<Box<RenderDispatcher>>>>

Used to dispatch functions to the main thread's event loop. Required to allow GLContext sharing in some implementations like WGL.

Methods

impl Renderer

fn new(options: RendererOptions) -> (Renderer, RenderApiSender)

Initializes webrender and creates a Renderer and RenderApiSender.

Examples

Initializes a Renderer with some reasonable values. For more information see RendererOptions.

let opts = webrender::RendererOptions {
   device_pixel_ratio: 1.0,
   resource_path: PathBuf::from("../webrender/res"),
   enable_aa: false,
   enable_msaa: false,
   enable_profiler: false,
};
let (renderer, sender) = Renderer::new(opts);

fn update_uniform_locations(&mut self)

fn set_render_notifier(&self, notifier: Box<RenderNotifier>)

Sets the new RenderNotifier.

The RenderNotifier will be called when processing e.g. of a (scrolling) frame is done, and therefore the screen should be updated.

fn set_main_thread_dispatcher(&self, dispatcher: Box<RenderDispatcher>)

Sets the new MainThreadDispatcher.

Allows to dispatch functions to the main thread's event loop.

fn current_epoch(&self, pipeline_id: PipelineId) -> Option<Epoch>

Returns the Epoch of the current frame in a pipeline.

fn update(&mut self)

Processes the result queue.

Should be called before render(), as texture cache updates are done here.

fn render(&mut self, framebuffer_size: Size2D<u32>)

Renders the current frame.

A Frame is supplied by calling set_root_stacking_context().

fn layers_are_bouncing_back(&self) -> bool

fn update_texture_cache(&mut self)

fn add_rect_to_raster_batch<F>(&mut self,
dest_texture_id: TextureId,
color_texture_id: TextureId,
program_id: ProgramId,
blur_direction: Option<AxisDirection>,
dest_rect: &Rect<u32>,
border_type: BorderType,
f: F) where F: Fn(&Rect<f32>) -> [PackedVertexForTextureCacheUpdate; 4]

fn flush_raster_batches(&mut self)

fn set_up_gl_state_for_texture_cache_update(&mut self,
target_texture_id: TextureId,
color_texture_id: TextureId,
program_id: ProgramId,
blur_direction: Option<AxisDirection>,
projection: &Matrix4D<f32>)

fn perform_gl_texture_cache_update(&mut self, batch: RasterBatch)

fn add_debug_rect(&mut self,
p0: TypedPoint2D<i32, DevicePixel>,
p1: TypedPoint2D<i32, DevicePixel>,
label: &str,
c: &ColorF)

fn draw_ubo_batch<T>(&mut self,
ubo_data: &[T],
shader: ProgramId,
quads_per_item: usize,
color_texture_id: TextureId,
mask_texture_id: TextureId,
max_prim_items: usize,
projection: &Matrix4D<f32>)

fn draw_target(&mut self,
render_target: Option<(TextureId, i32)>,
target: &RenderTarget,
target_size: &Size2D<f32>,
cache_texture: Option<TextureId>,
should_clear: bool)

fn draw_tile_frame(&mut self, frame: &mut Frame, framebuffer_size: &Size2D<u32>)

fn debug_renderer<'a>(&'a mut self) -> &'a mut DebugRenderer

fn get_profiler_enabled(&mut self) -> bool

fn set_profiler_enabled(&mut self, enabled: bool)