Struct webrender::renderer::Renderer

pub struct Renderer {
    result_rx: Receiver<ResultMsg>,
    device: Device,
    pending_texture_updates: Vec<TextureUpdateList>,
    pending_shader_updates: Vec<PathBuf>,
    current_frame: Option<RendererFrame>,
    cs_box_shadow: LazilyCompiledShader,
    cs_text_run: LazilyCompiledShader,
    cs_blur: LazilyCompiledShader,
    cs_clip_rectangle: LazilyCompiledShader,
    cs_clip_image: LazilyCompiledShader,
    ps_rectangle: PrimitiveShader,
    ps_rectangle_clip: PrimitiveShader,
    ps_text_run: PrimitiveShader,
    ps_text_run_subpixel: PrimitiveShader,
    ps_image: PrimitiveShader,
    ps_yuv_image: PrimitiveShader,
    ps_border: PrimitiveShader,
    ps_gradient: PrimitiveShader,
    ps_angle_gradient: PrimitiveShader,
    ps_radial_gradient: PrimitiveShader,
    ps_box_shadow: PrimitiveShader,
    ps_cache_image: PrimitiveShader,
    ps_blend: LazilyCompiledShader,
    ps_hw_composite: LazilyCompiledShader,
    ps_composite: LazilyCompiledShader,
    notifier: Arc<Mutex<Option<Box<RenderNotifier>>>>,
    enable_profiler: bool,
    clear_framebuffer: bool,
    clear_color: ColorF,
    debug: DebugRenderer,
    render_target_debug: bool,
    backend_profile_counters: BackendProfileCounters,
    profile_counters: RendererProfileCounters,
    profiler: Profiler,
    last_time: u64,
    render_targets: Vec<TextureId>,
    gpu_profile: GpuProfiler<GpuProfileTag>,
    prim_vao_id: VAOId,
    blur_vao_id: VAOId,
    clip_vao_id: VAOId,
    gdt_index: usize,
    gpu_data_textures: [GpuDataTextures; 5],
    pipeline_epoch_map: HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>>,
    main_thread_dispatcher: Arc<Mutex<Option<Box<RenderDispatcher>>>>,
    cache_texture_id_map: Vec<TextureId>,
    external_image_handler: Option<Box<ExternalImageHandler>>,
    external_images: HashMap<ExternalImageId, TextureId, BuildHasherDefault<FnvHasher>>,
    vr_compositor_handler: Arc<Mutex<Option<Box<VRCompositorHandler>>>>,
}

The renderer is responsible for submitting to the GPU the work prepared by the RenderBackend.

Fields
result_rx: Receiver<ResultMsg> device: Device pending_texture_updates: Vec<TextureUpdateList> pending_shader_updates: Vec<PathBuf> current_frame: Option<RendererFrame> cs_box_shadow: LazilyCompiledShader cs_text_run: LazilyCompiledShader cs_blur: LazilyCompiledShader cs_clip_rectangle: LazilyCompiledShader

These are "cache clip shaders". These shaders are used to draw clip instances into the cached clip mask. The results of these shaders are also used by the primitive shaders.

cs_clip_image: LazilyCompiledShader ps_rectangle: PrimitiveShader ps_rectangle_clip: PrimitiveShader ps_text_run: PrimitiveShader ps_text_run_subpixel: PrimitiveShader ps_image: PrimitiveShader ps_yuv_image: PrimitiveShader ps_border: PrimitiveShader ps_gradient: PrimitiveShader ps_angle_gradient: PrimitiveShader ps_radial_gradient: PrimitiveShader ps_box_shadow: PrimitiveShader ps_cache_image: PrimitiveShader ps_blend: LazilyCompiledShader ps_hw_composite: LazilyCompiledShader ps_composite: LazilyCompiledShader notifier: Arc<Mutex<Option<Box<RenderNotifier>>>> enable_profiler: bool clear_framebuffer: bool clear_color: ColorF debug: DebugRenderer render_target_debug: bool backend_profile_counters: BackendProfileCounters profile_counters: RendererProfileCounters profiler: Profiler last_time: u64 render_targets: Vec<TextureId> gpu_profile: GpuProfiler<GpuProfileTag> prim_vao_id: VAOId blur_vao_id: VAOId clip_vao_id: VAOId gdt_index: usize gpu_data_textures: [GpuDataTextures; 5] pipeline_epoch_map: HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>> main_thread_dispatcher: Arc<Mutex<Option<Box<RenderDispatcher>>>>

Used to dispatch functions to the main thread's event loop. Required to allow GLContext sharing in some implementations like WGL.

cache_texture_id_map: Vec<TextureId>

A vector for fast resolves of texture cache IDs to native texture IDs. This maps to a free-list managed by the backend thread / texture cache. We free the texture memory associated with a TextureId when its texture cache ID is freed by the texture cache, but reuse the TextureId when the texture caches's free list reuses the texture cache ID. This saves having to use a hashmap, and allows a flat vector for performance.

external_image_handler: Option<Box<ExternalImageHandler>>

Optional trait object that allows the client application to provide external buffers for image data.

external_images: HashMap<ExternalImageId, TextureId, BuildHasherDefault<FnvHasher>>

Map of external image IDs to native textures.

vr_compositor_handler: Arc<Mutex<Option<Box<VRCompositorHandler>>>>
Methods

impl Renderer

fn new(options: RendererOptions)
-> Result<(Renderer, RenderApiSender), InitError>

Initializes webrender and creates a Renderer and RenderApiSender.

Examples

Initializes a Renderer with some reasonable values. For more information see RendererOptions.

let opts = webrender::RendererOptions {
   device_pixel_ratio: 1.0,
   resource_override_path: None,
   enable_aa: false,
   enable_profiler: false,
};
let (renderer, sender) = Renderer::new(opts);

fn set_render_notifier(&self, notifier: Box<RenderNotifier>)

Sets the new RenderNotifier.

The RenderNotifier will be called when processing e.g. of a (scrolling) frame is done, and therefore the screen should be updated.

fn set_main_thread_dispatcher(&self, dispatcher: Box<RenderDispatcher>)

Sets the new MainThreadDispatcher.

Allows to dispatch functions to the main thread's event loop.

fn set_vr_compositor_handler(&self, creator: Box<VRCompositorHandler>)

Sets the VRCompositorHandler.

It's used to handle WebVR render commands. Some WebVR commands such as Vsync and SubmitFrame must be called in the WebGL render thread.

fn current_epoch(&self, pipeline_id: PipelineId) -> Option<Epoch>

Returns the Epoch of the current frame in a pipeline.

fn flush_rendered_epochs(&mut self)
-> HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>>

Returns a HashMap containing the pipeline ids that have been received by the renderer and their respective epochs since the last time the method was called.

fn update(&mut self)

Processes the result queue.

Should be called before render(), as texture cache updates are done here.

fn resolve_source_texture(&mut self, texture_id: &SourceTexture) -> TextureId

fn set_external_image_handler(&mut self, handler: Box<ExternalImageHandler>)

Set a callback for handling external images.

fn render(&mut self, framebuffer_size: DeviceUintSize)

Renders the current frame.

A Frame is supplied by calling set_root_stacking_context().

fn layers_are_bouncing_back(&self) -> bool

fn update_texture_cache(&mut self)

fn draw_instanced_batch<T>(&mut self,
data: &[T],
vao: VAOId,
shader: ProgramId,
textures: &BatchTextures,
projection: &Matrix4D<f32>)

fn submit_batch(&mut self,
batch: &PrimitiveBatch,
projection: &Matrix4D<f32>,
render_task_data: &Vec<RenderTaskData>,
cache_texture: Option<TextureId>,
render_target: Option<(TextureId, i32)>,
target_dimensions: DeviceUintSize)

fn draw_target(&mut self,
render_target: Option<(TextureId, i32)>,
target: &RenderTarget,
target_size: DeviceUintSize,
cache_texture: Option<TextureId>,
should_clear: bool,
background_color: Option<ColorF>,
render_task_data: &Vec<RenderTaskData>)

fn update_deferred_resolves(&mut self, frame: &mut Frame)

fn unlock_external_images(&mut self)

fn release_external_images(&mut self,
pending_external_image_updates: Vec<ExternalImageId>)

fn draw_tile_frame(&mut self,
frame: &mut Frame,
framebuffer_size: &DeviceUintSize)

fn debug_renderer<'a>(&'a mut self) -> &'a mut DebugRenderer

fn get_profiler_enabled(&mut self) -> bool

fn set_profiler_enabled(&mut self, enabled: bool)

fn save_cpu_profile(&self, filename: &str)

fn draw_render_target_debug(&mut self, framebuffer_size: &DeviceUintSize)