Struct webrender::renderer::Renderer

pub struct Renderer { /* fields omitted */ }

The renderer is responsible for submitting to the GPU the work prepared by the RenderBackend.

Methods

impl Renderer

fn new(options: RendererOptions) -> (Renderer, RenderApiSender)

Initializes webrender and creates a Renderer and RenderApiSender.

Examples

Initializes a Renderer with some reasonable values. For more information see RendererOptions.

let opts = webrender::RendererOptions {
   device_pixel_ratio: 1.0,
   resource_override_path: None,
   enable_aa: false,
   enable_profiler: false,
};
let (renderer, sender) = Renderer::new(opts);

fn set_render_notifier(&self, notifier: Box<RenderNotifier>)

Sets the new RenderNotifier.

The RenderNotifier will be called when processing e.g. of a (scrolling) frame is done, and therefore the screen should be updated.

fn set_main_thread_dispatcher(&self, dispatcher: Box<RenderDispatcher>)

Sets the new MainThreadDispatcher.

Allows to dispatch functions to the main thread's event loop.

fn set_vr_compositor_handler(&self, creator: Box<VRCompositorHandler>)

Sets the VRCompositorHandler.

It's used to handle WebVR render commands. Some WebVR commands such as Vsync and SubmitFrame must be called in the WebGL render thread.

fn current_epoch(&self, pipeline_id: PipelineId) -> Option<Epoch>

Returns the Epoch of the current frame in a pipeline.

fn flush_rendered_epochs(&mut self)
-> HashMap<PipelineId, Epoch, BuildHasherDefault<FnvHasher>>

Returns a HashMap containing the pipeline ids that have been received by the renderer and their respective epochs since the last time the method was called.

fn update(&mut self)

Processes the result queue.

Should be called before render(), as texture cache updates are done here.

fn set_external_image_handler(&mut self, handler: Box<ExternalImageHandler>)

Set a callback for handling external images.

fn render(&mut self, framebuffer_size: DeviceUintSize)

Renders the current frame.

A Frame is supplied by calling set_root_stacking_context().

fn layers_are_bouncing_back(&self) -> bool

fn debug_renderer<'a>(&'a mut self) -> &'a mut DebugRenderer

fn get_profiler_enabled(&mut self) -> bool

fn set_profiler_enabled(&mut self, enabled: bool)