Module webrender::renderer::upload

This module contains the convoluted logic that goes into uploading content into the texture cache’s textures.

We need to support various combinations of code paths depending on the quirks of each hardware/driver configuration:

direct upload,

staged upload via a pixel buffer object,

staged upload via a direct upload to a staging texture where PBO’s aren’t supported,

copy from the staging to destination textures, either via blits or batched draw calls.

Conceptually a lot of this logic should probably be in the device module, but some code here relies on submitting draw calls via the renderer.

Structs

BatchUploadBuffer 🔒

BatchUploadCopy 🔒

UploadStats 🔒

UploadTexturePool

A very basic pool to avoid reallocating staging textures as well as staging CPU side buffers.

Enums

StagingBufferKind 🔒

Constants

BATCH_UPLOAD_FORMAT_COUNT 🔒

BATCH_UPLOAD_TEXTURE_SIZE

Functions

copy_from_staging_to_cache 🔒

Copy from the staging PBOs or textures to texture cache textures using blit commands.

copy_from_staging_to_cache_using_draw_calls 🔒

Generate and submit composite shader batches to copy from the staging textures to the destination cache textures.

copy_into_staging_buffer 🔒

Copy an item into a batched upload staging buffer.

skip_staging_buffer 🔒

Take this code path instead of copying into a staging CPU buffer when the image we would copy is large enough that it’s unlikely anything else would fit in the buffer, therefore we might as well copy directly from the source image’s pixels.

upload_to_texture_cache

Upload a number of items to texture cache textures.