webrender::renderer::upload

Function upload_to_texture_cache

pub fn upload_to_texture_cache(
    renderer: &mut Renderer,
    update_list: FastHashMap<CacheTextureId, Vec<TextureCacheUpdate>>,
)

Upload a number of items to texture cache textures.

This is the main entry point of the texture cache upload code. See also the module documentation for more information.