webrender

Module resource_cache

Structs

CacheItem

CachedImageInfo ğŸ”’

CachedImageKey ğŸ”’

GlyphFetchResult

ImageGeneration

ImageProperties

ImageRequest

ImageResource ğŸ”’

ImageTemplates ğŸ”’

PlainCacheRef

PlainFontTemplate ğŸ”’

PlainImageTemplate ğŸ”’

PlainResources

RenderTarget ğŸ”’

Internal information about allocated render targets in the pool

ResourceCache

High-level container for resources managed by the RenderBackend.

ResourceClassCache

Resources ğŸ”’

Enums

CachedImageData

Represents the backing store of an image in the cache. This storage can take several forms.

ImageCacheError

ImageResult ğŸ”’

State ğŸ”’

Statics

NEXT_NATIVE_SURFACE_ID ğŸ”’

Functions

to_image_dirty_rect ğŸ”’

Type Aliases

GlyphDimensionsCache

ImageCache ğŸ”’

RasterizedBlob ğŸ”’

Post scene building state.