Module webrender::resource_cache

Structs

AsyncBlobImageInfo
Information attached to AsyncBlobImageRasterizer.

BlobImageClearParams
Stores parameters for clearing blob image tiles.

BlobImageRasterizerEpoch

BlobImageTemplate
Pre scene building state. We use this to generate the async blob rendering requests.

CacheItem

CachedImageInfo

CachedImageKey

GlyphFetchResult

ImageProperties

ImageRequest

ImageResource

ImageTemplates

ImageTiling

PlainCacheRef

PlainFontTemplate

PlainImageTemplate

PlainResources

ResourceCache
High-level container for resources managed by the RenderBackend.

ResourceClassCache

Resources

Enums

CachedImageData
Represents the backing store of an image in the cache. This storage can take several forms.

ImageCacheError

ImageResult

RasterizedBlob
Post scene building state.

State

Constants

DEFAULT_TILE_SIZE

_IMPL_SERIALIZE_FOR_CacheItem

_IMPL_SERIALIZE_FOR_CachedImageInfo

_IMPL_SERIALIZE_FOR_CachedImageKey

_IMPL_SERIALIZE_FOR_GlyphFetchResult

_IMPL_SERIALIZE_FOR_ImageCacheError

_IMPL_SERIALIZE_FOR_ImageProperties

_IMPL_SERIALIZE_FOR_ImageRequest

_IMPL_SERIALIZE_FOR_ImageResult

_IMPL_SERIALIZE_FOR_PlainCacheRef

_IMPL_SERIALIZE_FOR_PlainFontTemplate

_IMPL_SERIALIZE_FOR_PlainImageTemplate

_IMPL_SERIALIZE_FOR_PlainResources

_IMPL_SERIALIZE_FOR_ResourceClassCache

Functions

blob_rect

blob_size
For now the blob's coordinate space have the same pixel sizes as the rendered texture's device space (only a translation is applied). So going from one to the other is only a matter of casting away the unit for sizes.

get_blob_tiling

to_image_dirty_rect

Type Definitions

FontInstanceMap

GlyphDimensionsCache

ImageCache