Struct webrender::resource_cache::ResourceCache

pub struct ResourceCache {
    cached_glyphs: GlyphCache,
    cached_images: ResourceClassCache<ImageKey, ImageResult, ()>,
    cached_render_tasks: RenderTaskCache,
    resources: Resources,
    state: State,
    current_frame_id: FrameId,
    pub texture_cache: TextureCache,
    cached_glyph_dimensions: HashMap<(FontInstanceKey, GlyphIndex), Option<GlyphDimensions>, BuildHasherDefault<FxHasher>>,
    glyph_rasterizer: GlyphRasterizer,
    pending_image_requests: HashSet<ImageRequest, BuildHasherDefault<FxHasher>>,
    blob_image_handler: Option<Box<dyn BlobImageHandler>>,
    rasterized_blob_images: HashMap<BlobImageKey, RasterizedBlob, BuildHasherDefault<FxHasher>>,
    blob_image_templates: HashMap<BlobImageKey, BlobImageTemplate, BuildHasherDefault<FxHasher>>,
    missing_blob_images: Vec<BlobImageParams>,
    blob_image_rasterizer: Option<Box<dyn AsyncBlobImageRasterizer>>,
    blob_image_rasterizer_produced_epoch: BlobImageRasterizerEpoch,
    blob_image_rasterizer_consumed_epoch: BlobImageRasterizerEpoch,
    deleted_blob_keys: VecDeque<Vec<BlobImageKey>>,
    dirty_image_keys: HashSet<ImageKey, BuildHasherDefault<FxHasher>>,
    active_image_keys: HashSet<ImageKey, BuildHasherDefault<FxHasher>>,
}

High-level container for resources managed by the RenderBackend.

This includes a variety of things, including images, fonts, and glyphs, which may be stored as memory buffers, GPU textures, or handles to resources managed by the OS or other parts of WebRender.

Fields
cached_glyphs: GlyphCachecached_images: ResourceClassCache<ImageKey, ImageResult, ()>cached_render_tasks: RenderTaskCacheresources: Resourcesstate: Statecurrent_frame_id: FrameIdtexture_cache: TextureCachecached_glyph_dimensions: HashMap<(FontInstanceKey, GlyphIndex), Option<GlyphDimensions>, BuildHasherDefault<FxHasher>>

TODO(gw): We should expire (parts of) this cache semi-regularly!

glyph_rasterizer: GlyphRasterizerpending_image_requests: HashSet<ImageRequest, BuildHasherDefault<FxHasher>>

The set of images that aren't present or valid in the texture cache, and need to be rasterized and/or uploaded this frame. This includes both blobs and regular images.

blob_image_handler: Option<Box<dyn BlobImageHandler>>rasterized_blob_images: HashMap<BlobImageKey, RasterizedBlob, BuildHasherDefault<FxHasher>>blob_image_templates: HashMap<BlobImageKey, BlobImageTemplate, BuildHasherDefault<FxHasher>>missing_blob_images: Vec<BlobImageParams>

If while building a frame we encounter blobs that we didn't already rasterize, add them to this list and rasterize them synchronously.

blob_image_rasterizer: Option<Box<dyn AsyncBlobImageRasterizer>>

The rasterizer associated with the current scene.

blob_image_rasterizer_produced_epoch: BlobImageRasterizerEpoch

An epoch of the stored blob image rasterizer, used to skip the ones coming from low-priority scene builds if the current one is newer. This is to be removed when we get rid of the whole "missed" blob images concept. The produced one gets bumped whenever we produce a rasteriezer, which then travels through the scene building and eventually gets consumed back by us, bumping the consumed epoch.

blob_image_rasterizer_consumed_epoch: BlobImageRasterizerEpochdeleted_blob_keys: VecDeque<Vec<BlobImageKey>>

A log of the last three frames worth of deleted image keys kept for debugging purposes.

dirty_image_keys: HashSet<ImageKey, BuildHasherDefault<FxHasher>>

A set of the image keys that have been requested, and require updates to the texture cache. Images in this category trigger invalidations for picture caching tiles.

active_image_keys: HashSet<ImageKey, BuildHasherDefault<FxHasher>>

A set of the image keys that are used for render.

Methods

impl ResourceCache

pub fn new(
texture_cache: TextureCache,
glyph_rasterizer: GlyphRasterizer,
cached_glyphs: GlyphCache,
blob_image_handler: Option<Box<dyn BlobImageHandler>>
) -> Self

pub fn max_texture_size(&self) -> i32

fn should_tile(
limit: i32,
descriptor: &ImageDescriptor,
data: &CachedImageData
) -> bool

pub fn request_render_task<F>(
&mut self,
key: RenderTaskCacheKey,
gpu_cache: &mut GpuCache,
render_tasks: &mut RenderTaskGraph,
user_data: Option<[f32; 3]>,
is_opaque: bool,
f: F
) -> WeakFreeListHandle<RenderTaskCacheMarker> where
F: FnOnce(&mut RenderTaskGraph) -> RenderTaskId,

pub fn post_scene_building_update(
&mut self,
updates: Vec<ResourceUpdate>,
profile_counters: &mut ResourceProfileCounters
)

pub fn pre_scene_building_update(
&mut self,
updates: &mut Vec<ResourceUpdate>,
profile_counters: &mut ResourceProfileCounters
)

pub fn set_blob_rasterizer(
&mut self,
rasterizer: Box<dyn AsyncBlobImageRasterizer>,
supp: AsyncBlobImageInfo
)

pub fn add_rasterized_blob_images(
&mut self,
images: Vec<(BlobImageRequest, BlobImageResult)>,
texture_cache_profile: &mut TextureCacheProfileCounters
)

pub fn add_font_template(&mut self, font_key: FontKey, template: FontTemplate)

pub fn delete_font_template(&mut self, font_key: FontKey)

pub fn add_font_instance(
&mut self,
instance_key: FontInstanceKey,
font_key: FontKey,
size: Au,
options: Option<FontInstanceOptions>,
platform_options: Option<FontInstancePlatformOptions>,
variations: Vec<FontVariation>
)

pub fn delete_font_instance(&mut self, instance_key: FontInstanceKey)

pub fn get_font_instances(
&self
) -> Arc<RwLock<HashMap<FontInstanceKey, Arc<BaseFontInstance>, BuildHasherDefault<FxHasher>>>>

pub fn get_font_instance(
&self,
instance_key: FontInstanceKey
) -> Option<Arc<BaseFontInstance>>

pub fn add_image_template(
&mut self,
image_key: ImageKey,
descriptor: ImageDescriptor,
data: CachedImageData,
visible_rect: &DeviceIntRect,
tiling: Option<TileSize>
)

pub fn update_image_template(
&mut self,
image_key: ImageKey,
descriptor: ImageDescriptor,
data: CachedImageData,
dirty_rect: &ImageDirtyRect
)

pub fn add_blob_image(
&mut self,
key: BlobImageKey,
descriptor: &ImageDescriptor,
tiling: Option<TileSize>,
data: Arc<BlobImageData>,
visible_rect: &DeviceIntRect
)

pub fn update_blob_image(
&mut self,
key: BlobImageKey,
descriptor: Option<&ImageDescriptor>,
dirty_rect: Option<&BlobDirtyRect>,
data: Option<Arc<BlobImageData>>,
visible_rect: &DeviceIntRect
)

pub fn delete_image_template(&mut self, image_key: ImageKey)

pub fn is_image_dirty(&self, image_key: ImageKey) -> bool

Check if an image has changed since it was last requested.

pub fn is_image_active(&self, image_key: ImageKey) -> bool

pub fn set_image_active(&mut self, image_key: ImageKey)

pub fn request_image(&mut self, request: ImageRequest, gpu_cache: &mut GpuCache)

pub fn create_blob_scene_builder_requests(
&mut self,
keys: &[BlobImageKey]
) -> (Option<(Box<dyn AsyncBlobImageRasterizer>, AsyncBlobImageInfo)>, Vec<BlobImageParams>)

fn discard_tiles_outside_visible_area(
&mut self,
key: BlobImageKey,
area: &DeviceIntRect
)

fn set_image_visible_rect(&mut self, key: ImageKey, rect: &DeviceIntRect)

pub fn request_glyphs(
&mut self,
font: FontInstance,
glyph_keys: &[GlyphKey],
gpu_cache: &mut GpuCache,
render_task_tree: &mut RenderTaskGraph
)

pub fn pending_updates(&mut self) -> TextureUpdateList

pub fn fetch_glyphs<F>(
&self,
font: FontInstance,
glyph_keys: &[GlyphKey],
fetch_buffer: &mut Vec<GlyphFetchResult>,
gpu_cache: &mut GpuCache,
f: F
) where
F: FnMut(TextureSource, GlyphFormat, &[GlyphFetchResult]),

pub fn get_glyph_dimensions(
&mut self,
font: &FontInstance,
glyph_index: GlyphIndex
) -> Option<GlyphDimensions>

pub fn get_glyph_index(&mut self, font_key: FontKey, ch: char) -> Option<u32>

pub fn get_cached_image(&self, request: ImageRequest) -> Result<CacheItem, ()>

pub fn get_cached_render_task(
&self,
handle: &WeakFreeListHandle<RenderTaskCacheMarker>
) -> &RenderTaskCacheEntry

fn get_image_info(&self, request: ImageRequest) -> Result<&CachedImageInfo, ()>

pub fn get_texture_cache_item(
&self,
handle: &WeakFreeListHandle<CacheEntryMarker>
) -> CacheItem

pub fn get_image_properties(
&self,
image_key: ImageKey
) -> Option<ImageProperties>

pub fn prepare_for_frames(&mut self, time: SystemTime)

pub fn bookkeep_after_frames(&mut self)

pub fn requires_frame_build(&self) -> bool

pub fn begin_frame(&mut self, stamp: FrameStamp)

pub fn block_until_all_resources_added(
&mut self,
gpu_cache: &mut GpuCache,
render_tasks: &mut RenderTaskGraph,
texture_cache_profile: &mut TextureCacheProfileCounters
)

fn rasterize_missing_blob_images(
&mut self,
texture_cache_profile: &mut TextureCacheProfileCounters
)

fn update_texture_cache(&mut self, gpu_cache: &mut GpuCache)

pub fn end_frame(
&mut self,
texture_cache_profile: &mut TextureCacheProfileCounters
)

pub fn set_debug_flags(&mut self, flags: DebugFlags)

pub fn clear(&mut self, what: ClearCache)

pub fn clear_namespace(&mut self, namespace: IdNamespace)

pub fn report_memory(&self, op: VoidPtrToSizeFn) -> MemoryReport

Reports the CPU heap usage of this ResourceCache.

NB: It would be much better to use the derive(MallocSizeOf) machinery here, but the Arcs complicate things. The two ways to handle that would be to either (a) Implement MallocSizeOf manually for the things that own them and manually avoid double-counting, or (b) Use the "seen this pointer yet" machinery from the proper malloc_size_of crate. We can do this if/when more accurate memory reporting on these resources becomes a priority.

fn clear_images<F: Fn(&ImageKey) -> bool>(&mut self, f: F)

Properly deletes all images matching the predicate.

impl ResourceCache

pub fn save_capture(
&mut self,
root: &PathBuf
) -> (PlainResources, Vec<ExternalCaptureImage>)

pub fn save_caches(&self, _root: &PathBuf) -> PlainCacheRef