Module webrender::scene

Structs

BuiltScene

WebRender’s internal representation of the scene.

Scene

A complete representation of the layout bundling visible pipelines together.

ScenePipeline

A representation of the layout within the display port for a given document or iframe.

SceneProperties

Stores a map of the animated property bindings for the current display list. These can be used to animate the transform and/or opacity of a display list without re-submitting the display list itself.

SceneStats

Stores the allocation sizes of various arrays in the built scene. This is retrieved from the current frame builder and used to reserve an approximately correct capacity of the arrays for the next scene that is getting built.

Traits

StackingContextHelpers