Struct webrender::scene::SceneProperties

pub struct SceneProperties {
    transform_properties: HashMap<PropertyBindingId, LayoutTransform, BuildHasherDefault<FxHasher>>,
    float_properties: HashMap<PropertyBindingId, f32, BuildHasherDefault<FxHasher>>,
    current_properties: DynamicProperties,
    pending_properties: Option<DynamicProperties>,
}

Stores a map of the animated property bindings for the current display list. These can be used to animate the transform and/or opacity of a display list without re-submitting the display list itself.

Fields
transform_properties: HashMap<PropertyBindingId, LayoutTransform, BuildHasherDefault<FxHasher>>float_properties: HashMap<PropertyBindingId, f32, BuildHasherDefault<FxHasher>>current_properties: DynamicPropertiespending_properties: Option<DynamicProperties>
Methods

impl SceneProperties

pub fn new() -> Self

pub fn set_properties(&mut self, properties: DynamicProperties)

Set the current property list for this display list.

pub fn add_properties(&mut self, properties: DynamicProperties)

Add to the current property list for this display list.

pub fn flush_pending_updates(&mut self) -> bool

Flush any pending updates to the scene properties. Returns true if the properties have changed since the last flush was called. This code allows properties to be changed by multiple set_properties and add_properties calls during a single transaction, and still correctly determine if any properties have changed. This can have significant power saving implications, allowing a frame build to be skipped if the properties haven't changed in many cases.

pub fn resolve_layout_transform(
&self,
property: &PropertyBinding<LayoutTransform>
) -> LayoutTransform

Get the current value for a transform property.

pub fn resolve_float(&self, property: &PropertyBinding<f32>) -> f32

Get the current value for a float property.

pub fn float_properties(
&self
) -> &HashMap<PropertyBindingId, f32, BuildHasherDefault<FxHasher>>