Module webrender::scene_builder_thread

Structs

BuiltTransaction
Represent the remaining work associated to a transaction after the scene building phase as well as the result of scene building itself if applicable.

DisplayListUpdate

Document

InternerUpdates

Interners
This struct contains all items that can be shared between display lists. We want to intern and share the same clips, primitives and other things between display lists so that:

LowPrioritySceneBuilderThread
A scene builder thread which executes expensive operations such as blob rasterization with a lower priority than the normal scene builder thread.

SceneBuilderThread

SceneRequest
Contains the render backend data needed to build a scene.

Transaction
Represents the work associated to a transaction before scene building.

Enums

SceneBuilderRequest

SceneBuilderResult

SceneSwapResult

Constants

_IMPL_SERIALIZE_FOR_Interners