Struct webrender::scene_builder_thread::SceneBuilderThread

pub struct SceneBuilderThread {
    documents: FastHashMap<DocumentId, Document>,
    rx: Receiver<SceneBuilderRequest>,
    backend_rx: Receiver<BackendSceneBuilderRequest>,
    tx: Sender<SceneBuilderResult>,
    api_tx: Sender<ApiMsg>,
    config: FrameBuilderConfig,
    default_device_pixel_ratio: f32,
    font_instances: SharedFontInstanceMap,
    size_of_ops: Option<MallocSizeOfOps>,
    hooks: Option<Box<dyn SceneBuilderHooks + Send>>,
    simulate_slow_ms: u32,
    removed_pipelines: HashSet<PipelineId, BuildHasherDefault<FxHasher>>,
    capture_config: Option<CaptureConfig>,
}

Fields
documents: FastHashMap<DocumentId, Document>rx: Receiver<SceneBuilderRequest>backend_rx: Receiver<BackendSceneBuilderRequest>tx: Sender<SceneBuilderResult>api_tx: Sender<ApiMsg>config: FrameBuilderConfigdefault_device_pixel_ratio: f32font_instances: SharedFontInstanceMapsize_of_ops: Option<MallocSizeOfOps>hooks: Option<Box<dyn SceneBuilderHooks + Send>>simulate_slow_ms: u32removed_pipelines: HashSet<PipelineId, BuildHasherDefault<FxHasher>>capture_config: Option<CaptureConfig>
Implementations

impl SceneBuilderThread

pub fn new(
config: FrameBuilderConfig,
default_device_pixel_ratio: f32,
font_instances: SharedFontInstanceMap,
size_of_ops: Option<MallocSizeOfOps>,
hooks: Option<Box<dyn SceneBuilderHooks + Send>>,
channels: SceneBuilderThreadChannels
) -> Self

pub fn send(&self, msg: SceneBuilderResult)

Send a message to the render backend thread.

We first put something in the result queue and then send a wake-up message to the api queue that the render backend is blocking on.

pub fn run(&mut self)

The scene builder thread's event loop.

fn save_scene(&mut self, config: CaptureConfig)

fn save_capture_sequence(&mut self)

fn start_capture_sequence(&mut self, config: CaptureConfig)

fn get_docs_for_debugger(&self) -> String

fn process_transaction(
&mut self,
txn: &mut TransactionMsg
) -> Box<BuiltTransaction>

Do the bulk of the work of the scene builder thread.

fn forward_built_transactions(&mut self, txns: Vec<Box<BuiltTransaction>>)

Send the results of process_transaction back to the render backend.

fn report_memory(&mut self) -> MemoryReport

Reports CPU heap memory used by the SceneBuilder.