Module webrender::scene_building

Structs

CompositeOps

ExtendedPrimitiveInstance
A primitive instance + some extra information about the primitive. This is stored when constructing 3d rendering contexts, which involve cutting primitive lists.

FlattenedStackingContext
Properties of a stacking context that are maintained during creation of the scene. These structures are not persisted after the initial scene build.

NodeIdToIndexMapper
A data structure that keeps track of mapping between API Ids for spatials and the indices used internally in the SpatialTree to avoid having to do HashMap lookups for primitives and clips during frame building.

PendingPrimitive
A primitive that is added while a shadow context is active is stored as a pending primitive and only added to pictures during pop_all_shadows.

PendingShadow
As shadows are pushed, they are stored as pending shadows, and handled at once during pop_all_shadows.

ReferenceFrameMapper
Maps from stacking context layout coordinates into reference frame relative coordinates.

ReferenceFrameState
The offset stack for a given reference frame.

SceneBuilder
A structure that converts a serialized display list into a form that WebRender can use to later build a frame. This structure produces a BuiltScene. Public members are typically those that are destructured into the BuiltScene.

ScrollOffsetMapper
Offsets primitives (and clips) by the external scroll offset supplied to scroll nodes.

Slice
Information about a set of primitive clusters that will form a picture cache slice.

SliceFlags
Slice flags

Enums

ShadowItem

Traits

CreateShadow

IsVisible

Functions

add_clips

create_prim_instance

create_tile_cache
Given a PrimitiveList and scroll root, construct a tile cache primitive instance that wraps the primitive list.

filter_datas_for_compositing

filter_ops_for_compositing

filter_primitives_for_compositing

process_repeat_size