Struct webrender::scene_building::FlattenedStackingContext

struct FlattenedStackingContext {
    prim_list: PrimitiveList,
    prim_flags: PrimitiveFlags,
    requested_raster_space: RasterSpace,
    spatial_node_index: SpatialNodeIndex,
    clip_chain_id: ClipChainId,
    clip_id: Option<ClipId>,
    frame_output_pipeline_id: Option<PipelineId>,
    composite_ops: CompositeOps,
    blit_reason: BlitReason,
    pipeline_id: PipelineId,
    transform_style: TransformStyle,
    context_3d: Picture3DContext<ExtendedPrimitiveInstance>,
    is_backdrop_root: bool,
    is_redundant: bool,
    snap_to_device: SpaceSnapper,
}

Properties of a stacking context that are maintained during creation of the scene. These structures are not persisted after the initial scene build.

Fields
prim_list: PrimitiveList

The list of primitive instances added to this stacking context.

prim_flags: PrimitiveFlags

Primitive instance flags for compositing this stacking context

requested_raster_space: RasterSpace

Whether or not the caller wants this drawn in screen space (quality) or local space (performance)

spatial_node_index: SpatialNodeIndex

The positioning node for this stacking context

clip_chain_id: ClipChainId

The clip chain for this stacking context

clip_id: Option<ClipId>frame_output_pipeline_id: Option<PipelineId>

If set, this should be provided to caller as an output texture.

composite_ops: CompositeOps

The list of filters / mix-blend-mode for this stacking context.

blit_reason: BlitReason

Bitfield of reasons this stacking context needs to be an offscreen surface.

pipeline_id: PipelineId

Pipeline this stacking context belongs to.

transform_style: TransformStyle

CSS transform-style property.

context_3d: Picture3DContext<ExtendedPrimitiveInstance>

Defines the relationship to a preserve-3D hiearachy.

is_backdrop_root: bool

True if this stacking context is a backdrop root.

is_redundant: bool

True if this stacking context is redundant (i.e. doesn't require a surface)

snap_to_device: SpaceSnapper

A helper struct to snap local rects in device space. During frame building we may establish new raster roots, however typically that is in cases where we won't be applying snapping (e.g. has perspective), or in edge cases (e.g. SVG filter) where we can accept slightly incorrect behaviour in favour of getting the common case right.

Implementations

impl FlattenedStackingContext

pub fn is_3d(&self) -> bool

Return true if the stacking context has a valid preserve-3d property

fn init_picture_caching(
&mut self,
spatial_tree: &SpatialTree,
clip_store: &ClipStore,
quality_settings: &QualitySettings
) -> usize

Set up appropriate cluster flags for picture caching on this stacking context.

pub fn is_redundant(
context_3d: &Picture3DContext<ExtendedPrimitiveInstance>,
composite_ops: &CompositeOps,
prim_flags: PrimitiveFlags,
blit_reason: BlitReason,
requested_raster_space: RasterSpace,
parent: &FlattenedStackingContext
) -> bool

Return true if the stacking context isn't needed.

pub fn cut_item_sequence(
&mut self,
prim_store: &mut PrimitiveStore,
interners: &mut Interners,
composite_mode: Option<PictureCompositeMode>,
flat_items_context_3d: Picture3DContext<OrderedPictureChild>
) -> Option<(PictureIndex, PrimitiveInstance)>

Cut the sequence of the immediate children recorded so far and generate a picture from them.