Struct webrender::scene_building::FlattenedStackingContext

struct FlattenedStackingContext {
    prim_list: PrimitiveList,
    prim_flags: PrimitiveFlags,
    spatial_node_index: SpatialNodeIndex,
    clip_node_id: ClipNodeId,
    composite_ops: CompositeOps,
    blit_reason: BlitReason,
    transform_style: TransformStyle,
    context_3d: Picture3DContext<ExtendedPrimitiveInstance>,
    flags: StackingContextFlags,
    raster_space: RasterSpace,
    subregion_offset: LayoutVector2D,
}

Properties of a stacking context that are maintained during creation of the scene. These structures are not persisted after the initial scene build.

Fields
prim_list: PrimitiveList

The list of primitive instances added to this stacking context.

prim_flags: PrimitiveFlags

Primitive instance flags for compositing this stacking context

spatial_node_index: SpatialNodeIndex

The positioning node for this stacking context

clip_node_id: ClipNodeId

The clip chain for this stacking context

composite_ops: CompositeOps

The list of filters / mix-blend-mode for this stacking context.

blit_reason: BlitReason

Bitfield of reasons this stacking context needs to be an offscreen surface.

transform_style: TransformStyle

CSS transform-style property.

context_3d: Picture3DContext<ExtendedPrimitiveInstance>

Defines the relationship to a preserve-3D hiearachy.

flags: StackingContextFlags

Flags identifying the type of container (among other things) this stacking context is

raster_space: RasterSpace

Requested raster space for this stacking context

subregion_offset: LayoutVector2D

Offset to be applied to any filter sub-regions

Implementations

impl FlattenedStackingContext

pub fn is_3d(&self) -> bool

Return true if the stacking context has a valid preserve-3d property

pub fn is_redundant( context_3d: &Picture3DContext<ExtendedPrimitiveInstance>, composite_ops: &CompositeOps, blit_reason: BlitReason, parent: Option<&FlattenedStackingContext>, prim_flags: PrimitiveFlags, ) -> bool

Return true if the stacking context isn’t needed.

pub fn cut_item_sequence( &mut self, prim_store: &mut PrimitiveStore, interners: &mut Interners, composite_mode: Option<PictureCompositeMode>, flat_items_context_3d: Picture3DContext<OrderedPictureChild>, clip_tree_builder: &mut ClipTreeBuilder, ) -> Option<(PictureIndex, PrimitiveInstance)>

Cut the sequence of the immediate children recorded so far and generate a picture from them.