Struct webrender::scene_building::SceneBuilder

pub struct SceneBuilder<'a> {
    scene: &'a Scene,
    font_instances: Arc<RwLock<HashMap<FontInstanceKey, Arc<BaseFontInstance>, BuildHasherDefault<FxHasher>>>>,
    output_pipelines: &'a HashSet<PipelineId, BuildHasherDefault<FxHasher>>,
    id_to_index_mapper: NodeIdToIndexMapper,
    sc_stack: Vec<FlattenedStackingContext>,
    pending_shadow_items: VecDeque<ShadowItem>,
    pipeline_clip_chain_stack: Vec<ClipChainId>,
    pub clip_scroll_tree: ClipScrollTree,
    pub prim_store: PrimitiveStore,
    pub hit_testing_scene: HitTestingScene,
    pub clip_store: ClipStore,
    pub config: FrameBuilderConfig,
    interners: &'a mut Interners,
    pub root_pic_index: PictureIndex,
    rf_mapper: ReferenceFrameMapper,
    external_scroll_mapper: ScrollOffsetMapper,
    picture_caching_initialized: bool,
    iframe_depth: usize,
    content_slice_count: usize,
    picture_cache_spatial_nodes: HashSet<SpatialNodeIndex, BuildHasherDefault<FxHasher>>,
}

A structure that converts a serialized display list into a form that WebRender can use to later build a frame. This structure produces a BuiltScene. Public members are typically those that are destructured into the BuiltScene.

Fields
scene: &'a Scene

The scene that we are currently building.

font_instances: Arc<RwLock<HashMap<FontInstanceKey, Arc<BaseFontInstance>, BuildHasherDefault<FxHasher>>>>

The map of all font instances.

output_pipelines: &'a HashSet<PipelineId, BuildHasherDefault<FxHasher>>

A set of pipelines that the caller has requested be made available as output textures.

id_to_index_mapper: NodeIdToIndexMapper

The data structure that converts between ClipId/SpatialId and the various index types that the ClipScrollTree uses.

sc_stack: Vec<FlattenedStackingContext>

A stack of stacking context properties.

pending_shadow_items: VecDeque<ShadowItem>

Maintains state for any currently active shadows

pipeline_clip_chain_stack: Vec<ClipChainId>

The stack keeping track of the root clip chains associated with pipelines.

clip_scroll_tree: ClipScrollTree

The ClipScrollTree that we are currently building during building.

prim_store: PrimitiveStore

The store of primitives.

hit_testing_scene: HitTestingScene

Information about all primitives involved in hit testing.

clip_store: ClipStore

The store which holds all complex clipping information.

config: FrameBuilderConfig

The configuration to use for the FrameBuilder. We consult this in order to determine the default font.

interners: &'a mut Interners

Reference to the set of data that is interned across display lists.

root_pic_index: PictureIndex

The root picture index for this builder. This is the picture to start the culling phase from.

rf_mapper: ReferenceFrameMapper

Helper struct to map stacking context coords <-> reference frame coords.

external_scroll_mapper: ScrollOffsetMapper

Helper struct to map spatial nodes to external scroll offsets.

picture_caching_initialized: bool

If true, picture caching setup has already been completed.

iframe_depth: usize

The current recursion depth of iframes encountered. Used to restrict picture caching slices to only the top-level content frame.

content_slice_count: usize

The number of picture cache slices that were created for content.

picture_cache_spatial_nodes: HashSet<SpatialNodeIndex, BuildHasherDefault<FxHasher>>

A set of any spatial nodes that are attached to either a picture cache root, or a clip node on the picture cache primitive. These are used to detect cases where picture caching must be disabled. This is mostly a temporary workaround for some existing wrench tests. I don't think Gecko ever produces picture cache slices with complex transforms, so in future we should prevent this in the public API and remove this hack.

Methods

impl<'a> SceneBuilder<'a>

pub fn add_normal_border(
&mut self,
info: &LayoutPrimitiveInfo,
border: &ApiNormalBorder,
widths: LayoutSideOffsets,
clip_and_scroll: ScrollNodeAndClipChain
)

impl<'a> SceneBuilder<'a>

pub fn add_box_shadow(
&mut self,
clip_and_scroll: ScrollNodeAndClipChain,
prim_info: &LayoutPrimitiveInfo,
box_offset: &LayoutVector2D,
color: ColorF,
blur_radius: f32,
spread_radius: f32,
border_radius: BorderRadius,
clip_mode: BoxShadowClipMode
)

impl<'a> SceneBuilder<'a>

pub fn build(
scene: &Scene,
font_instances: Arc<RwLock<HashMap<FontInstanceKey, Arc<BaseFontInstance>, BuildHasherDefault<FxHasher>>>>,
view: &DocumentView,
output_pipelines: &HashSet<PipelineId, BuildHasherDefault<FxHasher>>,
frame_builder_config: &FrameBuilderConfig,
interners: &mut Interners,
stats: &SceneStats
) -> BuiltScene

fn current_offset(
&mut self,
spatial_node_index: SpatialNodeIndex
) -> LayoutVector2D

Retrieve the current offset to allow converting a stacking context relative coordinate to be relative to the owing reference frame, also considering any external scroll offset on the provided spatial node.

fn setup_picture_caching(&mut self, main_prim_list: &mut PrimitiveList)

Figure out the shape of the display list, and wrap various primitive clusters into tile cache primitive instances.

fn build_items(
&mut self,
traversal: &mut BuiltDisplayListIter<'a>,
pipeline_id: PipelineId,
apply_pipeline_clip: bool
)

fn build_sticky_frame(
&mut self,
info: &StickyFrameDisplayItem,
parent_node_index: SpatialNodeIndex
)

fn build_scroll_frame(
&mut self,
item: &DisplayItemRef,
info: &ScrollFrameDisplayItem,
parent_node_index: SpatialNodeIndex,
pipeline_id: PipelineId
)

fn build_reference_frame(
&mut self,
traversal: &mut BuiltDisplayListIter<'a>,
pipeline_id: PipelineId,
parent_spatial_node: SpatialNodeIndex,
origin: LayoutPoint,
reference_frame: &ReferenceFrame,
apply_pipeline_clip: bool
)

fn build_stacking_context(
&mut self,
traversal: &mut BuiltDisplayListIter<'a>,
pipeline_id: PipelineId,
stacking_context: &StackingContext,
spatial_node_index: SpatialNodeIndex,
origin: LayoutPoint,
filters: ItemRange<FilterOp>,
filter_datas: &[TempFilterData],
filter_primitives: ItemRange<FilterPrimitive>,
prim_flags: PrimitiveFlags,
apply_pipeline_clip: bool
)

fn build_iframe(
&mut self,
info: &IframeDisplayItem,
spatial_node_index: SpatialNodeIndex
)

fn get_space(&mut self, spatial_id: &SpatialId) -> SpatialNodeIndex

fn get_clip_and_scroll(
&mut self,
clip_id: &ClipId,
spatial_id: &SpatialId,
apply_pipeline_clip: bool
) -> ScrollNodeAndClipChain

fn process_common_properties(
&mut self,
common: &CommonItemProperties,
apply_pipeline_clip: bool
) -> (LayoutPrimitiveInfo, ScrollNodeAndClipChain)

fn process_common_properties_with_bounds(
&mut self,
common: &CommonItemProperties,
bounds: &LayoutRect,
apply_pipeline_clip: bool
) -> (LayoutPrimitiveInfo, LayoutRect, ScrollNodeAndClipChain)

pub fn snap_rect(
&mut self,
rect: &LayoutRect,
target_spatial_node: SpatialNodeIndex
) -> LayoutRect

fn build_item<'b>(
&'b mut self,
item: DisplayItemRef<'a, 'b>,
pipeline_id: PipelineId,
apply_pipeline_clip: bool
) -> Option<BuiltDisplayListIter<'a>>

fn build_clip_chain(
&mut self,
clip_items: Vec<ClipItemKey>,
parent_clip_chain_id: ClipChainId
) -> ClipChainId

fn create_primitive<P>(
&mut self,
info: &LayoutPrimitiveInfo,
clip_chain_id: ClipChainId,
spatial_node_index: SpatialNodeIndex,
prim: P
) -> PrimitiveInstance where
P: InternablePrimitive,
Interners: AsMut<Interner<P>>,

Create a primitive and add it to the prim store. This method doesn't add the primitive to the draw list, so can be used for creating sub-primitives.

TODO(djg): Can this inline into add_interned_prim_to_draw_list

pub fn add_primitive_to_hit_testing_list(
&mut self,
info: &LayoutPrimitiveInfo,
clip_and_scroll: ScrollNodeAndClipChain
)

pub fn add_primitive_to_draw_list(
&mut self,
prim_instance: PrimitiveInstance,
prim_size: LayoutSize,
spatial_node_index: SpatialNodeIndex,
flags: PrimitiveFlags
)

Add an already created primitive to the draw lists.

fn add_nonshadowable_primitive<P>(
&mut self,
clip_and_scroll: ScrollNodeAndClipChain,
info: &LayoutPrimitiveInfo,
clip_items: Vec<ClipItemKey>,
prim: P
) where
P: InternablePrimitive + IsVisible,
Interners: AsMut<Interner<P>>,

Convenience interface that creates a primitive entry and adds it to the draw list.

pub fn add_primitive<P>(
&mut self,
clip_and_scroll: ScrollNodeAndClipChain,
info: &LayoutPrimitiveInfo,
clip_items: Vec<ClipItemKey>,
prim: P
) where
P: InternablePrimitive + IsVisible,
Interners: AsMut<Interner<P>>,
ShadowItem: From<PendingPrimitive<P>>,

fn add_prim_to_draw_list<P>(
&mut self,
info: &LayoutPrimitiveInfo,
clip_chain_id: ClipChainId,
clip_and_scroll: ScrollNodeAndClipChain,
prim: P
) where
P: InternablePrimitive,
Interners: AsMut<Interner<P>>,

pub fn push_stacking_context(
&mut self,
pipeline_id: PipelineId,
composite_ops: CompositeOps,
transform_style: TransformStyle,
prim_flags: PrimitiveFlags,
create_tile_cache: bool,
spatial_node_index: SpatialNodeIndex,
clip_chain_id: ClipChainId,
requested_raster_space: RasterSpace,
is_backdrop_root: bool,
device_pixel_scale: DevicePixelScale
)

pub fn pop_stacking_context(&mut self)

pub fn push_reference_frame(
&mut self,
reference_frame_id: SpatialId,
parent_index: Option<SpatialNodeIndex>,
pipeline_id: PipelineId,
transform_style: TransformStyle,
source_transform: PropertyBinding<LayoutTransform>,
kind: ReferenceFrameKind,
origin_in_parent_reference_frame: LayoutVector2D
) -> SpatialNodeIndex

pub fn push_root(
&mut self,
pipeline_id: PipelineId,
viewport_size: &LayoutSize,
content_size: &LayoutSize,
device_pixel_scale: DevicePixelScale
)

pub fn add_clip_node<I>(
&mut self,
new_node_id: ClipId,
space_and_clip: &SpaceAndClipInfo,
clip_region: ClipRegion<I>
) -> ClipChainId where
I: IntoIterator<Item = ComplexClipRegion>,

pub fn add_scroll_frame(
&mut self,
new_node_id: SpatialId,
parent_node_index: SpatialNodeIndex,
external_id: Option<ExternalScrollId>,
pipeline_id: PipelineId,
frame_rect: &LayoutRect,
content_size: &LayoutSize,
scroll_sensitivity: ScrollSensitivity,
frame_kind: ScrollFrameKind,
external_scroll_offset: LayoutVector2D
) -> SpatialNodeIndex

pub fn push_shadow(
&mut self,
shadow: Shadow,
clip_and_scroll: ScrollNodeAndClipChain,
should_inflate: bool
)

pub fn pop_all_shadows(&mut self)

fn add_shadow_prim<P>(
&mut self,
pending_shadow: &PendingShadow,
pending_primitive: &PendingPrimitive<P>,
prim_list: &mut PrimitiveList
) where
P: InternablePrimitive + CreateShadow,
Interners: AsMut<Interner<P>>,

fn add_shadow_prim_to_draw_list<P>(
&mut self,
pending_primitive: PendingPrimitive<P>
) where
P: InternablePrimitive + IsVisible,
Interners: AsMut<Interner<P>>,

fn register_chase_primitive_by_rect(
&mut self,
rect: &LayoutRect,
prim_instance: &PrimitiveInstance
)

pub fn add_solid_rectangle(
&mut self,
clip_and_scroll: ScrollNodeAndClipChain,
info: &LayoutPrimitiveInfo,
color: ColorF
)

pub fn add_clear_rectangle(
&mut self,
clip_and_scroll: ScrollNodeAndClipChain,
info: &LayoutPrimitiveInfo
)

pub fn add_line(
&mut self,
clip_and_scroll: ScrollNodeAndClipChain,
info: &LayoutPrimitiveInfo,
wavy_line_thickness: f32,
orientation: LineOrientation,
color: ColorF,
style: LineStyle
)

pub fn add_border(
&mut self,
clip_and_scroll: ScrollNodeAndClipChain,
info: &LayoutPrimitiveInfo,
border_item: &BorderDisplayItem,
gradient_stops: ItemRange<GradientStop>
)

pub fn create_linear_gradient_prim(
&mut self,
info: &LayoutPrimitiveInfo,
start_point: LayoutPoint,
end_point: LayoutPoint,
stops: ItemRange<GradientStop>,
extend_mode: ExtendMode,
stretch_size: LayoutSize,
tile_spacing: LayoutSize,
nine_patch: Option<Box<NinePatchDescriptor>>
) -> Option<LinearGradient>

pub fn create_radial_gradient_prim(
&mut self,
info: &LayoutPrimitiveInfo,
center: LayoutPoint,
start_radius: f32,
end_radius: f32,
ratio_xy: f32,
stops: ItemRange<GradientStop>,
extend_mode: ExtendMode,
stretch_size: LayoutSize,
tile_spacing: LayoutSize,
nine_patch: Option<Box<NinePatchDescriptor>>
) -> RadialGradient

pub fn add_text(
&mut self,
clip_and_scroll: ScrollNodeAndClipChain,
prim_info: &LayoutPrimitiveInfo,
font_instance_key: &FontInstanceKey,
text_color: &ColorF,
glyph_range: ItemRange<GlyphInstance>,
glyph_options: Option<GlyphOptions>
)

pub fn add_image(
&mut self,
clip_and_scroll: ScrollNodeAndClipChain,
info: &LayoutPrimitiveInfo,
stretch_size: LayoutSize,
tile_spacing: LayoutSize,
sub_rect: Option<TexelRect>,
image_key: ImageKey,
image_rendering: ImageRendering,
alpha_type: AlphaType,
color: ColorF
)

pub fn add_yuv_image(
&mut self,
clip_and_scroll: ScrollNodeAndClipChain,
info: &LayoutPrimitiveInfo,
yuv_data: YuvData,
color_depth: ColorDepth,
color_space: YuvColorSpace,
color_range: ColorRange,
image_rendering: ImageRendering
)

fn add_primitive_instance_to_3d_root(&mut self, prim: ExtendedPrimitiveInstance)

pub fn add_backdrop_filter(
&mut self,
clip_and_scroll: ScrollNodeAndClipChain,
info: &LayoutPrimitiveInfo,
filters: Vec<Filter>,
filter_datas: Vec<FilterData>,
filter_primitives: Vec<FilterPrimitive>
)

pub fn cut_backdrop_picture(&mut self) -> Option<PictureIndex>

fn wrap_prim_with_filters(
&mut self,
cur_instance: PrimitiveInstance,
current_pic_index: PictureIndex,
filter_ops: Vec<Filter>,
filter_primitives: Vec<FilterPrimitive>,
filter_datas: Vec<FilterData>,
flags: PrimitiveFlags,
requested_raster_space: RasterSpace,
spatial_node_index: SpatialNodeIndex,
inflate_if_required: bool
) -> (PictureIndex, PrimitiveInstance)