Module webrender::spatial_node

Structs

ReferenceFrameInfo

Contains information about reference frames.

ScrollFrameInfo

SpatialNode

Contains information common among all types of SpatialTree nodes.

StickyFrameInfo

Enums

ScrollFrameKind

Defines whether we have an implicit scroll frame for a pipeline root, or an explicitly defined scroll frame from the display list.

SpatialNodeType

Functions

compute_offset_from

snap_offset

Snap an offset to be incorporated into a transform, where the local space may be considered the world space. We convert from world space to device space using the global device pixel scale, which may not always be correct if there are intermediate surfaces used, however those are either cases where snapping is not important (e.g. has perspective or is not axis aligned), or an edge case (e.g. SVG filters) which we can accept imperfection for now.