Function webrender::spatial_node::snap_offset

fn snap_offset<OffsetUnits, ScaleUnits>(
    offset: Vector2D<f32, OffsetUnits>, 
    scale: Vector2D<f32, ScaleUnits>, 
    global_device_pixel_scale: DevicePixelScale
) -> Vector2D<f32, OffsetUnits>

Snap an offset to be incorporated into a transform, where the local space may be considered the world space. We convert from world space to device space using the global device pixel scale, which may not always be correct if there are intermediate surfaces used, however those are either cases where snapping is not important (e.g. has perspective or is not axis aligned), or an edge case (e.g. SVG filters) which we can accept imperfection for now.