Struct webrender::spatial_node::SpatialNode

pub struct SpatialNode {
    pub viewport_transform: ScaleOffset,
    pub content_transform: ScaleOffset,
    pub snapping_transform: Option<ScaleOffset>,
    pub coordinate_system_id: CoordinateSystemId,
    pub transform_kind: TransformedRectKind,
    pub pipeline_id: PipelineId,
    pub parent: Option<SpatialNodeIndex>,
    pub children: Vec<SpatialNodeIndex>,
    pub node_type: SpatialNodeType,
    pub invertible: bool,
    pub is_async_zooming: bool,
    pub is_ancestor_or_self_zooming: bool,
}

Contains information common among all types of SpatialTree nodes.

Fields
viewport_transform: ScaleOffset

The scale/offset of the viewport for this spatial node, relative to the coordinate system. Includes any accumulated scrolling offsets from nodes between our reference frame and this node.

content_transform: ScaleOffset

Content scale/offset relative to the coordinate system.

snapping_transform: Option<ScaleOffset>

Snapping scale/offset relative to the coordinate system. If None, then we should not snap entities bound to this spatial node.

coordinate_system_id: CoordinateSystemId

The axis-aligned coordinate system id of this node.

transform_kind: TransformedRectKind

The current transform kind of this node.

pipeline_id: PipelineId

Pipeline that this layer belongs to

parent: Option<SpatialNodeIndex>

Parent layer. If this is None, we are the root node.

children: Vec<SpatialNodeIndex>

Child layers

node_type: SpatialNodeType

The type of this node and any data associated with that node type.

invertible: bool

True if this node is transformed by an invertible transform. If not, display items transformed by this node will not be displayed and display items not transformed by this node will not be clipped by clips that are transformed by this node.

is_async_zooming: bool

Whether this specific node is currently being async zoomed. Should be set when a SetIsTransformAsyncZooming FrameMsg is received.

is_ancestor_or_self_zooming: bool

Whether this node or any of its ancestors is being pinch zoomed. This is calculated in update(). This will be used to decide whether to override corresponding picture’s raster space as an optimisation.

Implementations

impl SpatialNode

pub fn new(
pipeline_id: PipelineId,
parent_index: Option<SpatialNodeIndex>,
node_type: SpatialNodeType
) -> Self

pub fn new_scroll_frame(
pipeline_id: PipelineId,
parent_index: SpatialNodeIndex,
external_id: Option<ExternalScrollId>,
frame_rect: &LayoutRect,
content_size: &LayoutSize,
scroll_sensitivity: ScrollSensitivity,
frame_kind: ScrollFrameKind,
external_scroll_offset: LayoutVector2D
) -> Self

pub fn new_reference_frame(
parent_index: Option<SpatialNodeIndex>,
transform_style: TransformStyle,
source_transform: PropertyBinding<LayoutTransform>,
kind: ReferenceFrameKind,
origin_in_parent_reference_frame: LayoutVector2D,
pipeline_id: PipelineId
) -> Self

pub fn new_sticky_frame(
parent_index: SpatialNodeIndex,
sticky_frame_info: StickyFrameInfo,
pipeline_id: PipelineId
) -> Self

pub fn add_child(&mut self, child: SpatialNodeIndex)

pub fn apply_old_scrolling_state(&mut self, old_scroll_info: &ScrollFrameInfo)

pub fn set_scroll_origin(
&mut self,
origin: &LayoutPoint,
clamp: ScrollClamping
) -> bool

pub fn mark_uninvertible(&mut self, state: &TransformUpdateState)

pub fn update(
&mut self,
state: &mut TransformUpdateState,
coord_systems: &mut Vec<CoordinateSystem>,
global_device_pixel_scale: DevicePixelScale,
scene_properties: &SceneProperties,
previous_spatial_nodes: &[SpatialNode]
)

pub fn update_transform(
&mut self,
state: &mut TransformUpdateState,
coord_systems: &mut Vec<CoordinateSystem>,
global_device_pixel_scale: DevicePixelScale,
scene_properties: &SceneProperties,
previous_spatial_nodes: &[SpatialNode]
)

fn calculate_sticky_offset(
&self,
viewport_scroll_offset: &LayoutVector2D,
viewport_rect: &LayoutRect
) -> LayoutVector2D

pub fn prepare_state_for_children(&self, state: &mut TransformUpdateState)

pub fn scroll(&mut self, scroll_location: ScrollLocation) -> bool

pub fn scroll_offset(&self) -> LayoutVector2D

pub fn matches_external_id(&self, external_id: ExternalScrollId) -> bool

pub fn update_snapping(&mut self, parent: Option<&SpatialNode>)

Updates the snapping transform.

pub fn is_transform_bound_to_property(&self, id: PropertyBindingId) -> bool

Returns true for ReferenceFrames whose source_transform is bound to the property binding id.