Module webrender::spatial_tree

Structs

CoordinateSystem
A node in the hierarchy of coordinate system transforms.

CoordinateSystemId
An id that identifies coordinate systems in the SpatialTree. Each coordinate system has an id and those ids will be shared when the coordinates system are the same or are in the same axis-aligned space. This allows for optimizing mask generation.

SpatialNodeIndex

SpatialTree

TransformUpdateState

Enums

CoordinateSpaceMapping
Transformation between two nodes in the spatial tree that can sometimes be encoded more efficiently than with a full matrix.

TransformScroll

VisibleFace

Constants

ROOT_SPATIAL_NODE_INDEX

TOPMOST_SCROLL_NODE_INDEX

Type Definitions

ScrollStates