Module webrender::spatial_tree

Structs

CoordinateSystem

A node in the hierarchy of coordinate system transforms.

CoordinateSystemId

An id that identifies coordinate systems in the SpatialTree. Each coordinate system has an id and those ids will be shared when the coordinates system are the same or are in the same axis-aligned space. This allows for optimizing mask generation.

SceneSpatialTree

The representation of the spatial tree during scene building, which is mostly write-only, with a small number of queries for snapping, picture cache building

SpatialNodeEntry 🔒

SpatialNodeIndex

SpatialTree

Represents the spatial tree during frame building, which is mostly read-only, apart from the tree update at the start of the frame

SpatialTreeUpdates

The delta updates to apply after building a new scene to the retained frame building tree.

Store 🔒

TransformUpdateState

Enums

CoordinateSpaceMapping

Transformation between two nodes in the spatial tree that can sometimes be encoded more efficiently than with a full matrix.

SpatialTreeUpdate

StoreElement 🔒

TransformScroll 🔒

VisibleFace

Constants

MIN_SCROLLABLE_AMOUNT 🔒

MIN_SCROLL_ROOT_SIZE 🔒

Traits

SpatialNodeContainer

Allows functions and methods to retrieve common information about a spatial node, whether during scene or frame building

Functions

calculate_snapping_transform 🔒

get_external_scroll_offset

Calculate the accumulated external scroll offset for a given spatial node.