Struct webrender::spatial_tree::SceneSpatialTree

pub struct SceneSpatialTree {
    spatial_nodes: Store<SceneSpatialNode>,
    spatial_node_map: FastHashMap<SpatialNodeUid, SpatialNodeEntry>,
    root_reference_frame_index: SpatialNodeIndex,
    frame_counter: u64,
    updates: SpatialTreeUpdates,
    spatial_nodes_set: HashSet<SpatialNodeUid, BuildHasherDefault<FxHasher>>,
}

The representation of the spatial tree during scene building, which is mostly write-only, with a small number of queries for snapping, picture cache building

Fields
spatial_nodes: Store<SceneSpatialNode>

Nodes which determine the positions (offsets and transforms) for primitives and clips.

spatial_node_map: FastHashMap<SpatialNodeUid, SpatialNodeEntry>

A set of the uids we’ve encountered for spatial nodes, used to assert that we’re not seeing duplicates. Likely to be removed once we rely on this feature.

root_reference_frame_index: SpatialNodeIndexframe_counter: u64updates: SpatialTreeUpdatesspatial_nodes_set: HashSet<SpatialNodeUid, BuildHasherDefault<FxHasher>>

A debug check that the caller never adds a spatial node with duplicate uid, since that can cause badness if it occurs (e.g. a malformed spatial tree and infinite loops in is_ancestor etc)

Implementations

impl SceneSpatialTree

pub fn new() -> Self

pub fn is_root_coord_system(&self, index: SpatialNodeIndex) -> bool

pub fn end_frame_and_get_pending_updates(&mut self) -> SpatialTreeUpdates

Complete building this scene, return the updates to apply to the frame spatial tree

pub fn is_ancestor( &self, maybe_parent: SpatialNodeIndex, maybe_child: SpatialNodeIndex, ) -> bool

Check if a given spatial node is an ancestor of another spatial node.

pub fn find_scroll_root( &self, spatial_node_index: SpatialNodeIndex, allow_sticky_frames: bool, ) -> SpatialNodeIndex

Find the spatial node that is the scroll root for a given spatial node. A scroll root is the first spatial node when found travelling up the spatial node tree that is an explicit scroll frame.

pub fn root_reference_frame_index(&self) -> SpatialNodeIndex

The root reference frame, which is the true root of the SpatialTree.

fn add_spatial_node( &mut self, node: SceneSpatialNode, uid: SpatialNodeUid, ) -> SpatialNodeIndex

pub fn add_reference_frame( &mut self, parent_index: SpatialNodeIndex, transform_style: TransformStyle, source_transform: PropertyBinding<LayoutTransform>, kind: ReferenceFrameKind, origin_in_parent_reference_frame: LayoutVector2D, pipeline_id: PipelineId, uid: SpatialNodeUid, ) -> SpatialNodeIndex

pub fn add_scroll_frame( &mut self, parent_index: SpatialNodeIndex, external_id: ExternalScrollId, pipeline_id: PipelineId, frame_rect: &LayoutRect, content_size: &LayoutSize, frame_kind: ScrollFrameKind, external_scroll_offset: LayoutVector2D, scroll_offset_generation: APZScrollGeneration, has_scroll_linked_effect: HasScrollLinkedEffect, uid: SpatialNodeUid, ) -> SpatialNodeIndex

pub fn add_sticky_frame( &mut self, parent_index: SpatialNodeIndex, sticky_frame_info: StickyFrameInfo, pipeline_id: PipelineId, key: SpatialTreeItemKey, instance_id: PipelineInstanceId, ) -> SpatialNodeIndex