Struct webrender::Renderer

pub struct Renderer {
    result_rx: Receiver<ResultMsg>,
    debug_server: Box<dyn DebugServer>,
    pub device: Device,
    pending_texture_updates: Vec<TextureUpdateList>,
    pending_gpu_cache_updates: Vec<GpuCacheUpdateList>,
    pending_gpu_cache_clear: bool,
    pending_shader_updates: Vec<PathBuf>,
    active_documents: Vec<(DocumentId, RenderedDocument)>,
    shaders: Rc<RefCell<Shaders>>,
    max_recorded_profiles: usize,
    clear_color: Option<ColorF>,
    enable_clear_scissor: bool,
    enable_advanced_blend_barriers: bool,
    debug: LazyInitializedDebugRenderer,
    debug_flags: DebugFlags,
    backend_profile_counters: BackendProfileCounters,
    profile_counters: RendererProfileCounters,
    resource_upload_time: u64,
    gpu_cache_upload_time: u64,
    profiler: Profiler,
    new_frame_indicator: ChangeIndicator,
    new_scene_indicator: ChangeIndicator,
    slow_frame_indicator: ChangeIndicator,
    last_time: u64,
    pub gpu_profile: GpuProfiler<GpuProfileTag>,
    vaos: RendererVAOs,
    prim_header_f_texture: VertexDataTexture<PrimitiveHeaderF>,
    prim_header_i_texture: VertexDataTexture<PrimitiveHeaderI>,
    transforms_texture: VertexDataTexture<TransformData>,
    render_task_texture: VertexDataTexture<RenderTaskData>,
    gpu_cache_texture: GpuCacheTexture,
    gpu_cache_debug_chunks: Vec<Vec<GpuCacheDebugChunk>>,
    gpu_cache_frame_id: FrameId,
    gpu_cache_overflow: bool,
    pipeline_info: PipelineInfo,
    texture_resolver: TextureResolver,
    texture_cache_upload_pbo: PBO,
    dither_matrix_texture: Option<Texture>,
    external_image_handler: Option<Box<dyn ExternalImageHandler>>,
    output_image_handler: Option<Box<dyn OutputImageHandler>>,
    size_of_ops: Option<MallocSizeOfOps>,
    output_targets: HashMap<u32, FrameOutput, BuildHasherDefault<FxHasher>>,
    pub renderer_errors: Vec<RendererError>,
    pub(in crate) async_frame_recorder: Option<AsyncScreenshotGrabber>,
    pub(in crate) async_screenshots: Option<AsyncScreenshotGrabber>,
    cpu_profiles: VecDeque<CpuProfile>,
    gpu_profiles: VecDeque<GpuProfile>,
    notifications: Vec<NotificationRequest>,
    device_size: Option<DeviceIntSize>,
    zoom_debug_texture: Option<Texture>,
    cursor_position: DeviceIntPoint,
    shared_texture_cache_cleared: bool,
    documents_seen: HashSet<DocumentId, BuildHasherDefault<FxHasher>>,
    read_fbo: FBOId,
    compositor_config: CompositorConfig,
    allocated_native_surfaces: HashSet<NativeSurfaceId, BuildHasherDefault<FxHasher>>,
    force_redraw: bool,
    debug_overlay_state: DebugOverlayState,
}

The renderer is responsible for submitting to the GPU the work prepared by the RenderBackend.

We have a separate Renderer instance for each instance of WebRender (generally one per OS window), and all instances share the same thread.

Fields
result_rx: Receiver<ResultMsg>debug_server: Box<dyn DebugServer>device: Devicepending_texture_updates: Vec<TextureUpdateList>pending_gpu_cache_updates: Vec<GpuCacheUpdateList>pending_gpu_cache_clear: boolpending_shader_updates: Vec<PathBuf>active_documents: Vec<(DocumentId, RenderedDocument)>shaders: Rc<RefCell<Shaders>>max_recorded_profiles: usizeclear_color: Option<ColorF>enable_clear_scissor: boolenable_advanced_blend_barriers: booldebug: LazyInitializedDebugRendererdebug_flags: DebugFlagsbackend_profile_counters: BackendProfileCountersprofile_counters: RendererProfileCountersresource_upload_time: u64gpu_cache_upload_time: u64profiler: Profilernew_frame_indicator: ChangeIndicatornew_scene_indicator: ChangeIndicatorslow_frame_indicator: ChangeIndicatorlast_time: u64gpu_profile: GpuProfiler<GpuProfileTag>vaos: RendererVAOsprim_header_f_texture: VertexDataTexture<PrimitiveHeaderF>prim_header_i_texture: VertexDataTexture<PrimitiveHeaderI>transforms_texture: VertexDataTexture<TransformData>render_task_texture: VertexDataTexture<RenderTaskData>gpu_cache_texture: GpuCacheTexturegpu_cache_debug_chunks: Vec<Vec<GpuCacheDebugChunk>>

When the GPU cache debugger is enabled, we keep track of the live blocks in the GPU cache so that we can use them for the debug display. This member stores those live blocks, indexed by row.

gpu_cache_frame_id: FrameIdgpu_cache_overflow: boolpipeline_info: PipelineInfotexture_resolver: TextureResolvertexture_cache_upload_pbo: PBOdither_matrix_texture: Option<Texture>external_image_handler: Option<Box<dyn ExternalImageHandler>>

Optional trait object that allows the client application to provide external buffers for image data.

output_image_handler: Option<Box<dyn OutputImageHandler>>

Optional trait object that allows the client application to provide a texture handle to copy the WR output to.

size_of_ops: Option<MallocSizeOfOps>

Optional function pointers for measuring memory used by a given heap-allocated pointer.

output_targets: HashMap<u32, FrameOutput, BuildHasherDefault<FxHasher>>renderer_errors: Vec<RendererError>async_frame_recorder: Option<AsyncScreenshotGrabber>async_screenshots: Option<AsyncScreenshotGrabber>cpu_profiles: VecDeque<CpuProfile>

List of profile results from previous frames. Can be retrieved via get_frame_profiles().

gpu_profiles: VecDeque<GpuProfile>notifications: Vec<NotificationRequest>

Notification requests to be fulfilled after rendering.

device_size: Option<DeviceIntSize>zoom_debug_texture: Option<Texture>

A lazily created texture for the zoom debugging widget.

cursor_position: DeviceIntPoint

The current mouse position. This is used for debugging functionality only, such as the debug zoom widget.

shared_texture_cache_cleared: bool

Guards to check if we might be rendering a frame with expired texture cache entries.

documents_seen: HashSet<DocumentId, BuildHasherDefault<FxHasher>>

The set of documents which we've seen a publish for since last render.

read_fbo: FBOIdcompositor_config: CompositorConfig

The compositing config, affecting how WR composites into the final scene.

allocated_native_surfaces: HashSet<NativeSurfaceId, BuildHasherDefault<FxHasher>>

Maintains a set of allocated native composite surfaces. This allows any currently allocated surfaces to be cleaned up as soon as deinit() is called (the normal bookkeeping for native surfaces exists in the render backend thread).

force_redraw: bool

If true, partial present state has been reset and everything needs to be drawn on the next render.

debug_overlay_state: DebugOverlayState

State related to the debug / profiling overlays

Methods

impl Renderer

pub fn new(
gl: Rc<dyn Gl>,
notifier: Box<dyn RenderNotifier>,
options: RendererOptions,
shaders: Option<&mut WrShaders>,
start_size: DeviceIntSize
) -> Result<(Self, RenderApiSender), RendererError>

Initializes WebRender and creates a Renderer and RenderApiSender.

Examples

Initializes a Renderer with some reasonable values. For more information see RendererOptions.

let opts = webrender::RendererOptions {
   device_pixel_ratio: 1.0,
   resource_override_path: None,
   enable_aa: false,
};
let (renderer, sender) = Renderer::new(opts);

pub fn device_size(&self) -> Option<DeviceIntSize>

pub fn set_cursor_position(&mut self, position: DeviceIntPoint)

Update the current position of the debug cursor.

pub fn get_max_texture_size(&self) -> i32

pub fn get_graphics_api_info(&self) -> GraphicsApiInfo

pub fn preferred_color_format(&self) -> ImageFormat

pub fn optimal_texture_stride_alignment(&self) -> usize

pub fn flush_pipeline_info(&mut self) -> PipelineInfo

pub fn current_epoch(
&self,
document_id: DocumentId,
pipeline_id: PipelineId
) -> Option<Epoch>

Returns the Epoch of the current frame in a pipeline.

pub fn update_program_cache(&mut self, cached_programs: Rc<ProgramCache>)

pub fn update(&mut self)

Processes the result queue.

Should be called before render(), as texture cache updates are done here.

fn get_screenshot_for_debugger(&mut self) -> String

fn get_passes_for_debugger(&self) -> String

fn get_render_tasks_for_debugger(&self) -> String

fn handle_debug_command(&mut self, command: DebugCommand)

pub fn set_external_image_handler(
&mut self,
handler: Box<dyn ExternalImageHandler>
)

Set a callback for handling external images.

pub fn set_output_image_handler(&mut self, handler: Box<dyn OutputImageHandler>)

Set a callback for handling external outputs.

pub fn get_frame_profiles(&mut self) -> (Vec<CpuProfile>, Vec<GpuProfile>)

Retrieve (and clear) the current list of recorded frame profiles.

pub fn notify_slow_frame(&mut self)

pub fn force_redraw(&mut self)

Reset the current partial present state. This forces the entire framebuffer to be refreshed next time render is called.

pub fn render(
&mut self,
device_size: DeviceIntSize
) -> Result<RenderResults, Vec<RendererError>>

Renders the current frame.

A Frame is supplied by calling [generate_frame()][webrender_api::Transaction::generate_frame].

fn update_debug_overlay(&mut self, framebuffer_size: DeviceIntSize)

Update the state of any debug / profiler overlays. This is currently only needed when running with the native compositor enabled.

fn bind_debug_overlay(&mut self)

Bind a draw target for the debug / profiler overlays, if required.

fn unbind_debug_overlay(&mut self)

Unbind the draw target for debug / profiler overlays, if required.

fn render_impl(
&mut self,
device_size: Option<DeviceIntSize>
) -> Result<RenderResults, Vec<RendererError>>

fn update_gpu_cache(&mut self)

fn prepare_gpu_cache(&mut self, frame: &Frame)

fn update_texture_cache(&mut self)

pub(crate) fn draw_instanced_batch<T>(
&mut self,
data: &[T],
vertex_array_kind: VertexArrayKind,
textures: &BatchTextures,
stats: &mut RendererStats
)

pub(crate) fn draw_instanced_batch_with_previously_bound_textures<T>(
&mut self,
data: &[T],
vertex_array_kind: VertexArrayKind,
stats: &mut RendererStats
)

fn handle_readback_composite(
&mut self,
draw_target: DrawTarget,
uses_scissor: bool,
source: &RenderTask,
backdrop: &RenderTask,
readback: &RenderTask
)

fn handle_blits(
&mut self,
blits: &[BlitJob],
render_tasks: &RenderTaskGraph,
draw_target: DrawTarget,
content_origin: &DeviceIntPoint
)

fn handle_scaling(
&mut self,
scalings: &HashMap<TextureSource, Vec<ScalingInstance>, BuildHasherDefault<FxHasher>>,
projection: &Transform3D<f32>,
stats: &mut RendererStats
)

fn handle_svg_filters(
&mut self,
textures: &BatchTextures,
svg_filters: &[SvgFilterInstance],
projection: &Transform3D<f32>,
stats: &mut RendererStats
)

fn draw_picture_cache_target(
&mut self,
target: &PictureCacheTarget,
draw_target: DrawTarget,
content_origin: DeviceIntPoint,
projection: &Transform3D<f32>,
render_tasks: &RenderTaskGraph,
stats: &mut RendererStats
)

fn draw_alpha_batch_container(
&mut self,
alpha_batch_container: &AlphaBatchContainer,
draw_target: DrawTarget,
content_origin: DeviceIntPoint,
framebuffer_kind: FramebufferKind,
projection: &Transform3D<f32>,
render_tasks: &RenderTaskGraph,
stats: &mut RendererStats
)

Draw an alpha batch container into a given draw target. This is used by both color and picture cache target kinds.

fn draw_tile_list<'a, I: Iterator<Item = &'a CompositeTile>>(
&mut self,
tiles_iter: I,
partial_present_mode: Option<PartialPresentMode>,
stats: &mut RendererStats
)

Draw a list of tiles to the framebuffer

fn composite_simple(
&mut self,
composite_state: &CompositeState,
clear_framebuffer: bool,
draw_target: DrawTarget,
projection: &Transform3D<f32>,
results: &mut RenderResults,
max_partial_present_rects: usize
)

Composite picture cache tiles into the framebuffer. This is currently the only way that picture cache tiles get drawn. In future, the tiles will often be handed to the OS compositor, and this method will be rarely used.

fn draw_color_target(
&mut self,
draw_target: DrawTarget,
target: &ColorRenderTarget,
content_origin: DeviceIntPoint,
clear_color: Option<[f32; 4]>,
clear_depth: Option<f32>,
render_tasks: &RenderTaskGraph,
projection: &Transform3D<f32>,
frame_id: GpuFrameId,
stats: &mut RendererStats
)

fn draw_clip_batch_list(
&mut self,
list: &ClipBatchList,
projection: &Transform3D<f32>,
stats: &mut RendererStats
)

Draw all the instances in a clip batcher list to the current target.

fn draw_alpha_target(
&mut self,
draw_target: DrawTarget,
target: &AlphaRenderTarget,
projection: &Transform3D<f32>,
render_tasks: &RenderTaskGraph,
stats: &mut RendererStats
)

fn draw_texture_cache_target(
&mut self,
texture: &CacheTextureId,
layer: usize,
target: &TextureCacheRenderTarget,
render_tasks: &RenderTaskGraph,
stats: &mut RendererStats
)

fn update_deferred_resolves(
&mut self,
deferred_resolves: &[DeferredResolve]
) -> Option<GpuCacheUpdateList>

fn unlock_external_images(&mut self)

fn allocate_target_texture<T: RenderTarget>(
&mut self,
list: &mut RenderTargetList<T>,
counters: &mut FrameProfileCounters
) -> Option<ActiveTexture>

Allocates a texture to be used as the output for a rendering pass.

We make an effort to reuse render targe textures across passes and across frames when the format and dimensions match. Because we use immutable storage, we can't resize textures.

We could consider approaches to re-use part of a larger target, if available. However, we'd need to be careful about eviction. Currently, render targets are freed if they haven't been used in 30 frames. If we used partial targets, we'd need to track how much of the target has been used in the last 30 frames, since we could otherwise end up keeping an enormous target alive indefinitely by constantly using it in situations where a much smaller target would suffice.

fn bind_frame_data(&mut self, frame: &mut Frame)

fn update_native_surfaces(&mut self, composite_state: &mut CompositeState)

fn draw_frame(
&mut self,
frame: &mut Frame,
device_size: Option<DeviceIntSize>,
frame_id: GpuFrameId,
results: &mut RenderResults,
clear_framebuffer: bool
)

pub fn init_pixel_local_storage(
&mut self,
task_rect: DeviceIntRect,
projection: &Transform3D<f32>,
stats: &mut RendererStats
)

Initialize the PLS block, by reading the current framebuffer color.

pub fn resolve_pixel_local_storage(
&mut self,
task_rect: DeviceIntRect,
projection: &Transform3D<f32>,
stats: &mut RendererStats
)

Resolve the current PLS structure, writing it to a fragment color output.

pub fn debug_renderer<'b>(&'b mut self) -> Option<&'b mut DebugRenderer>

pub fn get_debug_flags(&self) -> DebugFlags

pub fn set_debug_flags(&mut self, flags: DebugFlags)

pub fn save_cpu_profile(&self, filename: &str)

fn draw_frame_debug_items(&mut self, items: &[DebugItem])

fn draw_render_target_debug(&mut self, device_size: DeviceIntSize)

fn draw_zoom_debug(&mut self, device_size: DeviceIntSize)

fn draw_texture_cache_debug(&mut self, device_size: DeviceIntSize)

fn do_debug_blit(
device: &mut Device,
debug_renderer: &mut DebugRenderer,
textures: Vec<&Texture>,
device_size: DeviceIntSize,
bottom: i32,
select_color: &dyn Fn(&Texture) -> [f32; 4]
)

fn draw_epoch_debug(&mut self)

fn draw_gpu_cache_debug(&mut self, device_size: DeviceIntSize)

pub fn read_pixels_into(
&mut self,
rect: FramebufferIntRect,
format: ImageFormat,
output: &mut [u8]
)

Pass-through to Device::read_pixels_into, used by Gecko's WR bindings.

pub fn read_pixels_rgba8(&mut self, rect: FramebufferIntRect) -> Vec<u8>

pub fn read_gpu_cache(&mut self) -> (DeviceIntSize, Vec<u8>)

pub fn deinit(self)

fn size_of<T>(&self, ptr: *const T) -> usize

pub fn report_memory(&self) -> MemoryReport

Collects a memory report.

fn set_blend(&self, blend: bool, framebuffer_kind: FramebufferKind)

fn set_blend_mode_multiply(&self, framebuffer_kind: FramebufferKind)

fn set_blend_mode_premultiplied_alpha(&self, framebuffer_kind: FramebufferKind)

fn set_blend_mode_subpixel_with_bg_color_pass1(
&self,
framebuffer_kind: FramebufferKind
)

fn set_blend_mode_subpixel_with_bg_color_pass2(
&self,
framebuffer_kind: FramebufferKind
)

fn clear_texture(&mut self, texture: &Texture, color: [f32; 4])

Clears all the layers of a texture with a given color.

impl Renderer

fn save_texture(
texture: &Texture,
name: &str,
root: &PathBuf,
device: &mut Device
) -> PlainTexture

fn save_capture(
&mut self,
config: CaptureConfig,
deferred_images: Vec<ExternalCaptureImage>
)

impl Renderer

pub fn record_frame(
&mut self,
image_format: ImageFormat
) -> Option<(RecordedFrameHandle, DeviceIntSize)>

Record a frame for the Composition Recorder.

The returned handle can be passed to map_recorded_frame to copy it into a buffer. The returned size is the size of the frame.

pub fn map_recorded_frame(
&mut self,
handle: RecordedFrameHandle,
dst_buffer: &mut [u8],
dst_stride: usize
) -> bool

Map a frame captured for the composition recorder into the given buffer.

pub fn release_composition_recorder_structures(&mut self)

Free the data structures used by the composition recorder.

pub fn get_screenshot_async(
&mut self,
window_rect: DeviceIntRect,
buffer_size: DeviceIntSize,
image_format: ImageFormat
) -> (AsyncScreenshotHandle, DeviceIntSize)

Take a screenshot and scale it asynchronously.

The returned handle can be used to access the mapped screenshot data via map_and_recycle_screenshot.

The returned size is the size of the screenshot.

pub fn map_and_recycle_screenshot(
&mut self,
handle: AsyncScreenshotHandle,
dst_buffer: &mut [u8],
dst_stride: usize
) -> bool

Map the contents of the screenshot given by the handle and copy it into the given buffer.

pub fn release_profiler_structures(&mut self)

Release the screenshot grabbing structures that the profiler was using.