Struct webrender::Renderer

pub struct Renderer { /* fields omitted */ }

The renderer is responsible for submitting to the GPU the work prepared by the RenderBackend.

Methods

impl Renderer

pub fn new(
gl: Rc<Gl>,
notifier: Box<RenderNotifier>,
options: RendererOptions
) -> Result<(Renderer, RenderApiSender), RendererError>

Initializes webrender and creates a Renderer and RenderApiSender.

Examples

Initializes a Renderer with some reasonable values. For more information see RendererOptions.

let opts = webrender::RendererOptions {
   device_pixel_ratio: 1.0,
   resource_override_path: None,
   enable_aa: false,
};
let (renderer, sender) = Renderer::new(opts);

pub fn get_max_texture_size(&self) -> u32

pub fn get_graphics_api_info(&self) -> GraphicsApiInfo

pub fn current_epoch(&self, pipeline_id: PipelineId) -> Option<Epoch>

Returns the Epoch of the current frame in a pipeline.

pub fn flush_rendered_epochs(
&mut self
) -> HashMap<PipelineId, Epoch, BuildHasherDefault<FxHasher>>

Returns a HashMap containing the pipeline ids that have been received by the renderer and their respective epochs since the last time the method was called.

pub fn update_program_cache(&mut self, cached_programs: Rc<ProgramCache>)

pub fn update(&mut self)

Processes the result queue.

Should be called before render(), as texture cache updates are done here.

pub fn set_external_image_handler(&mut self, handler: Box<ExternalImageHandler>)

Set a callback for handling external images.

pub fn set_output_image_handler(&mut self, handler: Box<OutputImageHandler>)

Set a callback for handling external outputs.

pub fn get_frame_profiles(&mut self) -> (Vec<CpuProfile>, Vec<GpuProfile>)

Retrieve (and clear) the current list of recorded frame profiles.

pub fn render(
&mut self,
framebuffer_size: DeviceUintSize
) -> Result<RendererStats, Vec<RendererError>>

Renders the current frame.

A Frame is supplied by calling generate_frame().

pub fn layers_are_bouncing_back(&self) -> bool

pub fn debug_renderer<'b>(&'b mut self) -> &'b mut DebugRenderer

pub fn get_debug_flags(&self) -> DebugFlags

pub fn set_debug_flags(&mut self, flags: DebugFlags)

pub fn set_debug_flag(&mut self, flag: DebugFlags, enabled: bool)

pub fn toggle_debug_flags(&mut self, toggle: DebugFlags)

pub fn save_cpu_profile(&self, filename: &str)

pub fn read_pixels_rgba8(&self, rect: DeviceUintRect) -> Vec<u8>

pub fn read_gpu_cache(&mut self) -> (DeviceUintSize, Vec<u8>)

pub fn read_pixels_into(
&self,
rect: DeviceUintRect,
format: ReadPixelsFormat,
output: &mut [u8]
)

pub fn deinit(self)