Struct webrender::Shaders

pub struct Shaders {
    pub cs_blur_a8: LazilyCompiledShader,
    pub cs_blur_rgba8: LazilyCompiledShader,
    pub cs_border_segment: LazilyCompiledShader,
    pub cs_border_solid: LazilyCompiledShader,
    pub cs_scale: LazilyCompiledShader,
    pub cs_line_decoration: LazilyCompiledShader,
    pub cs_gradient: LazilyCompiledShader,
    pub cs_svg_filter: LazilyCompiledShader,
    brush_solid: BrushShader,
    brush_image: Vec<Option<BrushShader>>,
    brush_fast_image: Vec<Option<BrushShader>>,
    brush_blend: BrushShader,
    brush_mix_blend: BrushShader,
    brush_yuv_image: Vec<Option<BrushShader>>,
    brush_radial_gradient: BrushShader,
    brush_linear_gradient: BrushShader,
    pub cs_clip_rectangle_slow: LazilyCompiledShader,
    pub cs_clip_rectangle_fast: LazilyCompiledShader,
    pub cs_clip_box_shadow: LazilyCompiledShader,
    pub cs_clip_image: LazilyCompiledShader,
    pub ps_text_run: TextShader,
    pub ps_text_run_dual_source: TextShader,
    pub pls_init: LazilyCompiledShader,
    pub pls_resolve: LazilyCompiledShader,
    ps_split_composite: LazilyCompiledShader,
    pub composite: LazilyCompiledShader,
}

Fields
cs_blur_a8: LazilyCompiledShadercs_blur_rgba8: LazilyCompiledShadercs_border_segment: LazilyCompiledShadercs_border_solid: LazilyCompiledShadercs_scale: LazilyCompiledShadercs_line_decoration: LazilyCompiledShadercs_gradient: LazilyCompiledShadercs_svg_filter: LazilyCompiledShaderbrush_solid: BrushShaderbrush_image: Vec<Option<BrushShader>>brush_fast_image: Vec<Option<BrushShader>>brush_blend: BrushShaderbrush_mix_blend: BrushShaderbrush_yuv_image: Vec<Option<BrushShader>>brush_radial_gradient: BrushShaderbrush_linear_gradient: BrushShadercs_clip_rectangle_slow: LazilyCompiledShader

These are "cache clip shaders". These shaders are used to draw clip instances into the cached clip mask. The results of these shaders are also used by the primitive shaders.

cs_clip_rectangle_fast: LazilyCompiledShadercs_clip_box_shadow: LazilyCompiledShadercs_clip_image: LazilyCompiledShaderps_text_run: TextShaderps_text_run_dual_source: TextShaderpls_init: LazilyCompiledShaderpls_resolve: LazilyCompiledShaderps_split_composite: LazilyCompiledShadercomposite: LazilyCompiledShader
Methods

impl Shaders

pub fn new(
device: &mut Device,
gl_type: GlType,
options: &RendererOptions
) -> Result<Self, ShaderError>

fn get_yuv_shader_index(buffer_kind: ImageBufferKind) -> usize

pub fn get(
&mut self,
key: &BatchKey,
features: BatchFeatures,
debug_flags: DebugFlags
) -> &mut LazilyCompiledShader

pub fn deinit(self, device: &mut Device)