Struct webrender::Shaders

pub struct Shaders {

    pub cs_blur_a8: LazilyCompiledShader,
    pub cs_blur_rgba8: LazilyCompiledShader,
    pub cs_border_segment: LazilyCompiledShader,
    pub cs_border_solid: LazilyCompiledShader,
    pub cs_scale: Vec<Option<LazilyCompiledShader>>,
    pub cs_line_decoration: LazilyCompiledShader,
    pub cs_fast_linear_gradient: LazilyCompiledShader,
    pub cs_linear_gradient: LazilyCompiledShader,
    pub cs_radial_gradient: LazilyCompiledShader,
    pub cs_conic_gradient: LazilyCompiledShader,
    pub cs_svg_filter: LazilyCompiledShader,
    pub cs_svg_filter_node: LazilyCompiledShader,
    brush_solid: BrushShader,
    brush_image: Vec<Option<BrushShader>>,
    brush_fast_image: Vec<Option<BrushShader>>,
    brush_blend: BrushShader,
    brush_mix_blend: BrushShader,
    brush_yuv_image: Vec<Option<BrushShader>>,
    brush_linear_gradient: BrushShader,
    brush_opacity: BrushShader,
    brush_opacity_aa: BrushShader,
    pub cs_clip_rectangle_slow: LazilyCompiledShader,
    pub cs_clip_rectangle_fast: LazilyCompiledShader,
    pub cs_clip_box_shadow: LazilyCompiledShader,
    pub ps_text_run: TextShader,
    pub ps_text_run_dual_source: Option<TextShader>,
    ps_split_composite: LazilyCompiledShader,
    pub ps_quad_textured: LazilyCompiledShader,
    pub ps_quad_radial_gradient: LazilyCompiledShader,
    pub ps_quad_conic_gradient: LazilyCompiledShader,
    pub ps_mask: LazilyCompiledShader,
    pub ps_mask_fast: LazilyCompiledShader,
    pub ps_clear: LazilyCompiledShader,
    pub ps_copy: LazilyCompiledShader,
    pub composite: CompositorShaders,

}

Fields
cs_blur_a8: LazilyCompiledShadercs_blur_rgba8: LazilyCompiledShadercs_border_segment: LazilyCompiledShadercs_border_solid: LazilyCompiledShadercs_scale: Vec<Option<LazilyCompiledShader>>cs_line_decoration: LazilyCompiledShadercs_fast_linear_gradient: LazilyCompiledShadercs_linear_gradient: LazilyCompiledShadercs_radial_gradient: LazilyCompiledShadercs_conic_gradient: LazilyCompiledShadercs_svg_filter: LazilyCompiledShadercs_svg_filter_node: LazilyCompiledShaderbrush_solid: BrushShaderbrush_image: Vec<Option<BrushShader>>brush_fast_image: Vec<Option<BrushShader>>brush_blend: BrushShaderbrush_mix_blend: BrushShaderbrush_yuv_image: Vec<Option<BrushShader>>brush_linear_gradient: BrushShaderbrush_opacity: BrushShaderbrush_opacity_aa: BrushShadercs_clip_rectangle_slow: LazilyCompiledShader

These are “cache clip shaders”. These shaders are used to draw clip instances into the cached clip mask. The results of these shaders are also used by the primitive shaders.

cs_clip_rectangle_fast: LazilyCompiledShadercs_clip_box_shadow: LazilyCompiledShaderps_text_run: TextShaderps_text_run_dual_source: Option<TextShader>ps_split_composite: LazilyCompiledShaderps_quad_textured: LazilyCompiledShaderps_quad_radial_gradient: LazilyCompiledShaderps_quad_conic_gradient: LazilyCompiledShaderps_mask: LazilyCompiledShaderps_mask_fast: LazilyCompiledShaderps_clear: LazilyCompiledShaderps_copy: LazilyCompiledShadercomposite: CompositorShaders
Implementations

impl Shaders

pub fn new( device: &mut Device, gl_type: GlType, options: &WebRenderOptions, ) -> Result<Self, ShaderError>

fn get_compositing_shader_index(buffer_kind: ImageBufferKind) -> usize

pub fn get_composite_shader( &mut self, format: CompositeSurfaceFormat, buffer_kind: ImageBufferKind, features: CompositeFeatures, ) -> &mut LazilyCompiledShader

pub fn get_scale_shader( &mut self, buffer_kind: ImageBufferKind, ) -> &mut LazilyCompiledShader

pub fn get_quad_shader( &mut self, pattern: PatternKind, ) -> &mut LazilyCompiledShader

pub fn get( &mut self, key: &BatchKey, features: BatchFeatures, debug_flags: DebugFlags, device: &Device, ) -> &mut LazilyCompiledShader

pub fn deinit(self, device: &mut Device)