Struct webrender::texture_cache::TextureCache

pub struct TextureCache {

    shared_textures: SharedTextures,
    picture_textures: PictureTextures,
    max_texture_size: i32,
    max_texture_layers: usize,
    swizzle: Option<SwizzleSettings>,
    debug_flags: DebugFlags,
    next_id: CacheTextureId,
    pending_updates: TextureUpdateList,
    now: FrameStamp,
    picture_cache_handles: Vec<FreeListHandle<CacheEntryMarker>>,
    lru_cache: LRUCache<CacheEntry, CacheEntryMarker>,
    manual_handles: Vec<FreeListHandle<CacheEntryMarker>>,
    shared_bytes_allocated: usize,
    standalone_bytes_allocated: usize,
    eviction_threshold_bytes: usize,
    max_evictions_per_frame: usize,

}

General-purpose manager for images in GPU memory. This includes images, rasterized glyphs, rasterized blobs, cached render tasks, etc.

The texture cache is owned and managed by the RenderBackend thread, and produces a series of commands to manipulate the textures on the Renderer thread. These commands are executed before any rendering is performed for a given frame.

Entries in the texture cache are not guaranteed to live past the end of the frame in which they are requested, and may be evicted. The API supports querying whether an entry is still available.

The TextureCache is different from the GpuCache in that the former stores images, whereas the latter stores data and parameters for use in the shaders. This means that the texture cache can be visualized, which is a good way to understand how it works. Enabling gfx.webrender.debug.texture-cache shows a live view of its contents in Firefox.

Fields
shared_textures: SharedTextures

Set of texture arrays in different formats used for the shared cache.

picture_textures: PictureTextures

A texture array per tile size for picture caching.

max_texture_size: i32

Maximum texture size supported by hardware.

max_texture_layers: usize

Maximum number of texture layers supported by hardware.

swizzle: Option<SwizzleSettings>

Settings on using texture unit swizzling.

debug_flags: DebugFlags

The current set of debug flags.

next_id: CacheTextureId

The next unused virtual texture ID. Monotonically increasing.

pending_updates: TextureUpdateList

A list of allocations and updates that need to be applied to the texture cache in the rendering thread this frame.

now: FrameStamp

The current FrameStamp. Used for cache eviction policies.

picture_cache_handles: Vec<FreeListHandle<CacheEntryMarker>>

List of picture cache entries. These are maintained separately from regular texture cache entries.

lru_cache: LRUCache<CacheEntry, CacheEntryMarker>

Cache of texture cache handles with automatic lifetime management, evicted in a least-recently-used order (except those entries with manual eviction enabled).

manual_handles: Vec<FreeListHandle<CacheEntryMarker>>

A list of texture cache handles that have been set to explicitly have manual eviction policy enabled. The handles reference cache entries in the lru_cache above, but have opted in to manual lifetime management.

shared_bytes_allocated: usize

Estimated memory usage of allocated entries in all of the shared textures. This is used to decide when to evict old items from the cache.

standalone_bytes_allocated: usize

Number of bytes allocated in standalone textures. Used as an input to deciding when to run texture cache eviction.

eviction_threshold_bytes: usize

If the total bytes allocated in shared / standalone cache is less than this, then allow the cache to grow without forcing an eviction.

max_evictions_per_frame: usize

The maximum number of items that will be evicted per frame. This limit helps avoid jank on frames where we want to evict a large number of items. Instead, we’d prefer to drop the items incrementally over a number of frames, even if that means the total allocated size of the cache is above the desired threshold for a small number of frames.

Implementations

impl TextureCache

pub fn new(
max_texture_size: i32,
max_texture_layers: usize,
picture_tile_sizes: &[DeviceIntSize],
initial_size: DeviceIntSize,
color_formats: TextureFormatPair<ImageFormat>,
swizzle: Option<SwizzleSettings>,
eviction_threshold_bytes: usize,
max_evictions_per_frame: usize
) -> Self

pub fn set_debug_flags(&mut self, flags: DebugFlags)

pub fn clear_all(&mut self)

Clear all entries in the texture cache. This is a fairly drastic step that should only be called very rarely.

pub fn begin_frame(&mut self, stamp: FrameStamp)

Called at the beginning of each frame.

pub fn end_frame(
&mut self,
texture_cache_profile: &mut TextureCacheProfileCounters
)

pub fn request(
&mut self,
handle: &WeakFreeListHandle<CacheEntryMarker>,
gpu_cache: &mut GpuCache
) -> bool

pub fn needs_upload(
&self,
handle: &WeakFreeListHandle<CacheEntryMarker>
) -> bool

pub fn max_texture_size(&self) -> i32

pub fn pending_updates(&mut self) -> TextureUpdateList

pub fn update(
&mut self,
handle: &mut WeakFreeListHandle<CacheEntryMarker>,
descriptor: ImageDescriptor,
filter: TextureFilter,
data: Option<CachedImageData>,
user_data: [f32; 3],
dirty_rect: ImageDirtyRect,
gpu_cache: &mut GpuCache,
eviction_notice: Option<&EvictionNotice>,
uv_rect_kind: UvRectKind,
eviction: Eviction
)

pub fn is_allocated(
&self,
handle: &WeakFreeListHandle<CacheEntryMarker>
) -> bool

pub fn is_recently_used(
&self,
handle: &WeakFreeListHandle<CacheEntryMarker>,
margin: usize
) -> bool

pub fn get_allocated_size(
&self,
handle: &WeakFreeListHandle<CacheEntryMarker>
) -> Option<usize>

pub fn get(&self, handle: &WeakFreeListHandle<CacheEntryMarker>) -> CacheItem

pub fn get_cache_location(
&self,
handle: &WeakFreeListHandle<CacheEntryMarker>
) -> (CacheTextureId, usize, DeviceIntRect, Swizzle, GpuCacheHandle)

A more detailed version of get(). This allows access to the actual device rect of the cache allocation.

Returns a tuple identifying the texture, the layer, the region, and its GPU handle.

fn evict_impl(&mut self, handle: FreeListHandle<CacheEntryMarker>)

Internal helper function to evict a strong texture cache handle

pub fn evict_manual_handle(
&mut self,
handle: &WeakFreeListHandle<CacheEntryMarker>
)

Evict a texture cache handle that was previously set to be in manual eviction mode.

fn expire_old_picture_cache_tiles(&mut self)

Expire picture cache tiles that haven’t been referenced in the last frame. The picture cache code manually keeps tiles alive by calling request on them if it wants to retain a tile that is currently not visible.

fn evict_items_from_cache_if_required(&mut self)

Evict old items from the shared and standalone caches, if we’re over a threshold memory usage value

fn current_memory_estimate(&self) -> usize

Return the total used bytes in standalone and shared textures. This is used to determine how many textures need to be evicted to keep texture cache memory usage under a reasonable limit. Note that this does not include memory allocated to picture cache tiles, which are considered separately for the purposes of texture cache eviction.

fn free(&mut self, entry: &CacheEntry)

fn allocate_from_shared_cache(
&mut self,
params: &CacheAllocParams
) -> CacheEntry

Allocate a block from the shared cache.

pub fn is_allowed_in_shared_cache(
&self,
filter: TextureFilter,
descriptor: &ImageDescriptor
) -> bool

fn allocate_standalone_entry(&mut self, params: &CacheAllocParams) -> CacheEntry

Allocates a new standalone cache entry.

fn allocate_cache_entry(&mut self, params: &CacheAllocParams) -> CacheEntry

Allocates a cache entry appropriate for the given parameters.

This allocates from the shared cache unless the parameters do not meet the shared cache requirements, in which case a standalone texture is used.

fn allocate(
&mut self,
params: &CacheAllocParams,
handle: &mut WeakFreeListHandle<CacheEntryMarker>
)

Allocates a cache entry for the given parameters, and updates the provided handle to point to the new entry.

pub fn update_picture_cache(
&mut self,
tile_size: DeviceIntSize,
handle: &mut WeakFreeListHandle<CacheEntryMarker>,
gpu_cache: &mut GpuCache
)

pub fn shared_alpha_expected_format(&self) -> ImageFormat

pub fn shared_color_expected_format(&self) -> ImageFormat