Module webrender::tile_cache

Structs

DirtyTileDebugInfo

Debug information about a tile that was dirty and was rasterized

PictureCacheDebugInfo

Debug information about a set of picture cache slices, exposed via RenderResults

PrimarySlice 🔒

SliceDebugInfo

Debug information about a set of picture cache tiles, exposed via RenderResults

SliceDescriptor 🔒

TileCacheBuilder

Used during scene building to construct the list of pending tile caches.

TileCacheConfig

The output of a tile cache builder, containing all details needed to construct the tile cache(s) for the next scene, and retain tiles from the previous frame when sent send to the frame builder.

Enums

SliceKind 🔒

TileDebugInfo

Debug information about the state of a tile

Constants

MAX_CACHE_SLICES 🔒

Functions

create_tile_cache 🔒

Given a PrimitiveList and scroll root, construct a tile cache primitive instance that wraps the primitive list.

find_scroll_root 🔒

Find the scroll root for a given spatial node

find_shared_clip_root 🔒