Module webrender::tiling

Structs

AlphaBatchKey

AlphaBatchKeyFlags

AlphaBatchTask

AlphaBatcher
Encapsulates the logic of building batches for items that are blended.

AlphaRenderTask

BlurCommand

CacheClipInstance
A clipping primitive drawn into the clipping mask. Could be an image or a rectangle, which defines the way address is treated.

CacheMaskTask

ClipBatcher
Batcher managing draw calls into the clip mask (in the RT cache).

DynamicTaskInfo

Frame
A rendering-oriented representation of frame::Frame built by the render backend and presented to the renderer.

FrameBuilder

FrameBuilderConfig

PackedStackingContext

PrimitiveBatch

PrimitiveInstance

RenderPass
A render pass represents a set of rendering operations that don't depend on one another.

RenderPassIndex

RenderTarget
A render target represents a number of rendering operations on a surface.

RenderTargetContext

RenderTargetIndex

RenderTask

RenderTaskCollection

RenderTaskData

RenderTaskIndex

ScrollbarPrimitive

StackingContext

StackingContextIndex

Enums

AlphaBatchKind

AlphaRenderItem

BlurDirection

CompositeKind

MaskCacheKey

MaskGeometryKind

MaskResult

MaskSegment

PrimitiveBatchData

PrimitiveBatchItem

PrimitiveFlags

PrimitiveRunCmd

RenderTaskId

RenderTaskKey

RenderTaskKind

RenderTaskLocation

Constants

AXIS_ALIGNED

FLOATS_PER_RENDER_TASK_INFO

NEEDS_CLIPPING

OPAQUE_TASK_INDEX

Traits

AlphaBatchHelpers

Functions

textures_compatible

Type Definitions

AuxiliaryListsMap