Module webrender::util

Structs

ProfileScope

Traits

MatrixHelpers

RectHelpers

RectVaryings

VaryingElement

Functions

bilerp

bilerp_rect

distance

lerp

lerp_points

rect_center
Returns the center point of the given rect.

rect_is_empty

rect_is_well_formed_and_nonempty
Returns true if the rectangle's width and height are both strictly positive and false otherwise.

scale_color
Multiplies all non-alpha components of a color by the given value.

subdivide_rect_into_quadrants
Subdivides a rectangle into quadrants formed by a point. The quadrants are returned in the order of: top left, top right, bottom right, and bottom left.