Enum webrender::util::FastTransform

pub enum FastTransform<Src, Dst> {
    Offset(Vector2D<f32, Src>),
    Transform {
        transform: Transform3D<f32, Src, Dst>,
        inverse: Option<Transform3D<f32, Dst, Src>>,
        is_2d: bool,
    },
}

An enum that tries to avoid expensive transformation matrix calculations when possible when dealing with non-perspective axis-aligned transformations.

Variants

Offset

A simple offset, which can be used without doing any matrix math.

Tuple Fields of Offset

0: Vector2D<f32, Src>

Transform

A 2D transformation with an inverse.

Fields of Transform

transform: Transform3D<f32, Src, Dst>inverse: Option<Transform3D<f32, Dst, Src>>is_2d: bool

Implementations

impl<Src, Dst> FastTransform<Src, Dst>

pub fn identity() -> Self

pub fn with_vector(offset: Vector2D<f32, Src>) -> Self

pub fn with_scale_offset(scale_offset: ScaleOffset) -> Self

pub fn with_transform(transform: Transform3D<f32, Src, Dst>) -> Self

pub fn to_transform(&self) -> Cow<'_, Transform3D<f32, Src, Dst>>

pub fn is_identity(&self) -> bool

Return true if this is an identity transform

pub fn post_transform<NewDst>(
&self,
other: &FastTransform<Dst, NewDst>
) -> FastTransform<Src, NewDst>

pub fn pre_transform<NewSrc>(
&self,
other: &FastTransform<NewSrc, Src>
) -> FastTransform<NewSrc, Dst>

pub fn pre_translate(&self, other_offset: Vector2D<f32, Src>) -> Self

pub fn post_translate(&self, other_offset: Vector2D<f32, Dst>) -> Self

pub fn is_backface_visible(&self) -> bool

pub fn transform_point2d(
&self,
point: Point2D<f32, Src>
) -> Option<Point2D<f32, Dst>>

pub fn inverse(&self) -> Option<FastTransform<Dst, Src>>