Trait webrender::util::VecHelper

pub trait VecHelper<T> {
    fn alloc(&mut self) -> Allocation<T>;

    fn entry(&mut self, index: usize) -> VecEntry<T>;

    fn take(&mut self) -> Self;

    fn take_and_preallocate(&mut self) -> Self;
}

Required methods

fn alloc(&mut self) -> Allocation<T>

Growns the vector by a single entry, returning the allocation.

fn entry(&mut self, index: usize) -> VecEntry<T>

Either returns an existing elemenet, or grows the vector by one. Doesn't expect indices to be higher than the current length.

fn take(&mut self) -> Self

Equivalent to mem::replace(&mut vec, Vec::new())

fn take_and_preallocate(&mut self) -> Self

Functionally equivalent to mem::replace(&mut vec, Vec::new()) but tries to keep the allocation in the caller if it is empty or replace it with a pre-allocated vector.

Loading content...