webrender

Module visibility

Visibility pass

TODO: document what this pass does!

Structs

FrameVisibilityContext

FrameVisibilityState

PrimitiveVisibility

Information stored for a visible primitive about the visible rect and associated clip information.

PrimitiveVisibilityFlags

A set of bitflags that can be set in the visibility information for a primitive instance. This can be used to control how primitives are treated during batching.

Enums

VisibilityState

Contains the current state of the primitive’s visibility.

Functions

compute_conservative_visible_rect

update_prim_visibility